\name{CompactFactorNames}
\alias{CompactFactorNames}
\title{compact the factor names to camelCase}
\usage{CompactFactorNames(s)}
\description{\code{CompactFactorNames} compacts the factor names from separate words to camel case
}
\arguments{
  \item{s}{ string with the factor names' }
}
\value{string with names in camelCase }
\author{Jose Gama}
\examples{
CompactFactorNames('ton of refrigeration (12 000 BtuIT/h)')
}
\keyword{programming}
