\name{CompactPhysicalConstantNames}
\alias{CompactPhysicalConstantNames}
\title{compact the physical constant names to camelCase}
\usage{CompactPhysicalConstantNames(s)}
\description{\code{CompactPhysicalConstantNames} compacts the physical constant names from separate words to camel case
}
\arguments{
  \item{s}{ string with the physical constant names' }
}
\value{string with names in camelCase }
\author{Jose Gama}
\examples{
CompactPhysicalConstantNames('ton of refrigeration (12 000 BtuIT/h)')
}
\keyword{programming}
