\name{kNIST2010latticeSpacingOfSilicon}
\alias{kNIST2010latticeSpacingOfSilicon}
\alias{kNIST2010alphaParticleElectronMassRatio}
\alias{kNIST2010alphaParticleMass}
\alias{kNIST2010alphaParticleMassEnergyEquivalent}
\alias{kNIST2010alphaParticleMassEnergyEquivalentInMeV}
\alias{kNIST2010alphaParticleMassInU}
\alias{kNIST2010alphaParticleMolarMass}
\alias{kNIST2010alphaParticleProtonMassRatio}
\alias{kNIST2010AngstromStar}
\alias{kNIST2010atomicMassConstant}
\alias{kNIST2010atomicMassConstantEnergyEquivalent}
\alias{kNIST2010atomicMassConstantEnergyEquivalentInMeV}
\alias{kNIST2010atomicMassUnitElectronVoltRelationship}
\alias{kNIST2010atomicMassUnitHartreeRelationship}
\alias{kNIST2010atomicMassUnitHertzRelationship}
\alias{kNIST2010atomicMassUnitInverseMeterRelationship}
\alias{kNIST2010atomicMassUnitJouleRelationship}
\alias{kNIST2010atomicMassUnitKelvinRelationship}
\alias{kNIST2010atomicMassUnitKilogramRelationship}
\alias{kNIST2010atomicUnitOf1stHyperpolarizability}
\alias{kNIST2010atomicUnitOf2ndHyperpolarizability}
\alias{kNIST2010atomicUnitOfAction}
\alias{kNIST2010atomicUnitOfCharge}
\alias{kNIST2010atomicUnitOfChargeDensity}
\alias{kNIST2010atomicUnitOfCurrent}
\alias{kNIST2010atomicUnitOfElectricDipoleMom}
\alias{kNIST2010atomicUnitOfElectricField}
\alias{kNIST2010atomicUnitOfElectricFieldGradient}
\alias{kNIST2010atomicUnitOfElectricPolarizability}
\alias{kNIST2010atomicUnitOfElectricPotential}
\alias{kNIST2010atomicUnitOfElectricQuadrupoleMom}
\alias{kNIST2010atomicUnitOfEnergy}
\alias{kNIST2010atomicUnitOfForce}
\alias{kNIST2010atomicUnitOfLength}
\alias{kNIST2010atomicUnitOfMagDipoleMom}
\alias{kNIST2010atomicUnitOfMagFluxDensity}
\alias{kNIST2010atomicUnitOfMagnetizability}
\alias{kNIST2010atomicUnitOfMass}
\alias{kNIST2010atomicUnitOfMomUm}
\alias{kNIST2010atomicUnitOfPermittivity}
\alias{kNIST2010atomicUnitOfTime}
\alias{kNIST2010atomicUnitOfVelocity}
\alias{kNIST2010AvogadroConstant}
\alias{kNIST2010BohrMagneton}
\alias{kNIST2010BohrMagnetonInEVT}
\alias{kNIST2010BohrMagnetonInHzT}
\alias{kNIST2010BohrMagnetonInInverseMetersPerTesla}
\alias{kNIST2010BohrMagnetonInKT}
\alias{kNIST2010BohrRadius}
\alias{kNIST2010BoltzmannConstant}
\alias{kNIST2010BoltzmannConstantInEVK}
\alias{kNIST2010BoltzmannConstantInHzK}
\alias{kNIST2010BoltzmannConstantInInverseMetersPerKelvin}
\alias{kNIST2010characteristicImpedanceOfVacuum}
\alias{kNIST2010classicalElectronRadius}
\alias{kNIST2010ComptonWavelength}
\alias{kNIST2010ComptonWavelengthOver2Pi}
\alias{kNIST2010conductanceQuantum}
\alias{kNIST2010conventionalValueOfJosephsonConstant}
\alias{kNIST2010conventionalValueOfVonKlitzingConstant}
\alias{kNIST2010CuXUnit}
\alias{kNIST2010deuteronElectronMagMomRatio}
\alias{kNIST2010deuteronElectronMassRatio}
\alias{kNIST2010deuteronGFactor}
\alias{kNIST2010deuteronMagMom}
\alias{kNIST2010deuteronMagMomToBohrMagnetonRatio}
\alias{kNIST2010deuteronMagMomToNuclearMagnetonRatio}
\alias{kNIST2010deuteronMass}
\alias{kNIST2010deuteronMassEnergyEquivalent}
\alias{kNIST2010deuteronMassEnergyEquivalentInMeV}
\alias{kNIST2010deuteronMassInU}
\alias{kNIST2010deuteronMolarMass}
\alias{kNIST2010deuteronNeutronMagMomRatio}
\alias{kNIST2010deuteronProtonMagMomRatio}
\alias{kNIST2010deuteronProtonMassRatio}
\alias{kNIST2010deuteronRmsChargeRadius}
\alias{kNIST2010electricConstant}
\alias{kNIST2010electronChargeToMassQuotient}
\alias{kNIST2010electronDeuteronMagMomRatio}
\alias{kNIST2010electronDeuteronMassRatio}
\alias{kNIST2010electronGFactor}
\alias{kNIST2010electronGyromagRatio}
\alias{kNIST2010electronGyromagRatioOver2Pi}
\alias{kNIST2010electronHelionMassRatio}
\alias{kNIST2010electronMagMom}
\alias{kNIST2010electronMagMomAnomaly}
\alias{kNIST2010electronMagMomToBohrMagnetonRatio}
\alias{kNIST2010electronMagMomToNuclearMagnetonRatio}
\alias{kNIST2010electronMass}
\alias{kNIST2010electronMassEnergyEquivalent}
\alias{kNIST2010electronMassEnergyEquivalentInMeV}
\alias{kNIST2010electronMassInU}
\alias{kNIST2010electronMolarMass}
\alias{kNIST2010electronMuonMagMomRatio}
\alias{kNIST2010electronMuonMassRatio}
\alias{kNIST2010electronNeutronMagMomRatio}
\alias{kNIST2010electronNeutronMassRatio}
\alias{kNIST2010electronProtonMagMomRatio}
\alias{kNIST2010electronProtonMassRatio}
\alias{kNIST2010electronTauMassRatio}
\alias{kNIST2010electronToAlphaParticleMassRatio}
\alias{kNIST2010electronToShieldedHelionMagMomRatio}
\alias{kNIST2010electronToShieldedProtonMagMomRatio}
\alias{kNIST2010electronTritonMassRatio}
\alias{kNIST2010electronVolt}
\alias{kNIST2010electronVoltAtomicMassUnitRelationship}
\alias{kNIST2010electronVoltHartreeRelationship}
\alias{kNIST2010electronVoltHertzRelationship}
\alias{kNIST2010electronVoltInverseMeterRelationship}
\alias{kNIST2010electronVoltJouleRelationship}
\alias{kNIST2010electronVoltKelvinRelationship}
\alias{kNIST2010electronVoltKilogramRelationship}
\alias{kNIST2010elementaryCharge}
\alias{kNIST2010elementaryChargeOverH}
\alias{kNIST2010FaradayConstant}
\alias{kNIST2010FaradayConstantForConventionalElectricCurrent}
\alias{kNIST2010FermiCouplingConstant}
\alias{kNIST2010fineStructureConstant}
\alias{kNIST2010firstRadiationConstant}
\alias{kNIST2010firstRadiationConstantForSpectralRadiance}
\alias{kNIST2010hartreeAtomicMassUnitRelationship}
\alias{kNIST2010hartreeElectronVoltRelationship}
\alias{kNIST2010HartreeEnergy}
\alias{kNIST2010HartreeEnergyInEV}
\alias{kNIST2010hartreeHertzRelationship}
\alias{kNIST2010hartreeInverseMeterRelationship}
\alias{kNIST2010hartreeJouleRelationship}
\alias{kNIST2010hartreeKelvinRelationship}
\alias{kNIST2010hartreeKilogramRelationship}
\alias{kNIST2010helionElectronMassRatio}
\alias{kNIST2010helionGFactor}
\alias{kNIST2010helionMagMom}
\alias{kNIST2010helionMagMomToBohrMagnetonRatio}
\alias{kNIST2010helionMagMomToNuclearMagnetonRatio}
\alias{kNIST2010helionMass}
\alias{kNIST2010helionMassEnergyEquivalent}
\alias{kNIST2010helionMassEnergyEquivalentInMeV}
\alias{kNIST2010helionMassInU}
\alias{kNIST2010helionMolarMass}
\alias{kNIST2010helionProtonMassRatio}
\alias{kNIST2010hertzAtomicMassUnitRelationship}
\alias{kNIST2010hertzElectronVoltRelationship}
\alias{kNIST2010hertzHartreeRelationship}
\alias{kNIST2010hertzInverseMeterRelationship}
\alias{kNIST2010hertzJouleRelationship}
\alias{kNIST2010hertzKelvinRelationship}
\alias{kNIST2010hertzKilogramRelationship}
\alias{kNIST2010inverseFineStructureConstant}
\alias{kNIST2010inverseMeterAtomicMassUnitRelationship}
\alias{kNIST2010inverseMeterElectronVoltRelationship}
\alias{kNIST2010inverseMeterHartreeRelationship}
\alias{kNIST2010inverseMeterHertzRelationship}
\alias{kNIST2010inverseMeterJouleRelationship}
\alias{kNIST2010inverseMeterKelvinRelationship}
\alias{kNIST2010inverseMeterKilogramRelationship}
\alias{kNIST2010inverseOfConductanceQuantum}
\alias{kNIST2010JosephsonConstant}
\alias{kNIST2010jouleAtomicMassUnitRelationship}
\alias{kNIST2010jouleElectronVoltRelationship}
\alias{kNIST2010jouleHartreeRelationship}
\alias{kNIST2010jouleHertzRelationship}
\alias{kNIST2010jouleInverseMeterRelationship}
\alias{kNIST2010jouleKelvinRelationship}
\alias{kNIST2010jouleKilogramRelationship}
\alias{kNIST2010kelvinAtomicMassUnitRelationship}
\alias{kNIST2010kelvinElectronVoltRelationship}
\alias{kNIST2010kelvinHartreeRelationship}
\alias{kNIST2010kelvinHertzRelationship}
\alias{kNIST2010kelvinInverseMeterRelationship}
\alias{kNIST2010kelvinJouleRelationship}
\alias{kNIST2010kelvinKilogramRelationship}
\alias{kNIST2010kilogramAtomicMassUnitRelationship}
\alias{kNIST2010kilogramElectronVoltRelationship}
\alias{kNIST2010kilogramHartreeRelationship}
\alias{kNIST2010kilogramHertzRelationship}
\alias{kNIST2010kilogramInverseMeterRelationship}
\alias{kNIST2010kilogramJouleRelationship}
\alias{kNIST2010kilogramKelvinRelationship}
\alias{kNIST2010latticeParameterOfSilicon}
\alias{kNIST2010LoschmidtConstant}
\alias{kNIST2010LoschmidtConstant}
\alias{kNIST2010magConstant}
\alias{kNIST2010magFluxQuantum}
\alias{kNIST2010molarGasConstant}
\alias{kNIST2010molarMassConstant}
\alias{kNIST2010molarMassOfCarbon12}
\alias{kNIST2010molarPlanckConstant}
\alias{kNIST2010molarPlanckConstantTimesC}
\alias{kNIST2010molarVolumeOfIdealGas}
\alias{kNIST2010molarVolumeOfIdealGas}
\alias{kNIST2010molarVolumeOfSilicon}
\alias{kNIST2010MoXUnit}
\alias{kNIST2010muonComptonWavelength}
\alias{kNIST2010muonComptonWavelengthOver2Pi}
\alias{kNIST2010muonElectronMassRatio}
\alias{kNIST2010muonGFactor}
\alias{kNIST2010muonMagMom}
\alias{kNIST2010muonMagMomAnomaly}
\alias{kNIST2010muonMagMomToBohrMagnetonRatio}
\alias{kNIST2010muonMagMomToNuclearMagnetonRatio}
\alias{kNIST2010muonMass}
\alias{kNIST2010muonMassEnergyEquivalent}
\alias{kNIST2010muonMassEnergyEquivalentInMeV}
\alias{kNIST2010muonMassInU}
\alias{kNIST2010muonMolarMass}
\alias{kNIST2010muonNeutronMassRatio}
\alias{kNIST2010muonProtonMagMomRatio}
\alias{kNIST2010muonProtonMassRatio}
\alias{kNIST2010muonTauMassRatio}
\alias{kNIST2010naturalUnitOfAction}
\alias{kNIST2010naturalUnitOfActionInEVS}
\alias{kNIST2010naturalUnitOfEnergy}
\alias{kNIST2010naturalUnitOfEnergyInMeV}
\alias{kNIST2010naturalUnitOfLength}
\alias{kNIST2010naturalUnitOfMass}
\alias{kNIST2010naturalUnitOfMomUm}
\alias{kNIST2010naturalUnitOfMomUmInMeVc}
\alias{kNIST2010naturalUnitOfTime}
\alias{kNIST2010naturalUnitOfVelocity}
\alias{kNIST2010neutronComptonWavelength}
\alias{kNIST2010neutronComptonWavelengthOver2Pi}
\alias{kNIST2010neutronElectronMagMomRatio}
\alias{kNIST2010neutronElectronMassRatio}
\alias{kNIST2010neutronGFactor}
\alias{kNIST2010neutronGyromagRatio}
\alias{kNIST2010neutronGyromagRatioOver2Pi}
\alias{kNIST2010neutronMagMom}
\alias{kNIST2010neutronMagMomToBohrMagnetonRatio}
\alias{kNIST2010neutronMagMomToNuclearMagnetonRatio}
\alias{kNIST2010neutronMass}
\alias{kNIST2010neutronMassEnergyEquivalent}
\alias{kNIST2010neutronMassEnergyEquivalentInMeV}
\alias{kNIST2010neutronMassInU}
\alias{kNIST2010neutronMolarMass}
\alias{kNIST2010neutronMuonMassRatio}
\alias{kNIST2010neutronProtonMagMomRatio}
\alias{kNIST2010neutronProtonMassDifference}
\alias{kNIST2010neutronProtonMassDifferenceEnergyEquivalent}
\alias{kNIST2010neutronProtonMassDifferenceEnergyEquivalentInMeV}
\alias{kNIST2010neutronProtonMassDifferenceInU}
\alias{kNIST2010neutronProtonMassRatio}
\alias{kNIST2010neutronTauMassRatio}
\alias{kNIST2010neutronToShieldedProtonMagMomRatio}
\alias{kNIST2010NewtonianConstantOfGravitation}
\alias{kNIST2010NewtonianConstantOfGravitationOverHBarC}
\alias{kNIST2010nuclearMagneton}
\alias{kNIST2010nuclearMagnetonInEVT}
\alias{kNIST2010nuclearMagnetonInInverseMetersPerTesla}
\alias{kNIST2010nuclearMagnetonInKT}
\alias{kNIST2010nuclearMagnetonInMHzT}
\alias{kNIST2010PlanckConstant}
\alias{kNIST2010PlanckConstantInEVS}
\alias{kNIST2010PlanckConstantOver2Pi}
\alias{kNIST2010PlanckConstantOver2PiInEVS}
\alias{kNIST2010PlanckConstantOver2PiTimesCInMeVFm}
\alias{kNIST2010PlanckLength}
\alias{kNIST2010PlanckMass}
\alias{kNIST2010PlanckMassEnergyEquivalentInGeV}
\alias{kNIST2010PlanckTemperature}
\alias{kNIST2010PlanckTime}
\alias{kNIST2010protonChargeToMassQuotient}
\alias{kNIST2010protonComptonWavelength}
\alias{kNIST2010protonComptonWavelengthOver2Pi}
\alias{kNIST2010protonElectronMassRatio}
\alias{kNIST2010protonGFactor}
\alias{kNIST2010protonGyromagRatio}
\alias{kNIST2010protonGyromagRatioOver2Pi}
\alias{kNIST2010protonMagMom}
\alias{kNIST2010protonMagMomToBohrMagnetonRatio}
\alias{kNIST2010protonMagMomToNuclearMagnetonRatio}
\alias{kNIST2010protonMagShieldingCorrection}
\alias{kNIST2010protonMass}
\alias{kNIST2010protonMassEnergyEquivalent}
\alias{kNIST2010protonMassEnergyEquivalentInMeV}
\alias{kNIST2010protonMassInU}
\alias{kNIST2010protonMolarMass}
\alias{kNIST2010protonMuonMassRatio}
\alias{kNIST2010protonNeutronMagMomRatio}
\alias{kNIST2010protonNeutronMassRatio}
\alias{kNIST2010protonRmsChargeRadius}
\alias{kNIST2010protonTauMassRatio}
\alias{kNIST2010quantumOfCirculation}
\alias{kNIST2010quantumOfCirculationTimes2}
\alias{kNIST2010RydbergConstant}
\alias{kNIST2010RydbergConstantTimesCInHz}
\alias{kNIST2010RydbergConstantTimesHcInEV}
\alias{kNIST2010RydbergConstantTimesHcInJ}
\alias{kNIST2010SackurTetrodeConstant}
\alias{kNIST2010SackurTetrodeConstant}
\alias{kNIST2010secondRadiationConstant}
\alias{kNIST2010shieldedHelionGyromagRatio}
\alias{kNIST2010shieldedHelionGyromagRatioOver2Pi}
\alias{kNIST2010shieldedHelionMagMom}
\alias{kNIST2010shieldedHelionMagMomToBohrMagnetonRatio}
\alias{kNIST2010shieldedHelionMagMomToNuclearMagnetonRatio}
\alias{kNIST2010shieldedHelionToProtonMagMomRatio}
\alias{kNIST2010shieldedHelionToShieldedProtonMagMomRatio}
\alias{kNIST2010shieldedProtonGyromagRatio}
\alias{kNIST2010shieldedProtonGyromagRatioOver2Pi}
\alias{kNIST2010shieldedProtonMagMom}
\alias{kNIST2010shieldedProtonMagMomToBohrMagnetonRatio}
\alias{kNIST2010shieldedProtonMagMomToNuclearMagnetonRatio}
\alias{kNIST2010speedOfLightInVacuum}
\alias{kNIST2010standardAccelerationOfGravity}
\alias{kNIST2010standardAtmosphere}
\alias{kNIST2010standardStatePressure}
\alias{kNIST2010StefanBoltzmannConstant}
\alias{kNIST2010tauComptonWavelength}
\alias{kNIST2010tauComptonWavelengthOver2Pi}
\alias{kNIST2010tauElectronMassRatio}
\alias{kNIST2010tauMass}
\alias{kNIST2010tauMassEnergyEquivalent}
\alias{kNIST2010tauMassEnergyEquivalentInMeV}
\alias{kNIST2010tauMassInU}
\alias{kNIST2010tauMolarMass}
\alias{kNIST2010tauMuonMassRatio}
\alias{kNIST2010tauNeutronMassRatio}
\alias{kNIST2010tauProtonMassRatio}
\alias{kNIST2010ThomsonCrossSection}
\alias{kNIST2010tritonElectronMassRatio}
\alias{kNIST2010tritonGFactor}
\alias{kNIST2010tritonMagMom}
\alias{kNIST2010tritonMagMomToBohrMagnetonRatio}
\alias{kNIST2010tritonMagMomToNuclearMagnetonRatio}
\alias{kNIST2010tritonMass}
\alias{kNIST2010tritonMassEnergyEquivalent}
\alias{kNIST2010tritonMassEnergyEquivalentInMeV}
\alias{kNIST2010tritonMassInU}
\alias{kNIST2010tritonMolarMass}
\alias{kNIST2010tritonProtonMassRatio}
\alias{kNIST2010unifiedAtomicMassUnit}
\alias{kNIST2010vonKlitzingConstant}
\alias{kNIST2010weakMixingAngle}
\alias{kNIST2010WienFrequencyDisplacementLawConstant}
\alias{kNIST2010WienWavelengthDisplacementLawConstant}\title{NIST Physical Constants}
\usage{kNIST2010latticeSpacingOfSilicon}
\description{
\code{kNIST2010latticeSpacingOfSilicon} lattice spacing of silicon

\code{kNIST2010alphaParticleElectronMassRatio} alpha particle-electron mass ratio

\code{kNIST2010alphaParticleMass} alpha particle mass

\code{kNIST2010alphaParticleMassEnergyEquivalent} alpha particle mass energy equivalent

\code{kNIST2010alphaParticleMassEnergyEquivalentInMeV} alpha particle mass energy equivalent in MeV

\code{kNIST2010alphaParticleMassInU} alpha particle mass in u

\code{kNIST2010alphaParticleMolarMass} alpha particle molar mass

\code{kNIST2010alphaParticleProtonMassRatio} alpha particle-proton mass ratio

\code{kNIST2010AngstromStar} Angstrom star

\code{kNIST2010atomicMassConstant} atomic mass constant

\code{kNIST2010atomicMassConstantEnergyEquivalent} atomic mass constant energy equivalent

\code{kNIST2010atomicMassConstantEnergyEquivalentInMeV} atomic mass constant energy equivalent in MeV

\code{kNIST2010atomicMassUnitElectronVoltRelationship} atomic mass unit-electron volt relationship

\code{kNIST2010atomicMassUnitHartreeRelationship} atomic mass unit-hartree relationship

\code{kNIST2010atomicMassUnitHertzRelationship} atomic mass unit-hertz relationship

\code{kNIST2010atomicMassUnitInverseMeterRelationship} atomic mass unit-inverse meter relationship

\code{kNIST2010atomicMassUnitJouleRelationship} atomic mass unit-joule relationship

\code{kNIST2010atomicMassUnitKelvinRelationship} atomic mass unit-kelvin relationship

\code{kNIST2010atomicMassUnitKilogramRelationship} atomic mass unit-kilogram relationship

\code{kNIST2010atomicUnitOf1stHyperpolarizability} atomic unit of 1st hyperpolarizability

\code{kNIST2010atomicUnitOf2ndHyperpolarizability} atomic unit of 2nd hyperpolarizability

\code{kNIST2010atomicUnitOfAction} atomic unit of action

\code{kNIST2010atomicUnitOfCharge} atomic unit of charge

\code{kNIST2010atomicUnitOfChargeDensity} atomic unit of charge density

\code{kNIST2010atomicUnitOfCurrent} atomic unit of current

\code{kNIST2010atomicUnitOfElectricDipoleMom} atomic unit of electric dipole mom.

\code{kNIST2010atomicUnitOfElectricField} atomic unit of electric field

\code{kNIST2010atomicUnitOfElectricFieldGradient} atomic unit of electric field gradient

\code{kNIST2010atomicUnitOfElectricPolarizability} atomic unit of electric polarizability

\code{kNIST2010atomicUnitOfElectricPotential} atomic unit of electric potential

\code{kNIST2010atomicUnitOfElectricQuadrupoleMom} atomic unit of electric quadrupole mom.

\code{kNIST2010atomicUnitOfEnergy} atomic unit of energy

\code{kNIST2010atomicUnitOfForce} atomic unit of force

\code{kNIST2010atomicUnitOfLength} atomic unit of length

\code{kNIST2010atomicUnitOfMagDipoleMom} atomic unit of mag. dipole mom.

\code{kNIST2010atomicUnitOfMagFluxDensity} atomic unit of mag. flux density

\code{kNIST2010atomicUnitOfMagnetizability} atomic unit of magnetizability

\code{kNIST2010atomicUnitOfMass} atomic unit of mass

\code{kNIST2010atomicUnitOfMomUm} atomic unit of mom.um

\code{kNIST2010atomicUnitOfPermittivity} atomic unit of permittivity

\code{kNIST2010atomicUnitOfTime} atomic unit of time

\code{kNIST2010atomicUnitOfVelocity} atomic unit of velocity

\code{kNIST2010AvogadroConstant} Avogadro constant

\code{kNIST2010BohrMagneton} Bohr magneton

\code{kNIST2010BohrMagnetonInEVT} Bohr magneton in eV/T

\code{kNIST2010BohrMagnetonInHzT} Bohr magneton in Hz/T

\code{kNIST2010BohrMagnetonInInverseMetersPerTesla} Bohr magneton in inverse meters per tesla

\code{kNIST2010BohrMagnetonInKT} Bohr magneton in K/T

\code{kNIST2010BohrRadius} Bohr radius

\code{kNIST2010BoltzmannConstant} Boltzmann constant

\code{kNIST2010BoltzmannConstantInEVK} Boltzmann constant in eV/K

\code{kNIST2010BoltzmannConstantInHzK} Boltzmann constant in Hz/K

\code{kNIST2010BoltzmannConstantInInverseMetersPerKelvin} Boltzmann constant in inverse meters per kelvin

\code{kNIST2010characteristicImpedanceOfVacuum} characteristic impedance of vacuum

\code{kNIST2010classicalElectronRadius} classical electron radius

\code{kNIST2010ComptonWavelength} Compton wavelength

\code{kNIST2010ComptonWavelengthOver2Pi} Compton wavelength over 2 pi

\code{kNIST2010conductanceQuantum} conductance quantum

\code{kNIST2010conventionalValueOfJosephsonConstant} conventional value of Josephson constant

\code{kNIST2010conventionalValueOfVonKlitzingConstant} conventional value of von Klitzing constant

\code{kNIST2010CuXUnit} Cu x unit

\code{kNIST2010deuteronElectronMagMomRatio} deuteron-electron mag. mom. ratio

\code{kNIST2010deuteronElectronMassRatio} deuteron-electron mass ratio

\code{kNIST2010deuteronGFactor} deuteron g factor

\code{kNIST2010deuteronMagMom} deuteron mag. mom.

\code{kNIST2010deuteronMagMomToBohrMagnetonRatio} deuteron mag. mom. to Bohr magneton ratio

\code{kNIST2010deuteronMagMomToNuclearMagnetonRatio} deuteron mag. mom. to nuclear magneton ratio

\code{kNIST2010deuteronMass} deuteron mass

\code{kNIST2010deuteronMassEnergyEquivalent} deuteron mass energy equivalent

\code{kNIST2010deuteronMassEnergyEquivalentInMeV} deuteron mass energy equivalent in MeV

\code{kNIST2010deuteronMassInU} deuteron mass in u

\code{kNIST2010deuteronMolarMass} deuteron molar mass

\code{kNIST2010deuteronNeutronMagMomRatio} deuteron-neutron mag. mom. ratio

\code{kNIST2010deuteronProtonMagMomRatio} deuteron-proton mag. mom. ratio

\code{kNIST2010deuteronProtonMassRatio} deuteron-proton mass ratio

\code{kNIST2010deuteronRmsChargeRadius} deuteron rms charge radius

\code{kNIST2010electricConstant} electric constant

\code{kNIST2010electronChargeToMassQuotient} electron charge to mass quotient

\code{kNIST2010electronDeuteronMagMomRatio} electron-deuteron mag. mom. ratio

\code{kNIST2010electronDeuteronMassRatio} electron-deuteron mass ratio

\code{kNIST2010electronGFactor} electron g factor

\code{kNIST2010electronGyromagRatio} electron gyromag. ratio

\code{kNIST2010electronGyromagRatioOver2Pi} electron gyromag. ratio over 2 pi

\code{kNIST2010electronHelionMassRatio} electron-helion mass ratio

\code{kNIST2010electronMagMom} electron mag. mom.

\code{kNIST2010electronMagMomAnomaly} electron mag. mom. anomaly

\code{kNIST2010electronMagMomToBohrMagnetonRatio} electron mag. mom. to Bohr magneton ratio

\code{kNIST2010electronMagMomToNuclearMagnetonRatio} electron mag. mom. to nuclear magneton ratio

\code{kNIST2010electronMass} electron mass

\code{kNIST2010electronMassEnergyEquivalent} electron mass energy equivalent

\code{kNIST2010electronMassEnergyEquivalentInMeV} electron mass energy equivalent in MeV

\code{kNIST2010electronMassInU} electron mass in u

\code{kNIST2010electronMolarMass} electron molar mass

\code{kNIST2010electronMuonMagMomRatio} electron-muon mag. mom. ratio

\code{kNIST2010electronMuonMassRatio} electron-muon mass ratio

\code{kNIST2010electronNeutronMagMomRatio} electron-neutron mag. mom. ratio

\code{kNIST2010electronNeutronMassRatio} electron-neutron mass ratio

\code{kNIST2010electronProtonMagMomRatio} electron-proton mag. mom. ratio

\code{kNIST2010electronProtonMassRatio} electron-proton mass ratio

\code{kNIST2010electronTauMassRatio} electron-tau mass ratio

\code{kNIST2010electronToAlphaParticleMassRatio} electron to alpha particle mass ratio

\code{kNIST2010electronToShieldedHelionMagMomRatio} electron to shielded helion mag. mom. ratio

\code{kNIST2010electronToShieldedProtonMagMomRatio} electron to shielded proton mag. mom. ratio

\code{kNIST2010electronTritonMassRatio} electron-triton mass ratio

\code{kNIST2010electronVolt} electron volt

\code{kNIST2010electronVoltAtomicMassUnitRelationship} electron volt-atomic mass unit relationship

\code{kNIST2010electronVoltHartreeRelationship} electron volt-hartree relationship

\code{kNIST2010electronVoltHertzRelationship} electron volt-hertz relationship

\code{kNIST2010electronVoltInverseMeterRelationship} electron volt-inverse meter relationship

\code{kNIST2010electronVoltJouleRelationship} electron volt-joule relationship

\code{kNIST2010electronVoltKelvinRelationship} electron volt-kelvin relationship

\code{kNIST2010electronVoltKilogramRelationship} electron volt-kilogram relationship

\code{kNIST2010elementaryCharge} elementary charge

\code{kNIST2010elementaryChargeOverH} elementary charge over h

\code{kNIST2010FaradayConstant} Faraday constant

\code{kNIST2010FaradayConstantForConventionalElectricCurrent} Faraday constant for conventional electric current

\code{kNIST2010FermiCouplingConstant} Fermi coupling constant

\code{kNIST2010fineStructureConstant} fine-structure constant

\code{kNIST2010firstRadiationConstant} first radiation constant

\code{kNIST2010firstRadiationConstantForSpectralRadiance} first radiation constant for spectral radiance

\code{kNIST2010hartreeAtomicMassUnitRelationship} hartree-atomic mass unit relationship

\code{kNIST2010hartreeElectronVoltRelationship} hartree-electron volt relationship

\code{kNIST2010HartreeEnergy} Hartree energy

\code{kNIST2010HartreeEnergyInEV} Hartree energy in eV

\code{kNIST2010hartreeHertzRelationship} hartree-hertz relationship

\code{kNIST2010hartreeInverseMeterRelationship} hartree-inverse meter relationship

\code{kNIST2010hartreeJouleRelationship} hartree-joule relationship

\code{kNIST2010hartreeKelvinRelationship} hartree-kelvin relationship

\code{kNIST2010hartreeKilogramRelationship} hartree-kilogram relationship

\code{kNIST2010helionElectronMassRatio} helion-electron mass ratio

\code{kNIST2010helionGFactor} helion g factor

\code{kNIST2010helionMagMom} helion mag. mom.

\code{kNIST2010helionMagMomToBohrMagnetonRatio} helion mag. mom. to Bohr magneton ratio

\code{kNIST2010helionMagMomToNuclearMagnetonRatio} helion mag. mom. to nuclear magneton ratio

\code{kNIST2010helionMass} helion mass

\code{kNIST2010helionMassEnergyEquivalent} helion mass energy equivalent

\code{kNIST2010helionMassEnergyEquivalentInMeV} helion mass energy equivalent in MeV

\code{kNIST2010helionMassInU} helion mass in u

\code{kNIST2010helionMolarMass} helion molar mass

\code{kNIST2010helionProtonMassRatio} helion-proton mass ratio

\code{kNIST2010hertzAtomicMassUnitRelationship} hertz-atomic mass unit relationship

\code{kNIST2010hertzElectronVoltRelationship} hertz-electron volt relationship

\code{kNIST2010hertzHartreeRelationship} hertz-hartree relationship

\code{kNIST2010hertzInverseMeterRelationship} hertz-inverse meter relationship

\code{kNIST2010hertzJouleRelationship} hertz-joule relationship

\code{kNIST2010hertzKelvinRelationship} hertz-kelvin relationship

\code{kNIST2010hertzKilogramRelationship} hertz-kilogram relationship

\code{kNIST2010inverseFineStructureConstant} inverse fine-structure constant

\code{kNIST2010inverseMeterAtomicMassUnitRelationship} inverse meter-atomic mass unit relationship

\code{kNIST2010inverseMeterElectronVoltRelationship} inverse meter-electron volt relationship

\code{kNIST2010inverseMeterHartreeRelationship} inverse meter-hartree relationship

\code{kNIST2010inverseMeterHertzRelationship} inverse meter-hertz relationship

\code{kNIST2010inverseMeterJouleRelationship} inverse meter-joule relationship

\code{kNIST2010inverseMeterKelvinRelationship} inverse meter-kelvin relationship

\code{kNIST2010inverseMeterKilogramRelationship} inverse meter-kilogram relationship

\code{kNIST2010inverseOfConductanceQuantum} inverse of conductance quantum

\code{kNIST2010JosephsonConstant} Josephson constant

\code{kNIST2010jouleAtomicMassUnitRelationship} joule-atomic mass unit relationship

\code{kNIST2010jouleElectronVoltRelationship} joule-electron volt relationship

\code{kNIST2010jouleHartreeRelationship} joule-hartree relationship

\code{kNIST2010jouleHertzRelationship} joule-hertz relationship

\code{kNIST2010jouleInverseMeterRelationship} joule-inverse meter relationship

\code{kNIST2010jouleKelvinRelationship} joule-kelvin relationship

\code{kNIST2010jouleKilogramRelationship} joule-kilogram relationship

\code{kNIST2010kelvinAtomicMassUnitRelationship} kelvin-atomic mass unit relationship

\code{kNIST2010kelvinElectronVoltRelationship} kelvin-electron volt relationship

\code{kNIST2010kelvinHartreeRelationship} kelvin-hartree relationship

\code{kNIST2010kelvinHertzRelationship} kelvin-hertz relationship

\code{kNIST2010kelvinInverseMeterRelationship} kelvin-inverse meter relationship

\code{kNIST2010kelvinJouleRelationship} kelvin-joule relationship

\code{kNIST2010kelvinKilogramRelationship} kelvin-kilogram relationship

\code{kNIST2010kilogramAtomicMassUnitRelationship} kilogram-atomic mass unit relationship

\code{kNIST2010kilogramElectronVoltRelationship} kilogram-electron volt relationship

\code{kNIST2010kilogramHartreeRelationship} kilogram-hartree relationship

\code{kNIST2010kilogramHertzRelationship} kilogram-hertz relationship

\code{kNIST2010kilogramInverseMeterRelationship} kilogram-inverse meter relationship

\code{kNIST2010kilogramJouleRelationship} kilogram-joule relationship

\code{kNIST2010kilogramKelvinRelationship} kilogram-kelvin relationship

\code{kNIST2010latticeParameterOfSilicon} lattice parameter of silicon

\code{kNIST2010LoschmidtConstant} Loschmidt constant (273.15 K, 100 kPa)

\code{kNIST2010LoschmidtConstant} Loschmidt constant (273.15 K, 101.325 kPa)

\code{kNIST2010magConstant} mag. constant

\code{kNIST2010magFluxQuantum} mag. flux quantum

\code{kNIST2010molarGasConstant} molar gas constant

\code{kNIST2010molarMassConstant} molar mass constant

\code{kNIST2010molarMassOfCarbon12} molar mass of carbon-12

\code{kNIST2010molarPlanckConstant} molar Planck constant

\code{kNIST2010molarPlanckConstantTimesC} molar Planck constant times c

\code{kNIST2010molarVolumeOfIdealGas} molar volume of ideal gas (273.15 K, 100 kPa)

\code{kNIST2010molarVolumeOfIdealGas} molar volume of ideal gas (273.15 K, 101.325 kPa)

\code{kNIST2010molarVolumeOfSilicon} molar volume of silicon

\code{kNIST2010MoXUnit} Mo x unit

\code{kNIST2010muonComptonWavelength} muon Compton wavelength

\code{kNIST2010muonComptonWavelengthOver2Pi} muon Compton wavelength over 2 pi

\code{kNIST2010muonElectronMassRatio} muon-electron mass ratio

\code{kNIST2010muonGFactor} muon g factor

\code{kNIST2010muonMagMom} muon mag. mom.

\code{kNIST2010muonMagMomAnomaly} muon mag. mom. anomaly

\code{kNIST2010muonMagMomToBohrMagnetonRatio} muon mag. mom. to Bohr magneton ratio

\code{kNIST2010muonMagMomToNuclearMagnetonRatio} muon mag. mom. to nuclear magneton ratio

\code{kNIST2010muonMass} muon mass

\code{kNIST2010muonMassEnergyEquivalent} muon mass energy equivalent

\code{kNIST2010muonMassEnergyEquivalentInMeV} muon mass energy equivalent in MeV

\code{kNIST2010muonMassInU} muon mass in u

\code{kNIST2010muonMolarMass} muon molar mass

\code{kNIST2010muonNeutronMassRatio} muon-neutron mass ratio

\code{kNIST2010muonProtonMagMomRatio} muon-proton mag. mom. ratio

\code{kNIST2010muonProtonMassRatio} muon-proton mass ratio

\code{kNIST2010muonTauMassRatio} muon-tau mass ratio

\code{kNIST2010naturalUnitOfAction} natural unit of action

\code{kNIST2010naturalUnitOfActionInEVS} natural unit of action in eV s

\code{kNIST2010naturalUnitOfEnergy} natural unit of energy

\code{kNIST2010naturalUnitOfEnergyInMeV} natural unit of energy in MeV

\code{kNIST2010naturalUnitOfLength} natural unit of length

\code{kNIST2010naturalUnitOfMass} natural unit of mass

\code{kNIST2010naturalUnitOfMomUm} natural unit of mom.um

\code{kNIST2010naturalUnitOfMomUmInMeVc} natural unit of mom.um in MeV/c

\code{kNIST2010naturalUnitOfTime} natural unit of time

\code{kNIST2010naturalUnitOfVelocity} natural unit of velocity

\code{kNIST2010neutronComptonWavelength} neutron Compton wavelength

\code{kNIST2010neutronComptonWavelengthOver2Pi} neutron Compton wavelength over 2 pi

\code{kNIST2010neutronElectronMagMomRatio} neutron-electron mag. mom. ratio

\code{kNIST2010neutronElectronMassRatio} neutron-electron mass ratio

\code{kNIST2010neutronGFactor} neutron g factor

\code{kNIST2010neutronGyromagRatio} neutron gyromag. ratio

\code{kNIST2010neutronGyromagRatioOver2Pi} neutron gyromag. ratio over 2 pi

\code{kNIST2010neutronMagMom} neutron mag. mom.

\code{kNIST2010neutronMagMomToBohrMagnetonRatio} neutron mag. mom. to Bohr magneton ratio

\code{kNIST2010neutronMagMomToNuclearMagnetonRatio} neutron mag. mom. to nuclear magneton ratio

\code{kNIST2010neutronMass} neutron mass

\code{kNIST2010neutronMassEnergyEquivalent} neutron mass energy equivalent

\code{kNIST2010neutronMassEnergyEquivalentInMeV} neutron mass energy equivalent in MeV

\code{kNIST2010neutronMassInU} neutron mass in u

\code{kNIST2010neutronMolarMass} neutron molar mass

\code{kNIST2010neutronMuonMassRatio} neutron-muon mass ratio

\code{kNIST2010neutronProtonMagMomRatio} neutron-proton mag. mom. ratio

\code{kNIST2010neutronProtonMassDifference} neutron-proton mass difference

\code{kNIST2010neutronProtonMassDifferenceEnergyEquivalent} neutron-proton mass difference energy equivalent

\code{kNIST2010neutronProtonMassDifferenceEnergyEquivalentInMeV} neutron-proton mass difference energy equivalent in MeV

\code{kNIST2010neutronProtonMassDifferenceInU} neutron-proton mass difference in u

\code{kNIST2010neutronProtonMassRatio} neutron-proton mass ratio

\code{kNIST2010neutronTauMassRatio} neutron-tau mass ratio

\code{kNIST2010neutronToShieldedProtonMagMomRatio} neutron to shielded proton mag. mom. ratio

\code{kNIST2010NewtonianConstantOfGravitation} Newtonian constant of gravitation

\code{kNIST2010NewtonianConstantOfGravitationOverHBarC} Newtonian constant of gravitation over h-bar c

\code{kNIST2010nuclearMagneton} nuclear magneton

\code{kNIST2010nuclearMagnetonInEVT} nuclear magneton in eV/T

\code{kNIST2010nuclearMagnetonInInverseMetersPerTesla} nuclear magneton in inverse meters per tesla

\code{kNIST2010nuclearMagnetonInKT} nuclear magneton in K/T

\code{kNIST2010nuclearMagnetonInMHzT} nuclear magneton in MHz/T

\code{kNIST2010PlanckConstant} Planck constant

\code{kNIST2010PlanckConstantInEVS} Planck constant in eV s

\code{kNIST2010PlanckConstantOver2Pi} Planck constant over 2 pi

\code{kNIST2010PlanckConstantOver2PiInEVS} Planck constant over 2 pi in eV s

\code{kNIST2010PlanckConstantOver2PiTimesCInMeVFm} Planck constant over 2 pi times c in MeV fm

\code{kNIST2010PlanckLength} Planck length

\code{kNIST2010PlanckMass} Planck mass

\code{kNIST2010PlanckMassEnergyEquivalentInGeV} Planck mass energy equivalent in GeV

\code{kNIST2010PlanckTemperature} Planck temperature

\code{kNIST2010PlanckTime} Planck time

\code{kNIST2010protonChargeToMassQuotient} proton charge to mass quotient

\code{kNIST2010protonComptonWavelength} proton Compton wavelength

\code{kNIST2010protonComptonWavelengthOver2Pi} proton Compton wavelength over 2 pi

\code{kNIST2010protonElectronMassRatio} proton-electron mass ratio

\code{kNIST2010protonGFactor} proton g factor

\code{kNIST2010protonGyromagRatio} proton gyromag. ratio

\code{kNIST2010protonGyromagRatioOver2Pi} proton gyromag. ratio over 2 pi

\code{kNIST2010protonMagMom} proton mag. mom.

\code{kNIST2010protonMagMomToBohrMagnetonRatio} proton mag. mom. to Bohr magneton ratio

\code{kNIST2010protonMagMomToNuclearMagnetonRatio} proton mag. mom. to nuclear magneton ratio

\code{kNIST2010protonMagShieldingCorrection} proton mag. shielding correction

\code{kNIST2010protonMass} proton mass

\code{kNIST2010protonMassEnergyEquivalent} proton mass energy equivalent

\code{kNIST2010protonMassEnergyEquivalentInMeV} proton mass energy equivalent in MeV

\code{kNIST2010protonMassInU} proton mass in u

\code{kNIST2010protonMolarMass} proton molar mass

\code{kNIST2010protonMuonMassRatio} proton-muon mass ratio

\code{kNIST2010protonNeutronMagMomRatio} proton-neutron mag. mom. ratio

\code{kNIST2010protonNeutronMassRatio} proton-neutron mass ratio

\code{kNIST2010protonRmsChargeRadius} proton rms charge radius

\code{kNIST2010protonTauMassRatio} proton-tau mass ratio

\code{kNIST2010quantumOfCirculation} quantum of circulation

\code{kNIST2010quantumOfCirculationTimes2} quantum of circulation times 2

\code{kNIST2010RydbergConstant} Rydberg constant

\code{kNIST2010RydbergConstantTimesCInHz} Rydberg constant times c in Hz

\code{kNIST2010RydbergConstantTimesHcInEV} Rydberg constant times hc in eV

\code{kNIST2010RydbergConstantTimesHcInJ} Rydberg constant times hc in J

\code{kNIST2010SackurTetrodeConstant} Sackur-Tetrode constant (1 K, 100 kPa)

\code{kNIST2010SackurTetrodeConstant} Sackur-Tetrode constant (1 K, 101.325 kPa)

\code{kNIST2010secondRadiationConstant} second radiation constant

\code{kNIST2010shieldedHelionGyromagRatio} shielded helion gyromag. ratio

\code{kNIST2010shieldedHelionGyromagRatioOver2Pi} shielded helion gyromag. ratio over 2 pi

\code{kNIST2010shieldedHelionMagMom} shielded helion mag. mom.

\code{kNIST2010shieldedHelionMagMomToBohrMagnetonRatio} shielded helion mag. mom. to Bohr magneton ratio

\code{kNIST2010shieldedHelionMagMomToNuclearMagnetonRatio} shielded helion mag. mom. to nuclear magneton ratio

\code{kNIST2010shieldedHelionToProtonMagMomRatio} shielded helion to proton mag. mom. ratio

\code{kNIST2010shieldedHelionToShieldedProtonMagMomRatio} shielded helion to shielded proton mag. mom. ratio

\code{kNIST2010shieldedProtonGyromagRatio} shielded proton gyromag. ratio

\code{kNIST2010shieldedProtonGyromagRatioOver2Pi} shielded proton gyromag. ratio over 2 pi

\code{kNIST2010shieldedProtonMagMom} shielded proton mag. mom.

\code{kNIST2010shieldedProtonMagMomToBohrMagnetonRatio} shielded proton mag. mom. to Bohr magneton ratio

\code{kNIST2010shieldedProtonMagMomToNuclearMagnetonRatio} shielded proton mag. mom. to nuclear magneton ratio

\code{kNIST2010speedOfLightInVacuum} speed of light in vacuum

\code{kNIST2010standardAccelerationOfGravity} standard acceleration of gravity

\code{kNIST2010standardAtmosphere} standard atmosphere

\code{kNIST2010standardStatePressure} standard-state pressure

\code{kNIST2010StefanBoltzmannConstant} Stefan-Boltzmann constant

\code{kNIST2010tauComptonWavelength} tau Compton wavelength

\code{kNIST2010tauComptonWavelengthOver2Pi} tau Compton wavelength over 2 pi

\code{kNIST2010tauElectronMassRatio} tau-electron mass ratio

\code{kNIST2010tauMass} tau mass

\code{kNIST2010tauMassEnergyEquivalent} tau mass energy equivalent

\code{kNIST2010tauMassEnergyEquivalentInMeV} tau mass energy equivalent in MeV

\code{kNIST2010tauMassInU} tau mass in u

\code{kNIST2010tauMolarMass} tau molar mass

\code{kNIST2010tauMuonMassRatio} tau-muon mass ratio

\code{kNIST2010tauNeutronMassRatio} tau-neutron mass ratio

\code{kNIST2010tauProtonMassRatio} tau-proton mass ratio

\code{kNIST2010ThomsonCrossSection} Thomson cross section

\code{kNIST2010tritonElectronMassRatio} triton-electron mass ratio

\code{kNIST2010tritonGFactor} triton g factor

\code{kNIST2010tritonMagMom} triton mag. mom.

\code{kNIST2010tritonMagMomToBohrMagnetonRatio} triton mag. mom. to Bohr magneton ratio

\code{kNIST2010tritonMagMomToNuclearMagnetonRatio} triton mag. mom. to nuclear magneton ratio

\code{kNIST2010tritonMass} triton mass

\code{kNIST2010tritonMassEnergyEquivalent} triton mass energy equivalent

\code{kNIST2010tritonMassEnergyEquivalentInMeV} triton mass energy equivalent in MeV

\code{kNIST2010tritonMassInU} triton mass in u

\code{kNIST2010tritonMolarMass} triton molar mass

\code{kNIST2010tritonProtonMassRatio} triton-proton mass ratio

\code{kNIST2010unifiedAtomicMassUnit} unified atomic mass unit

\code{kNIST2010vonKlitzingConstant} von Klitzing constant

\code{kNIST2010weakMixingAngle} weak mixing angle

\code{kNIST2010WienFrequencyDisplacementLawConstant} Wien frequency displacement law constant

\code{kNIST2010WienWavelengthDisplacementLawConstant} Wien wavelength displacement law constant}
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
\dontrun{print(kNIST2010latticeSpacingOfSilicon)}
}
\keyword{datasets}
