\name{kNISTnonSIatomicUnitOf1stHyperpolarizability}
\alias{kNISTnonSIatomicUnitOf1stHyperpolarizability}
\alias{kNISTnonSIatomicUnitOf2ndHyperpolarizability}
\alias{kNISTnonSIatomicUnitOfAction}
\alias{kNISTnonSIatomicUnitOfCharge}
\alias{kNISTnonSIatomicUnitOfChargeDensity}
\alias{kNISTnonSIatomicUnitOfCurrent}
\alias{kNISTnonSIatomicUnitOfElectricDipoleMoment}
\alias{kNISTnonSIatomicUnitOfElectricField}
\alias{kNISTnonSIatomicUnitOfElectricFieldGradient}
\alias{kNISTnonSIatomicUnitOfElectricPolarizability}
\alias{kNISTnonSIatomicUnitOfElectricPotential}
\alias{kNISTnonSIatomicUnitOfElectricQuadrupoleMoment}
\alias{kNISTnonSIatomicUnitOfEnergy}
\alias{kNISTnonSIatomicUnitOfLength}
\alias{kNISTnonSIatomicUnitOfMagneticDipoleMoment}
\alias{kNISTnonSIatomicUnitOfMagneticFluxDensity}
\alias{kNISTnonSIatomicUnitOfMagnetizability}
\alias{kNISTnonSIatomicUnitOfMass}
\alias{kNISTnonSIatomicUnitOfMomentum}
\alias{kNISTnonSIatomicUnitOfPermittivity}
\alias{kNISTnonSIatomicUnitOfTime}
\alias{kNISTnonSIatomicUnitOfVelocity}
\alias{kNISTnonSIelectronVolt}
\alias{kNISTnonSInaturalUnitOfAction}
\alias{kNISTnonSInaturalUnitOfActionInEVS}
\alias{kNISTnonSInaturalUnitOfEnergy}
\alias{kNISTnonSInaturalUnitOfEnergyInMeV}
\alias{kNISTnonSInaturalUnitOfLength}
\alias{kNISTnonSInaturalUnitOfMass}
\alias{kNISTnonSInaturalUnitOfMomentum}
\alias{kNISTnonSInaturalUnitOfMomentumInMeVc}
\alias{kNISTnonSInaturalUnitOfTime}
\alias{kNISTnonSIPlanckConstantOver2PiTimesCInMeVFm}
\alias{kNISTnonSIspeedOfLightInVacuum}
\alias{kNISTnonSIunifiedAtomicMassUnit}\title{NIST Physical Constants}
\usage{kNISTnonSIatomicUnitOf1stHyperpolarizability}
\description{
\code{kNISTnonSIatomicUnitOf1stHyperpolarizability} atomic unit of 1st hyperpolarizability 

\code{kNISTnonSIatomicUnitOf2ndHyperpolarizability} atomic unit of 2nd hyperpolarizability 

\code{kNISTnonSIatomicUnitOfAction} atomic unit of action 

\code{kNISTnonSIatomicUnitOfCharge} atomic unit of charge 

\code{kNISTnonSIatomicUnitOfChargeDensity} atomic unit of charge density 

\code{kNISTnonSIatomicUnitOfCurrent} atomic unit of current 

\code{kNISTnonSIatomicUnitOfElectricDipoleMoment} atomic unit of electric dipole moment 

\code{kNISTnonSIatomicUnitOfElectricField} atomic unit of electric field 

\code{kNISTnonSIatomicUnitOfElectricFieldGradient} atomic unit of electric field gradient 

\code{kNISTnonSIatomicUnitOfElectricPolarizability} atomic unit of electric polarizability 

\code{kNISTnonSIatomicUnitOfElectricPotential} atomic unit of electric potential 

\code{kNISTnonSIatomicUnitOfElectricQuadrupoleMoment} atomic unit of electric quadrupole moment 

\code{kNISTnonSIatomicUnitOfEnergy} atomic unit of energy 

\code{kNISTnonSIatomicUnitOfLength} atomic unit of length 

\code{kNISTnonSIatomicUnitOfMagneticDipoleMoment} atomic unit of magnetic dipole moment 

\code{kNISTnonSIatomicUnitOfMagneticFluxDensity} atomic unit of magnetic flux density 

\code{kNISTnonSIatomicUnitOfMagnetizability} atomic unit of magnetizability 

\code{kNISTnonSIatomicUnitOfMass} atomic unit of mass 

\code{kNISTnonSIatomicUnitOfMomentum} atomic unit of momentum 

\code{kNISTnonSIatomicUnitOfPermittivity} atomic unit of permittivity 

\code{kNISTnonSIatomicUnitOfTime} atomic unit of time 

\code{kNISTnonSIatomicUnitOfVelocity} atomic unit of velocity 

\code{kNISTnonSIelectronVolt} electron volt 

\code{kNISTnonSInaturalUnitOfAction} natural unit of action 

\code{kNISTnonSInaturalUnitOfActionInEVS} natural unit of action in eV s 

\code{kNISTnonSInaturalUnitOfEnergy} natural unit of energy 

\code{kNISTnonSInaturalUnitOfEnergyInMeV} natural unit of energy in MeV 

\code{kNISTnonSInaturalUnitOfLength} natural unit of length 

\code{kNISTnonSInaturalUnitOfMass} natural unit of mass 

\code{kNISTnonSInaturalUnitOfMomentum} natural unit of momentum 

\code{kNISTnonSInaturalUnitOfMomentumInMeVc} natural unit of momentum in MeV/c 

\code{kNISTnonSInaturalUnitOfTime} natural unit of time 

\code{kNISTnonSIPlanckConstantOver2PiTimesCInMeVFm} Planck constant over 2 pi times c in MeV fm 

\code{kNISTnonSIspeedOfLightInVacuum} speed of light in vacuum 

\code{kNISTnonSIunifiedAtomicMassUnit} unified atomic mass unit}
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/cgi-bin/cuu/Value?auhypol|search_for=nonsi_in!}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/cgi-bin/cuu/Value?auhypol|search_for=nonsi_in!}
}
\author{Jose Gama}
\examples{
\dontrun{print(kNISTnonSIatomicUnitOf1stHyperpolarizability)}
}
\keyword{datasets}
