\name{NISTabampereTOampere}
\alias{NISTabampereTOampere}
\title{Convert abampere to ampere }
\usage{NISTabampereTOampere(abampere)}
\description{\code{NISTabampereTOampere} converts from abampere to ampere (A) }
\arguments{
  \item{abampere}{abampere }
}
\value{ampere (A) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTabampereTOampere(10)
}
\keyword{programming}