\name{NISTabcoulombTOcoulomb}
\alias{NISTabcoulombTOcoulomb}
\title{Convert abcoulomb to coulomb }
\usage{NISTabcoulombTOcoulomb(abcoulomb)}
\description{\code{NISTabcoulombTOcoulomb} converts from abcoulomb to coulomb (C) }
\arguments{
  \item{abcoulomb}{abcoulomb }
}
\value{coulomb (C) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTabcoulombTOcoulomb(10)
}
\keyword{programming}