\name{NISTabfaradTOfarad}
\alias{NISTabfaradTOfarad}
\title{Convert abfarad to farad }
\usage{NISTabfaradTOfarad(abfarad)}
\description{\code{NISTabfaradTOfarad} converts from abfarad to farad (F) }
\arguments{
  \item{abfarad}{abfarad }
}
\value{farad (F) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTabfaradTOfarad(10)
}
\keyword{programming}