\name{NISTabhenryTOhenry}
\alias{NISTabhenryTOhenry}
\title{Convert abhenry to henry }
\usage{NISTabhenryTOhenry(abhenry)}
\description{\code{NISTabhenryTOhenry} converts from abhenry to henry (H) }
\arguments{
  \item{abhenry}{abhenry }
}
\value{henry (H) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTabhenryTOhenry(10)
}
\keyword{programming}