\name{NISTabmhoTOsiemens}
\alias{NISTabmhoTOsiemens}
\title{Convert abmho to siemens }
\usage{NISTabmhoTOsiemens(abmho)}
\description{\code{NISTabmhoTOsiemens} converts from abmho to siemens (S) }
\arguments{
  \item{abmho}{abmho }
}
\value{siemens (S) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTabmhoTOsiemens(10)
}
\keyword{programming}