\name{NISTabohmTOohm}
\alias{NISTabohmTOohm}
\title{Convert abohm to ohm }
\usage{NISTabohmTOohm(abohm)}
\description{\code{NISTabohmTOohm} converts from abohm to ohm (ohm) }
\arguments{
  \item{abohm}{abohm }
}
\value{ohm (ohm) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTabohmTOohm(10)
}
\keyword{programming}