\name{NISTabvoltTOvolt}
\alias{NISTabvoltTOvolt}
\title{Convert abvolt to volt }
\usage{NISTabvoltTOvolt(abvolt)}
\description{\code{NISTabvoltTOvolt} converts from abvolt to volt (V) }
\arguments{
  \item{abvolt}{abvolt }
}
\value{volt (V) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTabvoltTOvolt(10)
}
\keyword{programming}