\name{NISTaccelOfFreeFallStdTOmeterPerSecSqrd}
\alias{NISTaccelOfFreeFallStdTOmeterPerSecSqrd}
\title{Convert acceleration of free fall, standard  to meter per second squared }
\usage{NISTaccelOfFreeFallStdTOmeterPerSecSqrd(accelOfFreeFallStd)}
\description{\code{NISTaccelOfFreeFallStdTOmeterPerSecSqrd} converts from acceleration of free fall, standard (gn) to meter per second squared (m / s2) }
\arguments{
  \item{accelOfFreeFallStd}{acceleration of free fall, standard (gn) }
}
\value{meter per second squared (m / s2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTaccelOfFreeFallStdTOmeterPerSecSqrd(10)
}
\keyword{programming}