\name{NISTacreFtTOcubMeter}
\alias{NISTacreFtTOcubMeter}
\title{Convert acre foot 7 to cubic meter }
\usage{NISTacreFtTOcubMeter(acreFt)}
\description{\code{NISTacreFtTOcubMeter} converts from acre foot (based on U.S. survey foot) 7 to cubic meter (m3) }
\arguments{
  \item{acreFt}{acre foot (based on U.S. survey foot) 7 }
}
\value{cubic meter (m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTacreFtTOcubMeter(10)
}
\keyword{programming}