\name{NISTacreTOsqrMeter}
\alias{NISTacreTOsqrMeter}
\title{Convert acre 7 to square meter }
\usage{NISTacreTOsqrMeter(acre)}
\description{\code{NISTacreTOsqrMeter} converts from acre (based on U.S. survey foot) 7 to square meter (m2) }
\arguments{
  \item{acre}{acre (based on U.S. survey foot) 7 }
}
\value{square meter (m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTacreTOsqrMeter(10)
}
\keyword{programming}