\name{NISTampereHourTOcoulomb}
\alias{NISTampereHourTOcoulomb}
\title{Convert ampere hour  to coulomb }
\usage{NISTampereHourTOcoulomb(ampereHour)}
\description{\code{NISTampereHourTOcoulomb} converts from ampere hour (A * h) to coulomb (C) }
\arguments{
  \item{ampereHour}{ampere hour (A * h) }
}
\value{coulomb (C) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTampereHourTOcoulomb(10)
}
\keyword{programming}