\name{NISTamperePerMeterTOoersted}
\alias{NISTamperePerMeterTOoersted}
\title{Convert ampere per meter  to oersted }
\usage{NISTamperePerMeterTOoersted(amperePerMeter)}
\description{\code{NISTamperePerMeterTOoersted} converts from ampere per meter (A/m) to oersted (Oe) }
\arguments{
  \item{amperePerMeter}{ampere per meter (A/m) }
}
\value{oersted (Oe) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTamperePerMeterTOoersted(10)
}
\keyword{programming}