\name{NISTampereTOabampere}
\alias{NISTampereTOabampere}
\title{Convert ampere  to abampere}
\usage{NISTampereTOabampere(ampere)}
\description{\code{NISTampereTOabampere} converts from ampere (A) to abampere }
\arguments{
  \item{ampere}{ampere (A) }
}
\value{abampere }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTampereTOabampere(10)
}
\keyword{programming}