\name{NISTampereTObiot}
\alias{NISTampereTObiot}
\title{Convert ampere  to biot }
\usage{NISTampereTObiot(ampere)}
\description{\code{NISTampereTObiot} converts from ampere (A) to biot (Bi) }
\arguments{
  \item{ampere}{ampere (A) }
}
\value{biot (Bi) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTampereTObiot(10)
}
\keyword{programming}