\name{NISTampereToEMUOfCurrent}
\alias{NISTampereToEMUOfCurrent}
\title{Convert ampere  to EMU of current }
\usage{NISTampereToEMUOfCurrent(ampere)}
\description{\code{NISTampereToEMUOfCurrent} converts from ampere (A) to EMU of current (abampere) }
\arguments{
  \item{ampere}{ampere (A) }
}
\value{EMU of current (abampere) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTampereToEMUOfCurrent(10)
}
\keyword{programming}