\name{NISTampereToESUOfCurrent}
\alias{NISTampereToESUOfCurrent}
\title{Convert ampere  to ESU of current }
\usage{NISTampereToESUOfCurrent(ampere)}
\description{\code{NISTampereToESUOfCurrent} converts from ampere (A) to ESU of current (statampere) }
\arguments{
  \item{ampere}{ampere (A) }
}
\value{ESU of current (statampere) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTampereToESUOfCurrent(10)
}
\keyword{programming}