\name{NISTampereTOgilbert}
\alias{NISTampereTOgilbert}
\title{Convert ampere  to gilbert }
\usage{NISTampereTOgilbert(ampere)}
\description{\code{NISTampereTOgilbert} converts from ampere (A) to gilbert (Gi) }
\arguments{
  \item{ampere}{ampere (A) }
}
\value{gilbert (Gi) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTampereTOgilbert(10)
}
\keyword{programming}