\name{NISTampereTOstatampere}
\alias{NISTampereTOstatampere}
\title{Convert ampere  to statampere}
\usage{NISTampereTOstatampere(ampere)}
\description{\code{NISTampereTOstatampere} converts from ampere (A) to statampere }
\arguments{
  \item{ampere}{ampere (A) }
}
\value{statampere }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTampereTOstatampere(10)
}
\keyword{programming}