\name{NISTangstromTOmeter}
\alias{NISTangstromTOmeter}
\title{Convert angstrom  to meter }
\usage{NISTangstromTOmeter(angstrom)}
\description{\code{NISTangstromTOmeter} converts from angstrom (a) to meter (m) }
\arguments{
  \item{angstrom}{angstrom (a) }
}
\value{meter (m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTangstromTOmeter(10)
}
\keyword{programming}