\name{NISTangstromTOnanometer}
\alias{NISTangstromTOnanometer}
\title{Convert angstrom  to nanometer }
\usage{NISTangstromTOnanometer(angstrom)}
\description{\code{NISTangstromTOnanometer} converts from angstrom (a) to nanometer (nm) }
\arguments{
  \item{angstrom}{angstrom (a) }
}
\value{nanometer (nm) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTangstromTOnanometer(10)
}
\keyword{programming}