\name{NISTareTOsqrMeter}
\alias{NISTareTOsqrMeter}
\title{Convert are  to square meter }
\usage{NISTareTOsqrMeter(are)}
\description{\code{NISTareTOsqrMeter} converts from are (a) to square meter (m2) }
\arguments{
  \item{are}{are (a) }
}
\value{square meter (m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTareTOsqrMeter(10)
}
\keyword{programming}