\name{NISTastronomicalUnTOmeter}
\alias{NISTastronomicalUnTOmeter}
\title{Convert astronomical unit  to meter }
\usage{NISTastronomicalUnTOmeter(astronomicalUn)}
\description{\code{NISTastronomicalUnTOmeter} converts from astronomical unit (ua) to meter (m) }
\arguments{
  \item{astronomicalUn}{astronomical unit (ua) }
}
\value{meter (m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTastronomicalUnTOmeter(10)
}
\keyword{programming}