\name{NISTatmosphereStdTOkpascal}
\alias{NISTatmosphereStdTOkpascal}
\title{Convert atmosphere, standard  to kilopascal }
\usage{NISTatmosphereStdTOkpascal(atmosphereStd)}
\description{\code{NISTatmosphereStdTOkpascal} converts from atmosphere, standard (atm) to kilopascal (kPa) }
\arguments{
  \item{atmosphereStd}{atmosphere, standard (atm) }
}
\value{kilopascal (kPa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTatmosphereStdTOkpascal(10)
}
\keyword{programming}