\name{NISTatmosphereTechnicalTOkpascal}
\alias{NISTatmosphereTechnicalTOkpascal}
\title{Convert atmosphere, technical 8 to kilopascal }
\usage{NISTatmosphereTechnicalTOkpascal(atmosphereTechnical)}
\description{\code{NISTatmosphereTechnicalTOkpascal} converts from atmosphere, technical (at) 8 to kilopascal (kPa) }
\arguments{
  \item{atmosphereTechnical}{atmosphere, technical (at) 8 }
}
\value{kilopascal (kPa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTatmosphereTechnicalTOkpascal(10)
}
\keyword{programming}