\name{NISTatmosphereTechnicalTOpascal}
\alias{NISTatmosphereTechnicalTOpascal}
\title{Convert atmosphere, technical 8 to pascal }
\usage{NISTatmosphereTechnicalTOpascal(atmosphereTechnical)}
\description{\code{NISTatmosphereTechnicalTOpascal} converts from atmosphere, technical (at) 8 to pascal (Pa) }
\arguments{
  \item{atmosphereTechnical}{atmosphere, technical (at) 8 }
}
\value{pascal (Pa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTatmosphereTechnicalTOpascal(10)
}
\keyword{programming}