\name{NISTbarnTOsqrMeter}
\alias{NISTbarnTOsqrMeter}
\title{Convert barn  to square meter }
\usage{NISTbarnTOsqrMeter(barn)}
\description{\code{NISTbarnTOsqrMeter} converts from barn (b) to square meter (m2) }
\arguments{
  \item{barn}{barn (b) }
}
\value{square meter (m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTbarnTOsqrMeter(10)
}
\keyword{programming}