\name{NISTbarrelTOcubMeter}
\alias{NISTbarrelTOcubMeter}
\title{Convert barrel  to cubic meter }
\usage{NISTbarrelTOcubMeter(barrel)}
\description{\code{NISTbarrelTOcubMeter} converts from barrel [for petroleum, 42 gallons (U.S.)] (bbl) to cubic meter (m3) }
\arguments{
  \item{barrel}{barrel [for petroleum, 42 gallons (U.S.)] (bbl) }
}
\value{cubic meter (m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTbarrelTOcubMeter(10)
}
\keyword{programming}