\name{NISTbarrelTOliter}
\alias{NISTbarrelTOliter}
\title{Convert barrel  to liter }
\usage{NISTbarrelTOliter(barrel)}
\description{\code{NISTbarrelTOliter} converts from barrel [for petroleum, 42 gallons (U.S.)] (bbl) to liter (L) }
\arguments{
  \item{barrel}{barrel [for petroleum, 42 gallons (U.S.)] (bbl) }
}
\value{liter (L) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTbarrelTOliter(10)
}
\keyword{programming}