\name{NISTbarTOkpascal}
\alias{NISTbarTOkpascal}
\title{Convert bar  to kilopascal }
\usage{NISTbarTOkpascal(bar)}
\description{\code{NISTbarTOkpascal} converts from bar (bar) to kilopascal (kPa) }
\arguments{
  \item{bar}{bar (bar) }
}
\value{kilopascal (kPa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTbarTOkpascal(10)
}
\keyword{programming}