\name{NISTbecquerelTOcurie}
\alias{NISTbecquerelTOcurie}
\title{Convert becquerel  to curie }
\usage{NISTbecquerelTOcurie(becquerel)}
\description{\code{NISTbecquerelTOcurie} converts from becquerel (Bq) to curie (Ci) }
\arguments{
  \item{becquerel}{becquerel (Bq) }
}
\value{curie (Ci) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTbecquerelTOcurie(10)
}
\keyword{programming}