\name{NISTbiotTOampere}
\alias{NISTbiotTOampere}
\title{Convert biot  to ampere }
\usage{NISTbiotTOampere(biot)}
\description{\code{NISTbiotTOampere} converts from biot (Bi) to ampere (A) }
\arguments{
  \item{biot}{biot (Bi) }
}
\value{ampere (A) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTbiotTOampere(10)
}
\keyword{programming}