\name{NISTbushelTOcubMeter}
\alias{NISTbushelTOcubMeter}
\title{Convert bushel  to cubic meter }
\usage{NISTbushelTOcubMeter(bushel)}
\description{\code{NISTbushelTOcubMeter} converts from bushel (U.S.) (bu) to cubic meter (m3) }
\arguments{
  \item{bushel}{bushel (U.S.) (bu) }
}
\value{cubic meter (m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTbushelTOcubMeter(10)
}
\keyword{programming}