\name{NISTbushelTOliter}
\alias{NISTbushelTOliter}
\title{Convert bushel  to liter }
\usage{NISTbushelTOliter(bushel)}
\description{\code{NISTbushelTOliter} converts from bushel (U.S.) (bu) to liter (L) }
\arguments{
  \item{bushel}{bushel (U.S.) (bu) }
}
\value{liter (L) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTbushelTOliter(10)
}
\keyword{programming}