\name{NISTcal15CtOjoule}
\alias{NISTcal15CtOjoule}
\title{Convert calorie  to joule }
\usage{NISTcal15CtOjoule(cal15C)}
\description{\code{NISTcal15CtOjoule} converts from calorie (15 C) (cal15) to joule (J) }
\arguments{
  \item{cal15C}{calorie (15 C) (cal15) }
}
\value{joule (J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcal15CtOjoule(10)
}
\keyword{programming}