\name{NISTcal20CtOjoule}
\alias{NISTcal20CtOjoule}
\title{Convert calorie  to joule }
\usage{NISTcal20CtOjoule(cal20C)}
\description{\code{NISTcal20CtOjoule} converts from calorie (20 C) (cal20) to joule (J) }
\arguments{
  \item{cal20C}{calorie (20 C) (cal20) }
}
\value{joule (J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcal20CtOjoule(10)
}
\keyword{programming}