\name{NISTcalITKgTOjoule}
\alias{NISTcalITKgTOjoule}
\title{Convert calorieIT, kilogram 11 to joule }
\usage{NISTcalITKgTOjoule(calITKg)}
\description{\code{NISTcalITKgTOjoule} converts from calorieIT, kilogram (nutrition) 11 to joule (J) }
\arguments{
  \item{calITKg}{calorieIT, kilogram (nutrition) 11 }
}
\value{joule (J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcalITKgTOjoule(10)
}
\keyword{programming}