\name{NISTcalITPerGramDegCtOjoulePerKgK}
\alias{NISTcalITPerGramDegCtOjoulePerKgK}
\title{Convert calorieIT per gram degree Celsius  to joule per kilogram kelvin }
\usage{NISTcalITPerGramDegCtOjoulePerKgK(calITPerGramDegC)}
\description{\code{NISTcalITPerGramDegCtOjoulePerKgK} converts from calorieIT per gram degree Celsius [calIT/(g * C)] to joule per kilogram kelvin [J/(kg * K)] }
\arguments{
  \item{calITPerGramDegC}{calorieIT per gram degree Celsius [calIT/(g * C)] }
}
\value{joule per kilogram kelvin [J/(kg * K)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcalITPerGramDegCtOjoulePerKgK(10)
}
\keyword{programming}