\name{NISTcalITPerGramKtOjoulePerKgK}
\alias{NISTcalITPerGramKtOjoulePerKgK}
\title{Convert calorieIT per gram kelvin  to joule per kilogram kelvin }
\usage{NISTcalITPerGramKtOjoulePerKgK(calITPerGramK)}
\description{\code{NISTcalITPerGramKtOjoulePerKgK} converts from calorieIT per gram kelvin [calIT/(g * K)] to joule per kilogram kelvin [J/(kg * K)] }
\arguments{
  \item{calITPerGramK}{calorieIT per gram kelvin [calIT/(g * K)] }
}
\value{joule per kilogram kelvin [J/(kg * K)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcalITPerGramKtOjoulePerKgK(10)
}
\keyword{programming}