\name{NISTcalITPerGramTOjoulePerKg}
\alias{NISTcalITPerGramTOjoulePerKg}
\title{Convert calorieIT per gram  to joule per kilogram }
\usage{NISTcalITPerGramTOjoulePerKg(calITPerGram)}
\description{\code{NISTcalITPerGramTOjoulePerKg} converts from calorieIT per gram (calIT/g) to joule per kilogram (J/kg) }
\arguments{
  \item{calITPerGram}{calorieIT per gram (calIT/g) }
}
\value{joule per kilogram (J/kg) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcalITPerGramTOjoulePerKg(10)
}
\keyword{programming}