\name{NISTcalITtOjoule}
\alias{NISTcalITtOjoule}
\title{Convert calorieIT 10 to joule }
\usage{NISTcalITtOjoule(calIT)}
\description{\code{NISTcalITtOjoule} converts from calorieIT (calIT) 10 to joule (J) }
\arguments{
  \item{calIT}{calorieIT (calIT) 10 }
}
\value{joule (J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcalITtOjoule(10)
}
\keyword{programming}