\name{NISTcalKgTOjoule}
\alias{NISTcalKgTOjoule}
\title{Convert calorie , kilogram 11 to joule }
\usage{NISTcalKgTOjoule(calKg)}
\description{\code{NISTcalKgTOjoule} converts from calorie (mean), kilogram (nutrition) 11 to joule (J) }
\arguments{
  \item{calKg}{calorie (mean), kilogram (nutrition) 11 }
}
\value{joule (J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcalKgTOjoule(10)
}
\keyword{programming}