\name{NISTcalthKgTOjoule}
\alias{NISTcalthKgTOjoule}
\title{Convert calorieth , kilogram 11 to joule }
\usage{NISTcalthKgTOjoule(calthKg)}
\description{\code{NISTcalthKgTOjoule} converts from calorieth , kilogram (nutrition) 11 to joule (J) }
\arguments{
  \item{calthKg}{calorieth , kilogram (nutrition) 11 }
}
\value{joule (J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcalthKgTOjoule(10)
}
\keyword{programming}