\name{NISTcalthPerCmSecDegCtOwattPerMeterK}
\alias{NISTcalthPerCmSecDegCtOwattPerMeterK}
\title{Convert calorieth per centimeter second degree Celsius  to watt per meter kelvin }
\usage{NISTcalthPerCmSecDegCtOwattPerMeterK(calthPerCmSecDegC)}
\description{\code{NISTcalthPerCmSecDegCtOwattPerMeterK} converts from calorieth per centimeter second degree Celsius [calth/(cm * s * C)] to watt per meter kelvin [W/(m * K)] }
\arguments{
  \item{calthPerCmSecDegC}{calorieth per centimeter second degree Celsius [calth/(cm * s * C)] }
}
\value{watt per meter kelvin [W/(m * K)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcalthPerCmSecDegCtOwattPerMeterK(10)
}
\keyword{programming}