\name{NISTcalthPerGramDegCtOjoulePerKgK}
\alias{NISTcalthPerGramDegCtOjoulePerKgK}
\title{Convert calorieth per gram degree Celsius  to joule per kilogram kelvin }
\usage{NISTcalthPerGramDegCtOjoulePerKgK(calthPerGramDegC)}
\description{\code{NISTcalthPerGramDegCtOjoulePerKgK} converts from calorieth per gram degree Celsius [calth/(g * C)] to joule per kilogram kelvin [J/(kg * K)] }
\arguments{
  \item{calthPerGramDegC}{calorieth per gram degree Celsius [calth/(g * C)] }
}
\value{joule per kilogram kelvin [J/(kg * K)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcalthPerGramDegCtOjoulePerKgK(10)
}
\keyword{programming}