\name{NISTcalthPerGramKtOjoulePerKgK}
\alias{NISTcalthPerGramKtOjoulePerKgK}
\title{Convert calorieth per gram kelvin  to joule per kilogram kelvin }
\usage{NISTcalthPerGramKtOjoulePerKgK(calthPerGramK)}
\description{\code{NISTcalthPerGramKtOjoulePerKgK} converts from calorieth per gram kelvin [calth/(g* K)] to joule per kilogram kelvin [J/(kg * K)] }
\arguments{
  \item{calthPerGramK}{calorieth per gram kelvin [calth/(g* K)] }
}
\value{joule per kilogram kelvin [J/(kg * K)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcalthPerGramKtOjoulePerKgK(10)
}
\keyword{programming}