\name{NISTcalthPerGramTOjoulePerKg}
\alias{NISTcalthPerGramTOjoulePerKg}
\title{Convert calorieth per gram  to joule per kilogram }
\usage{NISTcalthPerGramTOjoulePerKg(calthPerGram)}
\description{\code{NISTcalthPerGramTOjoulePerKg} converts from calorieth per gram (calth/g) to joule per kilogram (J/kg) }
\arguments{
  \item{calthPerGram}{calorieth per gram (calth/g) }
}
\value{joule per kilogram (J/kg) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcalthPerGramTOjoulePerKg(10)
}
\keyword{programming}