\name{NISTcalthPerMinTOwatt}
\alias{NISTcalthPerMinTOwatt}
\title{Convert calorieth per minute  to watt }
\usage{NISTcalthPerMinTOwatt(calthPerMin)}
\description{\code{NISTcalthPerMinTOwatt} converts from calorieth per minute (calth/min) to watt (W) }
\arguments{
  \item{calthPerMin}{calorieth per minute (calth/min) }
}
\value{watt (W) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcalthPerMinTOwatt(10)
}
\keyword{programming}