\name{NISTcalthPerSecTOwatt}
\alias{NISTcalthPerSecTOwatt}
\title{Convert calorieth per second  to watt }
\usage{NISTcalthPerSecTOwatt(calthPerSec)}
\description{\code{NISTcalthPerSecTOwatt} converts from calorieth per second (calth/s) to watt (W) }
\arguments{
  \item{calthPerSec}{calorieth per second (calth/s) }
}
\value{watt (W) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcalthPerSecTOwatt(10)
}
\keyword{programming}