\name{NISTcalthPerSqrCmMinTOwattPerSqrMeter}
\alias{NISTcalthPerSqrCmMinTOwattPerSqrMeter}
\title{Convert calorieth per square centimeter minute  to watt per square meter }
\usage{NISTcalthPerSqrCmMinTOwattPerSqrMeter(calthPerSqrCmMin)}
\description{\code{NISTcalthPerSqrCmMinTOwattPerSqrMeter} converts from calorieth per square centimeter minute [calth/(cm2 * min)] to watt per square meter (W/m2) }
\arguments{
  \item{calthPerSqrCmMin}{calorieth per square centimeter minute [calth/(cm2 * min)] }
}
\value{watt per square meter (W/m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcalthPerSqrCmMinTOwattPerSqrMeter(10)
}
\keyword{programming}