\name{NISTcalthPerSqrCmSecTOwattPerSqrMeter}
\alias{NISTcalthPerSqrCmSecTOwattPerSqrMeter}
\title{Convert calorieth per square centimeter second  to watt per square meter }
\usage{NISTcalthPerSqrCmSecTOwattPerSqrMeter(calthPerSqrCmSec)}
\description{\code{NISTcalthPerSqrCmSecTOwattPerSqrMeter} converts from calorieth per square centimeter second [calth/(cm2 * s)] to watt per square meter (W/m2) }
\arguments{
  \item{calthPerSqrCmSec}{calorieth per square centimeter second [calth/(cm2 * s)] }
}
\value{watt per square meter (W/m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcalthPerSqrCmSecTOwattPerSqrMeter(10)
}
\keyword{programming}