\name{NISTcalthPerSqrCmTOjoulePerSqrMeter}
\alias{NISTcalthPerSqrCmTOjoulePerSqrMeter}
\title{Convert calorieth per square centimeter  to joule per square meter }
\usage{NISTcalthPerSqrCmTOjoulePerSqrMeter(calthPerSqrCm)}
\description{\code{NISTcalthPerSqrCmTOjoulePerSqrMeter} converts from calorieth per square centimeter (calth/cm2) to joule per square meter (J/m2) }
\arguments{
  \item{calthPerSqrCm}{calorieth per square centimeter (calth/cm2) }
}
\value{joule per square meter (J/m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcalthPerSqrCmTOjoulePerSqrMeter(10)
}
\keyword{programming}