\name{NISTcalthTOjoule}
\alias{NISTcalthTOjoule}
\title{Convert calorieth 10 to joule }
\usage{NISTcalthTOjoule(calth)}
\description{\code{NISTcalthTOjoule} converts from calorieth (calth) 10 to joule (J) }
\arguments{
  \item{calth}{calorieth (calth) 10 }
}
\value{joule (J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcalthTOjoule(10)
}
\keyword{programming}