\name{NISTcalTOjoule}
\alias{NISTcalTOjoule}
\title{Convert calorie  to joule }
\usage{NISTcalTOjoule(cal)}
\description{\code{NISTcalTOjoule} converts from calorie (cal) (mean) to joule (J) }
\arguments{
  \item{cal}{calorie (cal) (mean) }
}
\value{joule (J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcalTOjoule(10)
}
\keyword{programming}