\name{NISTcandelaPerSqrInchTOcandelaPerSqrMeter}
\alias{NISTcandelaPerSqrInchTOcandelaPerSqrMeter}
\title{Convert candela per square inch  to candela per square meter }
\usage{NISTcandelaPerSqrInchTOcandelaPerSqrMeter(candelaPerSqrInch)}
\description{\code{NISTcandelaPerSqrInchTOcandelaPerSqrMeter} converts from candela per square inch (cd/in2) to candela per square meter (cd/m2) }
\arguments{
  \item{candelaPerSqrInch}{candela per square inch (cd/in2) }
}
\value{candela per square meter (cd/m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcandelaPerSqrInchTOcandelaPerSqrMeter(10)
}
\keyword{programming}