\name{NISTcandelaPerSqrMeterTOcandelaPerSqrInch}
\alias{NISTcandelaPerSqrMeterTOcandelaPerSqrInch}
\title{Convert candela per square meter  to candela per square inch }
\usage{NISTcandelaPerSqrMeterTOcandelaPerSqrInch(candelaPerSqrMeter)}
\description{\code{NISTcandelaPerSqrMeterTOcandelaPerSqrInch} converts from candela per square meter (cd/m2) to candela per square inch (cd/in2) }
\arguments{
  \item{candelaPerSqrMeter}{candela per square meter (cd/m2) }
}
\value{candela per square inch (cd/in2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcandelaPerSqrMeterTOcandelaPerSqrInch(10)
}
\keyword{programming}