\name{NISTcandelaPerSqrMeterTOftlambert}
\alias{NISTcandelaPerSqrMeterTOftlambert}
\title{Convert candela per square meter  to footlambert}
\usage{NISTcandelaPerSqrMeterTOftlambert(candelaPerSqrMeter)}
\description{\code{NISTcandelaPerSqrMeterTOftlambert} converts from candela per square meter (cd/m2) to footlambert }
\arguments{
  \item{candelaPerSqrMeter}{candela per square meter (cd/m2) }
}
\value{footlambert }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcandelaPerSqrMeterTOftlambert(10)
}
\keyword{programming}