\name{NISTcandelaPerSqrMeterTOlambert}
\alias{NISTcandelaPerSqrMeterTOlambert}
\title{Convert candela per square meter  to lambert 17}
\usage{NISTcandelaPerSqrMeterTOlambert(candelaPerSqrMeter)}
\description{\code{NISTcandelaPerSqrMeterTOlambert} converts from candela per square meter (cd/m2) to lambert 17 }
\arguments{
  \item{candelaPerSqrMeter}{candela per square meter (cd/m2) }
}
\value{lambert 17 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcandelaPerSqrMeterTOlambert(10)
}
\keyword{programming}