\name{NISTcandelaPerSqrMeterTOstilb}
\alias{NISTcandelaPerSqrMeterTOstilb}
\title{Convert candela per square meter  to stilb }
\usage{NISTcandelaPerSqrMeterTOstilb(candelaPerSqrMeter)}
\description{\code{NISTcandelaPerSqrMeterTOstilb} converts from candela per square meter (cd/m2) to stilb (sb) }
\arguments{
  \item{candelaPerSqrMeter}{candela per square meter (cd/m2) }
}
\value{stilb (sb) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcandelaPerSqrMeterTOstilb(10)
}
\keyword{programming}