\name{NISTcaratSItOgram}
\alias{NISTcaratSItOgram}
\title{Convert carat, metric to gram }
\usage{NISTcaratSItOgram(caratSI)}
\description{\code{NISTcaratSItOgram} converts from carat, metric to gram (g) }
\arguments{
  \item{caratSI}{carat, metric }
}
\value{gram (g) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcaratSItOgram(10)
}
\keyword{programming}