\name{NISTcaratSItOkg}
\alias{NISTcaratSItOkg}
\title{Convert carat, metric to kilogram }
\usage{NISTcaratSItOkg(caratSI)}
\description{\code{NISTcaratSItOkg} converts from carat, metric to kilogram (kg) }
\arguments{
  \item{caratSI}{carat, metric }
}
\value{kilogram (kg) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcaratSItOkg(10)
}
\keyword{programming}