\name{NISTcentipoiseTOpascalSec}
\alias{NISTcentipoiseTOpascalSec}
\title{Convert centipoise  to pascal second }
\usage{NISTcentipoiseTOpascalSec(centipoise)}
\description{\code{NISTcentipoiseTOpascalSec} converts from centipoise (cP) to pascal second (Pa * s) }
\arguments{
  \item{centipoise}{centipoise (cP) }
}
\value{pascal second (Pa * s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcentipoiseTOpascalSec(10)
}
\keyword{programming}