\name{NISTcentistokesTOmeterSqrdPerSec}
\alias{NISTcentistokesTOmeterSqrdPerSec}
\title{Convert centistokes  to meter squared per second }
\usage{NISTcentistokesTOmeterSqrdPerSec(centistokes)}
\description{\code{NISTcentistokesTOmeterSqrdPerSec} converts from centistokes (cSt) to meter squared per second (m2/s) }
\arguments{
  \item{centistokes}{centistokes (cSt) }
}
\value{meter squared per second (m2/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcentistokesTOmeterSqrdPerSec(10)
}
\keyword{programming}