\name{NISTchainTOmeter}
\alias{NISTchainTOmeter}
\title{Convert chain 7 to meter }
\usage{NISTchainTOmeter(chain)}
\description{\code{NISTchainTOmeter} converts from chain (based on U.S. survey foot) (ch) 7 to meter (m) }
\arguments{
  \item{chain}{chain (based on U.S. survey foot) (ch) 7 }
}
\value{meter (m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTchainTOmeter(10)
}
\keyword{programming}