\name{NISTcircularMilTOsqrMeter}
\alias{NISTcircularMilTOsqrMeter}
\title{Convert circular mil to square meter }
\usage{NISTcircularMilTOsqrMeter(circularMil)}
\description{\code{NISTcircularMilTOsqrMeter} converts from circular mil to square meter (m2) }
\arguments{
  \item{circularMil}{circular mil }
}
\value{square meter (m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcircularMilTOsqrMeter(10)
}
\keyword{programming}