\name{NISTcircularMilTOsqrMillimeter}
\alias{NISTcircularMilTOsqrMillimeter}
\title{Convert circular mil to square millimeter }
\usage{NISTcircularMilTOsqrMillimeter(circularMil)}
\description{\code{NISTcircularMilTOsqrMillimeter} converts from circular mil to square millimeter (mm2) }
\arguments{
  \item{circularMil}{circular mil }
}
\value{square millimeter (mm2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcircularMilTOsqrMillimeter(10)
}
\keyword{programming}