\name{NISTcloTOsqrMeterKPerWatt}
\alias{NISTcloTOsqrMeterKPerWatt}
\title{Convert clo to square meter kelvin per watt }
\usage{NISTcloTOsqrMeterKPerWatt(clo)}
\description{\code{NISTcloTOsqrMeterKPerWatt} converts from clo to square meter kelvin per watt (m2 * K/W) }
\arguments{
  \item{clo}{clo }
}
\value{square meter kelvin per watt (m2 * K/W) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcloTOsqrMeterKPerWatt(10)
}
\keyword{programming}