\name{NISTcmOfMercuryConvtnlTOkpascal}
\alias{NISTcmOfMercuryConvtnlTOkpascal}
\title{Convert centimeter of mercury, conventional 12 to kilopascal }
\usage{NISTcmOfMercuryConvtnlTOkpascal(cmOfMercuryConvtnl)}
\description{\code{NISTcmOfMercuryConvtnlTOkpascal} converts from centimeter of mercury, conventional (cmHg) 12 to kilopascal (kPa) }
\arguments{
  \item{cmOfMercuryConvtnl}{centimeter of mercury, conventional (cmHg) 12 }
}
\value{kilopascal (kPa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcmOfMercuryConvtnlTOkpascal(10)
}
\keyword{programming}