\name{NISTcmOfMercuryConvtnlTOpascal}
\alias{NISTcmOfMercuryConvtnlTOpascal}
\title{Convert centimeter of mercury, conventional 12 to pascal }
\usage{NISTcmOfMercuryConvtnlTOpascal(cmOfMercuryConvtnl)}
\description{\code{NISTcmOfMercuryConvtnlTOpascal} converts from centimeter of mercury, conventional (cmHg) 12 to pascal (Pa) }
\arguments{
  \item{cmOfMercuryConvtnl}{centimeter of mercury, conventional (cmHg) 12 }
}
\value{pascal (Pa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcmOfMercuryConvtnlTOpascal(10)
}
\keyword{programming}