\name{NISTcmOfMercuryTOkpascal}
\alias{NISTcmOfMercuryTOkpascal}
\title{Convert centimeter of mercury 12 to kilopascal }
\usage{NISTcmOfMercuryTOkpascal(cmOfMercury)}
\description{\code{NISTcmOfMercuryTOkpascal} converts from centimeter of mercury (0 C) 12 to kilopascal (kPa) }
\arguments{
  \item{cmOfMercury}{centimeter of mercury (0 C) 12 }
}
\value{kilopascal (kPa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcmOfMercuryTOkpascal(10)
}
\keyword{programming}