\name{NISTcmOfMercuryTOpascal}
\alias{NISTcmOfMercuryTOpascal}
\title{Convert centimeter of mercury 12 to pascal }
\usage{NISTcmOfMercuryTOpascal(cmOfMercury)}
\description{\code{NISTcmOfMercuryTOpascal} converts from centimeter of mercury (0 C) 12 to pascal (Pa) }
\arguments{
  \item{cmOfMercury}{centimeter of mercury (0 C) 12 }
}
\value{pascal (Pa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcmOfMercuryTOpascal(10)
}
\keyword{programming}