\name{NISTcmOfWaterTOpascal}
\alias{NISTcmOfWaterTOpascal}
\title{Convert centimeter of water 12 to pascal }
\usage{NISTcmOfWaterTOpascal(cmOfWater)}
\description{\code{NISTcmOfWaterTOpascal} converts from centimeter of water (4 C) 12 to pascal (Pa) }
\arguments{
  \item{cmOfWater}{centimeter of water (4 C) 12 }
}
\value{pascal (Pa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcmOfWaterTOpascal(10)
}
\keyword{programming}