\name{NISTcmTOinch}
\alias{NISTcmTOinch}
\title{Convert centimeter  to inch }
\usage{NISTcmTOinch(cm)}
\description{\code{NISTcmTOinch} converts from centimeter (cm) to inch (in) }
\arguments{
  \item{cm}{centimeter (cm) }
}
\value{inch (in) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcmTOinch(10)
}
\keyword{programming}