\name{NISTcordTOcubMeter}
\alias{NISTcordTOcubMeter}
\title{Convert cord  to cubic meter }
\usage{NISTcordTOcubMeter(cord)}
\description{\code{NISTcordTOcubMeter} converts from cord (128 ft3) to cubic meter (m3) }
\arguments{
  \item{cord}{cord (128 ft3) }
}
\value{cubic meter (m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcordTOcubMeter(10)
}
\keyword{programming}