\name{NISTcoulombMeterTOdebye}
\alias{NISTcoulombMeterTOdebye}
\title{Convert coulomb meter  to debye }
\usage{NISTcoulombMeterTOdebye(coulombMeter)}
\description{\code{NISTcoulombMeterTOdebye} converts from coulomb meter (C * m) to debye (D) }
\arguments{
  \item{coulombMeter}{coulomb meter (C * m) }
}
\value{debye (D) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcoulombMeterTOdebye(10)
}
\keyword{programming}