\name{NISTcoulombPerKgTOroentgen}
\alias{NISTcoulombPerKgTOroentgen}
\title{Convert coulomb per kilogram  to roentgen }
\usage{NISTcoulombPerKgTOroentgen(coulombPerKg)}
\description{\code{NISTcoulombPerKgTOroentgen} converts from coulomb per kilogram (C/kg) to roentgen (R) }
\arguments{
  \item{coulombPerKg}{coulomb per kilogram (C/kg) }
}
\value{roentgen (R) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcoulombPerKgTOroentgen(10)
}
\keyword{programming}