\name{NISTcoulombTOabcoulomb}
\alias{NISTcoulombTOabcoulomb}
\title{Convert coulomb  to abcoulomb}
\usage{NISTcoulombTOabcoulomb(coulomb)}
\description{\code{NISTcoulombTOabcoulomb} converts from coulomb (C) to abcoulomb }
\arguments{
  \item{coulomb}{coulomb (C) }
}
\value{abcoulomb }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcoulombTOabcoulomb(10)
}
\keyword{programming}