\name{NISTcoulombTOampereHour}
\alias{NISTcoulombTOampereHour}
\title{Convert coulomb  to ampere hour }
\usage{NISTcoulombTOampereHour(coulomb)}
\description{\code{NISTcoulombTOampereHour} converts from coulomb (C) to ampere hour (A * h) }
\arguments{
  \item{coulomb}{coulomb (C) }
}
\value{ampere hour (A * h) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcoulombTOampereHour(10)
}
\keyword{programming}