\name{NISTcoulombTOfranklin}
\alias{NISTcoulombTOfranklin}
\title{Convert coulomb  to franklin }
\usage{NISTcoulombTOfranklin(coulomb)}
\description{\code{NISTcoulombTOfranklin} converts from coulomb (C) to franklin (Fr) }
\arguments{
  \item{coulomb}{coulomb (C) }
}
\value{franklin (Fr) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcoulombTOfranklin(10)
}
\keyword{programming}