\name{NISTcoulombTOstatcoulomb}
\alias{NISTcoulombTOstatcoulomb}
\title{Convert coulomb  to statcoulomb}
\usage{NISTcoulombTOstatcoulomb(coulomb)}
\description{\code{NISTcoulombTOstatcoulomb} converts from coulomb (C) to statcoulomb }
\arguments{
  \item{coulomb}{coulomb (C) }
}
\value{statcoulomb }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcoulombTOstatcoulomb(10)
}
\keyword{programming}