\name{NISTcubFtPerMinTOcubMeterPerSec}
\alias{NISTcubFtPerMinTOcubMeterPerSec}
\title{Convert cubic foot per minute  to cubic meter per second }
\usage{NISTcubFtPerMinTOcubMeterPerSec(cubFtPerMin)}
\description{\code{NISTcubFtPerMinTOcubMeterPerSec} converts from cubic foot per minute (ft3/min) to cubic meter per second (m3/s) }
\arguments{
  \item{cubFtPerMin}{cubic foot per minute (ft3/min) }
}
\value{cubic meter per second (m3/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubFtPerMinTOcubMeterPerSec(10)
}
\keyword{programming}