\name{NISTcubFtPerMinTOliterPerSec}
\alias{NISTcubFtPerMinTOliterPerSec}
\title{Convert cubic foot per minute  to liter per second }
\usage{NISTcubFtPerMinTOliterPerSec(cubFtPerMin)}
\description{\code{NISTcubFtPerMinTOliterPerSec} converts from cubic foot per minute (ft3/min) to liter per second (L/s) }
\arguments{
  \item{cubFtPerMin}{cubic foot per minute (ft3/min) }
}
\value{liter per second (L/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubFtPerMinTOliterPerSec(10)
}
\keyword{programming}