\name{NISTcubFtPerSecTOcubMeterPerSec}
\alias{NISTcubFtPerSecTOcubMeterPerSec}
\title{Convert cubic foot per second  to cubic meter per second }
\usage{NISTcubFtPerSecTOcubMeterPerSec(cubFtPerSec)}
\description{\code{NISTcubFtPerSecTOcubMeterPerSec} converts from cubic foot per second (ft3/s) to cubic meter per second (m3/s) }
\arguments{
  \item{cubFtPerSec}{cubic foot per second (ft3/s) }
}
\value{cubic meter per second (m3/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubFtPerSecTOcubMeterPerSec(10)
}
\keyword{programming}