\name{NISTcubFtTOcubMeter}
\alias{NISTcubFtTOcubMeter}
\title{Convert cubic foot  to cubic meter }
\usage{NISTcubFtTOcubMeter(cubFt)}
\description{\code{NISTcubFtTOcubMeter} converts from cubic foot (ft3) to cubic meter (m3) }
\arguments{
  \item{cubFt}{cubic foot (ft3) }
}
\value{cubic meter (m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubFtTOcubMeter(10)
}
\keyword{programming}