\name{NISTcubInchPerMinTOcubMeterPerSec}
\alias{NISTcubInchPerMinTOcubMeterPerSec}
\title{Convert cubic inch per minute  to cubic meter per second }
\usage{NISTcubInchPerMinTOcubMeterPerSec(cubInchPerMin)}
\description{\code{NISTcubInchPerMinTOcubMeterPerSec} converts from cubic inch per minute (in3/min) to cubic meter per second (m3/s) }
\arguments{
  \item{cubInchPerMin}{cubic inch per minute (in3/min) }
}
\value{cubic meter per second (m3/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubInchPerMinTOcubMeterPerSec(10)
}
\keyword{programming}