\name{NISTcubInchTOcubMeter}
\alias{NISTcubInchTOcubMeter}
\title{Convert cubic inch 13 to cubic meter }
\usage{NISTcubInchTOcubMeter(cubInch)}
\description{\code{NISTcubInchTOcubMeter} converts from cubic inch (in3) 13 to cubic meter (m3) }
\arguments{
  \item{cubInch}{cubic inch (in3) 13 }
}
\value{cubic meter (m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubInchTOcubMeter(10)
}
\keyword{programming}