\name{NISTcubMeterPerJouleTOgallonPerHorsepowerHour}
\alias{NISTcubMeterPerJouleTOgallonPerHorsepowerHour}
\title{Convert cubic meter per joule  to gallon per horsepower hour }
\usage{NISTcubMeterPerJouleTOgallonPerHorsepowerHour(cubMeterPerJoule)}
\description{\code{NISTcubMeterPerJouleTOgallonPerHorsepowerHour} converts from cubic meter per joule (m3/J) to gallon (U.S.) per horsepower hour [gal/(hp * h)] }
\arguments{
  \item{cubMeterPerJoule}{cubic meter per joule (m3/J) }
}
\value{gallon (U.S.) per horsepower hour [gal/(hp * h)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubMeterPerJouleTOgallonPerHorsepowerHour(10)
}
\keyword{programming}