\name{NISTcubMeterPerSecTOcubFtPerSec}
\alias{NISTcubMeterPerSecTOcubFtPerSec}
\title{Convert cubic meter per second  to cubic foot per second }
\usage{NISTcubMeterPerSecTOcubFtPerSec(cubMeterPerSec)}
\description{\code{NISTcubMeterPerSecTOcubFtPerSec} converts from cubic meter per second (m3/s) to cubic foot per second (ft3/s) }
\arguments{
  \item{cubMeterPerSec}{cubic meter per second (m3/s) }
}
\value{cubic foot per second (ft3/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubMeterPerSecTOcubFtPerSec(10)
}
\keyword{programming}