\name{NISTcubMeterPerSecTOcubInchPerMin}
\alias{NISTcubMeterPerSecTOcubInchPerMin}
\title{Convert cubic meter per second  to cubic inch per minute }
\usage{NISTcubMeterPerSecTOcubInchPerMin(cubMeterPerSec)}
\description{\code{NISTcubMeterPerSecTOcubInchPerMin} converts from cubic meter per second (m3/s) to cubic inch per minute (in3/min) }
\arguments{
  \item{cubMeterPerSec}{cubic meter per second (m3/s) }
}
\value{cubic inch per minute (in3/min) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubMeterPerSecTOcubInchPerMin(10)
}
\keyword{programming}