\name{NISTcubMeterPerSecTOcubYardPerMin}
\alias{NISTcubMeterPerSecTOcubYardPerMin}
\title{Convert cubic meter per second  to cubic yard per minute }
\usage{NISTcubMeterPerSecTOcubYardPerMin(cubMeterPerSec)}
\description{\code{NISTcubMeterPerSecTOcubYardPerMin} converts from cubic meter per second (m3/s) to cubic yard per minute (yd3/min) }
\arguments{
  \item{cubMeterPerSec}{cubic meter per second (m3/s) }
}
\value{cubic yard per minute (yd3/min) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubMeterPerSecTOcubYardPerMin(10)
}
\keyword{programming}