\name{NISTcubMeterPerSecTOgallonPerDay}
\alias{NISTcubMeterPerSecTOgallonPerDay}
\title{Convert cubic meter per second  to gallon per day }
\usage{NISTcubMeterPerSecTOgallonPerDay(cubMeterPerSec)}
\description{\code{NISTcubMeterPerSecTOgallonPerDay} converts from cubic meter per second (m3/s) to gallon (U.S.) per day (gal/d) }
\arguments{
  \item{cubMeterPerSec}{cubic meter per second (m3/s) }
}
\value{gallon (U.S.) per day (gal/d) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubMeterPerSecTOgallonPerDay(10)
}
\keyword{programming}