\name{NISTcubMeterPerSecTOgallonPerMin}
\alias{NISTcubMeterPerSecTOgallonPerMin}
\title{Convert cubic meter per second  to gallon per minute }
\usage{NISTcubMeterPerSecTOgallonPerMin(cubMeterPerSec)}
\description{\code{NISTcubMeterPerSecTOgallonPerMin} converts from cubic meter per second (m3/s) to gallon (U.S.) per minute (gpm) (gal/min) }
\arguments{
  \item{cubMeterPerSec}{cubic meter per second (m3/s) }
}
\value{gallon (U.S.) per minute (gpm) (gal/min) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubMeterPerSecTOgallonPerMin(10)
}
\keyword{programming}