\name{NISTcubMeterTOacreFt}
\alias{NISTcubMeterTOacreFt}
\title{Convert cubic meter  to acre foot 7}
\usage{NISTcubMeterTOacreFt(cubMeter)}
\description{\code{NISTcubMeterTOacreFt} converts from cubic meter (m3) to acre foot (based on U.S. survey foot) 7 }
\arguments{
  \item{cubMeter}{cubic meter (m3) }
}
\value{acre foot (based on U.S. survey foot) 7 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubMeterTOacreFt(10)
}
\keyword{programming}