\name{NISTcubMeterTObarrel}
\alias{NISTcubMeterTObarrel}
\title{Convert cubic meter  to barrel }
\usage{NISTcubMeterTObarrel(cubMeter)}
\description{\code{NISTcubMeterTObarrel} converts from cubic meter (m3) to barrel [for petroleum, 42 gallons (U.S.)] (bbl) }
\arguments{
  \item{cubMeter}{cubic meter (m3) }
}
\value{barrel [for petroleum, 42 gallons (U.S.)] (bbl) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubMeterTObarrel(10)
}
\keyword{programming}