\name{NISTcubMeterTObushel}
\alias{NISTcubMeterTObushel}
\title{Convert cubic meter  to bushel }
\usage{NISTcubMeterTObushel(cubMeter)}
\description{\code{NISTcubMeterTObushel} converts from cubic meter (m3) to bushel (U.S.) (bu) }
\arguments{
  \item{cubMeter}{cubic meter (m3) }
}
\value{bushel (U.S.) (bu) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubMeterTObushel(10)
}
\keyword{programming}