\name{NISTcubMeterTOcord}
\alias{NISTcubMeterTOcord}
\title{Convert cubic meter  to cord }
\usage{NISTcubMeterTOcord(cubMeter)}
\description{\code{NISTcubMeterTOcord} converts from cubic meter (m3) to cord (128 ft3) }
\arguments{
  \item{cubMeter}{cubic meter (m3) }
}
\value{cord (128 ft3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubMeterTOcord(10)
}
\keyword{programming}