\name{NISTcubMeterTOcubFt}
\alias{NISTcubMeterTOcubFt}
\title{Convert cubic meter  to cubic foot }
\usage{NISTcubMeterTOcubFt(cubMeter)}
\description{\code{NISTcubMeterTOcubFt} converts from cubic meter (m3) to cubic foot (ft3) }
\arguments{
  \item{cubMeter}{cubic meter (m3) }
}
\value{cubic foot (ft3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubMeterTOcubFt(10)
}
\keyword{programming}