\name{NISTcubMeterTOcubInch}
\alias{NISTcubMeterTOcubInch}
\title{Convert cubic meter  to cubic inch 13}
\usage{NISTcubMeterTOcubInch(cubMeter)}
\description{\code{NISTcubMeterTOcubInch} converts from cubic meter (m3) to cubic inch (in3) 13 }
\arguments{
  \item{cubMeter}{cubic meter (m3) }
}
\value{cubic inch (in3) 13 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubMeterTOcubInch(10)
}
\keyword{programming}