\name{NISTcubMeterTOcubMile}
\alias{NISTcubMeterTOcubMile}
\title{Convert cubic meter  to cubic mile }
\usage{NISTcubMeterTOcubMile(cubMeter)}
\description{\code{NISTcubMeterTOcubMile} converts from cubic meter (m3) to cubic mile (mi3) }
\arguments{
  \item{cubMeter}{cubic meter (m3) }
}
\value{cubic mile (mi3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubMeterTOcubMile(10)
}
\keyword{programming}