\name{NISTcubMeterTOcubYard}
\alias{NISTcubMeterTOcubYard}
\title{Convert cubic meter  to cubic yard }
\usage{NISTcubMeterTOcubYard(cubMeter)}
\description{\code{NISTcubMeterTOcubYard} converts from cubic meter (m3) to cubic yard (yd3) }
\arguments{
  \item{cubMeter}{cubic meter (m3) }
}
\value{cubic yard (yd3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubMeterTOcubYard(10)
}
\keyword{programming}