\name{NISTcubMeterTOcup}
\alias{NISTcubMeterTOcup}
\title{Convert cubic meter  to cup }
\usage{NISTcubMeterTOcup(cubMeter)}
\description{\code{NISTcubMeterTOcup} converts from cubic meter (m3) to cup (U.S.) }
\arguments{
  \item{cubMeter}{cubic meter (m3) }
}
\value{cup (U.S.) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubMeterTOcup(10)
}
\keyword{programming}