\name{NISTcubMeterTOfluidOunce}
\alias{NISTcubMeterTOfluidOunce}
\title{Convert cubic meter  to fluid ounce }
\usage{NISTcubMeterTOfluidOunce(cubMeter)}
\description{\code{NISTcubMeterTOfluidOunce} converts from cubic meter (m3) to fluid ounce (U.S.) (fl oz) }
\arguments{
  \item{cubMeter}{cubic meter (m3) }
}
\value{fluid ounce (U.S.) (fl oz) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubMeterTOfluidOunce(10)
}
\keyword{programming}