\name{NISTcubMeterTOgallon}
\alias{NISTcubMeterTOgallon}
\title{Convert cubic meter  to gallon }
\usage{NISTcubMeterTOgallon(cubMeter)}
\description{\code{NISTcubMeterTOgallon} converts from cubic meter (m3) to gallon (U.S.) (gal) }
\arguments{
  \item{cubMeter}{cubic meter (m3) }
}
\value{gallon (U.S.) (gal) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubMeterTOgallon(10)
}
\keyword{programming}