\name{NISTcubMeterTOgallonImperial}
\alias{NISTcubMeterTOgallonImperial}
\title{Convert cubic meter  to gallon }
\usage{NISTcubMeterTOgallonImperial(cubMeter)}
\description{\code{NISTcubMeterTOgallonImperial} converts from cubic meter (m3) to gallon [Canadian and U.K. (Imperial)] (gal) }
\arguments{
  \item{cubMeter}{cubic meter (m3) }
}
\value{gallon [Canadian and U.K. (Imperial)] (gal) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubMeterTOgallonImperial(10)
}
\keyword{programming}