\name{NISTcubMeterTOgillImperial}
\alias{NISTcubMeterTOgillImperial}
\title{Convert cubic meter  to gill }
\usage{NISTcubMeterTOgillImperial(cubMeter)}
\description{\code{NISTcubMeterTOgillImperial} converts from cubic meter (m3) to gill [Canadian and U.K. (Imperial)] (gi) }
\arguments{
  \item{cubMeter}{cubic meter (m3) }
}
\value{gill [Canadian and U.K. (Imperial)] (gi) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubMeterTOgillImperial(10)
}
\keyword{programming}