\name{NISTcubMeterTOliter}
\alias{NISTcubMeterTOliter}
\title{Convert cubic meter  to liter 19}
\usage{NISTcubMeterTOliter(cubMeter)}
\description{\code{NISTcubMeterTOliter} converts from cubic meter (m3) to liter (L) 19 }
\arguments{
  \item{cubMeter}{cubic meter (m3) }
}
\value{liter (L) 19 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubMeterTOliter(10)
}
\keyword{programming}