\name{NISTcubMeterTOounce}
\alias{NISTcubMeterTOounce}
\title{Convert cubic meter  to ounce }
\usage{NISTcubMeterTOounce(cubMeter)}
\description{\code{NISTcubMeterTOounce} converts from cubic meter (m3) to ounce (U.S. fluid) (fl oz) }
\arguments{
  \item{cubMeter}{cubic meter (m3) }
}
\value{ounce (U.S. fluid) (fl oz) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubMeterTOounce(10)
}
\keyword{programming}