\name{NISTcubMeterTOounceImperial}
\alias{NISTcubMeterTOounceImperial}
\title{Convert cubic meter  to ounce }
\usage{NISTcubMeterTOounceImperial(cubMeter)}
\description{\code{NISTcubMeterTOounceImperial} converts from cubic meter (m3) to ounce [Canadian and U.K. fluid (Imperial)] (fl oz) }
\arguments{
  \item{cubMeter}{cubic meter (m3) }
}
\value{ounce [Canadian and U.K. fluid (Imperial)] (fl oz) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubMeterTOounceImperial(10)
}
\keyword{programming}