\name{NISTcubMeterTOpintUSdry}
\alias{NISTcubMeterTOpintUSdry}
\title{Convert cubic meter  to pint }
\usage{NISTcubMeterTOpintUSdry(cubMeter)}
\description{\code{NISTcubMeterTOpintUSdry} converts from cubic meter (m3) to pint (U.S. dry) (dry pt) }
\arguments{
  \item{cubMeter}{cubic meter (m3) }
}
\value{pint (U.S. dry) (dry pt) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubMeterTOpintUSdry(10)
}
\keyword{programming}