\name{NISTcubMeterTOpintUSliquid}
\alias{NISTcubMeterTOpintUSliquid}
\title{Convert cubic meter  to pint }
\usage{NISTcubMeterTOpintUSliquid(cubMeter)}
\description{\code{NISTcubMeterTOpintUSliquid} converts from cubic meter (m3) to pint (U.S. liquid) (liq pt) }
\arguments{
  \item{cubMeter}{cubic meter (m3) }
}
\value{pint (U.S. liquid) (liq pt) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubMeterTOpintUSliquid(10)
}
\keyword{programming}