\name{NISTcubMeterTOquartUSliquid}
\alias{NISTcubMeterTOquartUSliquid}
\title{Convert cubic meter  to quart }
\usage{NISTcubMeterTOquartUSliquid(cubMeter)}
\description{\code{NISTcubMeterTOquartUSliquid} converts from cubic meter (m3) to quart (U.S. liquid) (liq qt) }
\arguments{
  \item{cubMeter}{cubic meter (m3) }
}
\value{quart (U.S. liquid) (liq qt) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubMeterTOquartUSliquid(10)
}
\keyword{programming}