\name{NISTcubMeterTOstere}
\alias{NISTcubMeterTOstere}
\title{Convert cubic meter  to stere }
\usage{NISTcubMeterTOstere(cubMeter)}
\description{\code{NISTcubMeterTOstere} converts from cubic meter (m3) to stere (st) }
\arguments{
  \item{cubMeter}{cubic meter (m3) }
}
\value{stere (st) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubMeterTOstere(10)
}
\keyword{programming}