\name{NISTcubMeterTOteaspoon}
\alias{NISTcubMeterTOteaspoon}
\title{Convert cubic meter  to teaspoon}
\usage{NISTcubMeterTOteaspoon(cubMeter)}
\description{\code{NISTcubMeterTOteaspoon} converts from cubic meter (m3) to teaspoon }
\arguments{
  \item{cubMeter}{cubic meter (m3) }
}
\value{teaspoon }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubMeterTOteaspoon(10)
}
\keyword{programming}