\name{NISTcubMeterTOtonRegister}
\alias{NISTcubMeterTOtonRegister}
\title{Convert cubic meter  to ton, register}
\usage{NISTcubMeterTOtonRegister(cubMeter)}
\description{\code{NISTcubMeterTOtonRegister} converts from cubic meter (m3) to ton, register }
\arguments{
  \item{cubMeter}{cubic meter (m3) }
}
\value{ton, register }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubMeterTOtonRegister(10)
}
\keyword{programming}