\name{NISTcubMileTOcubMeter}
\alias{NISTcubMileTOcubMeter}
\title{Convert cubic mile  to cubic meter }
\usage{NISTcubMileTOcubMeter(cubMile)}
\description{\code{NISTcubMileTOcubMeter} converts from cubic mile (mi3) to cubic meter (m3) }
\arguments{
  \item{cubMile}{cubic mile (mi3) }
}
\value{cubic meter (m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubMileTOcubMeter(10)
}
\keyword{programming}