\name{NISTcubYardPerMinTOcubMeterPerSec}
\alias{NISTcubYardPerMinTOcubMeterPerSec}
\title{Convert cubic yard per minute  to cubic meter per second }
\usage{NISTcubYardPerMinTOcubMeterPerSec(cubYardPerMin)}
\description{\code{NISTcubYardPerMinTOcubMeterPerSec} converts from cubic yard per minute (yd3/min) to cubic meter per second (m3/s) }
\arguments{
  \item{cubYardPerMin}{cubic yard per minute (yd3/min) }
}
\value{cubic meter per second (m3/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubYardPerMinTOcubMeterPerSec(10)
}
\keyword{programming}