\name{NISTcubYardTOcubMeter}
\alias{NISTcubYardTOcubMeter}
\title{Convert cubic yard  to cubic meter }
\usage{NISTcubYardTOcubMeter(cubYard)}
\description{\code{NISTcubYardTOcubMeter} converts from cubic yard (yd3) to cubic meter (m3) }
\arguments{
  \item{cubYard}{cubic yard (yd3) }
}
\value{cubic meter (m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcubYardTOcubMeter(10)
}
\keyword{programming}