\name{NISTcupTOcubMeter}
\alias{NISTcupTOcubMeter}
\title{Convert cup  to cubic meter }
\usage{NISTcupTOcubMeter(cup)}
\description{\code{NISTcupTOcubMeter} converts from cup (U.S.) to cubic meter (m3) }
\arguments{
  \item{cup}{cup (U.S.) }
}
\value{cubic meter (m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcupTOcubMeter(10)
}
\keyword{programming}