\name{NISTcupTOliter}
\alias{NISTcupTOliter}
\title{Convert cup  to liter }
\usage{NISTcupTOliter(cup)}
\description{\code{NISTcupTOliter} converts from cup (U.S.) to liter (L) }
\arguments{
  \item{cup}{cup (U.S.) }
}
\value{liter (L) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcupTOliter(10)
}
\keyword{programming}