\name{NISTcupTOml}
\alias{NISTcupTOml}
\title{Convert cup  to milliliter }
\usage{NISTcupTOml(cup)}
\description{\code{NISTcupTOml} converts from cup (U.S.) to milliliter (mL) }
\arguments{
  \item{cup}{cup (U.S.) }
}
\value{milliliter (mL) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcupTOml(10)
}
\keyword{programming}