\name{NISTcurieTObecquerel}
\alias{NISTcurieTObecquerel}
\title{Convert curie  to becquerel }
\usage{NISTcurieTObecquerel(curie)}
\description{\code{NISTcurieTObecquerel} converts from curie (Ci) to becquerel (Bq) }
\arguments{
  \item{curie}{curie (Ci) }
}
\value{becquerel (Bq) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTcurieTObecquerel(10)
}
\keyword{programming}