\name{NISTdarcyTOmeterSqrd}
\alias{NISTdarcyTOmeterSqrd}
\title{Convert darcy 14 to meter squared }
\usage{NISTdarcyTOmeterSqrd(darcy)}
\description{\code{NISTdarcyTOmeterSqrd} converts from darcy 14 to meter squared (m2) }
\arguments{
  \item{darcy}{darcy 14 }
}
\value{meter squared (m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTdarcyTOmeterSqrd(10)
}
\keyword{programming}