\name{NISTdaySiderealTOsec}
\alias{NISTdaySiderealTOsec}
\title{Convert day  to second }
\usage{NISTdaySiderealTOsec(daySidereal)}
\description{\code{NISTdaySiderealTOsec} converts from day (sidereal) to second (s) }
\arguments{
  \item{daySidereal}{day (sidereal) }
}
\value{second (s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTdaySiderealTOsec(10)
}
\keyword{programming}