\name{NISTdayTOsec}
\alias{NISTdayTOsec}
\title{Convert day  to second }
\usage{NISTdayTOsec(day)}
\description{\code{NISTdayTOsec} converts from day (d) to second (s) }
\arguments{
  \item{day}{day (d) }
}
\value{second (s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTdayTOsec(10)
}
\keyword{programming}