\name{NISTdebyeTOcoulombMeter}
\alias{NISTdebyeTOcoulombMeter}
\title{Convert debye  to coulomb meter }
\usage{NISTdebyeTOcoulombMeter(debye)}
\description{\code{NISTdebyeTOcoulombMeter} converts from debye (D) to coulomb meter (C * m) }
\arguments{
  \item{debye}{debye (D) }
}
\value{coulomb meter (C * m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTdebyeTOcoulombMeter(10)
}
\keyword{programming}