\name{NISTdegCentigradeTOdegC}
\alias{NISTdegCentigradeTOdegC}
\title{Convert degree centigrade 15 to degree Celsius }
\usage{NISTdegCentigradeTOdegC(degCentigrade)}
\description{\code{NISTdegCentigradeTOdegC} converts from degree centigrade (temperature interval) 15 to degree Celsius (C) }
\arguments{
  \item{degCentigrade}{degree centigrade (temperature interval) 15 }
}
\value{degree Celsius (C) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTdegCentigradeTOdegC(10)
}
\keyword{programming}