\name{NISTdegCentigradeTOdegCinterval}
\alias{NISTdegCentigradeTOdegCinterval}
\title{Convert degree centigrade temperature interval to degree Celsius }
\usage{NISTdegCentigradeTOdegCinterval(degCentigrade)}
\description{\code{NISTdegCentigradeTOdegCinterval} converts from degree centigrade (temperature interval) to degree Celsius (C) }
\arguments{
  \item{degCentigrade}{degree centigrade (temperature interval) }
}
\value{degree Celsius (C) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTdegCentigradeTOdegCinterval(10)
}
\keyword{programming}
