\name{NISTdegCtOdegCentigrade}
\alias{NISTdegCtOdegCentigrade}
\title{Convert degree Celsius  to degree centigrade 15}
\usage{NISTdegCtOdegCentigrade(degC)}
\description{\code{NISTdegCtOdegCentigrade} converts from degree Celsius (C) to degree centigrade (temperature interval) 15 }
\arguments{
  \item{degC}{degree Celsius (C) }
}
\value{degree centigrade (temperature interval) 15 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTdegCtOdegCentigrade(10)
}
\keyword{programming}