\name{NISTdegCtOdegCentigradeInterval}
\alias{NISTdegCtOdegCentigradeInterval}
\title{Convert degree Celsius temperature interval to degree centigrade}
\usage{NISTdegCtOdegCentigradeInterval(degC)}
\description{\code{NISTdegCtOdegCentigradeInterval} converts from degree Celsius (C) to degree centigrade (temperature interval) }
\arguments{
  \item{degC}{degree Celsius (C) }
}
\value{degree centigrade (temperature interval) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTdegCtOdegCentigradeInterval(10)
}
\keyword{programming}
