\name{NISTdegCtOdegF}
\alias{NISTdegCtOdegF}
\title{Convert degree Celsius  to degree Fahrenheit }
\usage{NISTdegCtOdegF(degC)}
\description{\code{NISTdegCtOdegF} converts from degree Celsius (C) to degree Fahrenheit (temperature interval) (F) }
\arguments{
  \item{degC}{degree Celsius (C) }
}
\value{degree Fahrenheit (temperature interval) (F) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTdegCtOdegF(10)
}
\keyword{programming}