\name{NISTdegCtOdegFinterval}
\alias{NISTdegCtOdegFinterval}
\title{Convert degree Celsius temperature interval to degree Fahrenheit }
\usage{NISTdegCtOdegFinterval(degC)}
\description{\code{NISTdegCtOdegFinterval} converts from degree Celsius (C) to degree Fahrenheit (temperature interval) (F) }
\arguments{
  \item{degC}{degree Celsius (C) }
}
\value{degree Fahrenheit (temperature interval) (F) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTdegCtOdegFinterval(10)
}
\keyword{programming}
