\name{NISTdegCtOk}
\alias{NISTdegCtOk}
\title{Convert degree Celsius  to kelvin }
\usage{NISTdegCtOk(degC)}
\description{\code{NISTdegCtOk} converts from degree Celsius (C) to kelvin (K) }
\arguments{
  \item{degC}{degree Celsius (C) }
}
\value{kelvin (K) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTdegCtOk(10)
}
\keyword{programming}