\name{NISTdegCtOkInterval}
\alias{NISTdegCtOkInterval}
\title{Convert degree Celsius temperature interval to kelvin }
\usage{NISTdegCtOkInterval(degC)}
\description{\code{NISTdegCtOkInterval} converts from degree Celsius (C) to kelvin (temperature interval K) }
\arguments{
  \item{degC}{degree Celsius (C) }
}
\value{kelvin (K) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTdegCtOkInterval(10)
}
\keyword{programming}
