\name{NISTdegFHourPerUkThUnITtOkPerWatt}
\alias{NISTdegFHourPerUkThUnITtOkPerWatt}
\title{Convert degree Fahrenheit hour per British thermal unitIT  to kelvin per watt }
\usage{NISTdegFHourPerUkThUnITtOkPerWatt(degFHourPerUkThUnIT)}
\description{\code{NISTdegFHourPerUkThUnITtOkPerWatt} converts from degree Fahrenheit hour per British thermal unitIT (F * h/BtuIT) to kelvin per watt (K/W) }
\arguments{
  \item{degFHourPerUkThUnIT}{degree Fahrenheit hour per British thermal unitIT (F * h/BtuIT) }
}
\value{kelvin per watt (K/W) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTdegFHourPerUkThUnITtOkPerWatt(10)
}
\keyword{programming}