\name{NISTdegFHourPerUkThUnthTOkPerWatt}
\alias{NISTdegFHourPerUkThUnthTOkPerWatt}
\title{Convert degree Fahrenheit hour per British thermal unitth  to kelvin per watt }
\usage{NISTdegFHourPerUkThUnthTOkPerWatt(degFHourPerUkThUnth)}
\description{\code{NISTdegFHourPerUkThUnthTOkPerWatt} converts from degree Fahrenheit hour per British thermal unitth (F * h/Btuth) to kelvin per watt (K/W) }
\arguments{
  \item{degFHourPerUkThUnth}{degree Fahrenheit hour per British thermal unitth (F * h/Btuth) }
}
\value{kelvin per watt (K/W) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTdegFHourPerUkThUnthTOkPerWatt(10)
}
\keyword{programming}