\name{NISTdegFHourSqrFtPerUkThUnITInchTOmeterKPerWatt}
\alias{NISTdegFHourSqrFtPerUkThUnITInchTOmeterKPerWatt}
\title{Convert degree Fahrenheit hour square foot per British thermal unitIT inch  to meter kelvin per watt }
\usage{NISTdegFHourSqrFtPerUkThUnITInchTOmeterKPerWatt(degFHourSqrFtPerUkThUnITInch)}
\description{\code{NISTdegFHourSqrFtPerUkThUnITInchTOmeterKPerWatt} converts from degree Fahrenheit hour square foot per British thermal unitIT inch [F * h * ft2/(BtuIT * in)] to meter kelvin per watt (m * K/W) }
\arguments{
  \item{degFHourSqrFtPerUkThUnITInch}{degree Fahrenheit hour square foot per British thermal unitIT inch [F * h * ft2/(BtuIT * in)] }
}
\value{meter kelvin per watt (m * K/W) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTdegFHourSqrFtPerUkThUnITInchTOmeterKPerWatt(10)
}
\keyword{programming}