\name{NISTdegFHourSqrFtPerUkThUnITtOsqrMeterKPerWatt}
\alias{NISTdegFHourSqrFtPerUkThUnITtOsqrMeterKPerWatt}
\title{Convert degree Fahrenheit hour square foot per British thermal unitIT  to square meter kelvin per watt }
\usage{NISTdegFHourSqrFtPerUkThUnITtOsqrMeterKPerWatt(degFHourSqrFtPerUkThUnIT)}
\description{\code{NISTdegFHourSqrFtPerUkThUnITtOsqrMeterKPerWatt} converts from degree Fahrenheit hour square foot per British thermal unitIT (F * h * ft2/BtuIT) to square meter kelvin per watt (m2 * K/W) }
\arguments{
  \item{degFHourSqrFtPerUkThUnIT}{degree Fahrenheit hour square foot per British thermal unitIT (F * h * ft2/BtuIT) }
}
\value{square meter kelvin per watt (m2 * K/W) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTdegFHourSqrFtPerUkThUnITtOsqrMeterKPerWatt(10)
}
\keyword{programming}