\name{NISTdegFHourSqrFtPerUkThUnthInchTOmeterKPerWatt}
\alias{NISTdegFHourSqrFtPerUkThUnthInchTOmeterKPerWatt}
\title{Convert degree Fahrenheit hour square foot per British thermal unitth inch  to meter kelvin per watt }
\usage{NISTdegFHourSqrFtPerUkThUnthInchTOmeterKPerWatt(degFHourSqrFtPerUkThUnthInch)}
\description{\code{NISTdegFHourSqrFtPerUkThUnthInchTOmeterKPerWatt} converts from degree Fahrenheit hour square foot per British thermal unitth inch [F * h * ft2/(Btuth * in)] to meter kelvin per watt (m * K/W) }
\arguments{
  \item{degFHourSqrFtPerUkThUnthInch}{degree Fahrenheit hour square foot per British thermal unitth inch [F * h * ft2/(Btuth * in)] }
}
\value{meter kelvin per watt (m * K/W) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTdegFHourSqrFtPerUkThUnthInchTOmeterKPerWatt(10)
}
\keyword{programming}