\name{NISTdegFHourSqrFtPerUkThUnthTOsqrMeterKPerWatt}
\alias{NISTdegFHourSqrFtPerUkThUnthTOsqrMeterKPerWatt}
\title{Convert degree Fahrenheit hour square foot per British thermal unitth  to square meter kelvin per watt }
\usage{NISTdegFHourSqrFtPerUkThUnthTOsqrMeterKPerWatt(degFHourSqrFtPerUkThUnth)}
\description{\code{NISTdegFHourSqrFtPerUkThUnthTOsqrMeterKPerWatt} converts from degree Fahrenheit hour square foot per British thermal unitth (F * h * ft2/Btuth) to square meter kelvin per watt (m2 * K/W) }
\arguments{
  \item{degFHourSqrFtPerUkThUnth}{degree Fahrenheit hour square foot per British thermal unitth (F * h * ft2/Btuth) }
}
\value{square meter kelvin per watt (m2 * K/W) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTdegFHourSqrFtPerUkThUnthTOsqrMeterKPerWatt(10)
}
\keyword{programming}