\name{NISTdegFSecPerUkThUnITtOkPerWatt}
\alias{NISTdegFSecPerUkThUnITtOkPerWatt}
\title{Convert degree Fahrenheit second per British thermal unitIT  to kelvin per watt }
\usage{NISTdegFSecPerUkThUnITtOkPerWatt(degFSecPerUkThUnIT)}
\description{\code{NISTdegFSecPerUkThUnITtOkPerWatt} converts from degree Fahrenheit second per British thermal unitIT (F * s/BtuIT) to kelvin per watt (K/W) }
\arguments{
  \item{degFSecPerUkThUnIT}{degree Fahrenheit second per British thermal unitIT (F * s/BtuIT) }
}
\value{kelvin per watt (K/W) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTdegFSecPerUkThUnITtOkPerWatt(10)
}
\keyword{programming}