\name{NISTdegFSecPerUkThUnthTOkPerWatt}
\alias{NISTdegFSecPerUkThUnthTOkPerWatt}
\title{Convert degree Fahrenheit second per British thermal unitth  to kelvin per watt }
\usage{NISTdegFSecPerUkThUnthTOkPerWatt(degFSecPerUkThUnth)}
\description{\code{NISTdegFSecPerUkThUnthTOkPerWatt} converts from degree Fahrenheit second per British thermal unitth (F * s/Btuth) to kelvin per watt (K/W) }
\arguments{
  \item{degFSecPerUkThUnth}{degree Fahrenheit second per British thermal unitth (F * s/Btuth) }
}
\value{kelvin per watt (K/W) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTdegFSecPerUkThUnthTOkPerWatt(10)
}
\keyword{programming}