\name{NISTdegFtOdegC}
\alias{NISTdegFtOdegC}
\title{Convert degree Fahrenheit  to degree Celsius }
\usage{NISTdegFtOdegC(degF)}
\description{\code{NISTdegFtOdegC} converts from degree Fahrenheit (temperature interval) (F) to degree Celsius (C) }
\arguments{
  \item{degF}{degree Fahrenheit (temperature interval) (F) }
}
\value{degree Celsius (C) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTdegFtOdegC(10)
}
\keyword{programming}