\name{NISTdegFtOk}
\alias{NISTdegFtOk}
\title{Convert degree Fahrenheit  to kelvin }
\usage{NISTdegFtOk(degF)}
\description{\code{NISTdegFtOk} converts from degree Fahrenheit (temperature interval) (F) to kelvin (K) }
\arguments{
  \item{degF}{degree Fahrenheit (temperature interval) (F) }
}
\value{kelvin (K) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTdegFtOk(10)
}
\keyword{programming}