\name{NISTdegRankineTOk}
\alias{NISTdegRankineTOk}
\title{Convert degree Rankine  to kelvin }
\usage{NISTdegRankineTOk(degRankine)}
\description{\code{NISTdegRankineTOk} converts from degree Rankine (temperature interval) (R) to kelvin (K) }
\arguments{
  \item{degRankine}{degree Rankine (temperature interval) (R) }
}
\value{kelvin (K) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTdegRankineTOk(10)
}
\keyword{programming}