\name{NISTdegTOmil}
\alias{NISTdegTOmil}
\title{Convert degree  to mil }
\usage{NISTdegTOmil(deg)}
\description{\code{NISTdegTOmil} converts from degree () to mil (angle) }
\arguments{
  \item{deg}{degree () }
}
\value{mil (angle) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTdegTOmil(10)
}
\keyword{programming}