\name{NISTdegTOradian}
\alias{NISTdegTOradian}
\title{Convert degree  to radian }
\usage{NISTdegTOradian(deg)}
\description{\code{NISTdegTOradian} converts from degree (angle) () to radian (rad) }
\arguments{
  \item{deg}{degree (angle) () }
}
\value{radian (rad) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTdegTOradian(10)
}
\keyword{programming}