\name{NISTdenierTOkgPerMeter}
\alias{NISTdenierTOkgPerMeter}
\title{Convert denier to kilogram per meter }
\usage{NISTdenierTOkgPerMeter(denier)}
\description{\code{NISTdenierTOkgPerMeter} converts from denier to kilogram per meter (kg/m) }
\arguments{
  \item{denier}{denier }
}
\value{kilogram per meter (kg/m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTdenierTOkgPerMeter(10)
}
\keyword{programming}