\name{NISTdyneCmTOnewtonMeter}
\alias{NISTdyneCmTOnewtonMeter}
\title{Convert dyne centimeter  to newton meter }
\usage{NISTdyneCmTOnewtonMeter(dyneCm)}
\description{\code{NISTdyneCmTOnewtonMeter} converts from dyne centimeter (dyn * cm) to newton meter (N * m) }
\arguments{
  \item{dyneCm}{dyne centimeter (dyn * cm) }
}
\value{newton meter (N * m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTdyneCmTOnewtonMeter(10)
}
\keyword{programming}