\name{NISTdynePerSqrCmTOpascal}
\alias{NISTdynePerSqrCmTOpascal}
\title{Convert dyne per square centimeter  to pascal }
\usage{NISTdynePerSqrCmTOpascal(dynePerSqrCm)}
\description{\code{NISTdynePerSqrCmTOpascal} converts from dyne per square centimeter (dyn/cm2) to pascal (Pa) }
\arguments{
  \item{dynePerSqrCm}{dyne per square centimeter (dyn/cm2) }
}
\value{pascal (Pa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTdynePerSqrCmTOpascal(10)
}
\keyword{programming}