\name{NISTdyneTOnewton}
\alias{NISTdyneTOnewton}
\title{Convert dyne  to newton }
\usage{NISTdyneTOnewton(dyne)}
\description{\code{NISTdyneTOnewton} converts from dyne (dyn) to newton (N) }
\arguments{
  \item{dyne}{dyne (dyn) }
}
\value{newton (N) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTdyneTOnewton(10)
}
\keyword{programming}