\name{NISTelectronvoltTOjoule}
\alias{NISTelectronvoltTOjoule}
\title{Convert electronvolt  to joule }
\usage{NISTelectronvoltTOjoule(electronvolt)}
\description{\code{NISTelectronvoltTOjoule} converts from electronvolt (eV) to joule (J) }
\arguments{
  \item{electronvolt}{electronvolt (eV) }
}
\value{joule (J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTelectronvoltTOjoule(10)
}
\keyword{programming}