\name{NISTEMUOfCapacitanceTOfarad}
\alias{NISTEMUOfCapacitanceTOfarad}
\title{Convert EMU of capacitance  to farad }
\usage{NISTEMUOfCapacitanceTOfarad(EMUOfCapacitance)}
\description{\code{NISTEMUOfCapacitanceTOfarad} converts from EMU of capacitance (abfarad) to farad (F) }
\arguments{
  \item{EMUOfCapacitance}{EMU of capacitance (abfarad) }
}
\value{farad (F) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTEMUOfCapacitanceTOfarad(10)
}
\keyword{programming}