\name{NISTEMUOfCurrentTOampere}
\alias{NISTEMUOfCurrentTOampere}
\title{Convert EMU of current  to ampere }
\usage{NISTEMUOfCurrentTOampere(EMUOfCurrent)}
\description{\code{NISTEMUOfCurrentTOampere} converts from EMU of current (abampere) to ampere (A) }
\arguments{
  \item{EMUOfCurrent}{EMU of current (abampere) }
}
\value{ampere (A) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTEMUOfCurrentTOampere(10)
}
\keyword{programming}