\name{NISTEMUOfElectricPotentialTOvolt}
\alias{NISTEMUOfElectricPotentialTOvolt}
\title{Convert EMU of electric potential  to volt }
\usage{NISTEMUOfElectricPotentialTOvolt(EMUOfElectricPotential)}
\description{\code{NISTEMUOfElectricPotentialTOvolt} converts from EMU of electric potential (abvolt) to volt (V) }
\arguments{
  \item{EMUOfElectricPotential}{EMU of electric potential (abvolt) }
}
\value{volt (V) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTEMUOfElectricPotentialTOvolt(10)
}
\keyword{programming}