\name{NISTEMUOfInductanceTOhenry}
\alias{NISTEMUOfInductanceTOhenry}
\title{Convert EMU of inductance  to henry }
\usage{NISTEMUOfInductanceTOhenry(EMUOfInductance)}
\description{\code{NISTEMUOfInductanceTOhenry} converts from EMU of inductance (abhenry) to henry (H) }
\arguments{
  \item{EMUOfInductance}{EMU of inductance (abhenry) }
}
\value{henry (H) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTEMUOfInductanceTOhenry(10)
}
\keyword{programming}