\name{NISTEMUOfResistanceTOohm}
\alias{NISTEMUOfResistanceTOohm}
\title{Convert EMU of resistance  to ohm }
\usage{NISTEMUOfResistanceTOohm(EMUOfResistance)}
\description{\code{NISTEMUOfResistanceTOohm} converts from EMU of resistance (abohm) to ohm (ohm) }
\arguments{
  \item{EMUOfResistance}{EMU of resistance (abohm) }
}
\value{ohm (ohm) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTEMUOfResistanceTOohm(10)
}
\keyword{programming}