\name{NISTergPerSecTOwatt}
\alias{NISTergPerSecTOwatt}
\title{Convert erg per second  to watt }
\usage{NISTergPerSecTOwatt(ergPerSec)}
\description{\code{NISTergPerSecTOwatt} converts from erg per second (erg/s) to watt (W) }
\arguments{
  \item{ergPerSec}{erg per second (erg/s) }
}
\value{watt (W) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTergPerSecTOwatt(10)
}
\keyword{programming}