\name{NISTergPerSqrCmSecTOwattPerSqrMeter}
\alias{NISTergPerSqrCmSecTOwattPerSqrMeter}
\title{Convert erg per square centimeter second  to watt per square meter }
\usage{NISTergPerSqrCmSecTOwattPerSqrMeter(ergPerSqrCmSec)}
\description{\code{NISTergPerSqrCmSecTOwattPerSqrMeter} converts from erg per square centimeter second [erg/(cm2 * s)] to watt per square meter (W/m2) }
\arguments{
  \item{ergPerSqrCmSec}{erg per square centimeter second [erg/(cm2 * s)] }
}
\value{watt per square meter (W/m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTergPerSqrCmSecTOwattPerSqrMeter(10)
}
\keyword{programming}