\name{NISTergTOjoule}
\alias{NISTergTOjoule}
\title{Convert erg  to joule }
\usage{NISTergTOjoule(erg)}
\description{\code{NISTergTOjoule} converts from erg (erg) to joule (J) }
\arguments{
  \item{erg}{erg (erg) }
}
\value{joule (J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTergTOjoule(10)
}
\keyword{programming}