\name{NISTESUOfCapacitanceTOfarad}
\alias{NISTESUOfCapacitanceTOfarad}
\title{Convert ESU of capacitance  to farad }
\usage{NISTESUOfCapacitanceTOfarad(ESUOfCapacitance)}
\description{\code{NISTESUOfCapacitanceTOfarad} converts from ESU of capacitance (statfarad) to farad (F) }
\arguments{
  \item{ESUOfCapacitance}{ESU of capacitance (statfarad) }
}
\value{farad (F) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTESUOfCapacitanceTOfarad(10)
}
\keyword{programming}