\name{NISTESUOfCurrentTOampere}
\alias{NISTESUOfCurrentTOampere}
\title{Convert ESU of current  to ampere }
\usage{NISTESUOfCurrentTOampere(ESUOfCurrent)}
\description{\code{NISTESUOfCurrentTOampere} converts from ESU of current (statampere) to ampere (A) }
\arguments{
  \item{ESUOfCurrent}{ESU of current (statampere) }
}
\value{ampere (A) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTESUOfCurrentTOampere(10)
}
\keyword{programming}