\name{NISTESUOfElectricPotentialTOvolt}
\alias{NISTESUOfElectricPotentialTOvolt}
\title{Convert ESU of electric potential  to volt }
\usage{NISTESUOfElectricPotentialTOvolt(ESUOfElectricPotential)}
\description{\code{NISTESUOfElectricPotentialTOvolt} converts from ESU of electric potential (statvolt) to volt (V) }
\arguments{
  \item{ESUOfElectricPotential}{ESU of electric potential (statvolt) }
}
\value{volt (V) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTESUOfElectricPotentialTOvolt(10)
}
\keyword{programming}