\name{NISTESUOfInductanceTOhenry}
\alias{NISTESUOfInductanceTOhenry}
\title{Convert ESU of inductance  to henry }
\usage{NISTESUOfInductanceTOhenry(ESUOfInductance)}
\description{\code{NISTESUOfInductanceTOhenry} converts from ESU of inductance (stathenry) to henry (H) }
\arguments{
  \item{ESUOfInductance}{ESU of inductance (stathenry) }
}
\value{henry (H) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTESUOfInductanceTOhenry(10)
}
\keyword{programming}