\name{NISTESUOfResistanceTOohm}
\alias{NISTESUOfResistanceTOohm}
\title{Convert ESU of resistance  to ohm }
\usage{NISTESUOfResistanceTOohm(ESUOfResistance)}
\description{\code{NISTESUOfResistanceTOohm} converts from ESU of resistance (statohm) to ohm (ohm) }
\arguments{
  \item{ESUOfResistance}{ESU of resistance (statohm) }
}
\value{ohm (ohm) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTESUOfResistanceTOohm(10)
}
\keyword{programming}