\name{NISTfaradayTOcoulomb}
\alias{NISTfaradayTOcoulomb}
\title{Convert faraday  to coulomb }
\usage{NISTfaradayTOcoulomb(faraday)}
\description{\code{NISTfaradayTOcoulomb} converts from faraday (based on carbon 12) to coulomb (C) }
\arguments{
  \item{faraday}{faraday (based on carbon 12) }
}
\value{coulomb (C) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTfaradayTOcoulomb(10)
}
\keyword{programming}