\name{NISTfaradTOabfarad}
\alias{NISTfaradTOabfarad}
\title{Convert farad  to abfarad}
\usage{NISTfaradTOabfarad(farad)}
\description{\code{NISTfaradTOabfarad} converts from farad (F) to abfarad }
\arguments{
  \item{farad}{farad (F) }
}
\value{abfarad }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTfaradTOabfarad(10)
}
\keyword{programming}