\name{NISTfaradToEMUOfCapacitance}
\alias{NISTfaradToEMUOfCapacitance}
\title{Convert farad  to EMU of capacitance }
\usage{NISTfaradToEMUOfCapacitance(farad)}
\description{\code{NISTfaradToEMUOfCapacitance} converts from farad (F) to EMU of capacitance (abfarad) }
\arguments{
  \item{farad}{farad (F) }
}
\value{EMU of capacitance (abfarad) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTfaradToEMUOfCapacitance(10)
}
\keyword{programming}