\name{NISTfaradToESUOfCapacitance}
\alias{NISTfaradToESUOfCapacitance}
\title{Convert farad  to ESU of capacitance }
\usage{NISTfaradToESUOfCapacitance(farad)}
\description{\code{NISTfaradToESUOfCapacitance} converts from farad (F) to ESU of capacitance (statfarad) }
\arguments{
  \item{farad}{farad (F) }
}
\value{ESU of capacitance (statfarad) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTfaradToESUOfCapacitance(10)
}
\keyword{programming}