\name{NISTfaradTOstatfarad}
\alias{NISTfaradTOstatfarad}
\title{Convert farad  to statfarad}
\usage{NISTfaradTOstatfarad(farad)}
\description{\code{NISTfaradTOstatfarad} converts from farad (F) to statfarad }
\arguments{
  \item{farad}{farad (F) }
}
\value{statfarad }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTfaradTOstatfarad(10)
}
\keyword{programming}