\name{NISTfemtometerTOfermi}
\alias{NISTfemtometerTOfermi}
\title{Convert femtometer  to fermi}
\usage{NISTfemtometerTOfermi(femtometer)}
\description{\code{NISTfemtometerTOfermi} converts from femtometer (fm) to fermi }
\arguments{
  \item{femtometer}{femtometer (fm) }
}
\value{fermi }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTfemtometerTOfermi(10)
}
\keyword{programming}