\name{NISTfermiTOfemtometer}
\alias{NISTfermiTOfemtometer}
\title{Convert fermi to femtometer }
\usage{NISTfermiTOfemtometer(fermi)}
\description{\code{NISTfermiTOfemtometer} converts from fermi to femtometer (fm) }
\arguments{
  \item{fermi}{fermi }
}
\value{femtometer (fm) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTfermiTOfemtometer(10)
}
\keyword{programming}