\name{NISTfermiTOmeter}
\alias{NISTfermiTOmeter}
\title{Convert fermi to meter }
\usage{NISTfermiTOmeter(fermi)}
\description{\code{NISTfermiTOmeter} converts from fermi to meter (m) }
\arguments{
  \item{fermi}{fermi }
}
\value{meter (m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTfermiTOmeter(10)
}
\keyword{programming}