\name{NISTfluidOunceTOcubMeter}
\alias{NISTfluidOunceTOcubMeter}
\title{Convert fluid ounce  to cubic meter }
\usage{NISTfluidOunceTOcubMeter(fluidOunce)}
\description{\code{NISTfluidOunceTOcubMeter} converts from fluid ounce (U.S.) (fl oz) to cubic meter (m3) }
\arguments{
  \item{fluidOunce}{fluid ounce (U.S.) (fl oz) }
}
\value{cubic meter (m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTfluidOunceTOcubMeter(10)
}
\keyword{programming}