\name{NISTfluidOunceTOml}
\alias{NISTfluidOunceTOml}
\title{Convert fluid ounce  to milliliter }
\usage{NISTfluidOunceTOml(fluidOunce)}
\description{\code{NISTfluidOunceTOml} converts from fluid ounce (U.S.) (fl oz) to milliliter (mL) }
\arguments{
  \item{fluidOunce}{fluid ounce (U.S.) (fl oz) }
}
\value{milliliter (mL) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTfluidOunceTOml(10)
}
\keyword{programming}