\name{NISTfranklinTOcoulomb}
\alias{NISTfranklinTOcoulomb}
\title{Convert franklin  to coulomb }
\usage{NISTfranklinTOcoulomb(franklin)}
\description{\code{NISTfranklinTOcoulomb} converts from franklin (Fr) to coulomb (C) }
\arguments{
  \item{franklin}{franklin (Fr) }
}
\value{coulomb (C) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTfranklinTOcoulomb(10)
}
\keyword{programming}