\name{NISTft4thPowerTOmeter4thPower}
\alias{NISTft4thPowerTOmeter4thPower}
\title{Convert foot to the fourth power 16 to meter to the fourth power }
\usage{NISTft4thPowerTOmeter4thPower(ft4thPower)}
\description{\code{NISTft4thPowerTOmeter4thPower} converts from foot to the fourth power (ft4) 16 to meter to the fourth power (m4) }
\arguments{
  \item{ft4thPower}{foot to the fourth power (ft4) 16 }
}
\value{meter to the fourth power (m4) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTft4thPowerTOmeter4thPower(10)
}
\keyword{programming}