\name{NISTftcandleTOlux}
\alias{NISTftcandleTOlux}
\title{Convert footcandle to lux }
\usage{NISTftcandleTOlux(ftcandle)}
\description{\code{NISTftcandleTOlux} converts from footcandle to lux (lx) }
\arguments{
  \item{ftcandle}{footcandle }
}
\value{lux (lx) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTftcandleTOlux(10)
}
\keyword{programming}