\name{NISTftlambertTOcandelaPerSqrMeter}
\alias{NISTftlambertTOcandelaPerSqrMeter}
\title{Convert footlambert to candela per square meter }
\usage{NISTftlambertTOcandelaPerSqrMeter(ftlambert)}
\description{\code{NISTftlambertTOcandelaPerSqrMeter} converts from footlambert to candela per square meter (cd/m2) }
\arguments{
  \item{ftlambert}{footlambert }
}
\value{candela per square meter (cd/m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTftlambertTOcandelaPerSqrMeter(10)
}
\keyword{programming}