\name{NISTftOfMercuryConvtnlTOkpascal}
\alias{NISTftOfMercuryConvtnlTOkpascal}
\title{Convert foot of mercury, conventional 12 to kilopascal }
\usage{NISTftOfMercuryConvtnlTOkpascal(ftOfMercuryConvtnl)}
\description{\code{NISTftOfMercuryConvtnlTOkpascal} converts from foot of mercury, conventional (ftHg) 12 to kilopascal (kPa) }
\arguments{
  \item{ftOfMercuryConvtnl}{foot of mercury, conventional (ftHg) 12 }
}
\value{kilopascal (kPa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTftOfMercuryConvtnlTOkpascal(10)
}
\keyword{programming}