\name{NISTftOfMercuryConvtnlTOpascal}
\alias{NISTftOfMercuryConvtnlTOpascal}
\title{Convert foot of mercury, conventional 12 to pascal }
\usage{NISTftOfMercuryConvtnlTOpascal(ftOfMercuryConvtnl)}
\description{\code{NISTftOfMercuryConvtnlTOpascal} converts from foot of mercury, conventional (ftHg) 12 to pascal (Pa) }
\arguments{
  \item{ftOfMercuryConvtnl}{foot of mercury, conventional (ftHg) 12 }
}
\value{pascal (Pa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTftOfMercuryConvtnlTOpascal(10)
}
\keyword{programming}