\name{NISTftOfWaterConvtnlTOkpascal}
\alias{NISTftOfWaterConvtnlTOkpascal}
\title{Convert foot of water, conventional 12 to kilopascal }
\usage{NISTftOfWaterConvtnlTOkpascal(ftOfWaterConvtnl)}
\description{\code{NISTftOfWaterConvtnlTOkpascal} converts from foot of water, conventional (ftH2O) 12 to kilopascal (kPa) }
\arguments{
  \item{ftOfWaterConvtnl}{foot of water, conventional (ftH2O) 12 }
}
\value{kilopascal (kPa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTftOfWaterConvtnlTOkpascal(10)
}
\keyword{programming}