\name{NISTftOfWaterTOpascal}
\alias{NISTftOfWaterTOpascal}
\title{Convert foot of water 12 to pascal }
\usage{NISTftOfWaterTOpascal(ftOfWater)}
\description{\code{NISTftOfWaterTOpascal} converts from foot of water (39.2 F) 12 to pascal (Pa) }
\arguments{
  \item{ftOfWater}{foot of water (39.2 F) 12 }
}
\value{pascal (Pa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTftOfWaterTOpascal(10)
}
\keyword{programming}