\name{NISTftPerHourTOmeterPerSec}
\alias{NISTftPerHourTOmeterPerSec}
\title{Convert foot per hour  to meter per second }
\usage{NISTftPerHourTOmeterPerSec(ftPerHour)}
\description{\code{NISTftPerHourTOmeterPerSec} converts from foot per hour (ft/h) to meter per second (m/s) }
\arguments{
  \item{ftPerHour}{foot per hour (ft/h) }
}
\value{meter per second (m/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTftPerHourTOmeterPerSec(10)
}
\keyword{programming}