\name{NISTftPerMinTOmeterPerSec}
\alias{NISTftPerMinTOmeterPerSec}
\title{Convert foot per minute  to meter per second }
\usage{NISTftPerMinTOmeterPerSec(ftPerMin)}
\description{\code{NISTftPerMinTOmeterPerSec} converts from foot per minute (ft/min) to meter per second (m/s) }
\arguments{
  \item{ftPerMin}{foot per minute (ft/min) }
}
\value{meter per second (m/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTftPerMinTOmeterPerSec(10)
}
\keyword{programming}