\name{NISTftPerSecSqrdTOmeterPerSecSqrd}
\alias{NISTftPerSecSqrdTOmeterPerSecSqrd}
\title{Convert foot per second squared  to meter per second squared }
\usage{NISTftPerSecSqrdTOmeterPerSecSqrd(ftPerSecSqrd)}
\description{\code{NISTftPerSecSqrdTOmeterPerSecSqrd} converts from foot per second squared (ft/s2) to meter per second squared (m/s2) }
\arguments{
  \item{ftPerSecSqrd}{foot per second squared (ft/s2) }
}
\value{meter per second squared (m/s2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTftPerSecSqrdTOmeterPerSecSqrd(10)
}
\keyword{programming}