\name{NISTftPerSecTOmeterPerSec}
\alias{NISTftPerSecTOmeterPerSec}
\title{Convert foot per second  to meter per second }
\usage{NISTftPerSecTOmeterPerSec(ftPerSec)}
\description{\code{NISTftPerSecTOmeterPerSec} converts from foot per second (ft/s) to meter per second (m/s) }
\arguments{
  \item{ftPerSec}{foot per second (ft/s) }
}
\value{meter per second (m/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTftPerSecTOmeterPerSec(10)
}
\keyword{programming}