\name{NISTftPoundalTOjoule}
\alias{NISTftPoundalTOjoule}
\title{Convert foot poundal to joule }
\usage{NISTftPoundalTOjoule(ftPoundal)}
\description{\code{NISTftPoundalTOjoule} converts from foot poundal to joule (J) }
\arguments{
  \item{ftPoundal}{foot poundal }
}
\value{joule (J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTftPoundalTOjoule(10)
}
\keyword{programming}