\name{NISTftPoundForcePerHourTOwatt}
\alias{NISTftPoundForcePerHourTOwatt}
\title{Convert foot pound-force per hour  to watt }
\usage{NISTftPoundForcePerHourTOwatt(ftPoundForcePerHour)}
\description{\code{NISTftPoundForcePerHourTOwatt} converts from foot pound-force per hour (ft * lbf/h) to watt (W) }
\arguments{
  \item{ftPoundForcePerHour}{foot pound-force per hour (ft * lbf/h) }
}
\value{watt (W) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTftPoundForcePerHourTOwatt(10)
}
\keyword{programming}