\name{NISTftPoundForcePerMinTOwatt}
\alias{NISTftPoundForcePerMinTOwatt}
\title{Convert foot pound-force per minute  to watt }
\usage{NISTftPoundForcePerMinTOwatt(ftPoundForcePerMin)}
\description{\code{NISTftPoundForcePerMinTOwatt} converts from foot pound-force per minute (ft * lbf/min) to watt (W) }
\arguments{
  \item{ftPoundForcePerMin}{foot pound-force per minute (ft * lbf/min) }
}
\value{watt (W) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTftPoundForcePerMinTOwatt(10)
}
\keyword{programming}