\name{NISTftPoundForcePerSecTOwatt}
\alias{NISTftPoundForcePerSecTOwatt}
\title{Convert foot pound-force per second  to watt }
\usage{NISTftPoundForcePerSecTOwatt(ftPoundForcePerSec)}
\description{\code{NISTftPoundForcePerSecTOwatt} converts from foot pound-force per second (ft * lbf/s) to watt (W) }
\arguments{
  \item{ftPoundForcePerSec}{foot pound-force per second (ft * lbf/s) }
}
\value{watt (W) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTftPoundForcePerSecTOwatt(10)
}
\keyword{programming}