\name{NISTftPoundForceTOjoule}
\alias{NISTftPoundForceTOjoule}
\title{Convert foot pound-force  to joule }
\usage{NISTftPoundForceTOjoule(ftPoundForce)}
\description{\code{NISTftPoundForceTOjoule} converts from foot pound-force (ft * lbf) to joule (J) }
\arguments{
  \item{ftPoundForce}{foot pound-force (ft * lbf) }
}
\value{joule (J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTftPoundForceTOjoule(10)
}
\keyword{programming}