\name{NISTftTOmeter}
\alias{NISTftTOmeter}
\title{Convert foot  to meter }
\usage{NISTftTOmeter(ft)}
\description{\code{NISTftTOmeter} converts from foot (ft) to meter (m) }
\arguments{
  \item{ft}{foot (ft) }
}
\value{meter (m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTftTOmeter(10)
}
\keyword{programming}