\name{NISTftUSsurveyTOmeter}
\alias{NISTftUSsurveyTOmeter}
\title{Convert foot 7 to meter }
\usage{NISTftUSsurveyTOmeter(ftUSsurvey)}
\description{\code{NISTftUSsurveyTOmeter} converts from foot (U.S. survey) (ft) 7 to meter (m) }
\arguments{
  \item{ftUSsurvey}{foot (U.S. survey) (ft) 7 }
}
\value{meter (m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTftUSsurveyTOmeter(10)
}
\keyword{programming}