\name{NISTgallonImperialTOcubMeter}
\alias{NISTgallonImperialTOcubMeter}
\title{Convert gallon  to cubic meter }
\usage{NISTgallonImperialTOcubMeter(gallonImperial)}
\description{\code{NISTgallonImperialTOcubMeter} converts from gallon [Canadian and U.K. (Imperial)] (gal) to cubic meter (m3) }
\arguments{
  \item{gallonImperial}{gallon [Canadian and U.K. (Imperial)] (gal) }
}
\value{cubic meter (m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgallonImperialTOcubMeter(10)
}
\keyword{programming}