\name{NISTgallonImperialTOliter}
\alias{NISTgallonImperialTOliter}
\title{Convert gallon  to liter }
\usage{NISTgallonImperialTOliter(gallonImperial)}
\description{\code{NISTgallonImperialTOliter} converts from gallon [Canadian and U.K. (Imperial)] (gal) to liter (L) }
\arguments{
  \item{gallonImperial}{gallon [Canadian and U.K. (Imperial)] (gal) }
}
\value{liter (L) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgallonImperialTOliter(10)
}
\keyword{programming}