\name{NISTgallonPerDayTOcubMeterPerSec}
\alias{NISTgallonPerDayTOcubMeterPerSec}
\title{Convert gallon per day  to cubic meter per second }
\usage{NISTgallonPerDayTOcubMeterPerSec(gallonPerDay)}
\description{\code{NISTgallonPerDayTOcubMeterPerSec} converts from gallon (U.S.) per day (gal/d) to cubic meter per second (m3/s) }
\arguments{
  \item{gallonPerDay}{gallon (U.S.) per day (gal/d) }
}
\value{cubic meter per second (m3/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgallonPerDayTOcubMeterPerSec(10)
}
\keyword{programming}