\name{NISTgallonPerDayTOliterPerSec}
\alias{NISTgallonPerDayTOliterPerSec}
\title{Convert gallon per day  to liter per second }
\usage{NISTgallonPerDayTOliterPerSec(gallonPerDay)}
\description{\code{NISTgallonPerDayTOliterPerSec} converts from gallon (U.S.) per day (gal/d) to liter per second (L/s) }
\arguments{
  \item{gallonPerDay}{gallon (U.S.) per day (gal/d) }
}
\value{liter per second (L/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgallonPerDayTOliterPerSec(10)
}
\keyword{programming}