\name{NISTgallonPerHorsepowerHourTOcubMeterPerJoule}
\alias{NISTgallonPerHorsepowerHourTOcubMeterPerJoule}
\title{Convert gallon per horsepower hour  to cubic meter per joule }
\usage{NISTgallonPerHorsepowerHourTOcubMeterPerJoule(gallonPerHorsepowerHour)}
\description{\code{NISTgallonPerHorsepowerHourTOcubMeterPerJoule} converts from gallon (U.S.) per horsepower hour [gal/(hp * h)] to cubic meter per joule (m3/J) }
\arguments{
  \item{gallonPerHorsepowerHour}{gallon (U.S.) per horsepower hour [gal/(hp * h)] }
}
\value{cubic meter per joule (m3/J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgallonPerHorsepowerHourTOcubMeterPerJoule(10)
}
\keyword{programming}