\name{NISTgallonPerHorsepowerHourTOliterPerJoule}
\alias{NISTgallonPerHorsepowerHourTOliterPerJoule}
\title{Convert gallon per horsepower hour  to liter per joule }
\usage{NISTgallonPerHorsepowerHourTOliterPerJoule(gallonPerHorsepowerHour)}
\description{\code{NISTgallonPerHorsepowerHourTOliterPerJoule} converts from gallon (U.S.) per horsepower hour [gal/(hp * h)] to liter per joule (L/J) }
\arguments{
  \item{gallonPerHorsepowerHour}{gallon (U.S.) per horsepower hour [gal/(hp * h)] }
}
\value{liter per joule (L/J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgallonPerHorsepowerHourTOliterPerJoule(10)
}
\keyword{programming}