\name{NISTgallonPerMinTOcubMeterPerSec}
\alias{NISTgallonPerMinTOcubMeterPerSec}
\title{Convert gallon per minute  to cubic meter per second }
\usage{NISTgallonPerMinTOcubMeterPerSec(gallonPerMin)}
\description{\code{NISTgallonPerMinTOcubMeterPerSec} converts from gallon (U.S.) per minute (gpm) (gal/min) to cubic meter per second (m3/s) }
\arguments{
  \item{gallonPerMin}{gallon (U.S.) per minute (gpm) (gal/min) }
}
\value{cubic meter per second (m3/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgallonPerMinTOcubMeterPerSec(10)
}
\keyword{programming}