\name{NISTgallonPerMinTOliterPerSec}
\alias{NISTgallonPerMinTOliterPerSec}
\title{Convert gallon per minute  to liter per second }
\usage{NISTgallonPerMinTOliterPerSec(gallonPerMin)}
\description{\code{NISTgallonPerMinTOliterPerSec} converts from gallon (U.S.) per minute (gpm) (gal/min) to liter per second (L/s) }
\arguments{
  \item{gallonPerMin}{gallon (U.S.) per minute (gpm) (gal/min) }
}
\value{liter per second (L/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgallonPerMinTOliterPerSec(10)
}
\keyword{programming}