\name{NISTgallonTOcubMeter}
\alias{NISTgallonTOcubMeter}
\title{Convert gallon  to cubic meter }
\usage{NISTgallonTOcubMeter(gallon)}
\description{\code{NISTgallonTOcubMeter} converts from gallon (U.S.) (gal) to cubic meter (m3) }
\arguments{
  \item{gallon}{gallon (U.S.) (gal) }
}
\value{cubic meter (m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgallonTOcubMeter(10)
}
\keyword{programming}