\name{NISTgalTOmeterPerSecSqrd}
\alias{NISTgalTOmeterPerSecSqrd}
\title{Convert gal  to meter per second squared }
\usage{NISTgalTOmeterPerSecSqrd(gal)}
\description{\code{NISTgalTOmeterPerSecSqrd} converts from gal (Gal) to meter per second squared (m/s2) }
\arguments{
  \item{gal}{gal (Gal) }
}
\value{meter per second squared (m/s2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgalTOmeterPerSecSqrd(10)
}
\keyword{programming}