\name{NISTgammaTOtesla}
\alias{NISTgammaTOtesla}
\title{Convert gamma  to tesla }
\usage{NISTgammaTOtesla(gamma)}
\description{\code{NISTgammaTOtesla} converts from gamma (y) to tesla (T) }
\arguments{
  \item{gamma}{gamma (y) }
}
\value{tesla (T) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgammaTOtesla(10)
}
\keyword{programming}