\name{NISTgaussTOtesla}
\alias{NISTgaussTOtesla}
\title{Convert gauss  to tesla }
\usage{NISTgaussTOtesla(gauss)}
\description{\code{NISTgaussTOtesla} converts from gauss (Gs, G) to tesla (T) }
\arguments{
  \item{gauss}{gauss (Gs, G) }
}
\value{tesla (T) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgaussTOtesla(10)
}
\keyword{programming}