\name{NISTgilbertTOampere}
\alias{NISTgilbertTOampere}
\title{Convert gilbert  to ampere }
\usage{NISTgilbertTOampere(gilbert)}
\description{\code{NISTgilbertTOampere} converts from gilbert (Gi) to ampere (A) }
\arguments{
  \item{gilbert}{gilbert (Gi) }
}
\value{ampere (A) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgilbertTOampere(10)
}
\keyword{programming}