\name{NISTgillImperialTOliter}
\alias{NISTgillImperialTOliter}
\title{Convert gill  to liter }
\usage{NISTgillImperialTOliter(gillImperial)}
\description{\code{NISTgillImperialTOliter} converts from gill [Canadian and U.K. (Imperial)] (gi) to liter (L) }
\arguments{
  \item{gillImperial}{gill [Canadian and U.K. (Imperial)] (gi) }
}
\value{liter (L) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgillImperialTOliter(10)
}
\keyword{programming}