\name{NISTgillTOcubMeter}
\alias{NISTgillTOcubMeter}
\title{Convert gill  to cubic meter }
\usage{NISTgillTOcubMeter(gill)}
\description{\code{NISTgillTOcubMeter} converts from gill (U.S.) (gi) to cubic meter (m3) }
\arguments{
  \item{gill}{gill (U.S.) (gi) }
}
\value{cubic meter (m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgillTOcubMeter(10)
}
\keyword{programming}