\name{NISTgillTOliter}
\alias{NISTgillTOliter}
\title{Convert gill  to liter }
\usage{NISTgillTOliter(gill)}
\description{\code{NISTgillTOliter} converts from gill (U.S.) (gi) to liter (L) }
\arguments{
  \item{gill}{gill (U.S.) (gi) }
}
\value{liter (L) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgillTOliter(10)
}
\keyword{programming}