\name{NISTgonTOdeg}
\alias{NISTgonTOdeg}
\title{Convert gon  to degree }
\usage{NISTgonTOdeg(gon)}
\description{\code{NISTgonTOdeg} converts from gon (also called grade) (gon) to degree (angle) () }
\arguments{
  \item{gon}{gon (also called grade) (gon) }
}
\value{degree (angle) () }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgonTOdeg(10)
}
\keyword{programming}