\name{NISTgonTOradian}
\alias{NISTgonTOradian}
\title{Convert gon  to radian }
\usage{NISTgonTOradian(gon)}
\description{\code{NISTgonTOradian} converts from gon (also called grade) (gon) to radian (rad) }
\arguments{
  \item{gon}{gon (also called grade) (gon) }
}
\value{radian (rad) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgonTOradian(10)
}
\keyword{programming}