\name{NISTgrainPerGallonTOkgPerCubMeter}
\alias{NISTgrainPerGallonTOkgPerCubMeter}
\title{Convert grain per gallon  to kilogram per cubic meter }
\usage{NISTgrainPerGallonTOkgPerCubMeter(grainPerGallon)}
\description{\code{NISTgrainPerGallonTOkgPerCubMeter} converts from grain per gallon (U.S.) (gr/gal) to kilogram per cubic meter (kg/m3) }
\arguments{
  \item{grainPerGallon}{grain per gallon (U.S.) (gr/gal) }
}
\value{kilogram per cubic meter (kg/m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgrainPerGallonTOkgPerCubMeter(10)
}
\keyword{programming}