\name{NISTgrainPerGallonTOmilligramPerLiter}
\alias{NISTgrainPerGallonTOmilligramPerLiter}
\title{Convert grain per gallon  to milligram per liter }
\usage{NISTgrainPerGallonTOmilligramPerLiter(grainPerGallon)}
\description{\code{NISTgrainPerGallonTOmilligramPerLiter} converts from grain per gallon (U.S.) (gr/gal) to milligram per liter (mg/L) }
\arguments{
  \item{grainPerGallon}{grain per gallon (U.S.) (gr/gal) }
}
\value{milligram per liter (mg/L) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgrainPerGallonTOmilligramPerLiter(10)
}
\keyword{programming}