\name{NISTgrainTOkg}
\alias{NISTgrainTOkg}
\title{Convert grain  to kilogram }
\usage{NISTgrainTOkg(grain)}
\description{\code{NISTgrainTOkg} converts from grain (gr) to kilogram (kg) }
\arguments{
  \item{grain}{grain (gr) }
}
\value{kilogram (kg) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgrainTOkg(10)
}
\keyword{programming}