\name{NISTgrainTOmilligram}
\alias{NISTgrainTOmilligram}
\title{Convert grain  to milligram }
\usage{NISTgrainTOmilligram(grain)}
\description{\code{NISTgrainTOmilligram} converts from grain (gr) to milligram (mg) }
\arguments{
  \item{grain}{grain (gr) }
}
\value{milligram (mg) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgrainTOmilligram(10)
}
\keyword{programming}