\name{NISTgramForcePerSqrCmTOpascal}
\alias{NISTgramForcePerSqrCmTOpascal}
\title{Convert gram-force per square centimeter  to pascal }
\usage{NISTgramForcePerSqrCmTOpascal(gramForcePerSqrCm)}
\description{\code{NISTgramForcePerSqrCmTOpascal} converts from gram-force per square centimeter (gf/cm2) to pascal (Pa) }
\arguments{
  \item{gramForcePerSqrCm}{gram-force per square centimeter (gf/cm2) }
}
\value{pascal (Pa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgramForcePerSqrCmTOpascal(10)
}
\keyword{programming}