\name{NISTgramPerCubCmTOkgPerCubMeter}
\alias{NISTgramPerCubCmTOkgPerCubMeter}
\title{Convert gram per cubic centimeter  to kilogram per cubic meter }
\usage{NISTgramPerCubCmTOkgPerCubMeter(gramPerCubCm)}
\description{\code{NISTgramPerCubCmTOkgPerCubMeter} converts from gram per cubic centimeter (g/cm3) to kilogram per cubic meter (kg/m3) }
\arguments{
  \item{gramPerCubCm}{gram per cubic centimeter (g/cm3) }
}
\value{kilogram per cubic meter (kg/m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgramPerCubCmTOkgPerCubMeter(10)
}
\keyword{programming}