\name{NISTgramPerLiterTOouncePerGallon}
\alias{NISTgramPerLiterTOouncePerGallon}
\title{Convert gram per liter  to ounce per gallon }
\usage{NISTgramPerLiterTOouncePerGallon(gramPerLiter)}
\description{\code{NISTgramPerLiterTOouncePerGallon} converts from gram per liter (g/L) to ounce (avoirdupois) per gallon (U.S.) (oz/gal) }
\arguments{
  \item{gramPerLiter}{gram per liter (g/L) }
}
\value{ounce (avoirdupois) per gallon (U.S.) (oz/gal) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgramPerLiterTOouncePerGallon(10)
}
\keyword{programming}