\name{NISTgramPerMeterTOdenier}
\alias{NISTgramPerMeterTOdenier}
\title{Convert gram per meter  to denier}
\usage{NISTgramPerMeterTOdenier(gramPerMeter)}
\description{\code{NISTgramPerMeterTOdenier} converts from gram per meter (g/m) to denier }
\arguments{
  \item{gramPerMeter}{gram per meter (g/m) }
}
\value{denier }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgramPerMeterTOdenier(10)
}
\keyword{programming}