\name{NISTgramTOcaratSI}
\alias{NISTgramTOcaratSI}
\title{Convert gram  to carat, metric}
\usage{NISTgramTOcaratSI(gram)}
\description{\code{NISTgramTOcaratSI} converts from gram (g) to carat, metric }
\arguments{
  \item{gram}{gram (g) }
}
\value{carat, metric }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgramTOcaratSI(10)
}
\keyword{programming}