\name{NISTgramTOounceAvoirdupois}
\alias{NISTgramTOounceAvoirdupois}
\title{Convert gram  to ounce }
\usage{NISTgramTOounceAvoirdupois(gram)}
\description{\code{NISTgramTOounceAvoirdupois} converts from gram (g) to ounce (avoirdupois) (oz) }
\arguments{
  \item{gram}{gram (g) }
}
\value{ounce (avoirdupois) (oz) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgramTOounceAvoirdupois(10)
}
\keyword{programming}