\name{NISTgramTOounceTroy}
\alias{NISTgramTOounceTroy}
\title{Convert gram  to ounce }
\usage{NISTgramTOounceTroy(gram)}
\description{\code{NISTgramTOounceTroy} converts from gram (g) to ounce (troy or apothecary) (oz) }
\arguments{
  \item{gram}{gram (g) }
}
\value{ounce (troy or apothecary) (oz) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgramTOounceTroy(10)
}
\keyword{programming}