\name{NISTgramTOpennyweight}
\alias{NISTgramTOpennyweight}
\title{Convert gram  to pennyweight }
\usage{NISTgramTOpennyweight(gram)}
\description{\code{NISTgramTOpennyweight} converts from gram (g) to pennyweight (dwt) }
\arguments{
  \item{gram}{gram (g) }
}
\value{pennyweight (dwt) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgramTOpennyweight(10)
}
\keyword{programming}