\name{NISTgrayTOrad}
\alias{NISTgrayTOrad}
\title{Convert gray  to rad }
\usage{NISTgrayTOrad(gray)}
\description{\code{NISTgrayTOrad} converts from gray (Gy) to rad (absorbed dose) (rad) }
\arguments{
  \item{gray}{gray (Gy) }
}
\value{rad (absorbed dose) (rad) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTgrayTOrad(10)
}
\keyword{programming}