\name{NISThectareTOsqrMeter}
\alias{NISThectareTOsqrMeter}
\title{Convert hectare  to square meter }
\usage{NISThectareTOsqrMeter(hectare)}
\description{\code{NISThectareTOsqrMeter} converts from hectare (ha) to square meter (m2) }
\arguments{
  \item{hectare}{hectare (ha) }
}
\value{square meter (m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISThectareTOsqrMeter(10)
}
\keyword{programming}