\name{NISThenryTOabhenry}
\alias{NISThenryTOabhenry}
\title{Convert henry  to abhenry}
\usage{NISThenryTOabhenry(henry)}
\description{\code{NISThenryTOabhenry} converts from henry (H) to abhenry }
\arguments{
  \item{henry}{henry (H) }
}
\value{abhenry }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISThenryTOabhenry(10)
}
\keyword{programming}