\name{NISThenryToEMUOfInductance}
\alias{NISThenryToEMUOfInductance}
\title{Convert henry  to EMU of inductance }
\usage{NISThenryToEMUOfInductance(henry)}
\description{\code{NISThenryToEMUOfInductance} converts from henry (H) to EMU of inductance (abhenry) }
\arguments{
  \item{henry}{henry (H) }
}
\value{EMU of inductance (abhenry) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISThenryToEMUOfInductance(10)
}
\keyword{programming}