\name{NISThenryToESUOfInductance}
\alias{NISThenryToESUOfInductance}
\title{Convert henry  to ESU of inductance }
\usage{NISThenryToESUOfInductance(henry)}
\description{\code{NISThenryToESUOfInductance} converts from henry (H) to ESU of inductance (stathenry) }
\arguments{
  \item{henry}{henry (H) }
}
\value{ESU of inductance (stathenry) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISThenryToESUOfInductance(10)
}
\keyword{programming}