\name{NISThenryTOstathenry}
\alias{NISThenryTOstathenry}
\title{Convert henry  to stathenry}
\usage{NISThenryTOstathenry(henry)}
\description{\code{NISThenryTOstathenry} converts from henry (H) to stathenry }
\arguments{
  \item{henry}{henry (H) }
}
\value{stathenry }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISThenryTOstathenry(10)
}
\keyword{programming}