\name{NISThorsepowerBoilerTOwatt}
\alias{NISThorsepowerBoilerTOwatt}
\title{Convert horsepower  to watt }
\usage{NISThorsepowerBoilerTOwatt(horsepowerBoiler)}
\description{\code{NISThorsepowerBoilerTOwatt} converts from horsepower (boiler) to watt (W) }
\arguments{
  \item{horsepowerBoiler}{horsepower (boiler) }
}
\value{watt (W) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISThorsepowerBoilerTOwatt(10)
}
\keyword{programming}