\name{NISThorsepowerElectricTOwatt}
\alias{NISThorsepowerElectricTOwatt}
\title{Convert horsepower  to watt }
\usage{NISThorsepowerElectricTOwatt(horsepowerElectric)}
\description{\code{NISThorsepowerElectricTOwatt} converts from horsepower (electric) to watt (W) }
\arguments{
  \item{horsepowerElectric}{horsepower (electric) }
}
\value{watt (W) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISThorsepowerElectricTOwatt(10)
}
\keyword{programming}