\name{NISThorsepowerImperialTOwatt}
\alias{NISThorsepowerImperialTOwatt}
\title{Convert horsepower  to watt }
\usage{NISThorsepowerImperialTOwatt(horsepowerImperial)}
\description{\code{NISThorsepowerImperialTOwatt} converts from horsepower (U.K.) to watt (W) }
\arguments{
  \item{horsepowerImperial}{horsepower (U.K.) }
}
\value{watt (W) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISThorsepowerImperialTOwatt(10)
}
\keyword{programming}