\name{NISThorsepowerSItOwatt}
\alias{NISThorsepowerSItOwatt}
\title{Convert horsepower  to watt }
\usage{NISThorsepowerSItOwatt(horsepowerSI)}
\description{\code{NISThorsepowerSItOwatt} converts from horsepower (metric) to watt (W) }
\arguments{
  \item{horsepowerSI}{horsepower (metric) }
}
\value{watt (W) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISThorsepowerSItOwatt(10)
}
\keyword{programming}