\name{NISThorsepowerTOwatt}
\alias{NISThorsepowerTOwatt}
\title{Convert horsepower  to watt }
\usage{NISThorsepowerTOwatt(horsepower)}
\description{\code{NISThorsepowerTOwatt} converts from horsepower (550 ft * lbf/s) (hp) to watt (W) }
\arguments{
  \item{horsepower}{horsepower (550 ft * lbf/s) (hp) }
}
\value{watt (W) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISThorsepowerTOwatt(10)
}
\keyword{programming}