\name{NISThorsepowerWaterTOwatt}
\alias{NISThorsepowerWaterTOwatt}
\title{Convert horsepower  to watt }
\usage{NISThorsepowerWaterTOwatt(horsepowerWater)}
\description{\code{NISThorsepowerWaterTOwatt} converts from horsepower (water) to watt (W) }
\arguments{
  \item{horsepowerWater}{horsepower (water) }
}
\value{watt (W) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISThorsepowerWaterTOwatt(10)
}
\keyword{programming}