\name{NISThourSiderealTOsec}
\alias{NISThourSiderealTOsec}
\title{Convert hour  to second }
\usage{NISThourSiderealTOsec(hourSidereal)}
\description{\code{NISThourSiderealTOsec} converts from hour (sidereal) to second (s) }
\arguments{
  \item{hourSidereal}{hour (sidereal) }
}
\value{second (s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISThourSiderealTOsec(10)
}
\keyword{programming}