\name{NISThourTOsec}
\alias{NISThourTOsec}
\title{Convert hour  to second }
\usage{NISThourTOsec(hour)}
\description{\code{NISThourTOsec} converts from hour (h) to second (s) }
\arguments{
  \item{hour}{hour (h) }
}
\value{second (s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISThourTOsec(10)
}
\keyword{programming}