\name{NISThundredweightLongTOkg}
\alias{NISThundredweightLongTOkg}
\title{Convert hundredweight  to kilogram }
\usage{NISThundredweightLongTOkg(hundredweightLong)}
\description{\code{NISThundredweightLongTOkg} converts from hundredweight (long, 112 lb) to kilogram (kg) }
\arguments{
  \item{hundredweightLong}{hundredweight (long, 112 lb) }
}
\value{kilogram (kg) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISThundredweightLongTOkg(10)
}
\keyword{programming}