\name{NISThundredweightShortTOkg}
\alias{NISThundredweightShortTOkg}
\title{Convert hundredweight  to kilogram }
\usage{NISThundredweightShortTOkg(hundredweightShort)}
\description{\code{NISThundredweightShortTOkg} converts from hundredweight (short, 100 lb) to kilogram (kg) }
\arguments{
  \item{hundredweightShort}{hundredweight (short, 100 lb) }
}
\value{kilogram (kg) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISThundredweightShortTOkg(10)
}
\keyword{programming}