\name{NISTinch4thPowerTOmeter4thPower}
\alias{NISTinch4thPowerTOmeter4thPower}
\title{Convert inch to the fourth power 16 to meter to the fourth power }
\usage{NISTinch4thPowerTOmeter4thPower(inch4thPower)}
\description{\code{NISTinch4thPowerTOmeter4thPower} converts from inch to the fourth power (in4) 16 to meter to the fourth power (m4) }
\arguments{
  \item{inch4thPower}{inch to the fourth power (in4) 16 }
}
\value{meter to the fourth power (m4) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTinch4thPowerTOmeter4thPower(10)
}
\keyword{programming}