\name{NISTinchOfMercury32FtOkpascal}
\alias{NISTinchOfMercury32FtOkpascal}
\title{Convert inch of mercury 12 to kilopascal }
\usage{NISTinchOfMercury32FtOkpascal(inchOfMercury32F)}
\description{\code{NISTinchOfMercury32FtOkpascal} converts from inch of mercury (32 F) 12 to kilopascal (kPa) }
\arguments{
  \item{inchOfMercury32F}{inch of mercury (32 F) 12 }
}
\value{kilopascal (kPa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTinchOfMercury32FtOkpascal(10)
}
\keyword{programming}