\name{NISTinchOfMercury32FtOpascal}
\alias{NISTinchOfMercury32FtOpascal}
\title{Convert inch of mercury 12 to pascal }
\usage{NISTinchOfMercury32FtOpascal(inchOfMercury32F)}
\description{\code{NISTinchOfMercury32FtOpascal} converts from inch of mercury (32 F) 12 to pascal (Pa) }
\arguments{
  \item{inchOfMercury32F}{inch of mercury (32 F) 12 }
}
\value{pascal (Pa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTinchOfMercury32FtOpascal(10)
}
\keyword{programming}