\name{NISTinchOfMercury60FtOpascal}
\alias{NISTinchOfMercury60FtOpascal}
\title{Convert inch of mercury 12 to pascal }
\usage{NISTinchOfMercury60FtOpascal(inchOfMercury60F)}
\description{\code{NISTinchOfMercury60FtOpascal} converts from inch of mercury (60 F) 12 to pascal (Pa) }
\arguments{
  \item{inchOfMercury60F}{inch of mercury (60 F) 12 }
}
\value{pascal (Pa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTinchOfMercury60FtOpascal(10)
}
\keyword{programming}