\name{NISTinchOfMercuryConvtnlTOkpascal}
\alias{NISTinchOfMercuryConvtnlTOkpascal}
\title{Convert inch of mercury, conventional 12 to kilopascal }
\usage{NISTinchOfMercuryConvtnlTOkpascal(inchOfMercuryConvtnl)}
\description{\code{NISTinchOfMercuryConvtnlTOkpascal} converts from inch of mercury, conventional (inHg) 12 to kilopascal (kPa) }
\arguments{
  \item{inchOfMercuryConvtnl}{inch of mercury, conventional (inHg) 12 }
}
\value{kilopascal (kPa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTinchOfMercuryConvtnlTOkpascal(10)
}
\keyword{programming}