\name{NISTinchOfMercuryConvtnlTOpascal}
\alias{NISTinchOfMercuryConvtnlTOpascal}
\title{Convert inch of mercury, conventional 12 to pascal }
\usage{NISTinchOfMercuryConvtnlTOpascal(inchOfMercuryConvtnl)}
\description{\code{NISTinchOfMercuryConvtnlTOpascal} converts from inch of mercury, conventional (inHg) 12 to pascal (Pa) }
\arguments{
  \item{inchOfMercuryConvtnl}{inch of mercury, conventional (inHg) 12 }
}
\value{pascal (Pa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTinchOfMercuryConvtnlTOpascal(10)
}
\keyword{programming}