\name{NISTinchOfWater39.2FtOpascal}
\alias{NISTinchOfWater39.2FtOpascal}
\title{Convert inch of water 12 to pascal }
\usage{NISTinchOfWater39.2FtOpascal(inchOfWater39.2F)}
\description{\code{NISTinchOfWater39.2FtOpascal} converts from inch of water (39.2 F) 12 to pascal (Pa) }
\arguments{
  \item{inchOfWater39.2F}{inch of water (39.2 F) 12 }
}
\value{pascal (Pa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTinchOfWater39.2FtOpascal(10)
}
\keyword{programming}