\name{NISTinchOfWater60FtOpascal}
\alias{NISTinchOfWater60FtOpascal}
\title{Convert inch of water 12 to pascal }
\usage{NISTinchOfWater60FtOpascal(inchOfWater60F)}
\description{\code{NISTinchOfWater60FtOpascal} converts from inch of water (60 F) 12 to pascal (Pa) }
\arguments{
  \item{inchOfWater60F}{inch of water (60 F) 12 }
}
\value{pascal (Pa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTinchOfWater60FtOpascal(10)
}
\keyword{programming}