\name{NISTinchOfWaterConvtnlTOpascal}
\alias{NISTinchOfWaterConvtnlTOpascal}
\title{Convert inch of water, conventional 12 to pascal }
\usage{NISTinchOfWaterConvtnlTOpascal(inchOfWaterConvtnl)}
\description{\code{NISTinchOfWaterConvtnlTOpascal} converts from inch of water, conventional (inH2O) 12 to pascal (Pa) }
\arguments{
  \item{inchOfWaterConvtnl}{inch of water, conventional (inH2O) 12 }
}
\value{pascal (Pa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTinchOfWaterConvtnlTOpascal(10)
}
\keyword{programming}