\name{NISTinchPerSecSqrdTOmeterPerSecSqrd}
\alias{NISTinchPerSecSqrdTOmeterPerSecSqrd}
\title{Convert inch per second squared  to meter per second squared }
\usage{NISTinchPerSecSqrdTOmeterPerSecSqrd(inchPerSecSqrd)}
\description{\code{NISTinchPerSecSqrdTOmeterPerSecSqrd} converts from inch per second squared (in/s2) to meter per second squared (m/s2) }
\arguments{
  \item{inchPerSecSqrd}{inch per second squared (in/s2) }
}
\value{meter per second squared (m/s2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTinchPerSecSqrdTOmeterPerSecSqrd(10)
}
\keyword{programming}