\name{NISTinchPerSecTOmeterPerSec}
\alias{NISTinchPerSecTOmeterPerSec}
\title{Convert inch per second  to meter per second }
\usage{NISTinchPerSecTOmeterPerSec(inchPerSec)}
\description{\code{NISTinchPerSecTOmeterPerSec} converts from inch per second (in/s) to meter per second (m/s) }
\arguments{
  \item{inchPerSec}{inch per second (in/s) }
}
\value{meter per second (m/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTinchPerSecTOmeterPerSec(10)
}
\keyword{programming}