\name{NISTinchTOcm}
\alias{NISTinchTOcm}
\title{Convert inch  to centimeter }
\usage{NISTinchTOcm(inch)}
\description{\code{NISTinchTOcm} converts from inch (in) to centimeter (cm) }
\arguments{
  \item{inch}{inch (in) }
}
\value{centimeter (cm) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTinchTOcm(10)
}
\keyword{programming}