\name{NISTjoulePerCubMeterTOukThUnITPerCubFt}
\alias{NISTjoulePerCubMeterTOukThUnITPerCubFt}
\title{Convert joule per cubic meter  to British thermal unitIT per cubic foot }
\usage{NISTjoulePerCubMeterTOukThUnITPerCubFt(joulePerCubMeter)}
\description{\code{NISTjoulePerCubMeterTOukThUnITPerCubFt} converts from joule per cubic meter (J/m3) to British thermal unitIT per cubic foot (BtuIT/ft3) }
\arguments{
  \item{joulePerCubMeter}{joule per cubic meter (J/m3) }
}
\value{British thermal unitIT per cubic foot (BtuIT/ft3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjoulePerCubMeterTOukThUnITPerCubFt(10)
}
\keyword{programming}