\name{NISTjoulePerCubMeterTOukThUnthPerCubFt}
\alias{NISTjoulePerCubMeterTOukThUnthPerCubFt}
\title{Convert joule per cubic meter  to British thermal unitth per cubic foot }
\usage{NISTjoulePerCubMeterTOukThUnthPerCubFt(joulePerCubMeter)}
\description{\code{NISTjoulePerCubMeterTOukThUnthPerCubFt} converts from joule per cubic meter (J/m3) to British thermal unitth per cubic foot (Btuth/ft3) }
\arguments{
  \item{joulePerCubMeter}{joule per cubic meter (J/m3) }
}
\value{British thermal unitth per cubic foot (Btuth/ft3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjoulePerCubMeterTOukThUnthPerCubFt(10)
}
\keyword{programming}