\name{NISTjoulePerKgKtOcalITPerGramK}
\alias{NISTjoulePerKgKtOcalITPerGramK}
\title{Convert joule per kilogram kelvin  to calorieIT per gram kelvin }
\usage{NISTjoulePerKgKtOcalITPerGramK(joulePerKgK)}
\description{\code{NISTjoulePerKgKtOcalITPerGramK} converts from joule per kilogram kelvin [J/(kg * K)] to calorieIT per gram kelvin [calIT/(g * K)] }
\arguments{
  \item{joulePerKgK}{joule per kilogram kelvin [J/(kg * K)] }
}
\value{calorieIT per gram kelvin [calIT/(g * K)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjoulePerKgKtOcalITPerGramK(10)
}
\keyword{programming}