\name{NISTjoulePerKgKtOcalthPerGramDegC}
\alias{NISTjoulePerKgKtOcalthPerGramDegC}
\title{Convert joule per kilogram kelvin  to calorieth per gram degree Celsius }
\usage{NISTjoulePerKgKtOcalthPerGramDegC(joulePerKgK)}
\description{\code{NISTjoulePerKgKtOcalthPerGramDegC} converts from joule per kilogram kelvin [J/(kg * K)] to calorieth per gram degree Celsius [calth/(g * C)] }
\arguments{
  \item{joulePerKgK}{joule per kilogram kelvin [J/(kg * K)] }
}
\value{calorieth per gram degree Celsius [calth/(g * C)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjoulePerKgKtOcalthPerGramDegC(10)
}
\keyword{programming}