\name{NISTjoulePerKgKtOcalthPerGramK}
\alias{NISTjoulePerKgKtOcalthPerGramK}
\title{Convert joule per kilogram kelvin  to calorieth per gram kelvin }
\usage{NISTjoulePerKgKtOcalthPerGramK(joulePerKgK)}
\description{\code{NISTjoulePerKgKtOcalthPerGramK} converts from joule per kilogram kelvin [J/(kg * K)] to calorieth per gram kelvin [calth/(g* K)] }
\arguments{
  \item{joulePerKgK}{joule per kilogram kelvin [J/(kg * K)] }
}
\value{calorieth per gram kelvin [calth/(g* K)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjoulePerKgKtOcalthPerGramK(10)
}
\keyword{programming}