\name{NISTjoulePerKgKtOukThUnITPerPoundDegF}
\alias{NISTjoulePerKgKtOukThUnITPerPoundDegF}
\title{Convert joule per kilogram kelvin  to British thermal unitIT per pound degree Fahrenheit }
\usage{NISTjoulePerKgKtOukThUnITPerPoundDegF(joulePerKgK)}
\description{\code{NISTjoulePerKgKtOukThUnITPerPoundDegF} converts from joule per kilogram kelvin [J/(kg * K)] to British thermal unitIT per pound degree Fahrenheit [BtuIT/(lb * F)] }
\arguments{
  \item{joulePerKgK}{joule per kilogram kelvin [J/(kg * K)] }
}
\value{British thermal unitIT per pound degree Fahrenheit [BtuIT/(lb * F)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjoulePerKgKtOukThUnITPerPoundDegF(10)
}
\keyword{programming}