\name{NISTjoulePerKgKtOukThUnITPerPoundDegRankine}
\alias{NISTjoulePerKgKtOukThUnITPerPoundDegRankine}
\title{Convert joule per kilogram kelvin  to British thermal unitIT per pound degree Rankine }
\usage{NISTjoulePerKgKtOukThUnITPerPoundDegRankine(joulePerKgK)}
\description{\code{NISTjoulePerKgKtOukThUnITPerPoundDegRankine} converts from joule per kilogram kelvin [J/(kg * K)] to British thermal unitIT per pound degree Rankine [BtuIT/(lb * R)] }
\arguments{
  \item{joulePerKgK}{joule per kilogram kelvin [J/(kg * K)] }
}
\value{British thermal unitIT per pound degree Rankine [BtuIT/(lb * R)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjoulePerKgKtOukThUnITPerPoundDegRankine(10)
}
\keyword{programming}