\name{NISTjoulePerKgKtOukThUnthPerPoundDegF}
\alias{NISTjoulePerKgKtOukThUnthPerPoundDegF}
\title{Convert joule per kilogram kelvin  to British thermal unitth per pound degree Fahrenheit }
\usage{NISTjoulePerKgKtOukThUnthPerPoundDegF(joulePerKgK)}
\description{\code{NISTjoulePerKgKtOukThUnthPerPoundDegF} converts from joule per kilogram kelvin [J/(kg * K)] to British thermal unitth per pound degree Fahrenheit [Btuth/(lb * F)] }
\arguments{
  \item{joulePerKgK}{joule per kilogram kelvin [J/(kg * K)] }
}
\value{British thermal unitth per pound degree Fahrenheit [Btuth/(lb * F)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjoulePerKgKtOukThUnthPerPoundDegF(10)
}
\keyword{programming}