\name{NISTjoulePerKgTOcalITPerGram}
\alias{NISTjoulePerKgTOcalITPerGram}
\title{Convert joule per kilogram  to calorieIT per gram }
\usage{NISTjoulePerKgTOcalITPerGram(joulePerKg)}
\description{\code{NISTjoulePerKgTOcalITPerGram} converts from joule per kilogram (J/kg) to calorieIT per gram (calIT/g) }
\arguments{
  \item{joulePerKg}{joule per kilogram (J/kg) }
}
\value{calorieIT per gram (calIT/g) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjoulePerKgTOcalITPerGram(10)
}
\keyword{programming}