\name{NISTjoulePerKgTOcalthPerGram}
\alias{NISTjoulePerKgTOcalthPerGram}
\title{Convert joule per kilogram  to calorieth per gram }
\usage{NISTjoulePerKgTOcalthPerGram(joulePerKg)}
\description{\code{NISTjoulePerKgTOcalthPerGram} converts from joule per kilogram (J/kg) to calorieth per gram (calth/g) }
\arguments{
  \item{joulePerKg}{joule per kilogram (J/kg) }
}
\value{calorieth per gram (calth/g) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjoulePerKgTOcalthPerGram(10)
}
\keyword{programming}