\name{NISTjoulePerKgTOukThUnITPerPound}
\alias{NISTjoulePerKgTOukThUnITPerPound}
\title{Convert joule per kilogram  to British thermal unitIT per pound }
\usage{NISTjoulePerKgTOukThUnITPerPound(joulePerKg)}
\description{\code{NISTjoulePerKgTOukThUnITPerPound} converts from joule per kilogram (J/kg) to British thermal unitIT per pound (BtuIT/lb) }
\arguments{
  \item{joulePerKg}{joule per kilogram (J/kg) }
}
\value{British thermal unitIT per pound (BtuIT/lb) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjoulePerKgTOukThUnITPerPound(10)
}
\keyword{programming}