\name{NISTjoulePerKgTOukThUnthPerPound}
\alias{NISTjoulePerKgTOukThUnthPerPound}
\title{Convert joule per kilogram  to British thermal unitth per pound }
\usage{NISTjoulePerKgTOukThUnthPerPound(joulePerKg)}
\description{\code{NISTjoulePerKgTOukThUnthPerPound} converts from joule per kilogram (J/kg) to British thermal unitth per pound (Btuth/lb) }
\arguments{
  \item{joulePerKg}{joule per kilogram (J/kg) }
}
\value{British thermal unitth per pound (Btuth/lb) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjoulePerKgTOukThUnthPerPound(10)
}
\keyword{programming}