\name{NISTjoulePerKtOukThUnITPerDegF}
\alias{NISTjoulePerKtOukThUnITPerDegF}
\title{Convert joule per kelvin  to British thermal unitIT per degree Fahrenheit }
\usage{NISTjoulePerKtOukThUnITPerDegF(joulePerK)}
\description{\code{NISTjoulePerKtOukThUnITPerDegF} converts from joule per kelvin (J/K) to British thermal unitIT per degree Fahrenheit (BtuIT/F) }
\arguments{
  \item{joulePerK}{joule per kelvin (J/K) }
}
\value{British thermal unitIT per degree Fahrenheit (BtuIT/F) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjoulePerKtOukThUnITPerDegF(10)
}
\keyword{programming}