\name{NISTjoulePerKtOukThUnITPerDegRankine}
\alias{NISTjoulePerKtOukThUnITPerDegRankine}
\title{Convert joule per kelvin  to British thermal unitIT per degree Rankine }
\usage{NISTjoulePerKtOukThUnITPerDegRankine(joulePerK)}
\description{\code{NISTjoulePerKtOukThUnITPerDegRankine} converts from joule per kelvin (J/K) to British thermal unitIT per degree Rankine (BtuIT/R) }
\arguments{
  \item{joulePerK}{joule per kelvin (J/K) }
}
\value{British thermal unitIT per degree Rankine (BtuIT/R) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjoulePerKtOukThUnITPerDegRankine(10)
}
\keyword{programming}