\name{NISTjoulePerKtOukThUnthPerDegRankine}
\alias{NISTjoulePerKtOukThUnthPerDegRankine}
\title{Convert joule per kelvin  to British thermal unitth per degree Rankine }
\usage{NISTjoulePerKtOukThUnthPerDegRankine(joulePerK)}
\description{\code{NISTjoulePerKtOukThUnthPerDegRankine} converts from joule per kelvin (J/K) to British thermal unitth per degree Rankine (Btuth/R) }
\arguments{
  \item{joulePerK}{joule per kelvin (J/K) }
}
\value{British thermal unitth per degree Rankine (Btuth/R) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjoulePerKtOukThUnthPerDegRankine(10)
}
\keyword{programming}