\name{NISTjoulePerSqrMeterTOcalthPerSqrCm}
\alias{NISTjoulePerSqrMeterTOcalthPerSqrCm}
\title{Convert joule per square meter  to calorieth per square centimeter }
\usage{NISTjoulePerSqrMeterTOcalthPerSqrCm(joulePerSqrMeter)}
\description{\code{NISTjoulePerSqrMeterTOcalthPerSqrCm} converts from joule per square meter (J/m2) to calorieth per square centimeter (calth/cm2) }
\arguments{
  \item{joulePerSqrMeter}{joule per square meter (J/m2) }
}
\value{calorieth per square centimeter (calth/cm2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjoulePerSqrMeterTOcalthPerSqrCm(10)
}
\keyword{programming}