\name{NISTjoulePerSqrMeterTOukThUnthPerSqrFt}
\alias{NISTjoulePerSqrMeterTOukThUnthPerSqrFt}
\title{Convert joule per square meter  to British thermal unitth per square foot }
\usage{NISTjoulePerSqrMeterTOukThUnthPerSqrFt(joulePerSqrMeter)}
\description{\code{NISTjoulePerSqrMeterTOukThUnthPerSqrFt} converts from joule per square meter (J/m2) to British thermal unitth per square foot (Btuth/ft2) }
\arguments{
  \item{joulePerSqrMeter}{joule per square meter (J/m2) }
}
\value{British thermal unitth per square foot (Btuth/ft2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjoulePerSqrMeterTOukThUnthPerSqrFt(10)
}
\keyword{programming}