\name{NISTjouleTOcal}
\alias{NISTjouleTOcal}
\title{Convert joule  to calorie }
\usage{NISTjouleTOcal(joule)}
\description{\code{NISTjouleTOcal} converts from joule (J) to calorie (cal) (mean) }
\arguments{
  \item{joule}{joule (J) }
}
\value{calorie (cal) (mean) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjouleTOcal(10)
}
\keyword{programming}