\name{NISTjouleTOcal15C}
\alias{NISTjouleTOcal15C}
\title{Convert joule  to calorie }
\usage{NISTjouleTOcal15C(joule)}
\description{\code{NISTjouleTOcal15C} converts from joule (J) to calorie (15 C) (cal15) }
\arguments{
  \item{joule}{joule (J) }
}
\value{calorie (15 C) (cal15) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjouleTOcal15C(10)
}
\keyword{programming}