\name{NISTjouleTOcal20C}
\alias{NISTjouleTOcal20C}
\title{Convert joule  to calorie }
\usage{NISTjouleTOcal20C(joule)}
\description{\code{NISTjouleTOcal20C} converts from joule (J) to calorie (20 C) (cal20) }
\arguments{
  \item{joule}{joule (J) }
}
\value{calorie (20 C) (cal20) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjouleTOcal20C(10)
}
\keyword{programming}