\name{NISTjouleTOcalIT}
\alias{NISTjouleTOcalIT}
\title{Convert joule  to calorieIT 10}
\usage{NISTjouleTOcalIT(joule)}
\description{\code{NISTjouleTOcalIT} converts from joule (J) to calorieIT (calIT) 10 }
\arguments{
  \item{joule}{joule (J) }
}
\value{calorieIT (calIT) 10 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjouleTOcalIT(10)
}
\keyword{programming}