\name{NISTjouleTOcalITKg}
\alias{NISTjouleTOcalITKg}
\title{Convert joule  to calorieIT, kilogram 11}
\usage{NISTjouleTOcalITKg(joule)}
\description{\code{NISTjouleTOcalITKg} converts from joule (J) to calorieIT, kilogram (nutrition) 11 }
\arguments{
  \item{joule}{joule (J) }
}
\value{calorieIT, kilogram (nutrition) 11 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjouleTOcalITKg(10)
}
\keyword{programming}