\name{NISTjouleTOcalKg}
\alias{NISTjouleTOcalKg}
\title{Convert joule  to calorie , kilogram 11}
\usage{NISTjouleTOcalKg(joule)}
\description{\code{NISTjouleTOcalKg} converts from joule (J) to calorie (mean), kilogram (nutrition) 11 }
\arguments{
  \item{joule}{joule (J) }
}
\value{calorie (mean), kilogram (nutrition) 11 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjouleTOcalKg(10)
}
\keyword{programming}