\name{NISTjouleTOcalth}
\alias{NISTjouleTOcalth}
\title{Convert joule  to calorieth 10}
\usage{NISTjouleTOcalth(joule)}
\description{\code{NISTjouleTOcalth} converts from joule (J) to calorieth (calth) 10 }
\arguments{
  \item{joule}{joule (J) }
}
\value{calorieth (calth) 10 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjouleTOcalth(10)
}
\keyword{programming}