\name{NISTjouleTOcalthKg}
\alias{NISTjouleTOcalthKg}
\title{Convert joule  to calorieth , kilogram 11}
\usage{NISTjouleTOcalthKg(joule)}
\description{\code{NISTjouleTOcalthKg} converts from joule (J) to calorieth , kilogram (nutrition) 11 }
\arguments{
  \item{joule}{joule (J) }
}
\value{calorieth , kilogram (nutrition) 11 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjouleTOcalthKg(10)
}
\keyword{programming}