\name{NISTjouleTOelectronvolt}
\alias{NISTjouleTOelectronvolt}
\title{Convert joule  to electronvolt }
\usage{NISTjouleTOelectronvolt(joule)}
\description{\code{NISTjouleTOelectronvolt} converts from joule (J) to electronvolt (eV) }
\arguments{
  \item{joule}{joule (J) }
}
\value{electronvolt (eV) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjouleTOelectronvolt(10)
}
\keyword{programming}