\name{NISTjouleTOerg}
\alias{NISTjouleTOerg}
\title{Convert joule  to erg }
\usage{NISTjouleTOerg(joule)}
\description{\code{NISTjouleTOerg} converts from joule (J) to erg (erg) }
\arguments{
  \item{joule}{joule (J) }
}
\value{erg (erg) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjouleTOerg(10)
}
\keyword{programming}