\name{NISTjouleTOftPoundal}
\alias{NISTjouleTOftPoundal}
\title{Convert joule  to foot poundal}
\usage{NISTjouleTOftPoundal(joule)}
\description{\code{NISTjouleTOftPoundal} converts from joule (J) to foot poundal }
\arguments{
  \item{joule}{joule (J) }
}
\value{foot poundal }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjouleTOftPoundal(10)
}
\keyword{programming}