\name{NISTjouleTOftPoundForce}
\alias{NISTjouleTOftPoundForce}
\title{Convert joule  to foot pound-force }
\usage{NISTjouleTOftPoundForce(joule)}
\description{\code{NISTjouleTOftPoundForce} converts from joule (J) to foot pound-force (ft * lbf) }
\arguments{
  \item{joule}{joule (J) }
}
\value{foot pound-force (ft * lbf) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjouleTOftPoundForce(10)
}
\keyword{programming}