\name{NISTjouleTOkilocal}
\alias{NISTjouleTOkilocal}
\title{Convert joule  to kilocalorie }
\usage{NISTjouleTOkilocal(joule)}
\description{\code{NISTjouleTOkilocal} converts from joule (J) to kilocalorie (mean) (kcal) }
\arguments{
  \item{joule}{joule (J) }
}
\value{kilocalorie (mean) (kcal) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjouleTOkilocal(10)
}
\keyword{programming}