\name{NISTjouleTOkilocalIT}
\alias{NISTjouleTOkilocalIT}
\title{Convert joule  to kilocalorieIT }
\usage{NISTjouleTOkilocalIT(joule)}
\description{\code{NISTjouleTOkilocalIT} converts from joule (J) to kilocalorieIT (kcalIT) }
\arguments{
  \item{joule}{joule (J) }
}
\value{kilocalorieIT (kcalIT) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjouleTOkilocalIT(10)
}
\keyword{programming}