\name{NISTjouleTOkilocalth}
\alias{NISTjouleTOkilocalth}
\title{Convert joule  to kilocalorieth }
\usage{NISTjouleTOkilocalth(joule)}
\description{\code{NISTjouleTOkilocalth} converts from joule (J) to kilocalorieth (kcalth) }
\arguments{
  \item{joule}{joule (J) }
}
\value{kilocalorieth (kcalth) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjouleTOkilocalth(10)
}
\keyword{programming}