\name{NISTjouleTOkilowattHour}
\alias{NISTjouleTOkilowattHour}
\title{Convert joule  to kilowatt hour }
\usage{NISTjouleTOkilowattHour(joule)}
\description{\code{NISTjouleTOkilowattHour} converts from joule (J) to kilowatt hour (kW * h) }
\arguments{
  \item{joule}{joule (J) }
}
\value{kilowatt hour (kW * h) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjouleTOkilowattHour(10)
}
\keyword{programming}