\name{NISTjouleTOquad}
\alias{NISTjouleTOquad}
\title{Convert joule  to quad 11}
\usage{NISTjouleTOquad(joule)}
\description{\code{NISTjouleTOquad} converts from joule (J) to quad (1015 BtuIT) 11 }
\arguments{
  \item{joule}{joule (J) }
}
\value{quad (1015 BtuIT) 11 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjouleTOquad(10)
}
\keyword{programming}