\name{NISTjouleTOthermEC}
\alias{NISTjouleTOthermEC}
\title{Convert joule  to therm 24}
\usage{NISTjouleTOthermEC(joule)}
\description{\code{NISTjouleTOthermEC} converts from joule (J) to therm (EC) 24 }
\arguments{
  \item{joule}{joule (J) }
}
\value{therm (EC) 24 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjouleTOthermEC(10)
}
\keyword{programming}