\name{NISTjouleTOtonOfTNT}
\alias{NISTjouleTOtonOfTNT}
\title{Convert joule  to ton of TNT 25}
\usage{NISTjouleTOtonOfTNT(joule)}
\description{\code{NISTjouleTOtonOfTNT} converts from joule (J) to ton of TNT (energy equivalent) 25 }
\arguments{
  \item{joule}{joule (J) }
}
\value{ton of TNT (energy equivalent) 25 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjouleTOtonOfTNT(10)
}
\keyword{programming}