\name{NISTjouleTOukThUn59F}
\alias{NISTjouleTOukThUn59F}
\title{Convert joule  to British thermal unit }
\usage{NISTjouleTOukThUn59F(joule)}
\description{\code{NISTjouleTOukThUn59F} converts from joule (J) to British thermal unit (59 F) (Btu) }
\arguments{
  \item{joule}{joule (J) }
}
\value{British thermal unit (59 F) (Btu) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjouleTOukThUn59F(10)
}
\keyword{programming}