\name{NISTjouleTOukThUnIT}
\alias{NISTjouleTOukThUnIT}
\title{Convert joule  to British thermal unitIT 9}
\usage{NISTjouleTOukThUnIT(joule)}
\description{\code{NISTjouleTOukThUnIT} converts from joule (J) to British thermal unitIT (BtuIT) 9 }
\arguments{
  \item{joule}{joule (J) }
}
\value{British thermal unitIT (BtuIT) 9 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjouleTOukThUnIT(10)
}
\keyword{programming}