\name{NISTjouleTOukThUnth}
\alias{NISTjouleTOukThUnth}
\title{Convert joule  to British thermal unitth 9}
\usage{NISTjouleTOukThUnth(joule)}
\description{\code{NISTjouleTOukThUnth} converts from joule (J) to British thermal unitth (Btuth) 9 }
\arguments{
  \item{joule}{joule (J) }
}
\value{British thermal unitth (Btuth) 9 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjouleTOukThUnth(10)
}
\keyword{programming}