\name{NISTjouleTOwattHour}
\alias{NISTjouleTOwattHour}
\title{Convert joule  to watt hour }
\usage{NISTjouleTOwattHour(joule)}
\description{\code{NISTjouleTOwattHour} converts from joule (J) to watt hour (W * h) }
\arguments{
  \item{joule}{joule (J) }
}
\value{watt hour (W * h) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjouleTOwattHour(10)
}
\keyword{programming}