\name{NISTjouleTOwattSec}
\alias{NISTjouleTOwattSec}
\title{Convert joule  to watt second }
\usage{NISTjouleTOwattSec(joule)}
\description{\code{NISTjouleTOwattSec} converts from joule (J) to watt second (W * s) }
\arguments{
  \item{joule}{joule (J) }
}
\value{watt second (W * s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTjouleTOwattSec(10)
}
\keyword{programming}