\name{NISTkayserTOreciprocalMeter}
\alias{NISTkayserTOreciprocalMeter}
\title{Convert kayser  to reciprocal meter }
\usage{NISTkayserTOreciprocalMeter(kayser)}
\description{\code{NISTkayserTOreciprocalMeter} converts from kayser (K) to reciprocal meter (m-1) }
\arguments{
  \item{kayser}{kayser (K) }
}
\value{reciprocal meter (m-1) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkayserTOreciprocalMeter(10)
}
\keyword{programming}