\name{NISTkgForceMeterTOnewtonMeter}
\alias{NISTkgForceMeterTOnewtonMeter}
\title{Convert kilogram-force meter  to newton meter }
\usage{NISTkgForceMeterTOnewtonMeter(kgForceMeter)}
\description{\code{NISTkgForceMeterTOnewtonMeter} converts from kilogram-force meter (kgf * m) to newton meter (N * m) }
\arguments{
  \item{kgForceMeter}{kilogram-force meter (kgf * m) }
}
\value{newton meter (N * m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgForceMeterTOnewtonMeter(10)
}
\keyword{programming}