\name{NISTkgForcePerSqrCmTOkpascal}
\alias{NISTkgForcePerSqrCmTOkpascal}
\title{Convert kilogram-force per square centimeter  to kilopascal }
\usage{NISTkgForcePerSqrCmTOkpascal(kgForcePerSqrCm)}
\description{\code{NISTkgForcePerSqrCmTOkpascal} converts from kilogram-force per square centimeter (kgf/cm2) to kilopascal (kPa) }
\arguments{
  \item{kgForcePerSqrCm}{kilogram-force per square centimeter (kgf/cm2) }
}
\value{kilopascal (kPa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgForcePerSqrCmTOkpascal(10)
}
\keyword{programming}