\name{NISTkgForcePerSqrCmTOpascal}
\alias{NISTkgForcePerSqrCmTOpascal}
\title{Convert kilogram-force per square centimeter  to pascal }
\usage{NISTkgForcePerSqrCmTOpascal(kgForcePerSqrCm)}
\description{\code{NISTkgForcePerSqrCmTOpascal} converts from kilogram-force per square centimeter (kgf/cm2) to pascal (Pa) }
\arguments{
  \item{kgForcePerSqrCm}{kilogram-force per square centimeter (kgf/cm2) }
}
\value{pascal (Pa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgForcePerSqrCmTOpascal(10)
}
\keyword{programming}