\name{NISTkgForcePerSqrMeterTOpascal}
\alias{NISTkgForcePerSqrMeterTOpascal}
\title{Convert kilogram-force per square meter  to pascal }
\usage{NISTkgForcePerSqrMeterTOpascal(kgForcePerSqrMeter)}
\description{\code{NISTkgForcePerSqrMeterTOpascal} converts from kilogram-force per square meter (kgf/m2) to pascal (Pa) }
\arguments{
  \item{kgForcePerSqrMeter}{kilogram-force per square meter (kgf/m2) }
}
\value{pascal (Pa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgForcePerSqrMeterTOpascal(10)
}
\keyword{programming}