\name{NISTkgForcePerSqrMillimeterTOmpascal}
\alias{NISTkgForcePerSqrMillimeterTOmpascal}
\title{Convert kilogram-force per square millimeter  to megapascal }
\usage{NISTkgForcePerSqrMillimeterTOmpascal(kgForcePerSqrMillimeter)}
\description{\code{NISTkgForcePerSqrMillimeterTOmpascal} converts from kilogram-force per square millimeter (kgf/mm2) to megapascal (MPa) }
\arguments{
  \item{kgForcePerSqrMillimeter}{kilogram-force per square millimeter (kgf/mm2) }
}
\value{megapascal (MPa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgForcePerSqrMillimeterTOmpascal(10)
}
\keyword{programming}