\name{NISTkgForcePerSqrMillimeterTOpascal}
\alias{NISTkgForcePerSqrMillimeterTOpascal}
\title{Convert kilogram-force per square millimeter  to pascal }
\usage{NISTkgForcePerSqrMillimeterTOpascal(kgForcePerSqrMillimeter)}
\description{\code{NISTkgForcePerSqrMillimeterTOpascal} converts from kilogram-force per square millimeter (kgf/mm2) to pascal (Pa) }
\arguments{
  \item{kgForcePerSqrMillimeter}{kilogram-force per square millimeter (kgf/mm2) }
}
\value{pascal (Pa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgForcePerSqrMillimeterTOpascal(10)
}
\keyword{programming}