\name{NISTkgForceSecSqrdPerMeterTOkg}
\alias{NISTkgForceSecSqrdPerMeterTOkg}
\title{Convert kilogram-force second squared per meter  to kilogram }
\usage{NISTkgForceSecSqrdPerMeterTOkg(kgForceSecSqrdPerMeter)}
\description{\code{NISTkgForceSecSqrdPerMeterTOkg} converts from kilogram-force second squared per meter (kgf * s2/m) to kilogram (kg) }
\arguments{
  \item{kgForceSecSqrdPerMeter}{kilogram-force second squared per meter (kgf * s2/m) }
}
\value{kilogram (kg) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgForceSecSqrdPerMeterTOkg(10)
}
\keyword{programming}