\name{NISTkgForceTOnewton}
\alias{NISTkgForceTOnewton}
\title{Convert kilogram-force  to newton }
\usage{NISTkgForceTOnewton(kgForce)}
\description{\code{NISTkgForceTOnewton} converts from kilogram-force (kgf) to newton (N) }
\arguments{
  \item{kgForce}{kilogram-force (kgf) }
}
\value{newton (N) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgForceTOnewton(10)
}
\keyword{programming}