\name{NISTkgMeterSqrdTOpoundFtSqrd}
\alias{NISTkgMeterSqrdTOpoundFtSqrd}
\title{Convert kilogram meter squared  to pound foot squared }
\usage{NISTkgMeterSqrdTOpoundFtSqrd(kgMeterSqrd)}
\description{\code{NISTkgMeterSqrdTOpoundFtSqrd} converts from kilogram meter squared (kg * m2) to pound foot squared (lb * ft2) }
\arguments{
  \item{kgMeterSqrd}{kilogram meter squared (kg * m2) }
}
\value{pound foot squared (lb * ft2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgMeterSqrdTOpoundFtSqrd(10)
}
\keyword{programming}