\name{NISTkgMeterSqrdTOpoundInchSqrd}
\alias{NISTkgMeterSqrdTOpoundInchSqrd}
\title{Convert kilogram meter squared  to pound inch squared }
\usage{NISTkgMeterSqrdTOpoundInchSqrd(kgMeterSqrd)}
\description{\code{NISTkgMeterSqrdTOpoundInchSqrd} converts from kilogram meter squared (kg *m2) to pound inch squared (lb * in2) }
\arguments{
  \item{kgMeterSqrd}{kilogram meter squared (kg *m2) }
}
\value{pound inch squared (lb * in2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgMeterSqrdTOpoundInchSqrd(10)
}
\keyword{programming}