\name{NISTkgPerCubMeterTOgrainPerGallon}
\alias{NISTkgPerCubMeterTOgrainPerGallon}
\title{Convert kilogram per cubic meter  to grain per gallon }
\usage{NISTkgPerCubMeterTOgrainPerGallon(kgPerCubMeter)}
\description{\code{NISTkgPerCubMeterTOgrainPerGallon} converts from kilogram per cubic meter (kg/m3) to grain per gallon (U.S.) (gr/gal) }
\arguments{
  \item{kgPerCubMeter}{kilogram per cubic meter (kg/m3) }
}
\value{grain per gallon (U.S.) (gr/gal) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgPerCubMeterTOgrainPerGallon(10)
}
\keyword{programming}