\name{NISTkgPerCubMeterTOgramPerCubCm}
\alias{NISTkgPerCubMeterTOgramPerCubCm}
\title{Convert kilogram per cubic meter  to gram per cubic centimeter }
\usage{NISTkgPerCubMeterTOgramPerCubCm(kgPerCubMeter)}
\description{\code{NISTkgPerCubMeterTOgramPerCubCm} converts from kilogram per cubic meter (kg/m3) to gram per cubic centimeter (g/cm3) }
\arguments{
  \item{kgPerCubMeter}{kilogram per cubic meter (kg/m3) }
}
\value{gram per cubic centimeter (g/cm3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgPerCubMeterTOgramPerCubCm(10)
}
\keyword{programming}