\name{NISTkgPerCubMeterTOounceImperialPerGallon}
\alias{NISTkgPerCubMeterTOounceImperialPerGallon}
\title{Convert kilogram per cubic meter  to ounce per gallon }
\usage{NISTkgPerCubMeterTOounceImperialPerGallon(kgPerCubMeter)}
\description{\code{NISTkgPerCubMeterTOounceImperialPerGallon} converts from kilogram per cubic meter (kg/m3) to ounce (avoirdupois) per gallon [Canadian and U.K. (Imperial)] (oz/gal) }
\arguments{
  \item{kgPerCubMeter}{kilogram per cubic meter (kg/m3) }
}
\value{ounce (avoirdupois) per gallon [Canadian and U.K. (Imperial)] (oz/gal) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgPerCubMeterTOounceImperialPerGallon(10)
}
\keyword{programming}