\name{NISTkgPerCubMeterTOouncePerCubInch}
\alias{NISTkgPerCubMeterTOouncePerCubInch}
\title{Convert kilogram per cubic meter  to ounce per cubic inch }
\usage{NISTkgPerCubMeterTOouncePerCubInch(kgPerCubMeter)}
\description{\code{NISTkgPerCubMeterTOouncePerCubInch} converts from kilogram per cubic meter (kg/m3) to ounce (avoirdupois) per cubic inch (oz/in3) }
\arguments{
  \item{kgPerCubMeter}{kilogram per cubic meter (kg/m3) }
}
\value{ounce (avoirdupois) per cubic inch (oz/in3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgPerCubMeterTOouncePerCubInch(10)
}
\keyword{programming}