\name{NISTkgPerCubMeterTOpoundPerCubInch}
\alias{NISTkgPerCubMeterTOpoundPerCubInch}
\title{Convert kilogram per cubic meter  to pound per cubic inch }
\usage{NISTkgPerCubMeterTOpoundPerCubInch(kgPerCubMeter)}
\description{\code{NISTkgPerCubMeterTOpoundPerCubInch} converts from kilogram per cubic meter (kg/m3) to pound per cubic inch (lb/in3) }
\arguments{
  \item{kgPerCubMeter}{kilogram per cubic meter (kg/m3) }
}
\value{pound per cubic inch (lb/in3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgPerCubMeterTOpoundPerCubInch(10)
}
\keyword{programming}