\name{NISTkgPerCubMeterTOpoundPerGallonImperial}
\alias{NISTkgPerCubMeterTOpoundPerGallonImperial}
\title{Convert kilogram per cubic meter  to pound per gallon }
\usage{NISTkgPerCubMeterTOpoundPerGallonImperial(kgPerCubMeter)}
\description{\code{NISTkgPerCubMeterTOpoundPerGallonImperial} converts from kilogram per cubic meter (kg/m3) to pound per gallon [Canadian and U.K. (Imperial)] (lb/gal) }
\arguments{
  \item{kgPerCubMeter}{kilogram per cubic meter (kg/m3) }
}
\value{pound per gallon [Canadian and U.K. (Imperial)] (lb/gal) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgPerCubMeterTOpoundPerGallonImperial(10)
}
\keyword{programming}