\name{NISTkgPerCubMeterTOslugPerCubFt}
\alias{NISTkgPerCubMeterTOslugPerCubFt}
\title{Convert kilogram per cubic meter  to slug per cubic foot }
\usage{NISTkgPerCubMeterTOslugPerCubFt(kgPerCubMeter)}
\description{\code{NISTkgPerCubMeterTOslugPerCubFt} converts from kilogram per cubic meter (kg/m3) to slug per cubic foot (slug/ft3) }
\arguments{
  \item{kgPerCubMeter}{kilogram per cubic meter (kg/m3) }
}
\value{slug per cubic foot (slug/ft3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgPerCubMeterTOslugPerCubFt(10)
}
\keyword{programming}