\name{NISTkgPerCubMeterTOtonLongPerCubYard}
\alias{NISTkgPerCubMeterTOtonLongPerCubYard}
\title{Convert kilogram per cubic meter  to ton, long, per cubic yard}
\usage{NISTkgPerCubMeterTOtonLongPerCubYard(kgPerCubMeter)}
\description{\code{NISTkgPerCubMeterTOtonLongPerCubYard} converts from kilogram per cubic meter (kg/m3) to ton, long, per cubic yard }
\arguments{
  \item{kgPerCubMeter}{kilogram per cubic meter (kg/m3) }
}
\value{ton, long, per cubic yard }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgPerCubMeterTOtonLongPerCubYard(10)
}
\keyword{programming}