\name{NISTkgPerCubMeterTOtonShortPerCubYard}
\alias{NISTkgPerCubMeterTOtonShortPerCubYard}
\title{Convert kilogram per cubic meter  to ton, short, per cubic yard}
\usage{NISTkgPerCubMeterTOtonShortPerCubYard(kgPerCubMeter)}
\description{\code{NISTkgPerCubMeterTOtonShortPerCubYard} converts from kilogram per cubic meter (kg/m3) to ton, short, per cubic yard }
\arguments{
  \item{kgPerCubMeter}{kilogram per cubic meter (kg/m3) }
}
\value{ton, short, per cubic yard }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgPerCubMeterTOtonShortPerCubYard(10)
}
\keyword{programming}