\name{NISTkgPerJouleTOpoundPerHorsepowerHour}
\alias{NISTkgPerJouleTOpoundPerHorsepowerHour}
\title{Convert kilogram per joule  to pound per horsepower hour }
\usage{NISTkgPerJouleTOpoundPerHorsepowerHour(kgPerJoule)}
\description{\code{NISTkgPerJouleTOpoundPerHorsepowerHour} converts from kilogram per joule (kg/J) to pound per horsepower hour [lb/(hp * h)] }
\arguments{
  \item{kgPerJoule}{kilogram per joule (kg/J) }
}
\value{pound per horsepower hour [lb/(hp * h)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgPerJouleTOpoundPerHorsepowerHour(10)
}
\keyword{programming}