\name{NISTkgPerLiterTOpoundPerGallon}
\alias{NISTkgPerLiterTOpoundPerGallon}
\title{Convert kilogram per liter  to pound per gallon }
\usage{NISTkgPerLiterTOpoundPerGallon(kgPerLiter)}
\description{\code{NISTkgPerLiterTOpoundPerGallon} converts from kilogram per liter (kg/L) to pound per gallon (U.S.) (lb/gal) }
\arguments{
  \item{kgPerLiter}{kilogram per liter (kg/L) }
}
\value{pound per gallon (U.S.) (lb/gal) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgPerLiterTOpoundPerGallon(10)
}
\keyword{programming}