\name{NISTkgPerLiterTOpoundPerGallonImperial}
\alias{NISTkgPerLiterTOpoundPerGallonImperial}
\title{Convert kilogram per liter  to pound per gallon }
\usage{NISTkgPerLiterTOpoundPerGallonImperial(kgPerLiter)}
\description{\code{NISTkgPerLiterTOpoundPerGallonImperial} converts from kilogram per liter (kg/L) to pound per gallon [Canadian and U.K. (Imperial)] (lb/gal) }
\arguments{
  \item{kgPerLiter}{kilogram per liter (kg/L) }
}
\value{pound per gallon [Canadian and U.K. (Imperial)] (lb/gal) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgPerLiterTOpoundPerGallonImperial(10)
}
\keyword{programming}