\name{NISTkgPerMeterTOdenier}
\alias{NISTkgPerMeterTOdenier}
\title{Convert kilogram per meter  to denier}
\usage{NISTkgPerMeterTOdenier(kgPerMeter)}
\description{\code{NISTkgPerMeterTOdenier} converts from kilogram per meter (kg/m) to denier }
\arguments{
  \item{kgPerMeter}{kilogram per meter (kg/m) }
}
\value{denier }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgPerMeterTOdenier(10)
}
\keyword{programming}