\name{NISTkgPerMeterTOpoundPerFt}
\alias{NISTkgPerMeterTOpoundPerFt}
\title{Convert kilogram per meter  to pound per foot }
\usage{NISTkgPerMeterTOpoundPerFt(kgPerMeter)}
\description{\code{NISTkgPerMeterTOpoundPerFt} converts from kilogram per meter (kg/m) to pound per foot (lb/ft) }
\arguments{
  \item{kgPerMeter}{kilogram per meter (kg/m) }
}
\value{pound per foot (lb/ft) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgPerMeterTOpoundPerFt(10)
}
\keyword{programming}