\name{NISTkgPerMeterTOpoundPerInch}
\alias{NISTkgPerMeterTOpoundPerInch}
\title{Convert kilogram per meter  to pound per inch }
\usage{NISTkgPerMeterTOpoundPerInch(kgPerMeter)}
\description{\code{NISTkgPerMeterTOpoundPerInch} converts from kilogram per meter (kg/m) to pound per inch (lb/in) }
\arguments{
  \item{kgPerMeter}{kilogram per meter (kg/m) }
}
\value{pound per inch (lb/in) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgPerMeterTOpoundPerInch(10)
}
\keyword{programming}