\name{NISTkgPerMeterTOpoundPerYard}
\alias{NISTkgPerMeterTOpoundPerYard}
\title{Convert kilogram per meter  to pound per yard }
\usage{NISTkgPerMeterTOpoundPerYard(kgPerMeter)}
\description{\code{NISTkgPerMeterTOpoundPerYard} converts from kilogram per meter (kg/m) to pound per yard (lb/yd) }
\arguments{
  \item{kgPerMeter}{kilogram per meter (kg/m) }
}
\value{pound per yard (lb/yd) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgPerMeterTOpoundPerYard(10)
}
\keyword{programming}