\name{NISTkgPerMeterTOtex}
\alias{NISTkgPerMeterTOtex}
\title{Convert kilogram per meter  to tex}
\usage{NISTkgPerMeterTOtex(kgPerMeter)}
\description{\code{NISTkgPerMeterTOtex} converts from kilogram per meter (kg/m) to tex }
\arguments{
  \item{kgPerMeter}{kilogram per meter (kg/m) }
}
\value{tex }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgPerMeterTOtex(10)
}
\keyword{programming}