\name{NISTkgPerPascalSecMeterTOpermInch0C}
\alias{NISTkgPerPascalSecMeterTOpermInch0C}
\title{Convert kilogram per pascal second meter  to perm inch }
\usage{NISTkgPerPascalSecMeterTOpermInch0C(kgPerPascalSecMeter)}
\description{\code{NISTkgPerPascalSecMeterTOpermInch0C} converts from kilogram per pascal second meter [kg/(Pa*s*m)] to perm inch (0 C) }
\arguments{
  \item{kgPerPascalSecMeter}{kilogram per pascal second meter [kg/(Pa*s*m)] }
}
\value{perm inch (0 C) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgPerPascalSecMeterTOpermInch0C(10)
}
\keyword{programming}