\name{NISTkgPerPascalSecSqrMeterTOperm0C}
\alias{NISTkgPerPascalSecSqrMeterTOperm0C}
\title{Convert kilogram per pascal second square meter  to perm }
\usage{NISTkgPerPascalSecSqrMeterTOperm0C(kgPerPascalSecSqrMeter)}
\description{\code{NISTkgPerPascalSecSqrMeterTOperm0C} converts from kilogram per pascal second square meter [kg/(Pa * s * m2)] to perm (0 C) }
\arguments{
  \item{kgPerPascalSecSqrMeter}{kilogram per pascal second square meter [kg/(Pa * s * m2)] }
}
\value{perm (0 C) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgPerPascalSecSqrMeterTOperm0C(10)
}
\keyword{programming}