\name{NISTkgPerSecTOpoundPerMin}
\alias{NISTkgPerSecTOpoundPerMin}
\title{Convert kilogram per second  to pound per minute }
\usage{NISTkgPerSecTOpoundPerMin(kgPerSec)}
\description{\code{NISTkgPerSecTOpoundPerMin} converts from kilogram per second (kg/s) to pound per minute (lb/min) }
\arguments{
  \item{kgPerSec}{kilogram per second (kg/s) }
}
\value{pound per minute (lb/min) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgPerSecTOpoundPerMin(10)
}
\keyword{programming}