\name{NISTkgPerSecTOpoundPerSec}
\alias{NISTkgPerSecTOpoundPerSec}
\title{Convert kilogram per second  to pound per second }
\usage{NISTkgPerSecTOpoundPerSec(kgPerSec)}
\description{\code{NISTkgPerSecTOpoundPerSec} converts from kilogram per second (kg/s) to pound per second (lb/s) }
\arguments{
  \item{kgPerSec}{kilogram per second (kg/s) }
}
\value{pound per second (lb/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgPerSecTOpoundPerSec(10)
}
\keyword{programming}