\name{NISTkgPerSecTOtonShortPerHour}
\alias{NISTkgPerSecTOtonShortPerHour}
\title{Convert kilogram per second  to ton, short, per hour}
\usage{NISTkgPerSecTOtonShortPerHour(kgPerSec)}
\description{\code{NISTkgPerSecTOtonShortPerHour} converts from kilogram per second (kg/s) to ton, short, per hour }
\arguments{
  \item{kgPerSec}{kilogram per second (kg/s) }
}
\value{ton, short, per hour }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgPerSecTOtonShortPerHour(10)
}
\keyword{programming}