\name{NISTkgPerSqrMeterTOouncePerSqrInch}
\alias{NISTkgPerSqrMeterTOouncePerSqrInch}
\title{Convert kilogram per square meter  to ounce per square inch }
\usage{NISTkgPerSqrMeterTOouncePerSqrInch(kgPerSqrMeter)}
\description{\code{NISTkgPerSqrMeterTOouncePerSqrInch} converts from kilogram per square meter (kg/m2) to ounce (avoirdupois) per square inch (oz/in2) }
\arguments{
  \item{kgPerSqrMeter}{kilogram per square meter (kg/m2) }
}
\value{ounce (avoirdupois) per square inch (oz/in2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgPerSqrMeterTOouncePerSqrInch(10)
}
\keyword{programming}