\name{NISTkgPerSqrMeterTOpoundPerSqrInch}
\alias{NISTkgPerSqrMeterTOpoundPerSqrInch}
\title{Convert kilogram per square meter  to pound per square inch }
\usage{NISTkgPerSqrMeterTOpoundPerSqrInch(kgPerSqrMeter)}
\description{\code{NISTkgPerSqrMeterTOpoundPerSqrInch} converts from kilogram per square meter (kg/m2) to pound per square inch (not pound-force) (lb/in2) }
\arguments{
  \item{kgPerSqrMeter}{kilogram per square meter (kg/m2) }
}
\value{pound per square inch (not pound-force) (lb/in2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgPerSqrMeterTOpoundPerSqrInch(10)
}
\keyword{programming}