\name{NISTkgTOcaratSI}
\alias{NISTkgTOcaratSI}
\title{Convert kilogram  to carat, metric}
\usage{NISTkgTOcaratSI(kg)}
\description{\code{NISTkgTOcaratSI} converts from kilogram (kg) to carat, metric }
\arguments{
  \item{kg}{kilogram (kg) }
}
\value{carat, metric }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgTOcaratSI(10)
}
\keyword{programming}