\name{NISTkgTOgrain}
\alias{NISTkgTOgrain}
\title{Convert kilogram  to grain }
\usage{NISTkgTOgrain(kg)}
\description{\code{NISTkgTOgrain} converts from kilogram (kg) to grain (gr) }
\arguments{
  \item{kg}{kilogram (kg) }
}
\value{grain (gr) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgTOgrain(10)
}
\keyword{programming}