\name{NISTkgTOhundredweightShort}
\alias{NISTkgTOhundredweightShort}
\title{Convert kilogram  to hundredweight }
\usage{NISTkgTOhundredweightShort(kg)}
\description{\code{NISTkgTOhundredweightShort} converts from kilogram (kg) to hundredweight (short, 100 lb) }
\arguments{
  \item{kg}{kilogram (kg) }
}
\value{hundredweight (short, 100 lb) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgTOhundredweightShort(10)
}
\keyword{programming}