\name{NISTkgTOkgForceSecSqrdPerMeter}
\alias{NISTkgTOkgForceSecSqrdPerMeter}
\title{Convert kilogram  to kilogram-force second squared per meter }
\usage{NISTkgTOkgForceSecSqrdPerMeter(kg)}
\description{\code{NISTkgTOkgForceSecSqrdPerMeter} converts from kilogram (kg) to kilogram-force second squared per meter (kgf * s2/m) }
\arguments{
  \item{kg}{kilogram (kg) }
}
\value{kilogram-force second squared per meter (kgf * s2/m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgTOkgForceSecSqrdPerMeter(10)
}
\keyword{programming}