\name{NISTkgTOounceAvoirdupois}
\alias{NISTkgTOounceAvoirdupois}
\title{Convert kilogram  to ounce }
\usage{NISTkgTOounceAvoirdupois(kg)}
\description{\code{NISTkgTOounceAvoirdupois} converts from kilogram (kg) to ounce (avoirdupois) (oz) }
\arguments{
  \item{kg}{kilogram (kg) }
}
\value{ounce (avoirdupois) (oz) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgTOounceAvoirdupois(10)
}
\keyword{programming}