\name{NISTkgTOounceTroy}
\alias{NISTkgTOounceTroy}
\title{Convert kilogram  to ounce }
\usage{NISTkgTOounceTroy(kg)}
\description{\code{NISTkgTOounceTroy} converts from kilogram (kg) to ounce (troy or apothecary) (oz) }
\arguments{
  \item{kg}{kilogram (kg) }
}
\value{ounce (troy or apothecary) (oz) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgTOounceTroy(10)
}
\keyword{programming}