\name{NISTkgTOpennyweight}
\alias{NISTkgTOpennyweight}
\title{Convert kilogram  to pennyweight }
\usage{NISTkgTOpennyweight(kg)}
\description{\code{NISTkgTOpennyweight} converts from kilogram (kg) to pennyweight (dwt) }
\arguments{
  \item{kg}{kilogram (kg) }
}
\value{pennyweight (dwt) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgTOpennyweight(10)
}
\keyword{programming}