\name{NISTkgTOpoundAvoirdupois}
\alias{NISTkgTOpoundAvoirdupois}
\title{Convert kilogram  to pound 22}
\usage{NISTkgTOpoundAvoirdupois(kg)}
\description{\code{NISTkgTOpoundAvoirdupois} converts from kilogram (kg) to pound (avoirdupois) (lb) 22 }
\arguments{
  \item{kg}{kilogram (kg) }
}
\value{pound (avoirdupois) (lb) 22 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgTOpoundAvoirdupois(10)
}
\keyword{programming}