\name{NISTkgTOpoundTroy}
\alias{NISTkgTOpoundTroy}
\title{Convert kilogram  to pound }
\usage{NISTkgTOpoundTroy(kg)}
\description{\code{NISTkgTOpoundTroy} converts from kilogram (kg) to pound (troy or apothecary) (lb) }
\arguments{
  \item{kg}{kilogram (kg) }
}
\value{pound (troy or apothecary) (lb) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgTOpoundTroy(10)
}
\keyword{programming}