\name{NISTkgTOslug}
\alias{NISTkgTOslug}
\title{Convert kilogram  to slug }
\usage{NISTkgTOslug(kg)}
\description{\code{NISTkgTOslug} converts from kilogram (kg) to slug (slug) }
\arguments{
  \item{kg}{kilogram (kg) }
}
\value{slug (slug) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgTOslug(10)
}
\keyword{programming}