\name{NISTkgTOtonAssay}
\alias{NISTkgTOtonAssay}
\title{Convert kilogram  to ton, assay }
\usage{NISTkgTOtonAssay(kg)}
\description{\code{NISTkgTOtonAssay} converts from kilogram (kg) to ton, assay (AT) }
\arguments{
  \item{kg}{kilogram (kg) }
}
\value{ton, assay (AT) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgTOtonAssay(10)
}
\keyword{programming}