\name{NISTkgTOtonLong}
\alias{NISTkgTOtonLong}
\title{Convert kilogram  to ton, long }
\usage{NISTkgTOtonLong(kg)}
\description{\code{NISTkgTOtonLong} converts from kilogram (kg) to ton, long (2240 lb) }
\arguments{
  \item{kg}{kilogram (kg) }
}
\value{ton, long (2240 lb) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgTOtonLong(10)
}
\keyword{programming}