\name{NISTkgTOtonne}
\alias{NISTkgTOtonne}
\title{Convert kilogram  to tonne }
\usage{NISTkgTOtonne(kg)}
\description{\code{NISTkgTOtonne} converts from kilogram (kg) to tonne (called metric ton in U.S.) (t) }
\arguments{
  \item{kg}{kilogram (kg) }
}
\value{tonne (called metric ton in U.S.) (t) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgTOtonne(10)
}
\keyword{programming}