\name{NISTkgTOtonShort}
\alias{NISTkgTOtonShort}
\title{Convert kilogram  to ton, short }
\usage{NISTkgTOtonShort(kg)}
\description{\code{NISTkgTOtonShort} converts from kilogram (kg) to ton, short (2000 lb) }
\arguments{
  \item{kg}{kilogram (kg) }
}
\value{ton, short (2000 lb) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgTOtonShort(10)
}
\keyword{programming}