\name{NISTkgTOtonSI}
\alias{NISTkgTOtonSI}
\title{Convert kilogram  to ton, metric }
\usage{NISTkgTOtonSI(kg)}
\description{\code{NISTkgTOtonSI} converts from kilogram (kg) to ton, metric (t) }
\arguments{
  \item{kg}{kilogram (kg) }
}
\value{ton, metric (t) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkgTOtonSI(10)
}
\keyword{programming}