\name{NISTkilocalITtOjoule}
\alias{NISTkilocalITtOjoule}
\title{Convert kilocalorieIT  to joule }
\usage{NISTkilocalITtOjoule(kilocalIT)}
\description{\code{NISTkilocalITtOjoule} converts from kilocalorieIT (kcalIT) to joule (J) }
\arguments{
  \item{kilocalIT}{kilocalorieIT (kcalIT) }
}
\value{joule (J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkilocalITtOjoule(10)
}
\keyword{programming}