\name{NISTkilocalthPerMinTOwatt}
\alias{NISTkilocalthPerMinTOwatt}
\title{Convert kilocalorieth per minute  to watt }
\usage{NISTkilocalthPerMinTOwatt(kilocalthPerMin)}
\description{\code{NISTkilocalthPerMinTOwatt} converts from kilocalorieth per minute (kcalth/min) to watt (W) }
\arguments{
  \item{kilocalthPerMin}{kilocalorieth per minute (kcalth/min) }
}
\value{watt (W) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkilocalthPerMinTOwatt(10)
}
\keyword{programming}