\name{NISTkilocalthPerSecTOwatt}
\alias{NISTkilocalthPerSecTOwatt}
\title{Convert kilocalorieth per second  to watt }
\usage{NISTkilocalthPerSecTOwatt(kilocalthPerSec)}
\description{\code{NISTkilocalthPerSecTOwatt} converts from kilocalorieth per second (kcalth/s) to watt (W) }
\arguments{
  \item{kilocalthPerSec}{kilocalorieth per second (kcalth/s) }
}
\value{watt (W) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkilocalthPerSecTOwatt(10)
}
\keyword{programming}