\name{NISTkilocalthTOjoule}
\alias{NISTkilocalthTOjoule}
\title{Convert kilocalorieth  to joule }
\usage{NISTkilocalthTOjoule(kilocalth)}
\description{\code{NISTkilocalthTOjoule} converts from kilocalorieth (kcalth) to joule (J) }
\arguments{
  \item{kilocalth}{kilocalorieth (kcalth) }
}
\value{joule (J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkilocalthTOjoule(10)
}
\keyword{programming}