\name{NISTkilocalTOjoule}
\alias{NISTkilocalTOjoule}
\title{Convert kilocalorie  to joule }
\usage{NISTkilocalTOjoule(kilocal)}
\description{\code{NISTkilocalTOjoule} converts from kilocalorie (mean) (kcal) to joule (J) }
\arguments{
  \item{kilocal}{kilocalorie (mean) (kcal) }
}
\value{joule (J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkilocalTOjoule(10)
}
\keyword{programming}