\name{NISTkilonewtonTOkip}
\alias{NISTkilonewtonTOkip}
\title{Convert kilonewton  to kip }
\usage{NISTkilonewtonTOkip(kilonewton)}
\description{\code{NISTkilonewtonTOkip} converts from kilonewton (kN) to kip (1 kip= 1000 lbf) }
\arguments{
  \item{kilonewton}{kilonewton (kN) }
}
\value{kip (1 kip= 1000 lbf) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkilonewtonTOkip(10)
}
\keyword{programming}