\name{NISTkilonewtonTOtonForce}
\alias{NISTkilonewtonTOtonForce}
\title{Convert kilonewton  to ton-force }
\usage{NISTkilonewtonTOtonForce(kilonewton)}
\description{\code{NISTkilonewtonTOtonForce} converts from kilonewton (kN) to ton-force (2000 lbf) }
\arguments{
  \item{kilonewton}{kilonewton (kN) }
}
\value{ton-force (2000 lbf) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkilonewtonTOtonForce(10)
}
\keyword{programming}