\name{NISTkilowattHourTOjoule}
\alias{NISTkilowattHourTOjoule}
\title{Convert kilowatt hour  to joule }
\usage{NISTkilowattHourTOjoule(kilowattHour)}
\description{\code{NISTkilowattHourTOjoule} converts from kilowatt hour (kW * h) to joule (J) }
\arguments{
  \item{kilowattHour}{kilowatt hour (kW * h) }
}
\value{joule (J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkilowattHourTOjoule(10)
}
\keyword{programming}