\name{NISTkilowattHourTOmegajoule}
\alias{NISTkilowattHourTOmegajoule}
\title{Convert kilowatt hour  to megajoule }
\usage{NISTkilowattHourTOmegajoule(kilowattHour)}
\description{\code{NISTkilowattHourTOmegajoule} converts from kilowatt hour (kW * h) to megajoule (MJ) }
\arguments{
  \item{kilowattHour}{kilowatt hour (kW * h) }
}
\value{megajoule (MJ) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkilowattHourTOmegajoule(10)
}
\keyword{programming}