\name{NISTkipPerSqrInchTOkpascal}
\alias{NISTkipPerSqrInchTOkpascal}
\title{Convert kip per square inch  to kilopascal }
\usage{NISTkipPerSqrInchTOkpascal(kipPerSqrInch)}
\description{\code{NISTkipPerSqrInchTOkpascal} converts from kip per square inch (ksi) (kip/in2) to kilopascal (kPa) }
\arguments{
  \item{kipPerSqrInch}{kip per square inch (ksi) (kip/in2) }
}
\value{kilopascal (kPa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkipPerSqrInchTOkpascal(10)
}
\keyword{programming}