\name{NISTkipPerSqrInchTOpascal}
\alias{NISTkipPerSqrInchTOpascal}
\title{Convert kip per square inch  to pascal }
\usage{NISTkipPerSqrInchTOpascal(kipPerSqrInch)}
\description{\code{NISTkipPerSqrInchTOpascal} converts from kip per square inch (ksi) (kip/in2) to pascal (Pa) }
\arguments{
  \item{kipPerSqrInch}{kip per square inch (ksi) (kip/in2) }
}
\value{pascal (Pa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkipPerSqrInchTOpascal(10)
}
\keyword{programming}