\name{NISTkipTOkilonewton}
\alias{NISTkipTOkilonewton}
\title{Convert kip  to kilonewton }
\usage{NISTkipTOkilonewton(kip)}
\description{\code{NISTkipTOkilonewton} converts from kip (1 kip= 1000 lbf) to kilonewton (kN) }
\arguments{
  \item{kip}{kip (1 kip= 1000 lbf) }
}
\value{kilonewton (kN) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkipTOkilonewton(10)
}
\keyword{programming}