\name{NISTkipTOnewton}
\alias{NISTkipTOnewton}
\title{Convert kip  to newton }
\usage{NISTkipTOnewton(kip)}
\description{\code{NISTkipTOnewton} converts from kip (1 kip= 1000 lbf) to newton (N) }
\arguments{
  \item{kip}{kip (1 kip= 1000 lbf) }
}
\value{newton (N) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkipTOnewton(10)
}
\keyword{programming}