\name{NISTkmPerHourTOmeterPerSec}
\alias{NISTkmPerHourTOmeterPerSec}
\title{Convert kilometer per hour  to meter per second }
\usage{NISTkmPerHourTOmeterPerSec(kmPerHour)}
\description{\code{NISTkmPerHourTOmeterPerSec} converts from kilometer per hour (km/h) to meter per second (m/s) }
\arguments{
  \item{kmPerHour}{kilometer per hour (km/h) }
}
\value{meter per second (m/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkmPerHourTOmeterPerSec(10)
}
\keyword{programming}