\name{NISTkmPerHourTOmilePerHour}
\alias{NISTkmPerHourTOmilePerHour}
\title{Convert kilometer per hour  to mile per hour }
\usage{NISTkmPerHourTOmilePerHour(kmPerHour)}
\description{\code{NISTkmPerHourTOmilePerHour} converts from kilometer per hour (km/h) to mile per hour (mi/h) }
\arguments{
  \item{kmPerHour}{kilometer per hour (km/h) }
}
\value{mile per hour (mi/h) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkmPerHourTOmilePerHour(10)
}
\keyword{programming}