\name{NISTkmPerLiterTOmilePerGallon}
\alias{NISTkmPerLiterTOmilePerGallon}
\title{Convert kilometer per liter  to mile per gallon }
\usage{NISTkmPerLiterTOmilePerGallon(kmPerLiter)}
\description{\code{NISTkmPerLiterTOmilePerGallon} converts from kilometer per liter (km/L) to mile per gallon (U.S.) (mpg) (mi/gal) }
\arguments{
  \item{kmPerLiter}{kilometer per liter (km/L) }
}
\value{mile per gallon (U.S.) (mpg) (mi/gal) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkmPerLiterTOmilePerGallon(10)
}
\keyword{programming}