\name{NISTkmTOmileUSsurvey}
\alias{NISTkmTOmileUSsurvey}
\title{Convert kilometer  to mile 7}
\usage{NISTkmTOmileUSsurvey(km)}
\description{\code{NISTkmTOmileUSsurvey} converts from kilometer (km) to mile (based on U.S. survey foot) (mi) 7 }
\arguments{
  \item{km}{kilometer (km) }
}
\value{mile (based on U.S. survey foot) (mi) 7 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkmTOmileUSsurvey(10)
}
\keyword{programming}