\name{NISTknotTOmeterPerSec}
\alias{NISTknotTOmeterPerSec}
\title{Convert knot  to meter per second }
\usage{NISTknotTOmeterPerSec(knot)}
\description{\code{NISTknotTOmeterPerSec} converts from knot (nautical mile per hour) to meter per second (m/s) }
\arguments{
  \item{knot}{knot (nautical mile per hour) }
}
\value{meter per second (m/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTknotTOmeterPerSec(10)
}
\keyword{programming}