\name{NISTkpascalTOatmosphereStd}
\alias{NISTkpascalTOatmosphereStd}
\title{Convert kilopascal  to atmosphere, standard }
\usage{NISTkpascalTOatmosphereStd(kpascal)}
\description{\code{NISTkpascalTOatmosphereStd} converts from kilopascal (kPa) to atmosphere, standard (atm) }
\arguments{
  \item{kpascal}{kilopascal (kPa) }
}
\value{atmosphere, standard (atm) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkpascalTOatmosphereStd(10)
}
\keyword{programming}