\name{NISTkpascalTOatmosphereTechnical}
\alias{NISTkpascalTOatmosphereTechnical}
\title{Convert kilopascal  to atmosphere, technical 8}
\usage{NISTkpascalTOatmosphereTechnical(kpascal)}
\description{\code{NISTkpascalTOatmosphereTechnical} converts from kilopascal (kPa) to atmosphere, technical (at) 8 }
\arguments{
  \item{kpascal}{kilopascal (kPa) }
}
\value{atmosphere, technical (at) 8 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkpascalTOatmosphereTechnical(10)
}
\keyword{programming}