\name{NISTkpascalTObar}
\alias{NISTkpascalTObar}
\title{Convert kilopascal  to bar }
\usage{NISTkpascalTObar(kpascal)}
\description{\code{NISTkpascalTObar} converts from kilopascal (kPa) to bar (bar) }
\arguments{
  \item{kpascal}{kilopascal (kPa) }
}
\value{bar (bar) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkpascalTObar(10)
}
\keyword{programming}