\name{NISTkpascalTOcmOfMercury}
\alias{NISTkpascalTOcmOfMercury}
\title{Convert kilopascal  to centimeter of mercury 12}
\usage{NISTkpascalTOcmOfMercury(kpascal)}
\description{\code{NISTkpascalTOcmOfMercury} converts from kilopascal (kPa) to centimeter of mercury (0 C) 12 }
\arguments{
  \item{kpascal}{kilopascal (kPa) }
}
\value{centimeter of mercury (0 C) 12 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkpascalTOcmOfMercury(10)
}
\keyword{programming}