\name{NISTkpascalTOcmOfMercuryConvtnl}
\alias{NISTkpascalTOcmOfMercuryConvtnl}
\title{Convert kilopascal  to centimeter of mercury, conventional 12}
\usage{NISTkpascalTOcmOfMercuryConvtnl(kpascal)}
\description{\code{NISTkpascalTOcmOfMercuryConvtnl} converts from kilopascal (kPa) to centimeter of mercury, conventional (cmHg) 12 }
\arguments{
  \item{kpascal}{kilopascal (kPa) }
}
\value{centimeter of mercury, conventional (cmHg) 12 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkpascalTOcmOfMercuryConvtnl(10)
}
\keyword{programming}