\name{NISTkpascalTOftOfMercuryConvtnl}
\alias{NISTkpascalTOftOfMercuryConvtnl}
\title{Convert kilopascal  to foot of mercury, conventional 12}
\usage{NISTkpascalTOftOfMercuryConvtnl(kpascal)}
\description{\code{NISTkpascalTOftOfMercuryConvtnl} converts from kilopascal (kPa) to foot of mercury, conventional (ftHg) 12 }
\arguments{
  \item{kpascal}{kilopascal (kPa) }
}
\value{foot of mercury, conventional (ftHg) 12 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkpascalTOftOfMercuryConvtnl(10)
}
\keyword{programming}