\name{NISTkpascalTOftOfWater}
\alias{NISTkpascalTOftOfWater}
\title{Convert kilopascal  to foot of water 12}
\usage{NISTkpascalTOftOfWater(kpascal)}
\description{\code{NISTkpascalTOftOfWater} converts from kilopascal (kPa) to foot of water (39.2 F) 12 }
\arguments{
  \item{kpascal}{kilopascal (kPa) }
}
\value{foot of water (39.2 F) 12 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkpascalTOftOfWater(10)
}
\keyword{programming}