\name{NISTkpascalTOinchOfMercury32F}
\alias{NISTkpascalTOinchOfMercury32F}
\title{Convert kilopascal  to inch of mercury 12}
\usage{NISTkpascalTOinchOfMercury32F(kpascal)}
\description{\code{NISTkpascalTOinchOfMercury32F} converts from kilopascal (kPa) to inch of mercury (32 F) 12 }
\arguments{
  \item{kpascal}{kilopascal (kPa) }
}
\value{inch of mercury (32 F) 12 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkpascalTOinchOfMercury32F(10)
}
\keyword{programming}