\name{NISTkpascalTOinchOfMercury60F}
\alias{NISTkpascalTOinchOfMercury60F}
\title{Convert kilopascal  to inch of mercury 12}
\usage{NISTkpascalTOinchOfMercury60F(kpascal)}
\description{\code{NISTkpascalTOinchOfMercury60F} converts from kilopascal (kPa) to inch of mercury (60 F) 12 }
\arguments{
  \item{kpascal}{kilopascal (kPa) }
}
\value{inch of mercury (60 F) 12 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkpascalTOinchOfMercury60F(10)
}
\keyword{programming}