\name{NISTkpascalTOinchOfMercuryConvtnl}
\alias{NISTkpascalTOinchOfMercuryConvtnl}
\title{Convert kilopascal  to inch of mercury, conventional 12}
\usage{NISTkpascalTOinchOfMercuryConvtnl(kpascal)}
\description{\code{NISTkpascalTOinchOfMercuryConvtnl} converts from kilopascal (kPa) to inch of mercury, conventional (inHg) 12 }
\arguments{
  \item{kpascal}{kilopascal (kPa) }
}
\value{inch of mercury, conventional (inHg) 12 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkpascalTOinchOfMercuryConvtnl(10)
}
\keyword{programming}