\name{NISTkpascalTOkgForcePerSqrCm}
\alias{NISTkpascalTOkgForcePerSqrCm}
\title{Convert kilopascal  to kilogram-force per square centimeter }
\usage{NISTkpascalTOkgForcePerSqrCm(kpascal)}
\description{\code{NISTkpascalTOkgForcePerSqrCm} converts from kilopascal (kPa) to kilogram-force per square centimeter (kgf/cm2) }
\arguments{
  \item{kpascal}{kilopascal (kPa) }
}
\value{kilogram-force per square centimeter (kgf/cm2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkpascalTOkgForcePerSqrCm(10)
}
\keyword{programming}