\name{NISTkpascalTOkipPerSqrInch}
\alias{NISTkpascalTOkipPerSqrInch}
\title{Convert kilopascal  to kip per square inch }
\usage{NISTkpascalTOkipPerSqrInch(kpascal)}
\description{\code{NISTkpascalTOkipPerSqrInch} converts from kilopascal (kPa) to kip per square inch (ksi) (kip/in2) }
\arguments{
  \item{kpascal}{kilopascal (kPa) }
}
\value{kip per square inch (ksi) (kip/in2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkpascalTOkipPerSqrInch(10)
}
\keyword{programming}