\name{NISTkpascalTOmillibar}
\alias{NISTkpascalTOmillibar}
\title{Convert kilopascal  to millibar }
\usage{NISTkpascalTOmillibar(kpascal)}
\description{\code{NISTkpascalTOmillibar} converts from kilopascal (kPa) to millibar (mbar) }
\arguments{
  \item{kpascal}{kilopascal (kPa) }
}
\value{millibar (mbar) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkpascalTOmillibar(10)
}
\keyword{programming}