\name{NISTkpascalTOpoundForcePerSqrInch}
\alias{NISTkpascalTOpoundForcePerSqrInch}
\title{Convert kilopascal  to pound-force per square inch }
\usage{NISTkpascalTOpoundForcePerSqrInch(kpascal)}
\description{\code{NISTkpascalTOpoundForcePerSqrInch} converts from kilopascal (kPa) to pound-force per square inch (psi) (lbf/in2) }
\arguments{
  \item{kpascal}{kilopascal (kPa) }
}
\value{pound-force per square inch (psi) (lbf/in2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkpascalTOpoundForcePerSqrInch(10)
}
\keyword{programming}