\name{NISTkpascalTOpsi}
\alias{NISTkpascalTOpsi}
\title{Convert kilopascal  to psi }
\usage{NISTkpascalTOpsi(kpascal)}
\description{\code{NISTkpascalTOpsi} converts from kilopascal (kPa) to psi (pound-force per square inch) (lbf/in2) }
\arguments{
  \item{kpascal}{kilopascal (kPa) }
}
\value{psi (pound-force per square inch) (lbf/in2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkpascalTOpsi(10)
}
\keyword{programming}