\name{NISTkPerWattTOdegFHourPerUkThUnIT}
\alias{NISTkPerWattTOdegFHourPerUkThUnIT}
\title{Convert kelvin per watt  to degree Fahrenheit hour per British thermal unitIT }
\usage{NISTkPerWattTOdegFHourPerUkThUnIT(kPerWatt)}
\description{\code{NISTkPerWattTOdegFHourPerUkThUnIT} converts from kelvin per watt (K/W) to degree Fahrenheit hour per British thermal unitIT (F * h/BtuIT) }
\arguments{
  \item{kPerWatt}{kelvin per watt (K/W) }
}
\value{degree Fahrenheit hour per British thermal unitIT (F * h/BtuIT) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkPerWattTOdegFHourPerUkThUnIT(10)
}
\keyword{programming}