\name{NISTkPerWattTOdegFHourPerUkThUnth}
\alias{NISTkPerWattTOdegFHourPerUkThUnth}
\title{Convert kelvin per watt  to degree Fahrenheit hour per British thermal unitth }
\usage{NISTkPerWattTOdegFHourPerUkThUnth(kPerWatt)}
\description{\code{NISTkPerWattTOdegFHourPerUkThUnth} converts from kelvin per watt (K/W) to degree Fahrenheit hour per British thermal unitth (F * h/Btuth) }
\arguments{
  \item{kPerWatt}{kelvin per watt (K/W) }
}
\value{degree Fahrenheit hour per British thermal unitth (F * h/Btuth) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkPerWattTOdegFHourPerUkThUnth(10)
}
\keyword{programming}