\name{NISTkPerWattTOdegFSecPerUkThUnth}
\alias{NISTkPerWattTOdegFSecPerUkThUnth}
\title{Convert kelvin per watt  to degree Fahrenheit second per British thermal unitth }
\usage{NISTkPerWattTOdegFSecPerUkThUnth(kPerWatt)}
\description{\code{NISTkPerWattTOdegFSecPerUkThUnth} converts from kelvin per watt (K/W) to degree Fahrenheit second per British thermal unitth (F * s/Btuth) }
\arguments{
  \item{kPerWatt}{kelvin per watt (K/W) }
}
\value{degree Fahrenheit second per British thermal unitth (F * s/Btuth) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkPerWattTOdegFSecPerUkThUnth(10)
}
\keyword{programming}