\name{NISTkTOdegC}
\alias{NISTkTOdegC}
\title{Convert kelvin  to degree Celsius }
\usage{NISTkTOdegC(k)}
\description{\code{NISTkTOdegC} converts from kelvin (K) to degree Celsius (temperature interval) (C) }
\arguments{
  \item{k}{kelvin (K) }
}
\value{degree Celsius (temperature interval) (C) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkTOdegC(10)
}
\keyword{programming}