\name{NISTkTOdegCinterval}
\alias{NISTkTOdegCinterval}
\title{Convert kelvin temperature interval to degree Celsius }
\usage{NISTkTOdegCinterval(k)}
\description{\code{NISTkTOdegCinterval} converts from kelvin (K) to degree Celsius (temperature interval) (C) }
\arguments{
  \item{k}{kelvin (K) }
}
\value{degree Celsius (temperature interval) (C) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkTOdegCinterval(10)
}
\keyword{programming}
