\name{NISTkTOdegF}
\alias{NISTkTOdegF}
\title{Convert kelvin  to degree Fahrenheit }
\usage{NISTkTOdegF(k)}
\description{\code{NISTkTOdegF} converts from kelvin (K) to degree Fahrenheit (temperature interval) (F) }
\arguments{
  \item{k}{kelvin (K) }
}
\value{degree Fahrenheit (temperature interval) (F) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkTOdegF(10)
}
\keyword{programming}