\name{NISTkTOdegFinterval}
\alias{NISTkTOdegFinterval}
\title{Convert kelvin temperature interval to degree Fahrenheit }
\usage{NISTkTOdegFinterval(k)}
\description{\code{NISTkTOdegFinterval} converts from kelvin (K) to degree Fahrenheit (temperature interval) (F) }
\arguments{
  \item{k}{kelvin (K) }
}
\value{degree Fahrenheit (temperature interval) (F) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkTOdegFinterval(10)
}
\keyword{programming}
