\name{NISTkTOdegRankine}
\alias{NISTkTOdegRankine}
\title{Convert kelvin  to degree Rankine }
\usage{NISTkTOdegRankine(k)}
\description{\code{NISTkTOdegRankine} converts from kelvin (K) to degree Rankine (temperature interval) (R) }
\arguments{
  \item{k}{kelvin (K) }
}
\value{degree Rankine (temperature interval) (R) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTkTOdegRankine(10)
}
\keyword{programming}