\name{NISTlambertTOcandelaPerSqrMeter}
\alias{NISTlambertTOcandelaPerSqrMeter}
\title{Convert lambert 17 to candela per square meter }
\usage{NISTlambertTOcandelaPerSqrMeter(lambert)}
\description{\code{NISTlambertTOcandelaPerSqrMeter} converts from lambert 17 to candela per square meter (cd/m2) }
\arguments{
  \item{lambert}{lambert 17 }
}
\value{candela per square meter (cd/m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTlambertTOcandelaPerSqrMeter(10)
}
\keyword{programming}