\name{NISTlightYearTOmeter}
\alias{NISTlightYearTOmeter}
\title{Convert light year 18 to meter }
\usage{NISTlightYearTOmeter(lightYear)}
\description{\code{NISTlightYearTOmeter} converts from light year (l.y.) 18 to meter (m) }
\arguments{
  \item{lightYear}{light year (l.y.) 18 }
}
\value{meter (m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTlightYearTOmeter(10)
}
\keyword{programming}