\name{NISTliterPer100KmTOmilePerGallon}
\alias{NISTliterPer100KmTOmilePerGallon}
\title{Convert liter per 100 kilometer  to mile per gallon 21}
\usage{NISTliterPer100KmTOmilePerGallon(literPer100Km)}
\description{\code{NISTliterPer100KmTOmilePerGallon} converts from liter per 100 kilometer (L/100 km) to mile per gallon (U.S.) (mpg) (mi/gal) 21 }
\arguments{
  \item{literPer100Km}{liter per 100 kilometer (L/100 km) }
}
\value{mile per gallon (U.S.) (mpg) (mi/gal) 21 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTliterPer100KmTOmilePerGallon(10)
}
\keyword{programming}