\name{NISTliterPerJouleTOgallonPerHorsepowerHour}
\alias{NISTliterPerJouleTOgallonPerHorsepowerHour}
\title{Convert liter per joule  to gallon per horsepower hour }
\usage{NISTliterPerJouleTOgallonPerHorsepowerHour(literPerJoule)}
\description{\code{NISTliterPerJouleTOgallonPerHorsepowerHour} converts from liter per joule (L/J) to gallon (U.S.) per horsepower hour [gal/(hp * h)] }
\arguments{
  \item{literPerJoule}{liter per joule (L/J) }
}
\value{gallon (U.S.) per horsepower hour [gal/(hp * h)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTliterPerJouleTOgallonPerHorsepowerHour(10)
}
\keyword{programming}