\name{NISTliterPerSecTOcubFtPerMin}
\alias{NISTliterPerSecTOcubFtPerMin}
\title{Convert liter per second  to cubic foot per minute }
\usage{NISTliterPerSecTOcubFtPerMin(literPerSec)}
\description{\code{NISTliterPerSecTOcubFtPerMin} converts from liter per second (L/s) to cubic foot per minute (ft3/min) }
\arguments{
  \item{literPerSec}{liter per second (L/s) }
}
\value{cubic foot per minute (ft3/min) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTliterPerSecTOcubFtPerMin(10)
}
\keyword{programming}