\name{NISTliterPerSecTOgallonPerMin}
\alias{NISTliterPerSecTOgallonPerMin}
\title{Convert liter per second  to gallon per minute }
\usage{NISTliterPerSecTOgallonPerMin(literPerSec)}
\description{\code{NISTliterPerSecTOgallonPerMin} converts from liter per second (L/s) to gallon (U.S.) per minute (gpm) (gal/min) }
\arguments{
  \item{literPerSec}{liter per second (L/s) }
}
\value{gallon (U.S.) per minute (gpm) (gal/min) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTliterPerSecTOgallonPerMin(10)
}
\keyword{programming}