\name{NISTliterTObarrel}
\alias{NISTliterTObarrel}
\title{Convert liter  to barrel }
\usage{NISTliterTObarrel(liter)}
\description{\code{NISTliterTObarrel} converts from liter (L) to barrel [for petroleum, 42 gallons (U.S.)] (bbl) }
\arguments{
  \item{liter}{liter (L) }
}
\value{barrel [for petroleum, 42 gallons (U.S.)] (bbl) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTliterTObarrel(10)
}
\keyword{programming}