\name{NISTliterTObushel}
\alias{NISTliterTObushel}
\title{Convert liter  to bushel }
\usage{NISTliterTObushel(liter)}
\description{\code{NISTliterTObushel} converts from liter (L) to bushel (U.S.) (bu) }
\arguments{
  \item{liter}{liter (L) }
}
\value{bushel (U.S.) (bu) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTliterTObushel(10)
}
\keyword{programming}