\name{NISTliterTOcubMeter}
\alias{NISTliterTOcubMeter}
\title{Convert liter 19 to cubic meter }
\usage{NISTliterTOcubMeter(liter)}
\description{\code{NISTliterTOcubMeter} converts from liter (L) 19 to cubic meter (m3) }
\arguments{
  \item{liter}{liter (L) 19 }
}
\value{cubic meter (m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTliterTOcubMeter(10)
}
\keyword{programming}