\name{NISTliterTOcup}
\alias{NISTliterTOcup}
\title{Convert liter  to cup }
\usage{NISTliterTOcup(liter)}
\description{\code{NISTliterTOcup} converts from liter (L) to cup (U.S.) }
\arguments{
  \item{liter}{liter (L) }
}
\value{cup (U.S.) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTliterTOcup(10)
}
\keyword{programming}