\name{NISTliterTOgallon}
\alias{NISTliterTOgallon}
\title{Convert liter  to gallon }
\usage{NISTliterTOgallon(liter)}
\description{\code{NISTliterTOgallon} converts from liter (L) to gallon (U.S.) (gal) }
\arguments{
  \item{liter}{liter (L) }
}
\value{gallon (U.S.) (gal) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTliterTOgallon(10)
}
\keyword{programming}