\name{NISTliterTOgallonImperial}
\alias{NISTliterTOgallonImperial}
\title{Convert liter  to gallon }
\usage{NISTliterTOgallonImperial(liter)}
\description{\code{NISTliterTOgallonImperial} converts from liter (L) to gallon [Canadian and U.K. (Imperial)] (gal) }
\arguments{
  \item{liter}{liter (L) }
}
\value{gallon [Canadian and U.K. (Imperial)] (gal) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTliterTOgallonImperial(10)
}
\keyword{programming}