\name{NISTliterTOgill}
\alias{NISTliterTOgill}
\title{Convert liter  to gill }
\usage{NISTliterTOgill(liter)}
\description{\code{NISTliterTOgill} converts from liter (L) to gill (U.S.) (gi) }
\arguments{
  \item{liter}{liter (L) }
}
\value{gill (U.S.) (gi) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTliterTOgill(10)
}
\keyword{programming}