\name{NISTliterTOgillImperial}
\alias{NISTliterTOgillImperial}
\title{Convert liter  to gill }
\usage{NISTliterTOgillImperial(liter)}
\description{\code{NISTliterTOgillImperial} converts from liter (L) to gill [Canadian and U.K. (Imperial)] (gi) }
\arguments{
  \item{liter}{liter (L) }
}
\value{gill [Canadian and U.K. (Imperial)] (gi) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTliterTOgillImperial(10)
}
\keyword{programming}