\name{NISTliterTOpeck}
\alias{NISTliterTOpeck}
\title{Convert liter  to peck }
\usage{NISTliterTOpeck(liter)}
\description{\code{NISTliterTOpeck} converts from liter (L) to peck (U.S.) (pk) }
\arguments{
  \item{liter}{liter (L) }
}
\value{peck (U.S.) (pk) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTliterTOpeck(10)
}
\keyword{programming}