\name{NISTliterTOpintUSdry}
\alias{NISTliterTOpintUSdry}
\title{Convert liter  to pint }
\usage{NISTliterTOpintUSdry(liter)}
\description{\code{NISTliterTOpintUSdry} converts from liter (L) to pint (U.S. dry) (dry pt) }
\arguments{
  \item{liter}{liter (L) }
}
\value{pint (U.S. dry) (dry pt) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTliterTOpintUSdry(10)
}
\keyword{programming}