\name{NISTliterTOpintUSliquid}
\alias{NISTliterTOpintUSliquid}
\title{Convert liter  to pint }
\usage{NISTliterTOpintUSliquid(liter)}
\description{\code{NISTliterTOpintUSliquid} converts from liter (L) to pint (U.S. liquid) (liq pt) }
\arguments{
  \item{liter}{liter (L) }
}
\value{pint (U.S. liquid) (liq pt) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTliterTOpintUSliquid(10)
}
\keyword{programming}