\name{NISTliterTOquartUSdry}
\alias{NISTliterTOquartUSdry}
\title{Convert liter  to quart }
\usage{NISTliterTOquartUSdry(liter)}
\description{\code{NISTliterTOquartUSdry} converts from liter (L) to quart (U.S. dry) (dry qt) }
\arguments{
  \item{liter}{liter (L) }
}
\value{quart (U.S. dry) (dry qt) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTliterTOquartUSdry(10)
}
\keyword{programming}