\name{NISTlumenPerSqrFtTOlux}
\alias{NISTlumenPerSqrFtTOlux}
\title{Convert lumen per square foot  to lux }
\usage{NISTlumenPerSqrFtTOlux(lumenPerSqrFt)}
\description{\code{NISTlumenPerSqrFtTOlux} converts from lumen per square foot (lm/ft2) to lux (lx) }
\arguments{
  \item{lumenPerSqrFt}{lumen per square foot (lm/ft2) }
}
\value{lux (lx) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTlumenPerSqrFtTOlux(10)
}
\keyword{programming}