\name{NISTluxTOlumenPerSqrFt}
\alias{NISTluxTOlumenPerSqrFt}
\title{Convert lux  to lumen per square foot }
\usage{NISTluxTOlumenPerSqrFt(lux)}
\description{\code{NISTluxTOlumenPerSqrFt} converts from lux (lx) to lumen per square foot (lm/ft2) }
\arguments{
  \item{lux}{lux (lx) }
}
\value{lumen per square foot (lm/ft2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTluxTOlumenPerSqrFt(10)
}
\keyword{programming}