\name{NISTluxTOphot}
\alias{NISTluxTOphot}
\title{Convert lux  to phot }
\usage{NISTluxTOphot(lux)}
\description{\code{NISTluxTOphot} converts from lux (lx) to phot (ph) }
\arguments{
  \item{lux}{lux (lx) }
}
\value{phot (ph) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTluxTOphot(10)
}
\keyword{programming}