\name{NISTmaxwellTOweber}
\alias{NISTmaxwellTOweber}
\title{Convert maxwell  to weber }
\usage{NISTmaxwellTOweber(maxwell)}
\description{\code{NISTmaxwellTOweber} converts from maxwell (Mx) to weber (Wb) }
\arguments{
  \item{maxwell}{maxwell (Mx) }
}
\value{weber (Wb) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmaxwellTOweber(10)
}
\keyword{programming}