\name{NISTmegajouleTOkilowattHour}
\alias{NISTmegajouleTOkilowattHour}
\title{Convert megajoule  to kilowatt hour }
\usage{NISTmegajouleTOkilowattHour(megajoule)}
\description{\code{NISTmegajouleTOkilowattHour} converts from megajoule (MJ) to kilowatt hour (kW * h) }
\arguments{
  \item{megajoule}{megajoule (MJ) }
}
\value{kilowatt hour (kW * h) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmegajouleTOkilowattHour(10)
}
\keyword{programming}