\name{NISTmeter4thPowerTOft4thPower}
\alias{NISTmeter4thPowerTOft4thPower}
\title{Convert meter to the fourth power  to foot to the fourth power 16}
\usage{NISTmeter4thPowerTOft4thPower(meter4thPower)}
\description{\code{NISTmeter4thPowerTOft4thPower} converts from meter to the fourth power (m4) to foot to the fourth power (ft4) 16 }
\arguments{
  \item{meter4thPower}{meter to the fourth power (m4) }
}
\value{foot to the fourth power (ft4) 16 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeter4thPowerTOft4thPower(10)
}
\keyword{programming}