\name{NISTmeterKPerWattTOdegFHourSqrFtPerUkThUnITInch}
\alias{NISTmeterKPerWattTOdegFHourSqrFtPerUkThUnITInch}
\title{Convert meter kelvin per watt  to degree Fahrenheit hour square foot per British thermal unitIT inch }
\usage{NISTmeterKPerWattTOdegFHourSqrFtPerUkThUnITInch(meterKPerWatt)}
\description{\code{NISTmeterKPerWattTOdegFHourSqrFtPerUkThUnITInch} converts from meter kelvin per watt (m * K/W) to degree Fahrenheit hour square foot per British thermal unitIT inch [F * h * ft2/(BtuIT * in)] }
\arguments{
  \item{meterKPerWatt}{meter kelvin per watt (m * K/W) }
}
\value{degree Fahrenheit hour square foot per British thermal unitIT inch [F * h * ft2/(BtuIT * in)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterKPerWattTOdegFHourSqrFtPerUkThUnITInch(10)
}
\keyword{programming}