\name{NISTmeterPerCubMeterTOmilePerGallon}
\alias{NISTmeterPerCubMeterTOmilePerGallon}
\title{Convert meter per cubic meter  to mile per gallon }
\usage{NISTmeterPerCubMeterTOmilePerGallon(meterPerCubMeter)}
\description{\code{NISTmeterPerCubMeterTOmilePerGallon} converts from meter per cubic meter (m/m3) to mile per gallon (U.S.) (mpg) (mi/gal) }
\arguments{
  \item{meterPerCubMeter}{meter per cubic meter (m/m3) }
}
\value{mile per gallon (U.S.) (mpg) (mi/gal) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterPerCubMeterTOmilePerGallon(10)
}
\keyword{programming}