\name{NISTmeterPerSecSqrdTOaccelOfFreeFallStd}
\alias{NISTmeterPerSecSqrdTOaccelOfFreeFallStd}
\title{Convert meter per second squared  to acceleration of free fall, standard }
\usage{NISTmeterPerSecSqrdTOaccelOfFreeFallStd(meterPerSecSqrd)}
\description{\code{NISTmeterPerSecSqrdTOaccelOfFreeFallStd} converts from meter per second squared (m / s2) to acceleration of free fall, standard (gn) }
\arguments{
  \item{meterPerSecSqrd}{meter per second squared (m / s2) }
}
\value{acceleration of free fall, standard (gn) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterPerSecSqrdTOaccelOfFreeFallStd(10)
}
\keyword{programming}