\name{NISTmeterPerSecSqrdTOftPerSecSqrd}
\alias{NISTmeterPerSecSqrdTOftPerSecSqrd}
\title{Convert meter per second squared  to foot per second squared }
\usage{NISTmeterPerSecSqrdTOftPerSecSqrd(meterPerSecSqrd)}
\description{\code{NISTmeterPerSecSqrdTOftPerSecSqrd} converts from meter per second squared (m/s2) to foot per second squared (ft/s2) }
\arguments{
  \item{meterPerSecSqrd}{meter per second squared (m/s2) }
}
\value{foot per second squared (ft/s2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterPerSecSqrdTOftPerSecSqrd(10)
}
\keyword{programming}