\name{NISTmeterPerSecSqrdTOgal}
\alias{NISTmeterPerSecSqrdTOgal}
\title{Convert meter per second squared  to gal }
\usage{NISTmeterPerSecSqrdTOgal(meterPerSecSqrd)}
\description{\code{NISTmeterPerSecSqrdTOgal} converts from meter per second squared (m/s2) to gal (Gal) }
\arguments{
  \item{meterPerSecSqrd}{meter per second squared (m/s2) }
}
\value{gal (Gal) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterPerSecSqrdTOgal(10)
}
\keyword{programming}