\name{NISTmeterPerSecSqrdTOinchPerSecSqrd}
\alias{NISTmeterPerSecSqrdTOinchPerSecSqrd}
\title{Convert meter per second squared  to inch per second squared }
\usage{NISTmeterPerSecSqrdTOinchPerSecSqrd(meterPerSecSqrd)}
\description{\code{NISTmeterPerSecSqrdTOinchPerSecSqrd} converts from meter per second squared (m/s2) to inch per second squared (in/s2) }
\arguments{
  \item{meterPerSecSqrd}{meter per second squared (m/s2) }
}
\value{inch per second squared (in/s2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterPerSecSqrdTOinchPerSecSqrd(10)
}
\keyword{programming}