\name{NISTmeterPerSecTOftPerHour}
\alias{NISTmeterPerSecTOftPerHour}
\title{Convert meter per second  to foot per hour }
\usage{NISTmeterPerSecTOftPerHour(meterPerSec)}
\description{\code{NISTmeterPerSecTOftPerHour} converts from meter per second (m/s) to foot per hour (ft/h) }
\arguments{
  \item{meterPerSec}{meter per second (m/s) }
}
\value{foot per hour (ft/h) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterPerSecTOftPerHour(10)
}
\keyword{programming}