\name{NISTmeterPerSecTOftPerMin}
\alias{NISTmeterPerSecTOftPerMin}
\title{Convert meter per second  to foot per minute }
\usage{NISTmeterPerSecTOftPerMin(meterPerSec)}
\description{\code{NISTmeterPerSecTOftPerMin} converts from meter per second (m/s) to foot per minute (ft/min) }
\arguments{
  \item{meterPerSec}{meter per second (m/s) }
}
\value{foot per minute (ft/min) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterPerSecTOftPerMin(10)
}
\keyword{programming}