\name{NISTmeterPerSecTOftPerSec}
\alias{NISTmeterPerSecTOftPerSec}
\title{Convert meter per second  to foot per second }
\usage{NISTmeterPerSecTOftPerSec(meterPerSec)}
\description{\code{NISTmeterPerSecTOftPerSec} converts from meter per second (m/s) to foot per second (ft/s) }
\arguments{
  \item{meterPerSec}{meter per second (m/s) }
}
\value{foot per second (ft/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterPerSecTOftPerSec(10)
}
\keyword{programming}