\name{NISTmeterPerSecTOinchPerSec}
\alias{NISTmeterPerSecTOinchPerSec}
\title{Convert meter per second  to inch per second }
\usage{NISTmeterPerSecTOinchPerSec(meterPerSec)}
\description{\code{NISTmeterPerSecTOinchPerSec} converts from meter per second (m/s) to inch per second (in/s) }
\arguments{
  \item{meterPerSec}{meter per second (m/s) }
}
\value{inch per second (in/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterPerSecTOinchPerSec(10)
}
\keyword{programming}