\name{NISTmeterPerSecTOkmPerHour}
\alias{NISTmeterPerSecTOkmPerHour}
\title{Convert meter per second  to kilometer per hour }
\usage{NISTmeterPerSecTOkmPerHour(meterPerSec)}
\description{\code{NISTmeterPerSecTOkmPerHour} converts from meter per second (m/s) to kilometer per hour (km/h) }
\arguments{
  \item{meterPerSec}{meter per second (m/s) }
}
\value{kilometer per hour (km/h) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterPerSecTOkmPerHour(10)
}
\keyword{programming}