\name{NISTmeterPerSecTOknot}
\alias{NISTmeterPerSecTOknot}
\title{Convert meter per second  to knot }
\usage{NISTmeterPerSecTOknot(meterPerSec)}
\description{\code{NISTmeterPerSecTOknot} converts from meter per second (m/s) to knot (nautical mile per hour) }
\arguments{
  \item{meterPerSec}{meter per second (m/s) }
}
\value{knot (nautical mile per hour) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterPerSecTOknot(10)
}
\keyword{programming}