\name{NISTmeterPerSecTOmilePerHour}
\alias{NISTmeterPerSecTOmilePerHour}
\title{Convert meter per second  to mile per hour }
\usage{NISTmeterPerSecTOmilePerHour(meterPerSec)}
\description{\code{NISTmeterPerSecTOmilePerHour} converts from meter per second (m/s) to mile per hour (mi/h) }
\arguments{
  \item{meterPerSec}{meter per second (m/s) }
}
\value{mile per hour (mi/h) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterPerSecTOmilePerHour(10)
}
\keyword{programming}