\name{NISTmeterPerSecTOmilePerMin}
\alias{NISTmeterPerSecTOmilePerMin}
\title{Convert meter per second  to mile per minute }
\usage{NISTmeterPerSecTOmilePerMin(meterPerSec)}
\description{\code{NISTmeterPerSecTOmilePerMin} converts from meter per second (m/s) to mile per minute (mi/min) }
\arguments{
  \item{meterPerSec}{meter per second (m/s) }
}
\value{mile per minute (mi/min) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterPerSecTOmilePerMin(10)
}
\keyword{programming}