\name{NISTmeterPerSecTOmilePerSec}
\alias{NISTmeterPerSecTOmilePerSec}
\title{Convert meter per second  to mile per second }
\usage{NISTmeterPerSecTOmilePerSec(meterPerSec)}
\description{\code{NISTmeterPerSecTOmilePerSec} converts from meter per second (m/s) to mile per second (mi/s) }
\arguments{
  \item{meterPerSec}{meter per second (m/s) }
}
\value{mile per second (mi/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterPerSecTOmilePerSec(10)
}
\keyword{programming}