\name{NISTmeterSqrdPerSecTOcentistokes}
\alias{NISTmeterSqrdPerSecTOcentistokes}
\title{Convert meter squared per second  to centistokes }
\usage{NISTmeterSqrdPerSecTOcentistokes(meterSqrdPerSec)}
\description{\code{NISTmeterSqrdPerSecTOcentistokes} converts from meter squared per second (m2/s) to centistokes (cSt) }
\arguments{
  \item{meterSqrdPerSec}{meter squared per second (m2/s) }
}
\value{centistokes (cSt) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterSqrdPerSecTOcentistokes(10)
}
\keyword{programming}