\name{NISTmeterSqrdPerSecTOstokes}
\alias{NISTmeterSqrdPerSecTOstokes}
\title{Convert meter squared per second  to stokes }
\usage{NISTmeterSqrdPerSecTOstokes(meterSqrdPerSec)}
\description{\code{NISTmeterSqrdPerSecTOstokes} converts from meter squared per second (m2/s) to stokes (St) }
\arguments{
  \item{meterSqrdPerSec}{meter squared per second (m2/s) }
}
\value{stokes (St) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterSqrdPerSecTOstokes(10)
}
\keyword{programming}