\name{NISTmeterSqrdTOdarcy}
\alias{NISTmeterSqrdTOdarcy}
\title{Convert meter squared  to darcy 14}
\usage{NISTmeterSqrdTOdarcy(meterSqrd)}
\description{\code{NISTmeterSqrdTOdarcy} converts from meter squared (m2) to darcy 14 }
\arguments{
  \item{meterSqrd}{meter squared (m2) }
}
\value{darcy 14 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterSqrdTOdarcy(10)
}
\keyword{programming}