\name{NISTmeterTOangstrom}
\alias{NISTmeterTOangstrom}
\title{Convert meter  to angstrom }
\usage{NISTmeterTOangstrom(meter)}
\description{\code{NISTmeterTOangstrom} converts from meter (m) to angstrom (a) }
\arguments{
  \item{meter}{meter (m) }
}
\value{angstrom (a) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterTOangstrom(10)
}
\keyword{programming}