\name{NISTmeterTOastronomicalUn}
\alias{NISTmeterTOastronomicalUn}
\title{Convert meter  to astronomical unit }
\usage{NISTmeterTOastronomicalUn(meter)}
\description{\code{NISTmeterTOastronomicalUn} converts from meter (m) to astronomical unit (ua) }
\arguments{
  \item{meter}{meter (m) }
}
\value{astronomical unit (ua) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterTOastronomicalUn(10)
}
\keyword{programming}