\name{NISTmeterTOchain}
\alias{NISTmeterTOchain}
\title{Convert meter  to chain 7}
\usage{NISTmeterTOchain(meter)}
\description{\code{NISTmeterTOchain} converts from meter (m) to chain (based on U.S. survey foot) (ch) 7 }
\arguments{
  \item{meter}{meter (m) }
}
\value{chain (based on U.S. survey foot) (ch) 7 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterTOchain(10)
}
\keyword{programming}