\name{NISTmeterTOfathom}
\alias{NISTmeterTOfathom}
\title{Convert meter  to fathom 7}
\usage{NISTmeterTOfathom(meter)}
\description{\code{NISTmeterTOfathom} converts from meter (m) to fathom (based on U.S. survey foot) 7 }
\arguments{
  \item{meter}{meter (m) }
}
\value{fathom (based on U.S. survey foot) 7 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterTOfathom(10)
}
\keyword{programming}