\name{NISTmeterTOft}
\alias{NISTmeterTOft}
\title{Convert meter  to foot }
\usage{NISTmeterTOft(meter)}
\description{\code{NISTmeterTOft} converts from meter (m) to foot (ft) }
\arguments{
  \item{meter}{meter (m) }
}
\value{foot (ft) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterTOft(10)
}
\keyword{programming}