\name{NISTmeterTOftUSsurvey}
\alias{NISTmeterTOftUSsurvey}
\title{Convert meter  to foot 7}
\usage{NISTmeterTOftUSsurvey(meter)}
\description{\code{NISTmeterTOftUSsurvey} converts from meter (m) to foot (U.S. survey) (ft) 7 }
\arguments{
  \item{meter}{meter (m) }
}
\value{foot (U.S. survey) (ft) 7 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterTOftUSsurvey(10)
}
\keyword{programming}