\name{NISTmeterTOinch}
\alias{NISTmeterTOinch}
\title{Convert meter  to inch }
\usage{NISTmeterTOinch(meter)}
\description{\code{NISTmeterTOinch} converts from meter (m) to inch (in) }
\arguments{
  \item{meter}{meter (m) }
}
\value{inch (in) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterTOinch(10)
}
\keyword{programming}