\name{NISTmeterTOlightYear}
\alias{NISTmeterTOlightYear}
\title{Convert meter  to light year 18}
\usage{NISTmeterTOlightYear(meter)}
\description{\code{NISTmeterTOlightYear} converts from meter (m) to light year (l.y.) 18 }
\arguments{
  \item{meter}{meter (m) }
}
\value{light year (l.y.) 18 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterTOlightYear(10)
}
\keyword{programming}