\name{NISTmeterTOmicroinch}
\alias{NISTmeterTOmicroinch}
\title{Convert meter  to microinch}
\usage{NISTmeterTOmicroinch(meter)}
\description{\code{NISTmeterTOmicroinch} converts from meter (m) to microinch }
\arguments{
  \item{meter}{meter (m) }
}
\value{microinch }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterTOmicroinch(10)
}
\keyword{programming}