\name{NISTmeterTOmicron}
\alias{NISTmeterTOmicron}
\title{Convert meter  to micron }
\usage{NISTmeterTOmicron(meter)}
\description{\code{NISTmeterTOmicron} converts from meter (m) to micron (u) }
\arguments{
  \item{meter}{meter (m) }
}
\value{micron (u) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterTOmicron(10)
}
\keyword{programming}