\name{NISTmeterTOmil}
\alias{NISTmeterTOmil}
\title{Convert meter  to mil }
\usage{NISTmeterTOmil(meter)}
\description{\code{NISTmeterTOmil} converts from meter (m) to mil (0.001 in) }
\arguments{
  \item{meter}{meter (m) }
}
\value{mil (0.001 in) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterTOmil(10)
}
\keyword{programming}