\name{NISTmeterTOmile}
\alias{NISTmeterTOmile}
\title{Convert meter  to mile }
\usage{NISTmeterTOmile(meter)}
\description{\code{NISTmeterTOmile} converts from meter (m) to mile (mi) }
\arguments{
  \item{meter}{meter (m) }
}
\value{mile (mi) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterTOmile(10)
}
\keyword{programming}