\name{NISTmeterTOmileNautical}
\alias{NISTmeterTOmileNautical}
\title{Convert meter  to mile, nautical 20}
\usage{NISTmeterTOmileNautical(meter)}
\description{\code{NISTmeterTOmileNautical} converts from meter (m) to mile, nautical 20 }
\arguments{
  \item{meter}{meter (m) }
}
\value{mile, nautical 20 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterTOmileNautical(10)
}
\keyword{programming}