\name{NISTmeterTOmileUSsurvey}
\alias{NISTmeterTOmileUSsurvey}
\title{Convert meter  to mile 7}
\usage{NISTmeterTOmileUSsurvey(meter)}
\description{\code{NISTmeterTOmileUSsurvey} converts from meter (m) to mile (based on U.S. survey foot) (mi) 7 }
\arguments{
  \item{meter}{meter (m) }
}
\value{mile (based on U.S. survey foot) (mi) 7 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterTOmileUSsurvey(10)
}
\keyword{programming}