\name{NISTmeterTOparsec}
\alias{NISTmeterTOparsec}
\title{Convert meter  to parsec }
\usage{NISTmeterTOparsec(meter)}
\description{\code{NISTmeterTOparsec} converts from meter (m) to parsec (pc) }
\arguments{
  \item{meter}{meter (m) }
}
\value{parsec (pc) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterTOparsec(10)
}
\keyword{programming}