\name{NISTmeterTOpicaComputer}
\alias{NISTmeterTOpicaComputer}
\title{Convert meter  to pica }
\usage{NISTmeterTOpicaComputer(meter)}
\description{\code{NISTmeterTOpicaComputer} converts from meter (m) to pica (computer) (1/6 in) }
\arguments{
  \item{meter}{meter (m) }
}
\value{pica (computer) (1/6 in) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterTOpicaComputer(10)
}
\keyword{programming}