\name{NISTmeterTOpicaPrinter}
\alias{NISTmeterTOpicaPrinter}
\title{Convert meter  to pica }
\usage{NISTmeterTOpicaPrinter(meter)}
\description{\code{NISTmeterTOpicaPrinter} converts from meter (m) to pica (printer's) }
\arguments{
  \item{meter}{meter (m) }
}
\value{pica (printer's) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterTOpicaPrinter(10)
}
\keyword{programming}