\name{NISTmeterTOpointComputer}
\alias{NISTmeterTOpointComputer}
\title{Convert meter  to point }
\usage{NISTmeterTOpointComputer(meter)}
\description{\code{NISTmeterTOpointComputer} converts from meter (m) to point (computer) (1/72 in) }
\arguments{
  \item{meter}{meter (m) }
}
\value{point (computer) (1/72 in) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterTOpointComputer(10)
}
\keyword{programming}