\name{NISTmeterTOpointPrinter}
\alias{NISTmeterTOpointPrinter}
\title{Convert meter  to point }
\usage{NISTmeterTOpointPrinter(meter)}
\description{\code{NISTmeterTOpointPrinter} converts from meter (m) to point (printer's) }
\arguments{
  \item{meter}{meter (m) }
}
\value{point (printer's) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterTOpointPrinter(10)
}
\keyword{programming}