\name{NISTmeterTOrod}
\alias{NISTmeterTOrod}
\title{Convert meter  to rod 7}
\usage{NISTmeterTOrod(meter)}
\description{\code{NISTmeterTOrod} converts from meter (m) to rod (based on U.S. survey foot) (rd) 7 }
\arguments{
  \item{meter}{meter (m) }
}
\value{rod (based on U.S. survey foot) (rd) 7 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterTOrod(10)
}
\keyword{programming}