\name{NISTmeterTOyard}
\alias{NISTmeterTOyard}
\title{Convert meter  to yard }
\usage{NISTmeterTOyard(meter)}
\description{\code{NISTmeterTOyard} converts from meter (m) to yard (yd) }
\arguments{
  \item{meter}{meter (m) }
}
\value{yard (yd) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmeterTOyard(10)
}
\keyword{programming}