\name{NISTmhoTOsiemens}
\alias{NISTmhoTOsiemens}
\title{Convert mho to siemens }
\usage{NISTmhoTOsiemens(mho)}
\description{\code{NISTmhoTOsiemens} converts from mho to siemens (S) }
\arguments{
  \item{mho}{mho }
}
\value{siemens (S) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmhoTOsiemens(10)
}
\keyword{programming}