\name{NISTmicroinchTOmeter}
\alias{NISTmicroinchTOmeter}
\title{Convert microinch to meter }
\usage{NISTmicroinchTOmeter(microinch)}
\description{\code{NISTmicroinchTOmeter} converts from microinch to meter (m) }
\arguments{
  \item{microinch}{microinch }
}
\value{meter (m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmicroinchTOmeter(10)
}
\keyword{programming}