\name{NISTmicroinchTOmicrometer}
\alias{NISTmicroinchTOmicrometer}
\title{Convert microinch to micrometer }
\usage{NISTmicroinchTOmicrometer(microinch)}
\description{\code{NISTmicroinchTOmicrometer} converts from microinch to micrometer (um) }
\arguments{
  \item{microinch}{microinch }
}
\value{micrometer (um) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmicroinchTOmicrometer(10)
}
\keyword{programming}