\name{NISTmicrometerTOmicron}
\alias{NISTmicrometerTOmicron}
\title{Convert micrometer  to micron }
\usage{NISTmicrometerTOmicron(micrometer)}
\description{\code{NISTmicrometerTOmicron} converts from micrometer (um) to micron (u) }
\arguments{
  \item{micrometer}{micrometer (um) }
}
\value{micron (u) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmicrometerTOmicron(10)
}
\keyword{programming}