\name{NISTmicronTOmeter}
\alias{NISTmicronTOmeter}
\title{Convert micron  to meter }
\usage{NISTmicronTOmeter(micron)}
\description{\code{NISTmicronTOmeter} converts from micron (u) to meter (m) }
\arguments{
  \item{micron}{micron (u) }
}
\value{meter (m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmicronTOmeter(10)
}
\keyword{programming}