\name{NISTmileNauticalTOmeter}
\alias{NISTmileNauticalTOmeter}
\title{Convert mile, nautical 20 to meter }
\usage{NISTmileNauticalTOmeter(mileNautical)}
\description{\code{NISTmileNauticalTOmeter} converts from mile, nautical 20 to meter (m) }
\arguments{
  \item{mileNautical}{mile, nautical 20 }
}
\value{meter (m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmileNauticalTOmeter(10)
}
\keyword{programming}