\name{NISTmilePerGallonTOkmPerLiter}
\alias{NISTmilePerGallonTOkmPerLiter}
\title{Convert mile per gallon  to kilometer per liter }
\usage{NISTmilePerGallonTOkmPerLiter(milePerGallon)}
\description{\code{NISTmilePerGallonTOkmPerLiter} converts from mile per gallon (U.S.) (mpg) (mi/gal) to kilometer per liter (km/L) }
\arguments{
  \item{milePerGallon}{mile per gallon (U.S.) (mpg) (mi/gal) }
}
\value{kilometer per liter (km/L) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmilePerGallonTOkmPerLiter(10)
}
\keyword{programming}