\name{NISTmilePerGallonTOliterPer100Km}
\alias{NISTmilePerGallonTOliterPer100Km}
\title{Convert mile per gallon 21 to liter per 100 kilometer }
\usage{NISTmilePerGallonTOliterPer100Km(milePerGallon)}
\description{\code{NISTmilePerGallonTOliterPer100Km} converts from mile per gallon (U.S.) (mpg) (mi/gal) 21 to liter per 100 kilometer (L/100 km) }
\arguments{
  \item{milePerGallon}{mile per gallon (U.S.) (mpg) (mi/gal) 21 }
}
\value{liter per 100 kilometer (L/100 km) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmilePerGallonTOliterPer100Km(10)
}
\keyword{programming}