\name{NISTmilePerGallonTOmeterPerCubMeter}
\alias{NISTmilePerGallonTOmeterPerCubMeter}
\title{Convert mile per gallon  to meter per cubic meter }
\usage{NISTmilePerGallonTOmeterPerCubMeter(milePerGallon)}
\description{\code{NISTmilePerGallonTOmeterPerCubMeter} converts from mile per gallon (U.S.) (mpg) (mi/gal) to meter per cubic meter (m/m3) }
\arguments{
  \item{milePerGallon}{mile per gallon (U.S.) (mpg) (mi/gal) }
}
\value{meter per cubic meter (m/m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmilePerGallonTOmeterPerCubMeter(10)
}
\keyword{programming}