\name{NISTmilePerHourTOkmPerHour}
\alias{NISTmilePerHourTOkmPerHour}
\title{Convert mile per hour  to kilometer per hour }
\usage{NISTmilePerHourTOkmPerHour(milePerHour)}
\description{\code{NISTmilePerHourTOkmPerHour} converts from mile per hour (mi/h) to kilometer per hour (km/h) }
\arguments{
  \item{milePerHour}{mile per hour (mi/h) }
}
\value{kilometer per hour (km/h) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmilePerHourTOkmPerHour(10)
}
\keyword{programming}