\name{NISTmilePerHourTOmeterPerSec}
\alias{NISTmilePerHourTOmeterPerSec}
\title{Convert mile per hour  to meter per second }
\usage{NISTmilePerHourTOmeterPerSec(milePerHour)}
\description{\code{NISTmilePerHourTOmeterPerSec} converts from mile per hour (mi/h) to meter per second (m/s) }
\arguments{
  \item{milePerHour}{mile per hour (mi/h) }
}
\value{meter per second (m/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmilePerHourTOmeterPerSec(10)
}
\keyword{programming}