\name{NISTmilePerMinTOmeterPerSec}
\alias{NISTmilePerMinTOmeterPerSec}
\title{Convert mile per minute  to meter per second }
\usage{NISTmilePerMinTOmeterPerSec(milePerMin)}
\description{\code{NISTmilePerMinTOmeterPerSec} converts from mile per minute (mi/min) to meter per second (m/s) }
\arguments{
  \item{milePerMin}{mile per minute (mi/min) }
}
\value{meter per second (m/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmilePerMinTOmeterPerSec(10)
}
\keyword{programming}