\name{NISTmilePerSecTOmeterPerSec}
\alias{NISTmilePerSecTOmeterPerSec}
\title{Convert mile per second  to meter per second }
\usage{NISTmilePerSecTOmeterPerSec(milePerSec)}
\description{\code{NISTmilePerSecTOmeterPerSec} converts from mile per second (mi/s) to meter per second (m/s) }
\arguments{
  \item{milePerSec}{mile per second (mi/s) }
}
\value{meter per second (m/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmilePerSecTOmeterPerSec(10)
}
\keyword{programming}