\name{NISTmileTOkm}
\alias{NISTmileTOkm}
\title{Convert mile  to kilometer }
\usage{NISTmileTOkm(mile)}
\description{\code{NISTmileTOkm} converts from mile (mi) to kilometer (km) }
\arguments{
  \item{mile}{mile (mi) }
}
\value{kilometer (km) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmileTOkm(10)
}
\keyword{programming}