\name{NISTmileTOmeter}
\alias{NISTmileTOmeter}
\title{Convert mile  to meter }
\usage{NISTmileTOmeter(mile)}
\description{\code{NISTmileTOmeter} converts from mile (mi) to meter (m) }
\arguments{
  \item{mile}{mile (mi) }
}
\value{meter (m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmileTOmeter(10)
}
\keyword{programming}