\name{NISTmileUSsurveyTOkm}
\alias{NISTmileUSsurveyTOkm}
\title{Convert mile 7 to kilometer }
\usage{NISTmileUSsurveyTOkm(mileUSsurvey)}
\description{\code{NISTmileUSsurveyTOkm} converts from mile (based on U.S. survey foot) (mi) 7 to kilometer (km) }
\arguments{
  \item{mileUSsurvey}{mile (based on U.S. survey foot) (mi) 7 }
}
\value{kilometer (km) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmileUSsurveyTOkm(10)
}
\keyword{programming}