\name{NISTmileUSsurveyTOmeter}
\alias{NISTmileUSsurveyTOmeter}
\title{Convert mile 7 to meter }
\usage{NISTmileUSsurveyTOmeter(mileUSsurvey)}
\description{\code{NISTmileUSsurveyTOmeter} converts from mile (based on U.S. survey foot) (mi) 7 to meter (m) }
\arguments{
  \item{mileUSsurvey}{mile (based on U.S. survey foot) (mi) 7 }
}
\value{meter (m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmileUSsurveyTOmeter(10)
}
\keyword{programming}