\name{NISTmillibarTOkpascal}
\alias{NISTmillibarTOkpascal}
\title{Convert millibar  to kilopascal }
\usage{NISTmillibarTOkpascal(millibar)}
\description{\code{NISTmillibarTOkpascal} converts from millibar (mbar) to kilopascal (kPa) }
\arguments{
  \item{millibar}{millibar (mbar) }
}
\value{kilopascal (kPa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmillibarTOkpascal(10)
}
\keyword{programming}