\name{NISTmillibarTOpascal}
\alias{NISTmillibarTOpascal}
\title{Convert millibar  to pascal }
\usage{NISTmillibarTOpascal(millibar)}
\description{\code{NISTmillibarTOpascal} converts from millibar (mbar) to pascal (Pa) }
\arguments{
  \item{millibar}{millibar (mbar) }
}
\value{pascal (Pa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmillibarTOpascal(10)
}
\keyword{programming}