\name{NISTmilligramPerLiterTOgrainPerGallon}
\alias{NISTmilligramPerLiterTOgrainPerGallon}
\title{Convert milligram per liter  to grain per gallon }
\usage{NISTmilligramPerLiterTOgrainPerGallon(milligramPerLiter)}
\description{\code{NISTmilligramPerLiterTOgrainPerGallon} converts from milligram per liter (mg/L) to grain per gallon (U.S.) (gr/gal) }
\arguments{
  \item{milligramPerLiter}{milligram per liter (mg/L) }
}
\value{grain per gallon (U.S.) (gr/gal) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmilligramPerLiterTOgrainPerGallon(10)
}
\keyword{programming}