\name{NISTmilligramTOgrain}
\alias{NISTmilligramTOgrain}
\title{Convert milligram  to grain }
\usage{NISTmilligramTOgrain(milligram)}
\description{\code{NISTmilligramTOgrain} converts from milligram (mg) to grain (gr) }
\arguments{
  \item{milligram}{milligram (mg) }
}
\value{grain (gr) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmilligramTOgrain(10)
}
\keyword{programming}