\name{NISTmillimeterOfWaterConvtnlTOpascal}
\alias{NISTmillimeterOfWaterConvtnlTOpascal}
\title{Convert millimeter of water, conventional 12 to pascal }
\usage{NISTmillimeterOfWaterConvtnlTOpascal(millimeterOfWaterConvtnl)}
\description{\code{NISTmillimeterOfWaterConvtnlTOpascal} converts from millimeter of water, conventional (mm H2O) 12 to pascal (Pa) }
\arguments{
  \item{millimeterOfWaterConvtnl}{millimeter of water, conventional (mm H2O) 12 }
}
\value{pascal (Pa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmillimeterOfWaterConvtnlTOpascal(10)
}
\keyword{programming}