\name{NISTmillimeterTOmil}
\alias{NISTmillimeterTOmil}
\title{Convert millimeter  to mil }
\usage{NISTmillimeterTOmil(millimeter)}
\description{\code{NISTmillimeterTOmil} converts from millimeter (mm) to mil (0.001 in) }
\arguments{
  \item{millimeter}{millimeter (mm) }
}
\value{mil (0.001 in) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmillimeterTOmil(10)
}
\keyword{programming}