\name{NISTmillimeterTOpicaComputer}
\alias{NISTmillimeterTOpicaComputer}
\title{Convert millimeter  to pica }
\usage{NISTmillimeterTOpicaComputer(millimeter)}
\description{\code{NISTmillimeterTOpicaComputer} converts from millimeter (mm) to pica (computer) (1/6 in) }
\arguments{
  \item{millimeter}{millimeter (mm) }
}
\value{pica (computer) (1/6 in) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmillimeterTOpicaComputer(10)
}
\keyword{programming}