\name{NISTmillimeterTOpicaPrinter}
\alias{NISTmillimeterTOpicaPrinter}
\title{Convert millimeter  to pica }
\usage{NISTmillimeterTOpicaPrinter(millimeter)}
\description{\code{NISTmillimeterTOpicaPrinter} converts from millimeter (mm) to pica (printer's) }
\arguments{
  \item{millimeter}{millimeter (mm) }
}
\value{pica (printer's) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmillimeterTOpicaPrinter(10)
}
\keyword{programming}