\name{NISTmillimeterTOpointComputer}
\alias{NISTmillimeterTOpointComputer}
\title{Convert millimeter  to point }
\usage{NISTmillimeterTOpointComputer(millimeter)}
\description{\code{NISTmillimeterTOpointComputer} converts from millimeter (mm) to point (computer) (1/72 in) }
\arguments{
  \item{millimeter}{millimeter (mm) }
}
\value{point (computer) (1/72 in) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmillimeterTOpointComputer(10)
}
\keyword{programming}