\name{NISTmillimeterTOpointPrinter}
\alias{NISTmillimeterTOpointPrinter}
\title{Convert millimeter  to point }
\usage{NISTmillimeterTOpointPrinter(millimeter)}
\description{\code{NISTmillimeterTOpointPrinter} converts from millimeter (mm) to point (printer's) }
\arguments{
  \item{millimeter}{millimeter (mm) }
}
\value{point (printer's) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmillimeterTOpointPrinter(10)
}
\keyword{programming}