\name{NISTmillinewtonMeterTOounceForceInch}
\alias{NISTmillinewtonMeterTOounceForceInch}
\title{Convert millinewton meter  to ounce -force inch }
\usage{NISTmillinewtonMeterTOounceForceInch(millinewtonMeter)}
\description{\code{NISTmillinewtonMeterTOounceForceInch} converts from millinewton meter (mN * m) to ounce (avoirdupois)-force inch (ozf * in) }
\arguments{
  \item{millinewtonMeter}{millinewton meter (mN * m) }
}
\value{ounce (avoirdupois)-force inch (ozf * in) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmillinewtonMeterTOounceForceInch(10)
}
\keyword{programming}