\name{NISTmilTOdeg}
\alias{NISTmilTOdeg}
\title{Convert mil  to degree }
\usage{NISTmilTOdeg(mil)}
\description{\code{NISTmilTOdeg} converts from mil (angle) to degree () }
\arguments{
  \item{mil}{mil (angle) }
}
\value{degree () }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmilTOdeg(10)
}
\keyword{programming}