\name{NISTmilTOmeter}
\alias{NISTmilTOmeter}
\title{Convert mil  to meter }
\usage{NISTmilTOmeter(mil)}
\description{\code{NISTmilTOmeter} converts from mil (0.001 in) to meter (m) }
\arguments{
  \item{mil}{mil (0.001 in) }
}
\value{meter (m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmilTOmeter(10)
}
\keyword{programming}