\name{NISTmilTOmillimeter}
\alias{NISTmilTOmillimeter}
\title{Convert mil  to millimeter }
\usage{NISTmilTOmillimeter(mil)}
\description{\code{NISTmilTOmillimeter} converts from mil (0.001 in) to millimeter (mm) }
\arguments{
  \item{mil}{mil (0.001 in) }
}
\value{millimeter (mm) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmilTOmillimeter(10)
}
\keyword{programming}