\name{NISTmilTOradian}
\alias{NISTmilTOradian}
\title{Convert mil  to radian }
\usage{NISTmilTOradian(mil)}
\description{\code{NISTmilTOradian} converts from mil (angle) to radian (rad) }
\arguments{
  \item{mil}{mil (angle) }
}
\value{radian (rad) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmilTOradian(10)
}
\keyword{programming}