\name{NISTminSiderealTOsec}
\alias{NISTminSiderealTOsec}
\title{Convert minute  to second }
\usage{NISTminSiderealTOsec(minSidereal)}
\description{\code{NISTminSiderealTOsec} converts from minute (sidereal) to second (s) }
\arguments{
  \item{minSidereal}{minute (sidereal) }
}
\value{second (s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTminSiderealTOsec(10)
}
\keyword{programming}