\name{NISTminTOradian}
\alias{NISTminTOradian}
\title{Convert minute  to radian }
\usage{NISTminTOradian(min)}
\description{\code{NISTminTOradian} converts from minute (angle) (') to radian (rad) }
\arguments{
  \item{min}{minute (angle) (') }
}
\value{radian (rad) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTminTOradian(10)
}
\keyword{programming}