\name{NISTmlTOcup}
\alias{NISTmlTOcup}
\title{Convert milliliter  to cup }
\usage{NISTmlTOcup(ml)}
\description{\code{NISTmlTOcup} converts from milliliter (mL) to cup (U.S.) }
\arguments{
  \item{ml}{milliliter (mL) }
}
\value{cup (U.S.) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmlTOcup(10)
}
\keyword{programming}