\name{NISTmlTOfluidOunce}
\alias{NISTmlTOfluidOunce}
\title{Convert milliliter  to fluid ounce }
\usage{NISTmlTOfluidOunce(ml)}
\description{\code{NISTmlTOfluidOunce} converts from milliliter (mL) to fluid ounce (U.S.) (fl oz) }
\arguments{
  \item{ml}{milliliter (mL) }
}
\value{fluid ounce (U.S.) (fl oz) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmlTOfluidOunce(10)
}
\keyword{programming}