\name{NISTmlTOounce}
\alias{NISTmlTOounce}
\title{Convert milliliter  to ounce }
\usage{NISTmlTOounce(ml)}
\description{\code{NISTmlTOounce} converts from milliliter (mL) to ounce (U.S. fluid) (fl oz) }
\arguments{
  \item{ml}{milliliter (mL) }
}
\value{ounce (U.S. fluid) (fl oz) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmlTOounce(10)
}
\keyword{programming}