\name{NISTmlTOounceImperial}
\alias{NISTmlTOounceImperial}
\title{Convert milliliter  to ounce }
\usage{NISTmlTOounceImperial(ml)}
\description{\code{NISTmlTOounceImperial} converts from milliliter (mL) to ounce [Canadian and U.K. fluid (Imperial)] (fl oz) }
\arguments{
  \item{ml}{milliliter (mL) }
}
\value{ounce [Canadian and U.K. fluid (Imperial)] (fl oz) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmlTOounceImperial(10)
}
\keyword{programming}