\name{NISTmlTOteaspoon}
\alias{NISTmlTOteaspoon}
\title{Convert milliliter  to teaspoon}
\usage{NISTmlTOteaspoon(ml)}
\description{\code{NISTmlTOteaspoon} converts from milliliter (mL) to teaspoon }
\arguments{
  \item{ml}{milliliter (mL) }
}
\value{teaspoon }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmlTOteaspoon(10)
}
\keyword{programming}