\name{NISTmpascalTOkgForcePerSqrMillimeter}
\alias{NISTmpascalTOkgForcePerSqrMillimeter}
\title{Convert megapascal  to kilogram-force per square millimeter }
\usage{NISTmpascalTOkgForcePerSqrMillimeter(mpascal)}
\description{\code{NISTmpascalTOkgForcePerSqrMillimeter} converts from megapascal (MPa) to kilogram-force per square millimeter (kgf/mm2) }
\arguments{
  \item{mpascal}{megapascal (MPa) }
}
\value{kilogram-force per square millimeter (kgf/mm2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTmpascalTOkgForcePerSqrMillimeter(10)
}
\keyword{programming}