\name{NISTnanometerTOangstrom}
\alias{NISTnanometerTOangstrom}
\title{Convert nanometer  to angstrom }
\usage{NISTnanometerTOangstrom(nanometer)}
\description{\code{NISTnanometerTOangstrom} converts from nanometer (nm) to angstrom (a) }
\arguments{
  \item{nanometer}{nanometer (nm) }
}
\value{angstrom (a) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTnanometerTOangstrom(10)
}
\keyword{programming}