\name{NISTnanosecTOshake}
\alias{NISTnanosecTOshake}
\title{Convert nanosecond  to shake}
\usage{NISTnanosecTOshake(nanosec)}
\description{\code{NISTnanosecTOshake} converts from nanosecond (ns) to shake }
\arguments{
  \item{nanosec}{nanosecond (ns) }
}
\value{shake }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTnanosecTOshake(10)
}
\keyword{programming}