\name{NISTnewtonMeterPerMeterTOpoundForceFtPerInch}
\alias{NISTnewtonMeterPerMeterTOpoundForceFtPerInch}
\title{Convert newton meter per meter  to pound-force foot per inch }
\usage{NISTnewtonMeterPerMeterTOpoundForceFtPerInch(newtonMeterPerMeter)}
\description{\code{NISTnewtonMeterPerMeterTOpoundForceFtPerInch} converts from newton meter per meter (N * m/m) to pound-force foot per inch (lbf * ft/in) }
\arguments{
  \item{newtonMeterPerMeter}{newton meter per meter (N * m/m) }
}
\value{pound-force foot per inch (lbf * ft/in) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTnewtonMeterPerMeterTOpoundForceFtPerInch(10)
}
\keyword{programming}