\name{NISTnewtonMeterPerMeterTOpoundForceInchPerInch}
\alias{NISTnewtonMeterPerMeterTOpoundForceInchPerInch}
\title{Convert newton meter per meter  to pound-force inch per inch }
\usage{NISTnewtonMeterPerMeterTOpoundForceInchPerInch(newtonMeterPerMeter)}
\description{\code{NISTnewtonMeterPerMeterTOpoundForceInchPerInch} converts from newton meter per meter (N * m/m) to pound-force inch per inch (lbf * in/in) }
\arguments{
  \item{newtonMeterPerMeter}{newton meter per meter (N * m/m) }
}
\value{pound-force inch per inch (lbf * in/in) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTnewtonMeterPerMeterTOpoundForceInchPerInch(10)
}
\keyword{programming}