\name{NISTnewtonMeterTOdyneCm}
\alias{NISTnewtonMeterTOdyneCm}
\title{Convert newton meter  to dyne centimeter }
\usage{NISTnewtonMeterTOdyneCm(newtonMeter)}
\description{\code{NISTnewtonMeterTOdyneCm} converts from newton meter (N * m) to dyne centimeter (dyn * cm) }
\arguments{
  \item{newtonMeter}{newton meter (N * m) }
}
\value{dyne centimeter (dyn * cm) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTnewtonMeterTOdyneCm(10)
}
\keyword{programming}