\name{NISTnewtonMeterTOkgForceMeter}
\alias{NISTnewtonMeterTOkgForceMeter}
\title{Convert newton meter  to kilogram-force meter }
\usage{NISTnewtonMeterTOkgForceMeter(newtonMeter)}
\description{\code{NISTnewtonMeterTOkgForceMeter} converts from newton meter (N * m) to kilogram-force meter (kgf * m) }
\arguments{
  \item{newtonMeter}{newton meter (N * m) }
}
\value{kilogram-force meter (kgf * m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTnewtonMeterTOkgForceMeter(10)
}
\keyword{programming}