\name{NISTnewtonMeterTOounceForceInch}
\alias{NISTnewtonMeterTOounceForceInch}
\title{Convert newton meter  to ounce -force inch }
\usage{NISTnewtonMeterTOounceForceInch(newtonMeter)}
\description{\code{NISTnewtonMeterTOounceForceInch} converts from newton meter (N * m) to ounce (avoirdupois)-force inch (ozf * in) }
\arguments{
  \item{newtonMeter}{newton meter (N * m) }
}
\value{ounce (avoirdupois)-force inch (ozf * in) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTnewtonMeterTOounceForceInch(10)
}
\keyword{programming}