\name{NISTnewtonMeterTOpoundForceFt}
\alias{NISTnewtonMeterTOpoundForceFt}
\title{Convert newton meter  to pound-force foot }
\usage{NISTnewtonMeterTOpoundForceFt(newtonMeter)}
\description{\code{NISTnewtonMeterTOpoundForceFt} converts from newton meter (N * m) to pound-force foot (lbf * ft) }
\arguments{
  \item{newtonMeter}{newton meter (N * m) }
}
\value{pound-force foot (lbf * ft) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTnewtonMeterTOpoundForceFt(10)
}
\keyword{programming}