\name{NISTnewtonMeterTOpoundForceInch}
\alias{NISTnewtonMeterTOpoundForceInch}
\title{Convert newton meter  to pound-force inch }
\usage{NISTnewtonMeterTOpoundForceInch(newtonMeter)}
\description{\code{NISTnewtonMeterTOpoundForceInch} converts from newton meter (N * m) to pound-force inch (lbf * in) }
\arguments{
  \item{newtonMeter}{newton meter (N * m) }
}
\value{pound-force inch (lbf * in) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTnewtonMeterTOpoundForceInch(10)
}
\keyword{programming}