\name{NISTnewtonPerKgTOpoundForcePerPound}
\alias{NISTnewtonPerKgTOpoundForcePerPound}
\title{Convert newton per kilogram  to pound-force per pound }
\usage{NISTnewtonPerKgTOpoundForcePerPound(newtonPerKg)}
\description{\code{NISTnewtonPerKgTOpoundForcePerPound} converts from newton per kilogram (N/kg) to pound-force per pound (lbf/lb) (thrust to mass ratio) }
\arguments{
  \item{newtonPerKg}{newton per kilogram (N/kg) }
}
\value{pound-force per pound (lbf/lb) (thrust to mass ratio) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTnewtonPerKgTOpoundForcePerPound(10)
}
\keyword{programming}