\name{NISTnewtonPerMeterTOpoundForcePerFt}
\alias{NISTnewtonPerMeterTOpoundForcePerFt}
\title{Convert newton per meter  to pound-force per foot }
\usage{NISTnewtonPerMeterTOpoundForcePerFt(newtonPerMeter)}
\description{\code{NISTnewtonPerMeterTOpoundForcePerFt} converts from newton per meter (N/m) to pound-force per foot (lbf/ft) }
\arguments{
  \item{newtonPerMeter}{newton per meter (N/m) }
}
\value{pound-force per foot (lbf/ft) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTnewtonPerMeterTOpoundForcePerFt(10)
}
\keyword{programming}