\name{NISTnewtonPerMeterTOpoundForcePerInch}
\alias{NISTnewtonPerMeterTOpoundForcePerInch}
\title{Convert newton per meter  to pound-force per inch }
\usage{NISTnewtonPerMeterTOpoundForcePerInch(newtonPerMeter)}
\description{\code{NISTnewtonPerMeterTOpoundForcePerInch} converts from newton per meter (N/m) to pound-force per inch (lbf/in) }
\arguments{
  \item{newtonPerMeter}{newton per meter (N/m) }
}
\value{pound-force per inch (lbf/in) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTnewtonPerMeterTOpoundForcePerInch(10)
}
\keyword{programming}