\name{NISTnewtonTOdyne}
\alias{NISTnewtonTOdyne}
\title{Convert newton  to dyne }
\usage{NISTnewtonTOdyne(newton)}
\description{\code{NISTnewtonTOdyne} converts from newton (N) to dyne (dyn) }
\arguments{
  \item{newton}{newton (N) }
}
\value{dyne (dyn) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTnewtonTOdyne(10)
}
\keyword{programming}