\name{NISTnewtonTOkgForce}
\alias{NISTnewtonTOkgForce}
\title{Convert newton  to kilogram-force }
\usage{NISTnewtonTOkgForce(newton)}
\description{\code{NISTnewtonTOkgForce} converts from newton (N) to kilogram-force (kgf) }
\arguments{
  \item{newton}{newton (N) }
}
\value{kilogram-force (kgf) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTnewtonTOkgForce(10)
}
\keyword{programming}