\name{NISTnewtonTOkilopond}
\alias{NISTnewtonTOkilopond}
\title{Convert newton  to kilopond }
\usage{NISTnewtonTOkilopond(newton)}
\description{\code{NISTnewtonTOkilopond} converts from newton (N) to kilopond (kilogram-force) (kp) }
\arguments{
  \item{newton}{newton (N) }
}
\value{kilopond (kilogram-force) (kp) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTnewtonTOkilopond(10)
}
\keyword{programming}