\name{NISTnewtonTOounceForce}
\alias{NISTnewtonTOounceForce}
\title{Convert newton  to ounce -force }
\usage{NISTnewtonTOounceForce(newton)}
\description{\code{NISTnewtonTOounceForce} converts from newton (N) to ounce (avoirdupois)-force (ozf) }
\arguments{
  \item{newton}{newton (N) }
}
\value{ounce (avoirdupois)-force (ozf) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTnewtonTOounceForce(10)
}
\keyword{programming}