\name{NISTnewtonTOpoundal}
\alias{NISTnewtonTOpoundal}
\title{Convert newton  to poundal}
\usage{NISTnewtonTOpoundal(newton)}
\description{\code{NISTnewtonTOpoundal} converts from newton (N) to poundal }
\arguments{
  \item{newton}{newton (N) }
}
\value{poundal }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTnewtonTOpoundal(10)
}
\keyword{programming}