\name{NISTnewtonTOpoundForce}
\alias{NISTnewtonTOpoundForce}
\title{Convert newton  to pound-force 23}
\usage{NISTnewtonTOpoundForce(newton)}
\description{\code{NISTnewtonTOpoundForce} converts from newton (N) to pound-force (lbf) 23 }
\arguments{
  \item{newton}{newton (N) }
}
\value{pound-force (lbf) 23 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTnewtonTOpoundForce(10)
}
\keyword{programming}