\name{NISToerstedTOamperePerMeter}
\alias{NISToerstedTOamperePerMeter}
\title{Convert oersted  to ampere per meter }
\usage{NISToerstedTOamperePerMeter(oersted)}
\description{\code{NISToerstedTOamperePerMeter} converts from oersted (Oe) to ampere per meter (A/m) }
\arguments{
  \item{oersted}{oersted (Oe) }
}
\value{ampere per meter (A/m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISToerstedTOamperePerMeter(10)
}
\keyword{programming}