\name{NISTohmCircularMilPerFtTOohmMeter}
\alias{NISTohmCircularMilPerFtTOohmMeter}
\title{Convert ohm circular-mil per foot to ohm meter }
\usage{NISTohmCircularMilPerFtTOohmMeter(ohmCircularMilPerFt)}
\description{\code{NISTohmCircularMilPerFtTOohmMeter} converts from ohm circular-mil per foot to ohm meter (ohm * m) }
\arguments{
  \item{ohmCircularMilPerFt}{ohm circular-mil per foot }
}
\value{ohm meter (ohm * m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTohmCircularMilPerFtTOohmMeter(10)
}
\keyword{programming}