\name{NISTohmCircularMilPerFtTOohmSqrMillimeterPerMeter}
\alias{NISTohmCircularMilPerFtTOohmSqrMillimeterPerMeter}
\title{Convert ohm circular-mil per foot to ohm square millimeter per meter }
\usage{NISTohmCircularMilPerFtTOohmSqrMillimeterPerMeter(ohmCircularMilPerFt)}
\description{\code{NISTohmCircularMilPerFtTOohmSqrMillimeterPerMeter} converts from ohm circular-mil per foot to ohm square millimeter per meter (ohm * mm2/m) }
\arguments{
  \item{ohmCircularMilPerFt}{ohm circular-mil per foot }
}
\value{ohm square millimeter per meter (ohm * mm2/m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTohmCircularMilPerFtTOohmSqrMillimeterPerMeter(10)
}
\keyword{programming}