\name{NISTohmCmTOohmMeter}
\alias{NISTohmCmTOohmMeter}
\title{Convert ohm centimeter  to ohm meter }
\usage{NISTohmCmTOohmMeter(ohmCm)}
\description{\code{NISTohmCmTOohmMeter} converts from ohm centimeter (ohm * cm) to ohm meter (ohm * m) }
\arguments{
  \item{ohmCm}{ohm centimeter (ohm * cm) }
}
\value{ohm meter (ohm * m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTohmCmTOohmMeter(10)
}
\keyword{programming}