\name{NISTohmMeterTOohmCircularMilPerFt}
\alias{NISTohmMeterTOohmCircularMilPerFt}
\title{Convert ohm meter  to ohm circular-mil per foot}
\usage{NISTohmMeterTOohmCircularMilPerFt(ohmMeter)}
\description{\code{NISTohmMeterTOohmCircularMilPerFt} converts from ohm meter (ohm * m) to ohm circular-mil per foot }
\arguments{
  \item{ohmMeter}{ohm meter (ohm * m) }
}
\value{ohm circular-mil per foot }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTohmMeterTOohmCircularMilPerFt(10)
}
\keyword{programming}