\name{NISTohmMeterTOohmCm}
\alias{NISTohmMeterTOohmCm}
\title{Convert ohm meter  to ohm centimeter }
\usage{NISTohmMeterTOohmCm(ohmMeter)}
\description{\code{NISTohmMeterTOohmCm} converts from ohm meter (ohm * m) to ohm centimeter (ohm * cm) }
\arguments{
  \item{ohmMeter}{ohm meter (ohm * m) }
}
\value{ohm centimeter (ohm * cm) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTohmMeterTOohmCm(10)
}
\keyword{programming}