\name{NISTohmSqrMillimeterPerMeterTOohmCircularMilPerFt}
\alias{NISTohmSqrMillimeterPerMeterTOohmCircularMilPerFt}
\title{Convert ohm square millimeter per meter  to ohm circular-mil per foot}
\usage{NISTohmSqrMillimeterPerMeterTOohmCircularMilPerFt(ohmSqrMillimeterPerMeter)}
\description{\code{NISTohmSqrMillimeterPerMeterTOohmCircularMilPerFt} converts from ohm square millimeter per meter (ohm * mm2/m) to ohm circular-mil per foot }
\arguments{
  \item{ohmSqrMillimeterPerMeter}{ohm square millimeter per meter (ohm * mm2/m) }
}
\value{ohm circular-mil per foot }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTohmSqrMillimeterPerMeterTOohmCircularMilPerFt(10)
}
\keyword{programming}