\name{NISTohmTOabohm}
\alias{NISTohmTOabohm}
\title{Convert ohm  to abohm}
\usage{NISTohmTOabohm(ohm)}
\description{\code{NISTohmTOabohm} converts from ohm (ohm) to abohm }
\arguments{
  \item{ohm}{ohm (ohm) }
}
\value{abohm }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTohmTOabohm(10)
}
\keyword{programming}