\name{NISTohmToEMUOfResistance}
\alias{NISTohmToEMUOfResistance}
\title{Convert ohm  to EMU of resistance }
\usage{NISTohmToEMUOfResistance(ohm)}
\description{\code{NISTohmToEMUOfResistance} converts from ohm (ohm) to EMU of resistance (abohm) }
\arguments{
  \item{ohm}{ohm (ohm) }
}
\value{EMU of resistance (abohm) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTohmToEMUOfResistance(10)
}
\keyword{programming}