\name{NISTohmToESUOfResistance}
\alias{NISTohmToESUOfResistance}
\title{Convert ohm  to ESU of resistance }
\usage{NISTohmToESUOfResistance(ohm)}
\description{\code{NISTohmToESUOfResistance} converts from ohm (ohm) to ESU of resistance (statohm) }
\arguments{
  \item{ohm}{ohm (ohm) }
}
\value{ESU of resistance (statohm) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTohmToESUOfResistance(10)
}
\keyword{programming}