\name{NISTohmTOstatohm}
\alias{NISTohmTOstatohm}
\title{Convert ohm  to statohm}
\usage{NISTohmTOstatohm(ohm)}
\description{\code{NISTohmTOstatohm} converts from ohm (ohm) to statohm }
\arguments{
  \item{ohm}{ohm (ohm) }
}
\value{statohm }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTohmTOstatohm(10)
}
\keyword{programming}