\name{NISTounceAvoirdupoisTOgram}
\alias{NISTounceAvoirdupoisTOgram}
\title{Convert ounce  to gram }
\usage{NISTounceAvoirdupoisTOgram(ounceAvoirdupois)}
\description{\code{NISTounceAvoirdupoisTOgram} converts from ounce (avoirdupois) (oz) to gram (g) }
\arguments{
  \item{ounceAvoirdupois}{ounce (avoirdupois) (oz) }
}
\value{gram (g) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTounceAvoirdupoisTOgram(10)
}
\keyword{programming}