\name{NISTounceAvoirdupoisTOkg}
\alias{NISTounceAvoirdupoisTOkg}
\title{Convert ounce  to kilogram }
\usage{NISTounceAvoirdupoisTOkg(ounceAvoirdupois)}
\description{\code{NISTounceAvoirdupoisTOkg} converts from ounce (avoirdupois) (oz) to kilogram (kg) }
\arguments{
  \item{ounceAvoirdupois}{ounce (avoirdupois) (oz) }
}
\value{kilogram (kg) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTounceAvoirdupoisTOkg(10)
}
\keyword{programming}