\name{NISTounceForceInchTOmillinewtonMeter}
\alias{NISTounceForceInchTOmillinewtonMeter}
\title{Convert ounce -force inch  to millinewton meter }
\usage{NISTounceForceInchTOmillinewtonMeter(ounceForceInch)}
\description{\code{NISTounceForceInchTOmillinewtonMeter} converts from ounce (avoirdupois)-force inch (ozf * in) to millinewton meter (mN * m) }
\arguments{
  \item{ounceForceInch}{ounce (avoirdupois)-force inch (ozf * in) }
}
\value{millinewton meter (mN * m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTounceForceInchTOmillinewtonMeter(10)
}
\keyword{programming}