\name{NISTounceForceInchTOnewtonMeter}
\alias{NISTounceForceInchTOnewtonMeter}
\title{Convert ounce -force inch  to newton meter }
\usage{NISTounceForceInchTOnewtonMeter(ounceForceInch)}
\description{\code{NISTounceForceInchTOnewtonMeter} converts from ounce (avoirdupois)-force inch (ozf * in) to newton meter (N * m) }
\arguments{
  \item{ounceForceInch}{ounce (avoirdupois)-force inch (ozf * in) }
}
\value{newton meter (N * m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTounceForceInchTOnewtonMeter(10)
}
\keyword{programming}