\name{NISTounceForceTOnewton}
\alias{NISTounceForceTOnewton}
\title{Convert ounce -force  to newton }
\usage{NISTounceForceTOnewton(ounceForce)}
\description{\code{NISTounceForceTOnewton} converts from ounce (avoirdupois)-force (ozf) to newton (N) }
\arguments{
  \item{ounceForce}{ounce (avoirdupois)-force (ozf) }
}
\value{newton (N) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTounceForceTOnewton(10)
}
\keyword{programming}