\name{NISTounceImperialPerGallonTOgramPerLiter}
\alias{NISTounceImperialPerGallonTOgramPerLiter}
\title{Convert ounce per gallon  to gram per liter }
\usage{NISTounceImperialPerGallonTOgramPerLiter(ounceImperialPerGallon)}
\description{\code{NISTounceImperialPerGallonTOgramPerLiter} converts from ounce (avoirdupois) per gallon [Canadian and U.K. (Imperial)] (oz/gal) to gram per liter (g/L) }
\arguments{
  \item{ounceImperialPerGallon}{ounce (avoirdupois) per gallon [Canadian and U.K. (Imperial)] (oz/gal) }
}
\value{gram per liter (g/L) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTounceImperialPerGallonTOgramPerLiter(10)
}
\keyword{programming}