\name{NISTounceImperialTOcubMeter}
\alias{NISTounceImperialTOcubMeter}
\title{Convert ounce  to cubic meter }
\usage{NISTounceImperialTOcubMeter(ounceImperial)}
\description{\code{NISTounceImperialTOcubMeter} converts from ounce [Canadian and U.K. fluid (Imperial)] (fl oz) to cubic meter (m3) }
\arguments{
  \item{ounceImperial}{ounce [Canadian and U.K. fluid (Imperial)] (fl oz) }
}
\value{cubic meter (m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTounceImperialTOcubMeter(10)
}
\keyword{programming}