\name{NISTounceImperialTOml}
\alias{NISTounceImperialTOml}
\title{Convert ounce  to milliliter }
\usage{NISTounceImperialTOml(ounceImperial)}
\description{\code{NISTounceImperialTOml} converts from ounce [Canadian and U.K. fluid (Imperial)] (fl oz) to milliliter (mL) }
\arguments{
  \item{ounceImperial}{ounce [Canadian and U.K. fluid (Imperial)] (fl oz) }
}
\value{milliliter (mL) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTounceImperialTOml(10)
}
\keyword{programming}