\name{NISTouncePerCubInchTOkgPerCubMeter}
\alias{NISTouncePerCubInchTOkgPerCubMeter}
\title{Convert ounce per cubic inch  to kilogram per cubic meter }
\usage{NISTouncePerCubInchTOkgPerCubMeter(ouncePerCubInch)}
\description{\code{NISTouncePerCubInchTOkgPerCubMeter} converts from ounce (avoirdupois) per cubic inch (oz/in3) to kilogram per cubic meter (kg/m3) }
\arguments{
  \item{ouncePerCubInch}{ounce (avoirdupois) per cubic inch (oz/in3) }
}
\value{kilogram per cubic meter (kg/m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTouncePerCubInchTOkgPerCubMeter(10)
}
\keyword{programming}