\name{NISTouncePerGallonTOgramPerLiter}
\alias{NISTouncePerGallonTOgramPerLiter}
\title{Convert ounce per gallon  to gram per liter }
\usage{NISTouncePerGallonTOgramPerLiter(ouncePerGallon)}
\description{\code{NISTouncePerGallonTOgramPerLiter} converts from ounce (avoirdupois) per gallon (U.S.) (oz/gal) to gram per liter (g/L) }
\arguments{
  \item{ouncePerGallon}{ounce (avoirdupois) per gallon (U.S.) (oz/gal) }
}
\value{gram per liter (g/L) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTouncePerGallonTOgramPerLiter(10)
}
\keyword{programming}