\name{NISTouncePerGallonTOkgPerCubMeter}
\alias{NISTouncePerGallonTOkgPerCubMeter}
\title{Convert ounce per gallon  to kilogram per cubic meter }
\usage{NISTouncePerGallonTOkgPerCubMeter(ouncePerGallon)}
\description{\code{NISTouncePerGallonTOkgPerCubMeter} converts from ounce (avoirdupois) per gallon (U.S.) (oz/gal) to kilogram per cubic meter (kg/m3) }
\arguments{
  \item{ouncePerGallon}{ounce (avoirdupois) per gallon (U.S.) (oz/gal) }
}
\value{kilogram per cubic meter (kg/m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTouncePerGallonTOkgPerCubMeter(10)
}
\keyword{programming}