\name{NISTouncePerSqrFtTOkgPerSqrMeter}
\alias{NISTouncePerSqrFtTOkgPerSqrMeter}
\title{Convert ounce per square foot  to kilogram per square meter }
\usage{NISTouncePerSqrFtTOkgPerSqrMeter(ouncePerSqrFt)}
\description{\code{NISTouncePerSqrFtTOkgPerSqrMeter} converts from ounce (avoirdupois) per square foot (oz/ft2) to kilogram per square meter (kg/m2) }
\arguments{
  \item{ouncePerSqrFt}{ounce (avoirdupois) per square foot (oz/ft2) }
}
\value{kilogram per square meter (kg/m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTouncePerSqrFtTOkgPerSqrMeter(10)
}
\keyword{programming}