\name{NISTouncePerSqrInchTOkgPerSqrMeter}
\alias{NISTouncePerSqrInchTOkgPerSqrMeter}
\title{Convert ounce per square inch  to kilogram per square meter }
\usage{NISTouncePerSqrInchTOkgPerSqrMeter(ouncePerSqrInch)}
\description{\code{NISTouncePerSqrInchTOkgPerSqrMeter} converts from ounce (avoirdupois) per square inch (oz/in2) to kilogram per square meter (kg/m2) }
\arguments{
  \item{ouncePerSqrInch}{ounce (avoirdupois) per square inch (oz/in2) }
}
\value{kilogram per square meter (kg/m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTouncePerSqrInchTOkgPerSqrMeter(10)
}
\keyword{programming}