\name{NISTouncePerSqrYardTOkgPerSqrMeter}
\alias{NISTouncePerSqrYardTOkgPerSqrMeter}
\title{Convert ounce per square yard  to kilogram per square meter }
\usage{NISTouncePerSqrYardTOkgPerSqrMeter(ouncePerSqrYard)}
\description{\code{NISTouncePerSqrYardTOkgPerSqrMeter} converts from ounce (avoirdupois) per square yard (oz/yd2) to kilogram per square meter (kg/m2) }
\arguments{
  \item{ouncePerSqrYard}{ounce (avoirdupois) per square yard (oz/yd2) }
}
\value{kilogram per square meter (kg/m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTouncePerSqrYardTOkgPerSqrMeter(10)
}
\keyword{programming}