\name{NISTounceTOcubMeter}
\alias{NISTounceTOcubMeter}
\title{Convert ounce  to cubic meter }
\usage{NISTounceTOcubMeter(ounce)}
\description{\code{NISTounceTOcubMeter} converts from ounce (U.S. fluid) (fl oz) to cubic meter (m3) }
\arguments{
  \item{ounce}{ounce (U.S. fluid) (fl oz) }
}
\value{cubic meter (m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTounceTOcubMeter(10)
}
\keyword{programming}