\name{NISTounceTOml}
\alias{NISTounceTOml}
\title{Convert ounce  to milliliter }
\usage{NISTounceTOml(ounce)}
\description{\code{NISTounceTOml} converts from ounce (U.S. fluid) (fl oz) to milliliter (mL) }
\arguments{
  \item{ounce}{ounce (U.S. fluid) (fl oz) }
}
\value{milliliter (mL) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTounceTOml(10)
}
\keyword{programming}