\name{NISTounceTroyTOkg}
\alias{NISTounceTroyTOkg}
\title{Convert ounce  to kilogram }
\usage{NISTounceTroyTOkg(ounceTroy)}
\description{\code{NISTounceTroyTOkg} converts from ounce (troy or apothecary) (oz) to kilogram (kg) }
\arguments{
  \item{ounceTroy}{ounce (troy or apothecary) (oz) }
}
\value{kilogram (kg) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTounceTroyTOkg(10)
}
\keyword{programming}