\name{NISTparsecTOmeter}
\alias{NISTparsecTOmeter}
\title{Convert parsec  to meter }
\usage{NISTparsecTOmeter(parsec)}
\description{\code{NISTparsecTOmeter} converts from parsec (pc) to meter (m) }
\arguments{
  \item{parsec}{parsec (pc) }
}
\value{meter (m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTparsecTOmeter(10)
}
\keyword{programming}