\name{NISTpascalSecTOcentipoise}
\alias{NISTpascalSecTOcentipoise}
\title{Convert pascal second  to centipoise }
\usage{NISTpascalSecTOcentipoise(pascalSec)}
\description{\code{NISTpascalSecTOcentipoise} converts from pascal second (Pa * s) to centipoise (cP) }
\arguments{
  \item{pascalSec}{pascal second (Pa * s) }
}
\value{centipoise (cP) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalSecTOcentipoise(10)
}
\keyword{programming}