\name{NISTpascalSecTOpoise}
\alias{NISTpascalSecTOpoise}
\title{Convert pascal second  to poise }
\usage{NISTpascalSecTOpoise(pascalSec)}
\description{\code{NISTpascalSecTOpoise} converts from pascal second (Pa * s) to poise (P) }
\arguments{
  \item{pascalSec}{pascal second (Pa * s) }
}
\value{poise (P) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalSecTOpoise(10)
}
\keyword{programming}