\name{NISTpascalSecTOpoundalSecPerSqrFt}
\alias{NISTpascalSecTOpoundalSecPerSqrFt}
\title{Convert pascal second  to poundal second per square foot}
\usage{NISTpascalSecTOpoundalSecPerSqrFt(pascalSec)}
\description{\code{NISTpascalSecTOpoundalSecPerSqrFt} converts from pascal second (Pa * s) to poundal second per square foot }
\arguments{
  \item{pascalSec}{pascal second (Pa * s) }
}
\value{poundal second per square foot }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalSecTOpoundalSecPerSqrFt(10)
}
\keyword{programming}