\name{NISTpascalSecTOpoundForceSecPerSqrFt}
\alias{NISTpascalSecTOpoundForceSecPerSqrFt}
\title{Convert pascal second  to pound-force second per square foot }
\usage{NISTpascalSecTOpoundForceSecPerSqrFt(pascalSec)}
\description{\code{NISTpascalSecTOpoundForceSecPerSqrFt} converts from pascal second (Pa * s) to pound-force second per square foot (lbf * s/ft2) }
\arguments{
  \item{pascalSec}{pascal second (Pa * s) }
}
\value{pound-force second per square foot (lbf * s/ft2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalSecTOpoundForceSecPerSqrFt(10)
}
\keyword{programming}