\name{NISTpascalSecTOpoundPerFtHour}
\alias{NISTpascalSecTOpoundPerFtHour}
\title{Convert pascal second  to pound per foot hour }
\usage{NISTpascalSecTOpoundPerFtHour(pascalSec)}
\description{\code{NISTpascalSecTOpoundPerFtHour} converts from pascal second (Pa * s) to pound per foot hour [lb/(ft * h)] }
\arguments{
  \item{pascalSec}{pascal second (Pa * s) }
}
\value{pound per foot hour [lb/(ft * h)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalSecTOpoundPerFtHour(10)
}
\keyword{programming}