\name{NISTpascalSecTOpoundPerFtSec}
\alias{NISTpascalSecTOpoundPerFtSec}
\title{Convert pascal second  to pound per foot second }
\usage{NISTpascalSecTOpoundPerFtSec(pascalSec)}
\description{\code{NISTpascalSecTOpoundPerFtSec} converts from pascal second (Pa * s) to pound per foot second [lb/(ft * s)] }
\arguments{
  \item{pascalSec}{pascal second (Pa * s) }
}
\value{pound per foot second [lb/(ft * s)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalSecTOpoundPerFtSec(10)
}
\keyword{programming}