\name{NISTpascalSecTOslugPerFtSec}
\alias{NISTpascalSecTOslugPerFtSec}
\title{Convert pascal second  to slug per foot second }
\usage{NISTpascalSecTOslugPerFtSec(pascalSec)}
\description{\code{NISTpascalSecTOslugPerFtSec} converts from pascal second (Pa * s) to slug per foot second [slug/(ft * s)] }
\arguments{
  \item{pascalSec}{pascal second (Pa * s) }
}
\value{slug per foot second [slug/(ft * s)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalSecTOslugPerFtSec(10)
}
\keyword{programming}