\name{NISTpascalTOatmosphereStd}
\alias{NISTpascalTOatmosphereStd}
\title{Convert pascal  to atmosphere, standard }
\usage{NISTpascalTOatmosphereStd(pascal)}
\description{\code{NISTpascalTOatmosphereStd} converts from pascal (Pa) to atmosphere, standard (atm) }
\arguments{
  \item{pascal}{pascal (Pa) }
}
\value{atmosphere, standard (atm) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalTOatmosphereStd(10)
}
\keyword{programming}