\name{NISTpascalTOatmosphereTechnical}
\alias{NISTpascalTOatmosphereTechnical}
\title{Convert pascal  to atmosphere, technical 8}
\usage{NISTpascalTOatmosphereTechnical(pascal)}
\description{\code{NISTpascalTOatmosphereTechnical} converts from pascal (Pa) to atmosphere, technical (at) 8 }
\arguments{
  \item{pascal}{pascal (Pa) }
}
\value{atmosphere, technical (at) 8 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalTOatmosphereTechnical(10)
}
\keyword{programming}