\name{NISTpascalTObar}
\alias{NISTpascalTObar}
\title{Convert pascal  to bar }
\usage{NISTpascalTObar(pascal)}
\description{\code{NISTpascalTObar} converts from pascal (Pa) to bar (bar) }
\arguments{
  \item{pascal}{pascal (Pa) }
}
\value{bar (bar) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalTObar(10)
}
\keyword{programming}