\name{NISTpascalTOcmOfMercury}
\alias{NISTpascalTOcmOfMercury}
\title{Convert pascal  to centimeter of mercury 12}
\usage{NISTpascalTOcmOfMercury(pascal)}
\description{\code{NISTpascalTOcmOfMercury} converts from pascal (Pa) to centimeter of mercury (0 C) 12 }
\arguments{
  \item{pascal}{pascal (Pa) }
}
\value{centimeter of mercury (0 C) 12 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalTOcmOfMercury(10)
}
\keyword{programming}