\name{NISTpascalTOcmOfMercuryConvtnl}
\alias{NISTpascalTOcmOfMercuryConvtnl}
\title{Convert pascal  to centimeter of mercury, conventional 12}
\usage{NISTpascalTOcmOfMercuryConvtnl(pascal)}
\description{\code{NISTpascalTOcmOfMercuryConvtnl} converts from pascal (Pa) to centimeter of mercury, conventional (cmHg) 12 }
\arguments{
  \item{pascal}{pascal (Pa) }
}
\value{centimeter of mercury, conventional (cmHg) 12 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalTOcmOfMercuryConvtnl(10)
}
\keyword{programming}