\name{NISTpascalTOcmOfWater}
\alias{NISTpascalTOcmOfWater}
\title{Convert pascal  to centimeter of water 12}
\usage{NISTpascalTOcmOfWater(pascal)}
\description{\code{NISTpascalTOcmOfWater} converts from pascal (Pa) to centimeter of water (4 C) 12 }
\arguments{
  \item{pascal}{pascal (Pa) }
}
\value{centimeter of water (4 C) 12 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalTOcmOfWater(10)
}
\keyword{programming}