\name{NISTpascalTOcmOfWaterConvtnl}
\alias{NISTpascalTOcmOfWaterConvtnl}
\title{Convert pascal  to centimeter of water, conventional 12}
\usage{NISTpascalTOcmOfWaterConvtnl(pascal)}
\description{\code{NISTpascalTOcmOfWaterConvtnl} converts from pascal (Pa) to centimeter of water, conventional (cmH2O) 12 }
\arguments{
  \item{pascal}{pascal (Pa) }
}
\value{centimeter of water, conventional (cmH2O) 12 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalTOcmOfWaterConvtnl(10)
}
\keyword{programming}