\name{NISTpascalTOdynePerSqrCm}
\alias{NISTpascalTOdynePerSqrCm}
\title{Convert pascal  to dyne per square centimeter }
\usage{NISTpascalTOdynePerSqrCm(pascal)}
\description{\code{NISTpascalTOdynePerSqrCm} converts from pascal (Pa) to dyne per square centimeter (dyn/cm2) }
\arguments{
  \item{pascal}{pascal (Pa) }
}
\value{dyne per square centimeter (dyn/cm2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalTOdynePerSqrCm(10)
}
\keyword{programming}