\name{NISTpascalTOftOfMercuryConvtnl}
\alias{NISTpascalTOftOfMercuryConvtnl}
\title{Convert pascal  to foot of mercury, conventional 12}
\usage{NISTpascalTOftOfMercuryConvtnl(pascal)}
\description{\code{NISTpascalTOftOfMercuryConvtnl} converts from pascal (Pa) to foot of mercury, conventional (ftHg) 12 }
\arguments{
  \item{pascal}{pascal (Pa) }
}
\value{foot of mercury, conventional (ftHg) 12 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalTOftOfMercuryConvtnl(10)
}
\keyword{programming}