\name{NISTpascalTOftOfWater}
\alias{NISTpascalTOftOfWater}
\title{Convert pascal  to foot of water 12}
\usage{NISTpascalTOftOfWater(pascal)}
\description{\code{NISTpascalTOftOfWater} converts from pascal (Pa) to foot of water (39.2 F) 12 }
\arguments{
  \item{pascal}{pascal (Pa) }
}
\value{foot of water (39.2 F) 12 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalTOftOfWater(10)
}
\keyword{programming}