\name{NISTpascalTOftOfWaterConvtnl}
\alias{NISTpascalTOftOfWaterConvtnl}
\title{Convert pascal  to foot of water, conventional 12}
\usage{NISTpascalTOftOfWaterConvtnl(pascal)}
\description{\code{NISTpascalTOftOfWaterConvtnl} converts from pascal (Pa) to foot of water, conventional (ftH2O) 12 }
\arguments{
  \item{pascal}{pascal (Pa) }
}
\value{foot of water, conventional (ftH2O) 12 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalTOftOfWaterConvtnl(10)
}
\keyword{programming}