\name{NISTpascalTOgramForcePerSqrCm}
\alias{NISTpascalTOgramForcePerSqrCm}
\title{Convert pascal  to gram-force per square centimeter }
\usage{NISTpascalTOgramForcePerSqrCm(pascal)}
\description{\code{NISTpascalTOgramForcePerSqrCm} converts from pascal (Pa) to gram-force per square centimeter (gf/cm2) }
\arguments{
  \item{pascal}{pascal (Pa) }
}
\value{gram-force per square centimeter (gf/cm2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalTOgramForcePerSqrCm(10)
}
\keyword{programming}