\name{NISTpascalTOinchOfMercury60F}
\alias{NISTpascalTOinchOfMercury60F}
\title{Convert pascal  to inch of mercury 12}
\usage{NISTpascalTOinchOfMercury60F(pascal)}
\description{\code{NISTpascalTOinchOfMercury60F} converts from pascal (Pa) to inch of mercury (60 F) 12 }
\arguments{
  \item{pascal}{pascal (Pa) }
}
\value{inch of mercury (60 F) 12 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalTOinchOfMercury60F(10)
}
\keyword{programming}