\name{NISTpascalTOinchOfMercuryConvtnl}
\alias{NISTpascalTOinchOfMercuryConvtnl}
\title{Convert pascal  to inch of mercury, conventional 12}
\usage{NISTpascalTOinchOfMercuryConvtnl(pascal)}
\description{\code{NISTpascalTOinchOfMercuryConvtnl} converts from pascal (Pa) to inch of mercury, conventional (inHg) 12 }
\arguments{
  \item{pascal}{pascal (Pa) }
}
\value{inch of mercury, conventional (inHg) 12 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalTOinchOfMercuryConvtnl(10)
}
\keyword{programming}