\name{NISTpascalTOinchOfWater39.2F}
\alias{NISTpascalTOinchOfWater39.2F}
\title{Convert pascal  to inch of water 12}
\usage{NISTpascalTOinchOfWater39.2F(pascal)}
\description{\code{NISTpascalTOinchOfWater39.2F} converts from pascal (Pa) to inch of water (39.2 F) 12 }
\arguments{
  \item{pascal}{pascal (Pa) }
}
\value{inch of water (39.2 F) 12 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalTOinchOfWater39.2F(10)
}
\keyword{programming}