\name{NISTpascalTOinchOfWaterConvtnl}
\alias{NISTpascalTOinchOfWaterConvtnl}
\title{Convert pascal  to inch of water, conventional 12}
\usage{NISTpascalTOinchOfWaterConvtnl(pascal)}
\description{\code{NISTpascalTOinchOfWaterConvtnl} converts from pascal (Pa) to inch of water, conventional (inH2O) 12 }
\arguments{
  \item{pascal}{pascal (Pa) }
}
\value{inch of water, conventional (inH2O) 12 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalTOinchOfWaterConvtnl(10)
}
\keyword{programming}