\name{NISTpascalTOkgForcePerSqrCm}
\alias{NISTpascalTOkgForcePerSqrCm}
\title{Convert pascal  to kilogram-force per square centimeter }
\usage{NISTpascalTOkgForcePerSqrCm(pascal)}
\description{\code{NISTpascalTOkgForcePerSqrCm} converts from pascal (Pa) to kilogram-force per square centimeter (kgf/cm2) }
\arguments{
  \item{pascal}{pascal (Pa) }
}
\value{kilogram-force per square centimeter (kgf/cm2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalTOkgForcePerSqrCm(10)
}
\keyword{programming}