\name{NISTpascalTOkgForcePerSqrMeter}
\alias{NISTpascalTOkgForcePerSqrMeter}
\title{Convert pascal  to kilogram-force per square meter }
\usage{NISTpascalTOkgForcePerSqrMeter(pascal)}
\description{\code{NISTpascalTOkgForcePerSqrMeter} converts from pascal (Pa) to kilogram-force per square meter (kgf/m2) }
\arguments{
  \item{pascal}{pascal (Pa) }
}
\value{kilogram-force per square meter (kgf/m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalTOkgForcePerSqrMeter(10)
}
\keyword{programming}