\name{NISTpascalTOkgForcePerSqrMillimeter}
\alias{NISTpascalTOkgForcePerSqrMillimeter}
\title{Convert pascal  to kilogram-force per square millimeter }
\usage{NISTpascalTOkgForcePerSqrMillimeter(pascal)}
\description{\code{NISTpascalTOkgForcePerSqrMillimeter} converts from pascal (Pa) to kilogram-force per square millimeter (kgf/mm2) }
\arguments{
  \item{pascal}{pascal (Pa) }
}
\value{kilogram-force per square millimeter (kgf/mm2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalTOkgForcePerSqrMillimeter(10)
}
\keyword{programming}