\name{NISTpascalTOkipPerSqrInch}
\alias{NISTpascalTOkipPerSqrInch}
\title{Convert pascal  to kip per square inch }
\usage{NISTpascalTOkipPerSqrInch(pascal)}
\description{\code{NISTpascalTOkipPerSqrInch} converts from pascal (Pa) to kip per square inch (ksi) (kip/in2) }
\arguments{
  \item{pascal}{pascal (Pa) }
}
\value{kip per square inch (ksi) (kip/in2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalTOkipPerSqrInch(10)
}
\keyword{programming}