\name{NISTpascalTOmillibar}
\alias{NISTpascalTOmillibar}
\title{Convert pascal  to millibar }
\usage{NISTpascalTOmillibar(pascal)}
\description{\code{NISTpascalTOmillibar} converts from pascal (Pa) to millibar (mbar) }
\arguments{
  \item{pascal}{pascal (Pa) }
}
\value{millibar (mbar) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalTOmillibar(10)
}
\keyword{programming}