\name{NISTpascalTOmillimeterOfMercuryConvtnl}
\alias{NISTpascalTOmillimeterOfMercuryConvtnl}
\title{Convert pascal  to millimeter of mercury, conventional 12}
\usage{NISTpascalTOmillimeterOfMercuryConvtnl(pascal)}
\description{\code{NISTpascalTOmillimeterOfMercuryConvtnl} converts from pascal (Pa) to millimeter of mercury, conventional (mmHg) 12 }
\arguments{
  \item{pascal}{pascal (Pa) }
}
\value{millimeter of mercury, conventional (mmHg) 12 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalTOmillimeterOfMercuryConvtnl(10)
}
\keyword{programming}