\name{NISTpascalTOmillimeterOfWaterConvtnl}
\alias{NISTpascalTOmillimeterOfWaterConvtnl}
\title{Convert pascal  to millimeter of water, conventional 12}
\usage{NISTpascalTOmillimeterOfWaterConvtnl(pascal)}
\description{\code{NISTpascalTOmillimeterOfWaterConvtnl} converts from pascal (Pa) to millimeter of water, conventional (mm H2O) 12 }
\arguments{
  \item{pascal}{pascal (Pa) }
}
\value{millimeter of water, conventional (mm H2O) 12 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalTOmillimeterOfWaterConvtnl(10)
}
\keyword{programming}