\name{NISTpascalTOpoundalPerSqrFt}
\alias{NISTpascalTOpoundalPerSqrFt}
\title{Convert pascal  to poundal per square foot}
\usage{NISTpascalTOpoundalPerSqrFt(pascal)}
\description{\code{NISTpascalTOpoundalPerSqrFt} converts from pascal (Pa) to poundal per square foot }
\arguments{
  \item{pascal}{pascal (Pa) }
}
\value{poundal per square foot }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalTOpoundalPerSqrFt(10)
}
\keyword{programming}