\name{NISTpascalTOpoundForcePerSqrFt}
\alias{NISTpascalTOpoundForcePerSqrFt}
\title{Convert pascal  to pound-force per square foot }
\usage{NISTpascalTOpoundForcePerSqrFt(pascal)}
\description{\code{NISTpascalTOpoundForcePerSqrFt} converts from pascal (Pa) to pound-force per square foot (lbf/ft2) }
\arguments{
  \item{pascal}{pascal (Pa) }
}
\value{pound-force per square foot (lbf/ft2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalTOpoundForcePerSqrFt(10)
}
\keyword{programming}