\name{NISTpascalTOpoundForcePerSqrInch}
\alias{NISTpascalTOpoundForcePerSqrInch}
\title{Convert pascal  to pound-force per square inch }
\usage{NISTpascalTOpoundForcePerSqrInch(pascal)}
\description{\code{NISTpascalTOpoundForcePerSqrInch} converts from pascal (Pa) to pound-force per square inch (psi) (lbf/in2) }
\arguments{
  \item{pascal}{pascal (Pa) }
}
\value{pound-force per square inch (psi) (lbf/in2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalTOpoundForcePerSqrInch(10)
}
\keyword{programming}