\name{NISTpascalTOpsi}
\alias{NISTpascalTOpsi}
\title{Convert pascal  to psi }
\usage{NISTpascalTOpsi(pascal)}
\description{\code{NISTpascalTOpsi} converts from pascal (Pa) to psi (pound-force per square inch) (lbf/in2) }
\arguments{
  \item{pascal}{pascal (Pa) }
}
\value{psi (pound-force per square inch) (lbf/in2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalTOpsi(10)
}
\keyword{programming}