\name{NISTpascalTOtorr}
\alias{NISTpascalTOtorr}
\title{Convert pascal  to torr }
\usage{NISTpascalTOtorr(pascal)}
\description{\code{NISTpascalTOtorr} converts from pascal (Pa) to torr (Torr) }
\arguments{
  \item{pascal}{pascal (Pa) }
}
\value{torr (Torr) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpascalTOtorr(10)
}
\keyword{programming}