\name{NISTpeckTOcubMeter}
\alias{NISTpeckTOcubMeter}
\title{Convert peck  to cubic meter }
\usage{NISTpeckTOcubMeter(peck)}
\description{\code{NISTpeckTOcubMeter} converts from peck (U.S.) (pk) to cubic meter (m3) }
\arguments{
  \item{peck}{peck (U.S.) (pk) }
}
\value{cubic meter (m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpeckTOcubMeter(10)
}
\keyword{programming}