\name{NISTpeckTOliter}
\alias{NISTpeckTOliter}
\title{Convert peck  to liter }
\usage{NISTpeckTOliter(peck)}
\description{\code{NISTpeckTOliter} converts from peck (U.S.) (pk) to liter (L) }
\arguments{
  \item{peck}{peck (U.S.) (pk) }
}
\value{liter (L) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpeckTOliter(10)
}
\keyword{programming}