\name{NISTpennyweightTOgram}
\alias{NISTpennyweightTOgram}
\title{Convert pennyweight  to gram }
\usage{NISTpennyweightTOgram(pennyweight)}
\description{\code{NISTpennyweightTOgram} converts from pennyweight (dwt) to gram (g) }
\arguments{
  \item{pennyweight}{pennyweight (dwt) }
}
\value{gram (g) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpennyweightTOgram(10)
}
\keyword{programming}