\name{NISTpennyweightTOkg}
\alias{NISTpennyweightTOkg}
\title{Convert pennyweight  to kilogram }
\usage{NISTpennyweightTOkg(pennyweight)}
\description{\code{NISTpennyweightTOkg} converts from pennyweight (dwt) to kilogram (kg) }
\arguments{
  \item{pennyweight}{pennyweight (dwt) }
}
\value{kilogram (kg) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpennyweightTOkg(10)
}
\keyword{programming}