\name{NISTperm0CtOkgPerPascalSecSqrMeter}
\alias{NISTperm0CtOkgPerPascalSecSqrMeter}
\title{Convert perm  to kilogram per pascal second square meter }
\usage{NISTperm0CtOkgPerPascalSecSqrMeter(perm0C)}
\description{\code{NISTperm0CtOkgPerPascalSecSqrMeter} converts from perm (0 C) to kilogram per pascal second square meter [kg/(Pa * s * m2)] }
\arguments{
  \item{perm0C}{perm (0 C) }
}
\value{kilogram per pascal second square meter [kg/(Pa * s * m2)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTperm0CtOkgPerPascalSecSqrMeter(10)
}
\keyword{programming}