\name{NISTperm23CtOkgPerPascalSecSqrMeter}
\alias{NISTperm23CtOkgPerPascalSecSqrMeter}
\title{Convert perm  to kilogram per pascal second square meter }
\usage{NISTperm23CtOkgPerPascalSecSqrMeter(perm23C)}
\description{\code{NISTperm23CtOkgPerPascalSecSqrMeter} converts from perm (23 C) to kilogram per pascal second square meter [kg/(Pa * s * m2)] }
\arguments{
  \item{perm23C}{perm (23 C) }
}
\value{kilogram per pascal second square meter [kg/(Pa * s * m2)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTperm23CtOkgPerPascalSecSqrMeter(10)
}
\keyword{programming}