\name{NISTpermInch0CtOkgPerPascalSecMeter}
\alias{NISTpermInch0CtOkgPerPascalSecMeter}
\title{Convert perm inch  to kilogram per pascal second meter }
\usage{NISTpermInch0CtOkgPerPascalSecMeter(permInch0C)}
\description{\code{NISTpermInch0CtOkgPerPascalSecMeter} converts from perm inch (0 C) to kilogram per pascal second meter [kg/(Pa*s*m)] }
\arguments{
  \item{permInch0C}{perm inch (0 C) }
}
\value{kilogram per pascal second meter [kg/(Pa*s*m)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpermInch0CtOkgPerPascalSecMeter(10)
}
\keyword{programming}