\name{NISTpermInch23CtOkgPerPascalSecMeter}
\alias{NISTpermInch23CtOkgPerPascalSecMeter}
\title{Convert perm inch  to kilogram per pascal second meter }
\usage{NISTpermInch23CtOkgPerPascalSecMeter(permInch23C)}
\description{\code{NISTpermInch23CtOkgPerPascalSecMeter} converts from perm inch (23 C) to kilogram per pascal second meter [kg/(Pa*s*m)] }
\arguments{
  \item{permInch23C}{perm inch (23 C) }
}
\value{kilogram per pascal second meter [kg/(Pa*s*m)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpermInch23CtOkgPerPascalSecMeter(10)
}
\keyword{programming}