\name{NISTphotTOlux}
\alias{NISTphotTOlux}
\title{Convert phot  to lux }
\usage{NISTphotTOlux(phot)}
\description{\code{NISTphotTOlux} converts from phot (ph) to lux (lx) }
\arguments{
  \item{phot}{phot (ph) }
}
\value{lux (lx) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTphotTOlux(10)
}
\keyword{programming}