\name{NISTpicaComputerTOmeter}
\alias{NISTpicaComputerTOmeter}
\title{Convert pica  to meter }
\usage{NISTpicaComputerTOmeter(picaComputer)}
\description{\code{NISTpicaComputerTOmeter} converts from pica (computer) (1/6 in) to meter (m) }
\arguments{
  \item{picaComputer}{pica (computer) (1/6 in) }
}
\value{meter (m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpicaComputerTOmeter(10)
}
\keyword{programming}