\name{NISTpicaComputerTOmillimeter}
\alias{NISTpicaComputerTOmillimeter}
\title{Convert pica  to millimeter }
\usage{NISTpicaComputerTOmillimeter(picaComputer)}
\description{\code{NISTpicaComputerTOmillimeter} converts from pica (computer) (1/6 in) to millimeter (mm) }
\arguments{
  \item{picaComputer}{pica (computer) (1/6 in) }
}
\value{millimeter (mm) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpicaComputerTOmillimeter(10)
}
\keyword{programming}