\name{NISTpicaPrinterTOmeter}
\alias{NISTpicaPrinterTOmeter}
\title{Convert pica  to meter }
\usage{NISTpicaPrinterTOmeter(picaPrinter)}
\description{\code{NISTpicaPrinterTOmeter} converts from pica (printer's) to meter (m) }
\arguments{
  \item{picaPrinter}{pica (printer's) }
}
\value{meter (m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpicaPrinterTOmeter(10)
}
\keyword{programming}