\name{NISTpicaPrinterTOmillimeter}
\alias{NISTpicaPrinterTOmillimeter}
\title{Convert pica  to millimeter }
\usage{NISTpicaPrinterTOmillimeter(picaPrinter)}
\description{\code{NISTpicaPrinterTOmillimeter} converts from pica (printer's) to millimeter (mm) }
\arguments{
  \item{picaPrinter}{pica (printer's) }
}
\value{millimeter (mm) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpicaPrinterTOmillimeter(10)
}
\keyword{programming}