\name{NISTpintUSdryTOcubMeter}
\alias{NISTpintUSdryTOcubMeter}
\title{Convert pint  to cubic meter }
\usage{NISTpintUSdryTOcubMeter(pintUSdry)}
\description{\code{NISTpintUSdryTOcubMeter} converts from pint (U.S. dry) (dry pt) to cubic meter (m3) }
\arguments{
  \item{pintUSdry}{pint (U.S. dry) (dry pt) }
}
\value{cubic meter (m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpintUSdryTOcubMeter(10)
}
\keyword{programming}