\name{NISTpintUSdryTOliter}
\alias{NISTpintUSdryTOliter}
\title{Convert pint  to liter }
\usage{NISTpintUSdryTOliter(pintUSdry)}
\description{\code{NISTpintUSdryTOliter} converts from pint (U.S. dry) (dry pt) to liter (L) }
\arguments{
  \item{pintUSdry}{pint (U.S. dry) (dry pt) }
}
\value{liter (L) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpintUSdryTOliter(10)
}
\keyword{programming}