\name{NISTpintUSliquidTOcubMeter}
\alias{NISTpintUSliquidTOcubMeter}
\title{Convert pint  to cubic meter }
\usage{NISTpintUSliquidTOcubMeter(pintUSliquid)}
\description{\code{NISTpintUSliquidTOcubMeter} converts from pint (U.S. liquid) (liq pt) to cubic meter (m3) }
\arguments{
  \item{pintUSliquid}{pint (U.S. liquid) (liq pt) }
}
\value{cubic meter (m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpintUSliquidTOcubMeter(10)
}
\keyword{programming}