\name{NISTpintUSliquidTOliter}
\alias{NISTpintUSliquidTOliter}
\title{Convert pint  to liter }
\usage{NISTpintUSliquidTOliter(pintUSliquid)}
\description{\code{NISTpintUSliquidTOliter} converts from pint (U.S. liquid) (liq pt) to liter (L) }
\arguments{
  \item{pintUSliquid}{pint (U.S. liquid) (liq pt) }
}
\value{liter (L) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpintUSliquidTOliter(10)
}
\keyword{programming}