\name{NISTpointComputerTOmeter}
\alias{NISTpointComputerTOmeter}
\title{Convert point  to meter }
\usage{NISTpointComputerTOmeter(pointComputer)}
\description{\code{NISTpointComputerTOmeter} converts from point (computer) (1/72 in) to meter (m) }
\arguments{
  \item{pointComputer}{point (computer) (1/72 in) }
}
\value{meter (m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpointComputerTOmeter(10)
}
\keyword{programming}