\name{NISTpointComputerTOmillimeter}
\alias{NISTpointComputerTOmillimeter}
\title{Convert point  to millimeter }
\usage{NISTpointComputerTOmillimeter(pointComputer)}
\description{\code{NISTpointComputerTOmillimeter} converts from point (computer) (1/72 in) to millimeter (mm) }
\arguments{
  \item{pointComputer}{point (computer) (1/72 in) }
}
\value{millimeter (mm) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpointComputerTOmillimeter(10)
}
\keyword{programming}