\name{NISTpointPrinterTOmeter}
\alias{NISTpointPrinterTOmeter}
\title{Convert point  to meter }
\usage{NISTpointPrinterTOmeter(pointPrinter)}
\description{\code{NISTpointPrinterTOmeter} converts from point (printer's) to meter (m) }
\arguments{
  \item{pointPrinter}{point (printer's) }
}
\value{meter (m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpointPrinterTOmeter(10)
}
\keyword{programming}