\name{NISTpointPrinterTOmillimeter}
\alias{NISTpointPrinterTOmillimeter}
\title{Convert point  to millimeter }
\usage{NISTpointPrinterTOmillimeter(pointPrinter)}
\description{\code{NISTpointPrinterTOmillimeter} converts from point (printer's) to millimeter (mm) }
\arguments{
  \item{pointPrinter}{point (printer's) }
}
\value{millimeter (mm) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpointPrinterTOmillimeter(10)
}
\keyword{programming}