\name{NISTpoiseTOpascalSec}
\alias{NISTpoiseTOpascalSec}
\title{Convert poise  to pascal second }
\usage{NISTpoiseTOpascalSec(poise)}
\description{\code{NISTpoiseTOpascalSec} converts from poise (P) to pascal second (Pa * s) }
\arguments{
  \item{poise}{poise (P) }
}
\value{pascal second (Pa * s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoiseTOpascalSec(10)
}
\keyword{programming}