\name{NISTpoundalPerSqrFtTOpascal}
\alias{NISTpoundalPerSqrFtTOpascal}
\title{Convert poundal per square foot to pascal }
\usage{NISTpoundalPerSqrFtTOpascal(poundalPerSqrFt)}
\description{\code{NISTpoundalPerSqrFtTOpascal} converts from poundal per square foot to pascal (Pa) }
\arguments{
  \item{poundalPerSqrFt}{poundal per square foot }
}
\value{pascal (Pa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundalPerSqrFtTOpascal(10)
}
\keyword{programming}