\name{NISTpoundalSecPerSqrFtTOpascalSec}
\alias{NISTpoundalSecPerSqrFtTOpascalSec}
\title{Convert poundal second per square foot to pascal second }
\usage{NISTpoundalSecPerSqrFtTOpascalSec(poundalSecPerSqrFt)}
\description{\code{NISTpoundalSecPerSqrFtTOpascalSec} converts from poundal second per square foot to pascal second (Pa * s) }
\arguments{
  \item{poundalSecPerSqrFt}{poundal second per square foot }
}
\value{pascal second (Pa * s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundalSecPerSqrFtTOpascalSec(10)
}
\keyword{programming}