\name{NISTpoundAvoirdupoisTOkg}
\alias{NISTpoundAvoirdupoisTOkg}
\title{Convert pound 22 to kilogram }
\usage{NISTpoundAvoirdupoisTOkg(poundAvoirdupois)}
\description{\code{NISTpoundAvoirdupoisTOkg} converts from pound (avoirdupois) (lb) 22 to kilogram (kg) }
\arguments{
  \item{poundAvoirdupois}{pound (avoirdupois) (lb) 22 }
}
\value{kilogram (kg) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundAvoirdupoisTOkg(10)
}
\keyword{programming}