\name{NISTpoundForceFtPerInchTOnewtonMeterPerMeter}
\alias{NISTpoundForceFtPerInchTOnewtonMeterPerMeter}
\title{Convert pound-force foot per inch  to newton meter per meter }
\usage{NISTpoundForceFtPerInchTOnewtonMeterPerMeter(poundForceFtPerInch)}
\description{\code{NISTpoundForceFtPerInchTOnewtonMeterPerMeter} converts from pound-force foot per inch (lbf * ft/in) to newton meter per meter (N * m/m) }
\arguments{
  \item{poundForceFtPerInch}{pound-force foot per inch (lbf * ft/in) }
}
\value{newton meter per meter (N * m/m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundForceFtPerInchTOnewtonMeterPerMeter(10)
}
\keyword{programming}