\name{NISTpoundForceFtTOnewtonMeter}
\alias{NISTpoundForceFtTOnewtonMeter}
\title{Convert pound-force foot  to newton meter }
\usage{NISTpoundForceFtTOnewtonMeter(poundForceFt)}
\description{\code{NISTpoundForceFtTOnewtonMeter} converts from pound-force foot (lbf * ft) to newton meter (N * m) }
\arguments{
  \item{poundForceFt}{pound-force foot (lbf * ft) }
}
\value{newton meter (N * m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundForceFtTOnewtonMeter(10)
}
\keyword{programming}