\name{NISTpoundForceInchPerInchTOnewtonMeterPerMeter}
\alias{NISTpoundForceInchPerInchTOnewtonMeterPerMeter}
\title{Convert pound-force inch per inch  to newton meter per meter }
\usage{NISTpoundForceInchPerInchTOnewtonMeterPerMeter(poundForceInchPerInch)}
\description{\code{NISTpoundForceInchPerInchTOnewtonMeterPerMeter} converts from pound-force inch per inch (lbf * in/in) to newton meter per meter (N * m/m) }
\arguments{
  \item{poundForceInchPerInch}{pound-force inch per inch (lbf * in/in) }
}
\value{newton meter per meter (N * m/m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundForceInchPerInchTOnewtonMeterPerMeter(10)
}
\keyword{programming}