\name{NISTpoundForceInchTOnewtonMeter}
\alias{NISTpoundForceInchTOnewtonMeter}
\title{Convert pound-force inch  to newton meter }
\usage{NISTpoundForceInchTOnewtonMeter(poundForceInch)}
\description{\code{NISTpoundForceInchTOnewtonMeter} converts from pound-force inch (lbf * in) to newton meter (N * m) }
\arguments{
  \item{poundForceInch}{pound-force inch (lbf * in) }
}
\value{newton meter (N * m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundForceInchTOnewtonMeter(10)
}
\keyword{programming}