\name{NISTpoundForcePerFtTOnewtonPerMeter}
\alias{NISTpoundForcePerFtTOnewtonPerMeter}
\title{Convert pound-force per foot  to newton per meter }
\usage{NISTpoundForcePerFtTOnewtonPerMeter(poundForcePerFt)}
\description{\code{NISTpoundForcePerFtTOnewtonPerMeter} converts from pound-force per foot (lbf/ft) to newton per meter (N/m) }
\arguments{
  \item{poundForcePerFt}{pound-force per foot (lbf/ft) }
}
\value{newton per meter (N/m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundForcePerFtTOnewtonPerMeter(10)
}
\keyword{programming}