\name{NISTpoundForcePerInchTOnewtonPerMeter}
\alias{NISTpoundForcePerInchTOnewtonPerMeter}
\title{Convert pound-force per inch  to newton per meter }
\usage{NISTpoundForcePerInchTOnewtonPerMeter(poundForcePerInch)}
\description{\code{NISTpoundForcePerInchTOnewtonPerMeter} converts from pound-force per inch (lbf/in) to newton per meter (N/m) }
\arguments{
  \item{poundForcePerInch}{pound-force per inch (lbf/in) }
}
\value{newton per meter (N/m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundForcePerInchTOnewtonPerMeter(10)
}
\keyword{programming}