\name{NISTpoundForcePerPoundTOnewtonPerKg}
\alias{NISTpoundForcePerPoundTOnewtonPerKg}
\title{Convert pound-force per pound  to newton per kilogram }
\usage{NISTpoundForcePerPoundTOnewtonPerKg(poundForcePerPound)}
\description{\code{NISTpoundForcePerPoundTOnewtonPerKg} converts from pound-force per pound (lbf/lb) (thrust to mass ratio) to newton per kilogram (N/kg) }
\arguments{
  \item{poundForcePerPound}{pound-force per pound (lbf/lb) (thrust to mass ratio) }
}
\value{newton per kilogram (N/kg) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundForcePerPoundTOnewtonPerKg(10)
}
\keyword{programming}