\name{NISTpoundForcePerSqrFtTOpascal}
\alias{NISTpoundForcePerSqrFtTOpascal}
\title{Convert pound-force per square foot  to pascal }
\usage{NISTpoundForcePerSqrFtTOpascal(poundForcePerSqrFt)}
\description{\code{NISTpoundForcePerSqrFtTOpascal} converts from pound-force per square foot (lbf/ft2) to pascal (Pa) }
\arguments{
  \item{poundForcePerSqrFt}{pound-force per square foot (lbf/ft2) }
}
\value{pascal (Pa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundForcePerSqrFtTOpascal(10)
}
\keyword{programming}