\name{NISTpoundForcePerSqrInchTOkpascal}
\alias{NISTpoundForcePerSqrInchTOkpascal}
\title{Convert pound-force per square inch  to kilopascal }
\usage{NISTpoundForcePerSqrInchTOkpascal(poundForcePerSqrInch)}
\description{\code{NISTpoundForcePerSqrInchTOkpascal} converts from pound-force per square inch (psi) (lbf/in2) to kilopascal (kPa) }
\arguments{
  \item{poundForcePerSqrInch}{pound-force per square inch (psi) (lbf/in2) }
}
\value{kilopascal (kPa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundForcePerSqrInchTOkpascal(10)
}
\keyword{programming}