\name{NISTpoundForcePerSqrInchTOpascal}
\alias{NISTpoundForcePerSqrInchTOpascal}
\title{Convert pound-force per square inch  to pascal }
\usage{NISTpoundForcePerSqrInchTOpascal(poundForcePerSqrInch)}
\description{\code{NISTpoundForcePerSqrInchTOpascal} converts from pound-force per square inch (psi) (lbf/in2) to pascal (Pa) }
\arguments{
  \item{poundForcePerSqrInch}{pound-force per square inch (psi) (lbf/in2) }
}
\value{pascal (Pa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundForcePerSqrInchTOpascal(10)
}
\keyword{programming}