\name{NISTpoundForceSecPerSqrFtTOpascalSec}
\alias{NISTpoundForceSecPerSqrFtTOpascalSec}
\title{Convert pound-force second per square foot  to pascal second }
\usage{NISTpoundForceSecPerSqrFtTOpascalSec(poundForceSecPerSqrFt)}
\description{\code{NISTpoundForceSecPerSqrFtTOpascalSec} converts from pound-force second per square foot (lbf * s/ft2) to pascal second (Pa * s) }
\arguments{
  \item{poundForceSecPerSqrFt}{pound-force second per square foot (lbf * s/ft2) }
}
\value{pascal second (Pa * s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundForceSecPerSqrFtTOpascalSec(10)
}
\keyword{programming}