\name{NISTpoundForceSecPerSqrInchTOpascalSec}
\alias{NISTpoundForceSecPerSqrInchTOpascalSec}
\title{Convert pound-force second per square inch  to pascal second }
\usage{NISTpoundForceSecPerSqrInchTOpascalSec(poundForceSecPerSqrInch)}
\description{\code{NISTpoundForceSecPerSqrInchTOpascalSec} converts from pound-force second per square inch (lbf * s/in2) to pascal second (Pa * s) }
\arguments{
  \item{poundForceSecPerSqrInch}{pound-force second per square inch (lbf * s/in2) }
}
\value{pascal second (Pa * s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundForceSecPerSqrInchTOpascalSec(10)
}
\keyword{programming}