\name{NISTpoundForceTOnewton}
\alias{NISTpoundForceTOnewton}
\title{Convert pound-force 23 to newton }
\usage{NISTpoundForceTOnewton(poundForce)}
\description{\code{NISTpoundForceTOnewton} converts from pound-force (lbf) 23 to newton (N) }
\arguments{
  \item{poundForce}{pound-force (lbf) 23 }
}
\value{newton (N) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundForceTOnewton(10)
}
\keyword{programming}