\name{NISTpoundFtSqrdTOkgMeterSqrd}
\alias{NISTpoundFtSqrdTOkgMeterSqrd}
\title{Convert pound foot squared  to kilogram meter squared }
\usage{NISTpoundFtSqrdTOkgMeterSqrd(poundFtSqrd)}
\description{\code{NISTpoundFtSqrdTOkgMeterSqrd} converts from pound foot squared (lb * ft2) to kilogram meter squared (kg * m2) }
\arguments{
  \item{poundFtSqrd}{pound foot squared (lb * ft2) }
}
\value{kilogram meter squared (kg * m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundFtSqrdTOkgMeterSqrd(10)
}
\keyword{programming}