\name{NISTpoundInchSqrdTOkgMeterSqrd}
\alias{NISTpoundInchSqrdTOkgMeterSqrd}
\title{Convert pound inch squared  to kilogram meter squared }
\usage{NISTpoundInchSqrdTOkgMeterSqrd(poundInchSqrd)}
\description{\code{NISTpoundInchSqrdTOkgMeterSqrd} converts from pound inch squared (lb * in2) to kilogram meter squared (kg *m2) }
\arguments{
  \item{poundInchSqrd}{pound inch squared (lb * in2) }
}
\value{kilogram meter squared (kg *m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundInchSqrdTOkgMeterSqrd(10)
}
\keyword{programming}