\name{NISTpoundPerCubFtTOkgPerCubMeter}
\alias{NISTpoundPerCubFtTOkgPerCubMeter}
\title{Convert pound per cubic foot  to kilogram per cubic meter }
\usage{NISTpoundPerCubFtTOkgPerCubMeter(poundPerCubFt)}
\description{\code{NISTpoundPerCubFtTOkgPerCubMeter} converts from pound per cubic foot (lb/ft3) to kilogram per cubic meter (kg/m3) }
\arguments{
  \item{poundPerCubFt}{pound per cubic foot (lb/ft3) }
}
\value{kilogram per cubic meter (kg/m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundPerCubFtTOkgPerCubMeter(10)
}
\keyword{programming}