\name{NISTpoundPerCubYardTOkgPerCubMeter}
\alias{NISTpoundPerCubYardTOkgPerCubMeter}
\title{Convert pound per cubic yard  to kilogram per cubic meter }
\usage{NISTpoundPerCubYardTOkgPerCubMeter(poundPerCubYard)}
\description{\code{NISTpoundPerCubYardTOkgPerCubMeter} converts from pound per cubic yard (lb/yd3) to kilogram per cubic meter (kg/m3) }
\arguments{
  \item{poundPerCubYard}{pound per cubic yard (lb/yd3) }
}
\value{kilogram per cubic meter (kg/m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundPerCubYardTOkgPerCubMeter(10)
}
\keyword{programming}