\name{NISTpoundPerFtHourTOpascalSec}
\alias{NISTpoundPerFtHourTOpascalSec}
\title{Convert pound per foot hour  to pascal second }
\usage{NISTpoundPerFtHourTOpascalSec(poundPerFtHour)}
\description{\code{NISTpoundPerFtHourTOpascalSec} converts from pound per foot hour [lb/(ft * h)] to pascal second (Pa * s) }
\arguments{
  \item{poundPerFtHour}{pound per foot hour [lb/(ft * h)] }
}
\value{pascal second (Pa * s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundPerFtHourTOpascalSec(10)
}
\keyword{programming}