\name{NISTpoundPerFtTOkgPerMeter}
\alias{NISTpoundPerFtTOkgPerMeter}
\title{Convert pound per foot  to kilogram per meter }
\usage{NISTpoundPerFtTOkgPerMeter(poundPerFt)}
\description{\code{NISTpoundPerFtTOkgPerMeter} converts from pound per foot (lb/ft) to kilogram per meter (kg/m) }
\arguments{
  \item{poundPerFt}{pound per foot (lb/ft) }
}
\value{kilogram per meter (kg/m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundPerFtTOkgPerMeter(10)
}
\keyword{programming}