\name{NISTpoundPerGallonImperialTOkgPerCubMeter}
\alias{NISTpoundPerGallonImperialTOkgPerCubMeter}
\title{Convert pound per gallon  to kilogram per cubic meter }
\usage{NISTpoundPerGallonImperialTOkgPerCubMeter(poundPerGallonImperial)}
\description{\code{NISTpoundPerGallonImperialTOkgPerCubMeter} converts from pound per gallon [Canadian and U.K. (Imperial)] (lb/gal) to kilogram per cubic meter (kg/m3) }
\arguments{
  \item{poundPerGallonImperial}{pound per gallon [Canadian and U.K. (Imperial)] (lb/gal) }
}
\value{kilogram per cubic meter (kg/m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundPerGallonImperialTOkgPerCubMeter(10)
}
\keyword{programming}