\name{NISTpoundPerGallonImperialTOkgPerLiter}
\alias{NISTpoundPerGallonImperialTOkgPerLiter}
\title{Convert pound per gallon  to kilogram per liter }
\usage{NISTpoundPerGallonImperialTOkgPerLiter(poundPerGallonImperial)}
\description{\code{NISTpoundPerGallonImperialTOkgPerLiter} converts from pound per gallon [Canadian and U.K. (Imperial)] (lb/gal) to kilogram per liter (kg/L) }
\arguments{
  \item{poundPerGallonImperial}{pound per gallon [Canadian and U.K. (Imperial)] (lb/gal) }
}
\value{kilogram per liter (kg/L) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundPerGallonImperialTOkgPerLiter(10)
}
\keyword{programming}