\name{NISTpoundPerGallonTOkgPerCubMeter}
\alias{NISTpoundPerGallonTOkgPerCubMeter}
\title{Convert pound per gallon  to kilogram per cubic meter }
\usage{NISTpoundPerGallonTOkgPerCubMeter(poundPerGallon)}
\description{\code{NISTpoundPerGallonTOkgPerCubMeter} converts from pound per gallon (U.S.) (lb/gal) to kilogram per cubic meter (kg/m3) }
\arguments{
  \item{poundPerGallon}{pound per gallon (U.S.) (lb/gal) }
}
\value{kilogram per cubic meter (kg/m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundPerGallonTOkgPerCubMeter(10)
}
\keyword{programming}