\name{NISTpoundPerGallonTOkgPerLiter}
\alias{NISTpoundPerGallonTOkgPerLiter}
\title{Convert pound per gallon  to kilogram per liter }
\usage{NISTpoundPerGallonTOkgPerLiter(poundPerGallon)}
\description{\code{NISTpoundPerGallonTOkgPerLiter} converts from pound per gallon (U.S.) (lb/gal) to kilogram per liter (kg/L) }
\arguments{
  \item{poundPerGallon}{pound per gallon (U.S.) (lb/gal) }
}
\value{kilogram per liter (kg/L) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundPerGallonTOkgPerLiter(10)
}
\keyword{programming}