\name{NISTpoundPerHorsepowerHourTOkgPerJoule}
\alias{NISTpoundPerHorsepowerHourTOkgPerJoule}
\title{Convert pound per horsepower hour  to kilogram per joule }
\usage{NISTpoundPerHorsepowerHourTOkgPerJoule(poundPerHorsepowerHour)}
\description{\code{NISTpoundPerHorsepowerHourTOkgPerJoule} converts from pound per horsepower hour [lb/(hp * h)] to kilogram per joule (kg/J) }
\arguments{
  \item{poundPerHorsepowerHour}{pound per horsepower hour [lb/(hp * h)] }
}
\value{kilogram per joule (kg/J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundPerHorsepowerHourTOkgPerJoule(10)
}
\keyword{programming}