\name{NISTpoundPerHourTOkgPerSec}
\alias{NISTpoundPerHourTOkgPerSec}
\title{Convert pound per hour  to kilogram per second }
\usage{NISTpoundPerHourTOkgPerSec(poundPerHour)}
\description{\code{NISTpoundPerHourTOkgPerSec} converts from pound per hour (lb/h) to kilogram per second (kg/s) }
\arguments{
  \item{poundPerHour}{pound per hour (lb/h) }
}
\value{kilogram per second (kg/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundPerHourTOkgPerSec(10)
}
\keyword{programming}