\name{NISTpoundPerInchTOkgPerMeter}
\alias{NISTpoundPerInchTOkgPerMeter}
\title{Convert pound per inch  to kilogram per meter }
\usage{NISTpoundPerInchTOkgPerMeter(poundPerInch)}
\description{\code{NISTpoundPerInchTOkgPerMeter} converts from pound per inch (lb/in) to kilogram per meter (kg/m) }
\arguments{
  \item{poundPerInch}{pound per inch (lb/in) }
}
\value{kilogram per meter (kg/m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundPerInchTOkgPerMeter(10)
}
\keyword{programming}