\name{NISTpoundPerMinTOkgPerSec}
\alias{NISTpoundPerMinTOkgPerSec}
\title{Convert pound per minute  to kilogram per second }
\usage{NISTpoundPerMinTOkgPerSec(poundPerMin)}
\description{\code{NISTpoundPerMinTOkgPerSec} converts from pound per minute (lb/min) to kilogram per second (kg/s) }
\arguments{
  \item{poundPerMin}{pound per minute (lb/min) }
}
\value{kilogram per second (kg/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundPerMinTOkgPerSec(10)
}
\keyword{programming}