\name{NISTpoundPerSecTOkgPerSec}
\alias{NISTpoundPerSecTOkgPerSec}
\title{Convert pound per second  to kilogram per second }
\usage{NISTpoundPerSecTOkgPerSec(poundPerSec)}
\description{\code{NISTpoundPerSecTOkgPerSec} converts from pound per second (lb/s) to kilogram per second (kg/s) }
\arguments{
  \item{poundPerSec}{pound per second (lb/s) }
}
\value{kilogram per second (kg/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundPerSecTOkgPerSec(10)
}
\keyword{programming}