\name{NISTpoundPerSqrFtTOkgPerSqrMeter}
\alias{NISTpoundPerSqrFtTOkgPerSqrMeter}
\title{Convert pound per square foot  to kilogram per square meter }
\usage{NISTpoundPerSqrFtTOkgPerSqrMeter(poundPerSqrFt)}
\description{\code{NISTpoundPerSqrFtTOkgPerSqrMeter} converts from pound per square foot (lb/ft2) to kilogram per square meter (kg/m2) }
\arguments{
  \item{poundPerSqrFt}{pound per square foot (lb/ft2) }
}
\value{kilogram per square meter (kg/m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundPerSqrFtTOkgPerSqrMeter(10)
}
\keyword{programming}