\name{NISTpoundPerSqrInchTOkgPerSqrMeter}
\alias{NISTpoundPerSqrInchTOkgPerSqrMeter}
\title{Convert pound per square inch  to kilogram per square meter }
\usage{NISTpoundPerSqrInchTOkgPerSqrMeter(poundPerSqrInch)}
\description{\code{NISTpoundPerSqrInchTOkgPerSqrMeter} converts from pound per square inch (not pound-force) (lb/in2) to kilogram per square meter (kg/m2) }
\arguments{
  \item{poundPerSqrInch}{pound per square inch (not pound-force) (lb/in2) }
}
\value{kilogram per square meter (kg/m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundPerSqrInchTOkgPerSqrMeter(10)
}
\keyword{programming}