\name{NISTpoundPerYardTOkgPerMeter}
\alias{NISTpoundPerYardTOkgPerMeter}
\title{Convert pound per yard  to kilogram per meter }
\usage{NISTpoundPerYardTOkgPerMeter(poundPerYard)}
\description{\code{NISTpoundPerYardTOkgPerMeter} converts from pound per yard (lb/yd) to kilogram per meter (kg/m) }
\arguments{
  \item{poundPerYard}{pound per yard (lb/yd) }
}
\value{kilogram per meter (kg/m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundPerYardTOkgPerMeter(10)
}
\keyword{programming}