\name{NISTpoundTroyTOkg}
\alias{NISTpoundTroyTOkg}
\title{Convert pound  to kilogram }
\usage{NISTpoundTroyTOkg(poundTroy)}
\description{\code{NISTpoundTroyTOkg} converts from pound (troy or apothecary) (lb) to kilogram (kg) }
\arguments{
  \item{poundTroy}{pound (troy or apothecary) (lb) }
}
\value{kilogram (kg) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpoundTroyTOkg(10)
}
\keyword{programming}