\name{NISTpsiTOkpascal}
\alias{NISTpsiTOkpascal}
\title{Convert psi  to kilopascal }
\usage{NISTpsiTOkpascal(psi)}
\description{\code{NISTpsiTOkpascal} converts from psi (pound-force per square inch) (lbf/in2) to kilopascal (kPa) }
\arguments{
  \item{psi}{psi (pound-force per square inch) (lbf/in2) }
}
\value{kilopascal (kPa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpsiTOkpascal(10)
}
\keyword{programming}