\name{NISTpsiTOpascal}
\alias{NISTpsiTOpascal}
\title{Convert psi  to pascal }
\usage{NISTpsiTOpascal(psi)}
\description{\code{NISTpsiTOpascal} converts from psi (pound-force per square inch) (lbf/in2) to pascal (Pa) }
\arguments{
  \item{psi}{psi (pound-force per square inch) (lbf/in2) }
}
\value{pascal (Pa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTpsiTOpascal(10)
}
\keyword{programming}