\name{NISTquadTOjoule}
\alias{NISTquadTOjoule}
\title{Convert quad 11 to joule }
\usage{NISTquadTOjoule(quad)}
\description{\code{NISTquadTOjoule} converts from quad (1015 BtuIT) 11 to joule (J) }
\arguments{
  \item{quad}{quad (1015 BtuIT) 11 }
}
\value{joule (J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTquadTOjoule(10)
}
\keyword{programming}