\name{NISTquartUSdryTOcubMeter}
\alias{NISTquartUSdryTOcubMeter}
\title{Convert quart  to cubic meter }
\usage{NISTquartUSdryTOcubMeter(quartUSdry)}
\description{\code{NISTquartUSdryTOcubMeter} converts from quart (U.S. dry) (dry qt) to cubic meter (m3) }
\arguments{
  \item{quartUSdry}{quart (U.S. dry) (dry qt) }
}
\value{cubic meter (m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTquartUSdryTOcubMeter(10)
}
\keyword{programming}