\name{NISTquartUSdryTOliter}
\alias{NISTquartUSdryTOliter}
\title{Convert quart  to liter }
\usage{NISTquartUSdryTOliter(quartUSdry)}
\description{\code{NISTquartUSdryTOliter} converts from quart (U.S. dry) (dry qt) to liter (L) }
\arguments{
  \item{quartUSdry}{quart (U.S. dry) (dry qt) }
}
\value{liter (L) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTquartUSdryTOliter(10)
}
\keyword{programming}