\name{NISTquartUSliquidTOcubMeter}
\alias{NISTquartUSliquidTOcubMeter}
\title{Convert quart  to cubic meter }
\usage{NISTquartUSliquidTOcubMeter(quartUSliquid)}
\description{\code{NISTquartUSliquidTOcubMeter} converts from quart (U.S. liquid) (liq qt) to cubic meter (m3) }
\arguments{
  \item{quartUSliquid}{quart (U.S. liquid) (liq qt) }
}
\value{cubic meter (m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTquartUSliquidTOcubMeter(10)
}
\keyword{programming}