\name{NISTquartUSliquidTOliter}
\alias{NISTquartUSliquidTOliter}
\title{Convert quart  to liter }
\usage{NISTquartUSliquidTOliter(quartUSliquid)}
\description{\code{NISTquartUSliquidTOliter} converts from quart (U.S. liquid) (liq qt) to liter (L) }
\arguments{
  \item{quartUSliquid}{quart (U.S. liquid) (liq qt) }
}
\value{liter (L) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTquartUSliquidTOliter(10)
}
\keyword{programming}