\name{NISTradianPerSecTOrevolutionPerMin}
\alias{NISTradianPerSecTOrevolutionPerMin}
\title{Convert radian per second  to revolution per minute }
\usage{NISTradianPerSecTOrevolutionPerMin(radianPerSec)}
\description{\code{NISTradianPerSecTOrevolutionPerMin} converts from radian per second (rad/s) to revolution per minute (rpm) (r/min) }
\arguments{
  \item{radianPerSec}{radian per second (rad/s) }
}
\value{revolution per minute (rpm) (r/min) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTradianPerSecTOrevolutionPerMin(10)
}
\keyword{programming}