\name{NISTradianPerSecTOrpm}
\alias{NISTradianPerSecTOrpm}
\title{Convert radian per second  to rpm }
\usage{NISTradianPerSecTOrpm(radianPerSec)}
\description{\code{NISTradianPerSecTOrpm} converts from radian per second (rad/s) to rpm (revolution per minute) (r/min) }
\arguments{
  \item{radianPerSec}{radian per second (rad/s) }
}
\value{rpm (revolution per minute) (r/min) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTradianPerSecTOrpm(10)
}
\keyword{programming}