\name{NISTradianTOdeg}
\alias{NISTradianTOdeg}
\title{Convert radian  to degree }
\usage{NISTradianTOdeg(radian)}
\description{\code{NISTradianTOdeg} converts from radian (rad) to degree (angle) () }
\arguments{
  \item{radian}{radian (rad) }
}
\value{degree (angle) () }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTradianTOdeg(10)
}
\keyword{programming}