\name{NISTradianTOgon}
\alias{NISTradianTOgon}
\title{Convert radian  to gon }
\usage{NISTradianTOgon(radian)}
\description{\code{NISTradianTOgon} converts from radian (rad) to gon (also called grade) (gon) }
\arguments{
  \item{radian}{radian (rad) }
}
\value{gon (also called grade) (gon) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTradianTOgon(10)
}
\keyword{programming}