\name{NISTradianTOmil}
\alias{NISTradianTOmil}
\title{Convert radian  to mil }
\usage{NISTradianTOmil(radian)}
\description{\code{NISTradianTOmil} converts from radian (rad) to mil (angle) }
\arguments{
  \item{radian}{radian (rad) }
}
\value{mil (angle) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTradianTOmil(10)
}
\keyword{programming}