\name{NISTradianTOrevolution}
\alias{NISTradianTOrevolution}
\title{Convert radian  to revolution }
\usage{NISTradianTOrevolution(radian)}
\description{\code{NISTradianTOrevolution} converts from radian (rad) to revolution (r) }
\arguments{
  \item{radian}{radian (rad) }
}
\value{revolution (r) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTradianTOrevolution(10)
}
\keyword{programming}