\name{NISTradTOgray}
\alias{NISTradTOgray}
\title{Convert rad  to gray }
\usage{NISTradTOgray(rad)}
\description{\code{NISTradTOgray} converts from rad (absorbed dose) (rad) to gray (Gy) }
\arguments{
  \item{rad}{rad (absorbed dose) (rad) }
}
\value{gray (Gy) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTradTOgray(10)
}
\keyword{programming}