\name{NISTreciprocalMeterTOkayser}
\alias{NISTreciprocalMeterTOkayser}
\title{Convert reciprocal meter  to kayser }
\usage{NISTreciprocalMeterTOkayser(reciprocalMeter)}
\description{\code{NISTreciprocalMeterTOkayser} converts from reciprocal meter (m-1) to kayser (K) }
\arguments{
  \item{reciprocalMeter}{reciprocal meter (m-1) }
}
\value{kayser (K) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTreciprocalMeterTOkayser(10)
}
\keyword{programming}