\name{NISTreciprocalPascalSec1TOrhe}
\alias{NISTreciprocalPascalSec1TOrhe}
\title{Convert reciprocal pascal second -1 to rhe}
\usage{NISTreciprocalPascalSec1TOrhe(reciprocalPascalSec1)}
\description{\code{NISTreciprocalPascalSec1TOrhe} converts from reciprocal pascal second (Pa * s)-1 to rhe }
\arguments{
  \item{reciprocalPascalSec1}{reciprocal pascal second (Pa * s)-1 }
}
\value{rhe }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTreciprocalPascalSec1TOrhe(10)
}
\keyword{programming}