\name{NISTremTOsievert}
\alias{NISTremTOsievert}
\title{Convert rem  to sievert }
\usage{NISTremTOsievert(rem)}
\description{\code{NISTremTOsievert} converts from rem (rem) to sievert (Sv) }
\arguments{
  \item{rem}{rem (rem) }
}
\value{sievert (Sv) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTremTOsievert(10)
}
\keyword{programming}