\name{NISTrevolutionPerMinTOradianPerSec}
\alias{NISTrevolutionPerMinTOradianPerSec}
\title{Convert revolution per minute  to radian per second }
\usage{NISTrevolutionPerMinTOradianPerSec(revolutionPerMin)}
\description{\code{NISTrevolutionPerMinTOradianPerSec} converts from revolution per minute (rpm) (r/min) to radian per second (rad/s) }
\arguments{
  \item{revolutionPerMin}{revolution per minute (rpm) (r/min) }
}
\value{radian per second (rad/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTrevolutionPerMinTOradianPerSec(10)
}
\keyword{programming}