\name{NISTrevolutionTOradian}
\alias{NISTrevolutionTOradian}
\title{Convert revolution  to radian }
\usage{NISTrevolutionTOradian(revolution)}
\description{\code{NISTrevolutionTOradian} converts from revolution (r) to radian (rad) }
\arguments{
  \item{revolution}{revolution (r) }
}
\value{radian (rad) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTrevolutionTOradian(10)
}
\keyword{programming}