\name{NISTrheTOreciprocalPascalSec1}
\alias{NISTrheTOreciprocalPascalSec1}
\title{Convert rhe to reciprocal pascal second -1}
\usage{NISTrheTOreciprocalPascalSec1(rhe)}
\description{\code{NISTrheTOreciprocalPascalSec1} converts from rhe to reciprocal pascal second (Pa * s)-1 }
\arguments{
  \item{rhe}{rhe }
}
\value{reciprocal pascal second (Pa * s)-1 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTrheTOreciprocalPascalSec1(10)
}
\keyword{programming}