\name{NISTrodTOmeter}
\alias{NISTrodTOmeter}
\title{Convert rod 7 to meter }
\usage{NISTrodTOmeter(rod)}
\description{\code{NISTrodTOmeter} converts from rod (based on U.S. survey foot) (rd) 7 to meter (m) }
\arguments{
  \item{rod}{rod (based on U.S. survey foot) (rd) 7 }
}
\value{meter (m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTrodTOmeter(10)
}
\keyword{programming}