\name{NISTroentgenTOcoulombPerKg}
\alias{NISTroentgenTOcoulombPerKg}
\title{Convert roentgen  to coulomb per kilogram }
\usage{NISTroentgenTOcoulombPerKg(roentgen)}
\description{\code{NISTroentgenTOcoulombPerKg} converts from roentgen (R) to coulomb per kilogram (C/kg) }
\arguments{
  \item{roentgen}{roentgen (R) }
}
\value{coulomb per kilogram (C/kg) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTroentgenTOcoulombPerKg(10)
}
\keyword{programming}