\name{NISTrpmTOradianPerSec}
\alias{NISTrpmTOradianPerSec}
\title{Convert rpm  to radian per second }
\usage{NISTrpmTOradianPerSec(rpm)}
\description{\code{NISTrpmTOradianPerSec} converts from rpm (revolution per minute) (r/min) to radian per second (rad/s) }
\arguments{
  \item{rpm}{rpm (revolution per minute) (r/min) }
}
\value{radian per second (rad/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTrpmTOradianPerSec(10)
}
\keyword{programming}