\name{NISTsecTOday}
\alias{NISTsecTOday}
\title{Convert second  to day }
\usage{NISTsecTOday(sec)}
\description{\code{NISTsecTOday} converts from second (s) to day (d) }
\arguments{
  \item{sec}{second (s) }
}
\value{day (d) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsecTOday(10)
}
\keyword{programming}