\name{NISTsecTOdaySidereal}
\alias{NISTsecTOdaySidereal}
\title{Convert second  to day }
\usage{NISTsecTOdaySidereal(sec)}
\description{\code{NISTsecTOdaySidereal} converts from second (s) to day (sidereal) }
\arguments{
  \item{sec}{second (s) }
}
\value{day (sidereal) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsecTOdaySidereal(10)
}
\keyword{programming}