\name{NISTsecTOhour}
\alias{NISTsecTOhour}
\title{Convert second  to hour }
\usage{NISTsecTOhour(sec)}
\description{\code{NISTsecTOhour} converts from second (s) to hour (h) }
\arguments{
  \item{sec}{second (s) }
}
\value{hour (h) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsecTOhour(10)
}
\keyword{programming}