\name{NISTsecTOhourSidereal}
\alias{NISTsecTOhourSidereal}
\title{Convert second  to hour }
\usage{NISTsecTOhourSidereal(sec)}
\description{\code{NISTsecTOhourSidereal} converts from second (s) to hour (sidereal) }
\arguments{
  \item{sec}{second (s) }
}
\value{hour (sidereal) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsecTOhourSidereal(10)
}
\keyword{programming}