\name{NISTsecTOmin}
\alias{NISTsecTOmin}
\title{Convert second  to minute }
\usage{NISTsecTOmin(sec)}
\description{\code{NISTsecTOmin} converts from second (s) to minute (min) }
\arguments{
  \item{sec}{second (s) }
}
\value{minute (min) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsecTOmin(10)
}
\keyword{programming}