\name{NISTsecTOminSidereal}
\alias{NISTsecTOminSidereal}
\title{Convert second  to minute }
\usage{NISTsecTOminSidereal(sec)}
\description{\code{NISTsecTOminSidereal} converts from second (s) to minute (sidereal) }
\arguments{
  \item{sec}{second (s) }
}
\value{minute (sidereal) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsecTOminSidereal(10)
}
\keyword{programming}