\name{NISTsecTOradian}
\alias{NISTsecTOradian}
\title{Convert second  to radian }
\usage{NISTsecTOradian(sec)}
\description{\code{NISTsecTOradian} converts from second (angle) ('') to radian (rad) }
\arguments{
  \item{sec}{second (angle) ('') }
}
\value{radian (rad) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsecTOradian(10)
}
\keyword{programming}