\name{NISTsecTOshake}
\alias{NISTsecTOshake}
\title{Convert second  to shake}
\usage{NISTsecTOshake(sec)}
\description{\code{NISTsecTOshake} converts from second (s) to shake }
\arguments{
  \item{sec}{second (s) }
}
\value{shake }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsecTOshake(10)
}
\keyword{programming}