\name{NISTsecTOyearSidereal}
\alias{NISTsecTOyearSidereal}
\title{Convert second  to year }
\usage{NISTsecTOyearSidereal(sec)}
\description{\code{NISTsecTOyearSidereal} converts from second (s) to year (sidereal) }
\arguments{
  \item{sec}{second (s) }
}
\value{year (sidereal) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsecTOyearSidereal(10)
}
\keyword{programming}