\name{NISTsecTOyearTropical}
\alias{NISTsecTOyearTropical}
\title{Convert second  to year }
\usage{NISTsecTOyearTropical(sec)}
\description{\code{NISTsecTOyearTropical} converts from second (s) to year (tropical) }
\arguments{
  \item{sec}{second (s) }
}
\value{year (tropical) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsecTOyearTropical(10)
}
\keyword{programming}