\name{NISTshakeTOnanosec}
\alias{NISTshakeTOnanosec}
\title{Convert shake to nanosecond }
\usage{NISTshakeTOnanosec(shake)}
\description{\code{NISTshakeTOnanosec} converts from shake to nanosecond (ns) }
\arguments{
  \item{shake}{shake }
}
\value{nanosecond (ns) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTshakeTOnanosec(10)
}
\keyword{programming}