\name{NISTshakeTOsec}
\alias{NISTshakeTOsec}
\title{Convert shake to second }
\usage{NISTshakeTOsec(shake)}
\description{\code{NISTshakeTOsec} converts from shake to second (s) }
\arguments{
  \item{shake}{shake }
}
\value{second (s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTshakeTOsec(10)
}
\keyword{programming}