\name{NISTsiemensTOabmho}
\alias{NISTsiemensTOabmho}
\title{Convert siemens  to abmho}
\usage{NISTsiemensTOabmho(siemens)}
\description{\code{NISTsiemensTOabmho} converts from siemens (S) to abmho }
\arguments{
  \item{siemens}{siemens (S) }
}
\value{abmho }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsiemensTOabmho(10)
}
\keyword{programming}