\name{NISTsiemensTOmho}
\alias{NISTsiemensTOmho}
\title{Convert siemens  to mho}
\usage{NISTsiemensTOmho(siemens)}
\description{\code{NISTsiemensTOmho} converts from siemens (S) to mho }
\arguments{
  \item{siemens}{siemens (S) }
}
\value{mho }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsiemensTOmho(10)
}
\keyword{programming}