\name{NISTsiemensTOstatmho}
\alias{NISTsiemensTOstatmho}
\title{Convert siemens  to statmho}
\usage{NISTsiemensTOstatmho(siemens)}
\description{\code{NISTsiemensTOstatmho} converts from siemens (S) to statmho }
\arguments{
  \item{siemens}{siemens (S) }
}
\value{statmho }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsiemensTOstatmho(10)
}
\keyword{programming}