\name{NISTsievertTOrem}
\alias{NISTsievertTOrem}
\title{Convert sievert  to rem }
\usage{NISTsievertTOrem(sievert)}
\description{\code{NISTsievertTOrem} converts from sievert (Sv) to rem (rem) }
\arguments{
  \item{sievert}{sievert (Sv) }
}
\value{rem (rem) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsievertTOrem(10)
}
\keyword{programming}