\name{NISTslugPerCubFtTOkgPerCubMeter}
\alias{NISTslugPerCubFtTOkgPerCubMeter}
\title{Convert slug per cubic foot  to kilogram per cubic meter }
\usage{NISTslugPerCubFtTOkgPerCubMeter(slugPerCubFt)}
\description{\code{NISTslugPerCubFtTOkgPerCubMeter} converts from slug per cubic foot (slug/ft3) to kilogram per cubic meter (kg/m3) }
\arguments{
  \item{slugPerCubFt}{slug per cubic foot (slug/ft3) }
}
\value{kilogram per cubic meter (kg/m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTslugPerCubFtTOkgPerCubMeter(10)
}
\keyword{programming}