\name{NISTslugPerFtSecTOpascalSec}
\alias{NISTslugPerFtSecTOpascalSec}
\title{Convert slug per foot second  to pascal second }
\usage{NISTslugPerFtSecTOpascalSec(slugPerFtSec)}
\description{\code{NISTslugPerFtSecTOpascalSec} converts from slug per foot second [slug/(ft * s)] to pascal second (Pa * s) }
\arguments{
  \item{slugPerFtSec}{slug per foot second [slug/(ft * s)] }
}
\value{pascal second (Pa * s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTslugPerFtSecTOpascalSec(10)
}
\keyword{programming}