\name{NISTslugTOkg}
\alias{NISTslugTOkg}
\title{Convert slug  to kilogram }
\usage{NISTslugTOkg(slug)}
\description{\code{NISTslugTOkg} converts from slug (slug) to kilogram (kg) }
\arguments{
  \item{slug}{slug (slug) }
}
\value{kilogram (kg) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTslugTOkg(10)
}
\keyword{programming}