\name{NISTsqrCmTOsqrInch}
\alias{NISTsqrCmTOsqrInch}
\title{Convert square centimeter  to square inch }
\usage{NISTsqrCmTOsqrInch(sqrCm)}
\description{\code{NISTsqrCmTOsqrInch} converts from square centimeter (cm2) to square inch (in2) }
\arguments{
  \item{sqrCm}{square centimeter (cm2) }
}
\value{square inch (in2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsqrCmTOsqrInch(10)
}
\keyword{programming}