\name{NISTsqrFtPerHourTOsqrMeterPerSec}
\alias{NISTsqrFtPerHourTOsqrMeterPerSec}
\title{Convert square foot per hour  to square meter per second }
\usage{NISTsqrFtPerHourTOsqrMeterPerSec(sqrFtPerHour)}
\description{\code{NISTsqrFtPerHourTOsqrMeterPerSec} converts from square foot per hour (ft2/h) to square meter per second (m2/s) }
\arguments{
  \item{sqrFtPerHour}{square foot per hour (ft2/h) }
}
\value{square meter per second (m2/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsqrFtPerHourTOsqrMeterPerSec(10)
}
\keyword{programming}