\name{NISTsqrFtPerSecTOsqrMeterPerSec}
\alias{NISTsqrFtPerSecTOsqrMeterPerSec}
\title{Convert square foot per second  to square meter per second }
\usage{NISTsqrFtPerSecTOsqrMeterPerSec(sqrFtPerSec)}
\description{\code{NISTsqrFtPerSecTOsqrMeterPerSec} converts from square foot per second (ft2/s) to square meter per second (m2/s) }
\arguments{
  \item{sqrFtPerSec}{square foot per second (ft2/s) }
}
\value{square meter per second (m2/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsqrFtPerSecTOsqrMeterPerSec(10)
}
\keyword{programming}