\name{NISTsqrFtTOsqrMeter}
\alias{NISTsqrFtTOsqrMeter}
\title{Convert square foot  to square meter }
\usage{NISTsqrFtTOsqrMeter(sqrFt)}
\description{\code{NISTsqrFtTOsqrMeter} converts from square foot (ft2) to square meter (m2) }
\arguments{
  \item{sqrFt}{square foot (ft2) }
}
\value{square meter (m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsqrFtTOsqrMeter(10)
}
\keyword{programming}