\name{NISTsqrInchTOsqrCm}
\alias{NISTsqrInchTOsqrCm}
\title{Convert square inch  to square centimeter }
\usage{NISTsqrInchTOsqrCm(sqrInch)}
\description{\code{NISTsqrInchTOsqrCm} converts from square inch (in2) to square centimeter (cm2) }
\arguments{
  \item{sqrInch}{square inch (in2) }
}
\value{square centimeter (cm2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsqrInchTOsqrCm(10)
}
\keyword{programming}