\name{NISTsqrInchTOsqrMeter}
\alias{NISTsqrInchTOsqrMeter}
\title{Convert square inch  to square meter }
\usage{NISTsqrInchTOsqrMeter(sqrInch)}
\description{\code{NISTsqrInchTOsqrMeter} converts from square inch (in2) to square meter (m2) }
\arguments{
  \item{sqrInch}{square inch (in2) }
}
\value{square meter (m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsqrInchTOsqrMeter(10)
}
\keyword{programming}