\name{NISTsqrKmTOsqrMile}
\alias{NISTsqrKmTOsqrMile}
\title{Convert square kilometer  to square mile }
\usage{NISTsqrKmTOsqrMile(sqrKm)}
\description{\code{NISTsqrKmTOsqrMile} converts from square kilometer (km2) to square mile (mi2) }
\arguments{
  \item{sqrKm}{square kilometer (km2) }
}
\value{square mile (mi2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsqrKmTOsqrMile(10)
}
\keyword{programming}