\name{NISTsqrKmTOsqrMileUSsurvey}
\alias{NISTsqrKmTOsqrMileUSsurvey}
\title{Convert square kilometer  to square mile 7}
\usage{NISTsqrKmTOsqrMileUSsurvey(sqrKm)}
\description{\code{NISTsqrKmTOsqrMileUSsurvey} converts from square kilometer (km2) to square mile (based on U.S. survey foot) (mi2) 7 }
\arguments{
  \item{sqrKm}{square kilometer (km2) }
}
\value{square mile (based on U.S. survey foot) (mi2) 7 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsqrKmTOsqrMileUSsurvey(10)
}
\keyword{programming}