\name{NISTsqrMeterKPerWattTOclo}
\alias{NISTsqrMeterKPerWattTOclo}
\title{Convert square meter kelvin per watt  to clo}
\usage{NISTsqrMeterKPerWattTOclo(sqrMeterKPerWatt)}
\description{\code{NISTsqrMeterKPerWattTOclo} converts from square meter kelvin per watt (m2 * K/W) to clo }
\arguments{
  \item{sqrMeterKPerWatt}{square meter kelvin per watt (m2 * K/W) }
}
\value{clo }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsqrMeterKPerWattTOclo(10)
}
\keyword{programming}