\name{NISTsqrMeterKPerWattTOdegFHourSqrFtPerUkThUnIT}
\alias{NISTsqrMeterKPerWattTOdegFHourSqrFtPerUkThUnIT}
\title{Convert square meter kelvin per watt  to degree Fahrenheit hour square foot per British thermal unitIT }
\usage{NISTsqrMeterKPerWattTOdegFHourSqrFtPerUkThUnIT(sqrMeterKPerWatt)}
\description{\code{NISTsqrMeterKPerWattTOdegFHourSqrFtPerUkThUnIT} converts from square meter kelvin per watt (m2 * K/W) to degree Fahrenheit hour square foot per British thermal unitIT (F * h * ft2/BtuIT) }
\arguments{
  \item{sqrMeterKPerWatt}{square meter kelvin per watt (m2 * K/W) }
}
\value{degree Fahrenheit hour square foot per British thermal unitIT (F * h * ft2/BtuIT) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsqrMeterKPerWattTOdegFHourSqrFtPerUkThUnIT(10)
}
\keyword{programming}