\name{NISTsqrMeterKPerWattTOdegFHourSqrFtPerUkThUnth}
\alias{NISTsqrMeterKPerWattTOdegFHourSqrFtPerUkThUnth}
\title{Convert square meter kelvin per watt  to degree Fahrenheit hour square foot per British thermal unitth }
\usage{NISTsqrMeterKPerWattTOdegFHourSqrFtPerUkThUnth(sqrMeterKPerWatt)}
\description{\code{NISTsqrMeterKPerWattTOdegFHourSqrFtPerUkThUnth} converts from square meter kelvin per watt (m2 * K/W) to degree Fahrenheit hour square foot per British thermal unitth (F * h * ft2/Btuth) }
\arguments{
  \item{sqrMeterKPerWatt}{square meter kelvin per watt (m2 * K/W) }
}
\value{degree Fahrenheit hour square foot per British thermal unitth (F * h * ft2/Btuth) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsqrMeterKPerWattTOdegFHourSqrFtPerUkThUnth(10)
}
\keyword{programming}