\name{NISTsqrMeterPerSecTOsqrFtPerHour}
\alias{NISTsqrMeterPerSecTOsqrFtPerHour}
\title{Convert square meter per second  to square foot per hour }
\usage{NISTsqrMeterPerSecTOsqrFtPerHour(sqrMeterPerSec)}
\description{\code{NISTsqrMeterPerSecTOsqrFtPerHour} converts from square meter per second (m2/s) to square foot per hour (ft2/h) }
\arguments{
  \item{sqrMeterPerSec}{square meter per second (m2/s) }
}
\value{square foot per hour (ft2/h) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsqrMeterPerSecTOsqrFtPerHour(10)
}
\keyword{programming}