\name{NISTsqrMeterPerSecTOsqrFtPerSec}
\alias{NISTsqrMeterPerSecTOsqrFtPerSec}
\title{Convert square meter per second  to square foot per second }
\usage{NISTsqrMeterPerSecTOsqrFtPerSec(sqrMeterPerSec)}
\description{\code{NISTsqrMeterPerSecTOsqrFtPerSec} converts from square meter per second (m2/s) to square foot per second (ft2/s) }
\arguments{
  \item{sqrMeterPerSec}{square meter per second (m2/s) }
}
\value{square foot per second (ft2/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsqrMeterPerSecTOsqrFtPerSec(10)
}
\keyword{programming}