\name{NISTsqrMeterTOacre}
\alias{NISTsqrMeterTOacre}
\title{Convert square meter  to acre 7}
\usage{NISTsqrMeterTOacre(sqrMeter)}
\description{\code{NISTsqrMeterTOacre} converts from square meter (m2) to acre (based on U.S. survey foot) 7 }
\arguments{
  \item{sqrMeter}{square meter (m2) }
}
\value{acre (based on U.S. survey foot) 7 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsqrMeterTOacre(10)
}
\keyword{programming}