\name{NISTsqrMeterTOare}
\alias{NISTsqrMeterTOare}
\title{Convert square meter  to are }
\usage{NISTsqrMeterTOare(sqrMeter)}
\description{\code{NISTsqrMeterTOare} converts from square meter (m2) to are (a) }
\arguments{
  \item{sqrMeter}{square meter (m2) }
}
\value{are (a) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsqrMeterTOare(10)
}
\keyword{programming}