\name{NISTsqrMeterTObarn}
\alias{NISTsqrMeterTObarn}
\title{Convert square meter  to barn }
\usage{NISTsqrMeterTObarn(sqrMeter)}
\description{\code{NISTsqrMeterTObarn} converts from square meter (m2) to barn (b) }
\arguments{
  \item{sqrMeter}{square meter (m2) }
}
\value{barn (b) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsqrMeterTObarn(10)
}
\keyword{programming}