\name{NISTsqrMeterTOcircularMil}
\alias{NISTsqrMeterTOcircularMil}
\title{Convert square meter  to circular mil}
\usage{NISTsqrMeterTOcircularMil(sqrMeter)}
\description{\code{NISTsqrMeterTOcircularMil} converts from square meter (m2) to circular mil }
\arguments{
  \item{sqrMeter}{square meter (m2) }
}
\value{circular mil }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsqrMeterTOcircularMil(10)
}
\keyword{programming}