\name{NISTsqrMeterTOhectare}
\alias{NISTsqrMeterTOhectare}
\title{Convert square meter  to hectare }
\usage{NISTsqrMeterTOhectare(sqrMeter)}
\description{\code{NISTsqrMeterTOhectare} converts from square meter (m2) to hectare (ha) }
\arguments{
  \item{sqrMeter}{square meter (m2) }
}
\value{hectare (ha) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsqrMeterTOhectare(10)
}
\keyword{programming}