\name{NISTsqrMeterTOsqrFt}
\alias{NISTsqrMeterTOsqrFt}
\title{Convert square meter  to square foot }
\usage{NISTsqrMeterTOsqrFt(sqrMeter)}
\description{\code{NISTsqrMeterTOsqrFt} converts from square meter (m2) to square foot (ft2) }
\arguments{
  \item{sqrMeter}{square meter (m2) }
}
\value{square foot (ft2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsqrMeterTOsqrFt(10)
}
\keyword{programming}