\name{NISTsqrMeterTOsqrInch}
\alias{NISTsqrMeterTOsqrInch}
\title{Convert square meter  to square inch }
\usage{NISTsqrMeterTOsqrInch(sqrMeter)}
\description{\code{NISTsqrMeterTOsqrInch} converts from square meter (m2) to square inch (in2) }
\arguments{
  \item{sqrMeter}{square meter (m2) }
}
\value{square inch (in2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsqrMeterTOsqrInch(10)
}
\keyword{programming}