\name{NISTsqrMeterTOsqrMile}
\alias{NISTsqrMeterTOsqrMile}
\title{Convert square meter  to square mile }
\usage{NISTsqrMeterTOsqrMile(sqrMeter)}
\description{\code{NISTsqrMeterTOsqrMile} converts from square meter (m2) to square mile (mi2) }
\arguments{
  \item{sqrMeter}{square meter (m2) }
}
\value{square mile (mi2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsqrMeterTOsqrMile(10)
}
\keyword{programming}