\name{NISTsqrMeterTOsqrMileUSsurvey}
\alias{NISTsqrMeterTOsqrMileUSsurvey}
\title{Convert square meter  to square mile 7}
\usage{NISTsqrMeterTOsqrMileUSsurvey(sqrMeter)}
\description{\code{NISTsqrMeterTOsqrMileUSsurvey} converts from square meter (m2) to square mile (based on U.S. survey foot) (mi2) 7 }
\arguments{
  \item{sqrMeter}{square meter (m2) }
}
\value{square mile (based on U.S. survey foot) (mi2) 7 }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsqrMeterTOsqrMileUSsurvey(10)
}
\keyword{programming}