\name{NISTsqrMeterTOsqrYard}
\alias{NISTsqrMeterTOsqrYard}
\title{Convert square meter  to square yard }
\usage{NISTsqrMeterTOsqrYard(sqrMeter)}
\description{\code{NISTsqrMeterTOsqrYard} converts from square meter (m2) to square yard (yd2) }
\arguments{
  \item{sqrMeter}{square meter (m2) }
}
\value{square yard (yd2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsqrMeterTOsqrYard(10)
}
\keyword{programming}