\name{NISTsqrMileTOsqrKm}
\alias{NISTsqrMileTOsqrKm}
\title{Convert square mile  to square kilometer }
\usage{NISTsqrMileTOsqrKm(sqrMile)}
\description{\code{NISTsqrMileTOsqrKm} converts from square mile (mi2) to square kilometer (km2) }
\arguments{
  \item{sqrMile}{square mile (mi2) }
}
\value{square kilometer (km2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsqrMileTOsqrKm(10)
}
\keyword{programming}