\name{NISTsqrMileTOsqrMeter}
\alias{NISTsqrMileTOsqrMeter}
\title{Convert square mile  to square meter }
\usage{NISTsqrMileTOsqrMeter(sqrMile)}
\description{\code{NISTsqrMileTOsqrMeter} converts from square mile (mi2) to square meter (m2) }
\arguments{
  \item{sqrMile}{square mile (mi2) }
}
\value{square meter (m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsqrMileTOsqrMeter(10)
}
\keyword{programming}