\name{NISTsqrMileUSsurveyTOsqrKm}
\alias{NISTsqrMileUSsurveyTOsqrKm}
\title{Convert square mile 7 to square kilometer }
\usage{NISTsqrMileUSsurveyTOsqrKm(sqrMileUSsurvey)}
\description{\code{NISTsqrMileUSsurveyTOsqrKm} converts from square mile (based on U.S. survey foot) (mi2) 7 to square kilometer (km2) }
\arguments{
  \item{sqrMileUSsurvey}{square mile (based on U.S. survey foot) (mi2) 7 }
}
\value{square kilometer (km2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsqrMileUSsurveyTOsqrKm(10)
}
\keyword{programming}