\name{NISTsqrMillimeterTOcircularMil}
\alias{NISTsqrMillimeterTOcircularMil}
\title{Convert square millimeter  to circular mil}
\usage{NISTsqrMillimeterTOcircularMil(sqrMillimeter)}
\description{\code{NISTsqrMillimeterTOcircularMil} converts from square millimeter (mm2) to circular mil }
\arguments{
  \item{sqrMillimeter}{square millimeter (mm2) }
}
\value{circular mil }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsqrMillimeterTOcircularMil(10)
}
\keyword{programming}