\name{NISTsqrYardTOsqrMeter}
\alias{NISTsqrYardTOsqrMeter}
\title{Convert square yard  to square meter }
\usage{NISTsqrYardTOsqrMeter(sqrYard)}
\description{\code{NISTsqrYardTOsqrMeter} converts from square yard (yd2) to square meter (m2) }
\arguments{
  \item{sqrYard}{square yard (yd2) }
}
\value{square meter (m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTsqrYardTOsqrMeter(10)
}
\keyword{programming}