\name{NISTstatampereTOampere}
\alias{NISTstatampereTOampere}
\title{Convert statampere to ampere }
\usage{NISTstatampereTOampere(statampere)}
\description{\code{NISTstatampereTOampere} converts from statampere to ampere (A) }
\arguments{
  \item{statampere}{statampere }
}
\value{ampere (A) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTstatampereTOampere(10)
}
\keyword{programming}