\name{NISTstatcoulombTOcoulomb}
\alias{NISTstatcoulombTOcoulomb}
\title{Convert statcoulomb to coulomb }
\usage{NISTstatcoulombTOcoulomb(statcoulomb)}
\description{\code{NISTstatcoulombTOcoulomb} converts from statcoulomb to coulomb (C) }
\arguments{
  \item{statcoulomb}{statcoulomb }
}
\value{coulomb (C) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTstatcoulombTOcoulomb(10)
}
\keyword{programming}