\name{NISTstathenryTOhenry}
\alias{NISTstathenryTOhenry}
\title{Convert stathenry to henry }
\usage{NISTstathenryTOhenry(stathenry)}
\description{\code{NISTstathenryTOhenry} converts from stathenry to henry (H) }
\arguments{
  \item{stathenry}{stathenry }
}
\value{henry (H) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTstathenryTOhenry(10)
}
\keyword{programming}