\name{NISTstatmhoTOsiemens}
\alias{NISTstatmhoTOsiemens}
\title{Convert statmho to siemens }
\usage{NISTstatmhoTOsiemens(statmho)}
\description{\code{NISTstatmhoTOsiemens} converts from statmho to siemens (S) }
\arguments{
  \item{statmho}{statmho }
}
\value{siemens (S) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTstatmhoTOsiemens(10)
}
\keyword{programming}