\name{NISTstatohmTOohm}
\alias{NISTstatohmTOohm}
\title{Convert statohm to ohm }
\usage{NISTstatohmTOohm(statohm)}
\description{\code{NISTstatohmTOohm} converts from statohm to ohm (ohm) }
\arguments{
  \item{statohm}{statohm }
}
\value{ohm (ohm) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTstatohmTOohm(10)
}
\keyword{programming}