\name{NISTstatvoltTOvolt}
\alias{NISTstatvoltTOvolt}
\title{Convert statvolt to volt }
\usage{NISTstatvoltTOvolt(statvolt)}
\description{\code{NISTstatvoltTOvolt} converts from statvolt to volt (V) }
\arguments{
  \item{statvolt}{statvolt }
}
\value{volt (V) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTstatvoltTOvolt(10)
}
\keyword{programming}