\name{NISTstereTOcubMeter}
\alias{NISTstereTOcubMeter}
\title{Convert stere  to cubic meter }
\usage{NISTstereTOcubMeter(stere)}
\description{\code{NISTstereTOcubMeter} converts from stere (st) to cubic meter (m3) }
\arguments{
  \item{stere}{stere (st) }
}
\value{cubic meter (m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTstereTOcubMeter(10)
}
\keyword{programming}