\name{NISTstilbTOcandelaPerSqrMeter}
\alias{NISTstilbTOcandelaPerSqrMeter}
\title{Convert stilb  to candela per square meter }
\usage{NISTstilbTOcandelaPerSqrMeter(stilb)}
\description{\code{NISTstilbTOcandelaPerSqrMeter} converts from stilb (sb) to candela per square meter (cd/m2) }
\arguments{
  \item{stilb}{stilb (sb) }
}
\value{candela per square meter (cd/m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTstilbTOcandelaPerSqrMeter(10)
}
\keyword{programming}