\name{NISTstokesTOmeterSqrdPerSec}
\alias{NISTstokesTOmeterSqrdPerSec}
\title{Convert stokes  to meter squared per second }
\usage{NISTstokesTOmeterSqrdPerSec(stokes)}
\description{\code{NISTstokesTOmeterSqrdPerSec} converts from stokes (St) to meter squared per second (m2/s) }
\arguments{
  \item{stokes}{stokes (St) }
}
\value{meter squared per second (m2/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTstokesTOmeterSqrdPerSec(10)
}
\keyword{programming}