\name{NISTteaspoonTOcubMeter}
\alias{NISTteaspoonTOcubMeter}
\title{Convert teaspoon to cubic meter }
\usage{NISTteaspoonTOcubMeter(teaspoon)}
\description{\code{NISTteaspoonTOcubMeter} converts from teaspoon to cubic meter (m3) }
\arguments{
  \item{teaspoon}{teaspoon }
}
\value{cubic meter (m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTteaspoonTOcubMeter(10)
}
\keyword{programming}