\name{NISTteslaTOgamma}
\alias{NISTteslaTOgamma}
\title{Convert tesla  to gamma }
\usage{NISTteslaTOgamma(tesla)}
\description{\code{NISTteslaTOgamma} converts from tesla (T) to gamma (y) }
\arguments{
  \item{tesla}{tesla (T) }
}
\value{gamma (y) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTteslaTOgamma(10)
}
\keyword{programming}