\name{NISTteslaTOgauss}
\alias{NISTteslaTOgauss}
\title{Convert tesla  to gauss }
\usage{NISTteslaTOgauss(tesla)}
\description{\code{NISTteslaTOgauss} converts from tesla (T) to gauss (Gs, G) }
\arguments{
  \item{tesla}{tesla (T) }
}
\value{gauss (Gs, G) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTteslaTOgauss(10)
}
\keyword{programming}