\name{NISTtexTOkgPerMeter}
\alias{NISTtexTOkgPerMeter}
\title{Convert tex to kilogram per meter }
\usage{NISTtexTOkgPerMeter(tex)}
\description{\code{NISTtexTOkgPerMeter} converts from tex to kilogram per meter (kg/m) }
\arguments{
  \item{tex}{tex }
}
\value{kilogram per meter (kg/m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTtexTOkgPerMeter(10)
}
\keyword{programming}