\name{NISTthermECtOjoule}
\alias{NISTthermECtOjoule}
\title{Convert therm 24 to joule }
\usage{NISTthermECtOjoule(thermEC)}
\description{\code{NISTthermECtOjoule} converts from therm (EC) 24 to joule (J) }
\arguments{
  \item{thermEC}{therm (EC) 24 }
}
\value{joule (J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTthermECtOjoule(10)
}
\keyword{programming}