\name{NISTtonAssayTOgram}
\alias{NISTtonAssayTOgram}
\title{Convert ton, assay  to gram }
\usage{NISTtonAssayTOgram(tonAssay)}
\description{\code{NISTtonAssayTOgram} converts from ton, assay (AT) to gram (g) }
\arguments{
  \item{tonAssay}{ton, assay (AT) }
}
\value{gram (g) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTtonAssayTOgram(10)
}
\keyword{programming}