\name{NISTtonAssayTOkg}
\alias{NISTtonAssayTOkg}
\title{Convert ton, assay  to kilogram }
\usage{NISTtonAssayTOkg(tonAssay)}
\description{\code{NISTtonAssayTOkg} converts from ton, assay (AT) to kilogram (kg) }
\arguments{
  \item{tonAssay}{ton, assay (AT) }
}
\value{kilogram (kg) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTtonAssayTOkg(10)
}
\keyword{programming}