\name{NISTtonForceTOnewton}
\alias{NISTtonForceTOnewton}
\title{Convert ton-force  to newton }
\usage{NISTtonForceTOnewton(tonForce)}
\description{\code{NISTtonForceTOnewton} converts from ton-force (2000 lbf) to newton (N) }
\arguments{
  \item{tonForce}{ton-force (2000 lbf) }
}
\value{newton (N) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTtonForceTOnewton(10)
}
\keyword{programming}