\name{NISTtonLongTOkg}
\alias{NISTtonLongTOkg}
\title{Convert ton, long  to kilogram }
\usage{NISTtonLongTOkg(tonLong)}
\description{\code{NISTtonLongTOkg} converts from ton, long (2240 lb) to kilogram (kg) }
\arguments{
  \item{tonLong}{ton, long (2240 lb) }
}
\value{kilogram (kg) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTtonLongTOkg(10)
}
\keyword{programming}