\name{NISTtonneTOkg}
\alias{NISTtonneTOkg}
\title{Convert tonne  to kilogram }
\usage{NISTtonneTOkg(tonne)}
\description{\code{NISTtonneTOkg} converts from tonne (called metric ton in U.S.) (t) to kilogram (kg) }
\arguments{
  \item{tonne}{tonne (called metric ton in U.S.) (t) }
}
\value{kilogram (kg) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTtonneTOkg(10)
}
\keyword{programming}