\name{NISTtonOfRefrigerationTOwatt}
\alias{NISTtonOfRefrigerationTOwatt}
\title{Convert ton of refrigeration  to watt }
\usage{NISTtonOfRefrigerationTOwatt(tonOfRefrigeration)}
\description{\code{NISTtonOfRefrigerationTOwatt} converts from ton of refrigeration (12 000 BtuIT/h) to watt (W) }
\arguments{
  \item{tonOfRefrigeration}{ton of refrigeration (12 000 BtuIT/h) }
}
\value{watt (W) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTtonOfRefrigerationTOwatt(10)
}
\keyword{programming}