\name{NISTtonOfTNTtOjoule}
\alias{NISTtonOfTNTtOjoule}
\title{Convert ton of TNT 25 to joule }
\usage{NISTtonOfTNTtOjoule(tonOfTNT)}
\description{\code{NISTtonOfTNTtOjoule} converts from ton of TNT (energy equivalent) 25 to joule (J) }
\arguments{
  \item{tonOfTNT}{ton of TNT (energy equivalent) 25 }
}
\value{joule (J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTtonOfTNTtOjoule(10)
}
\keyword{programming}