\name{NISTtonRegisterTOcubMeter}
\alias{NISTtonRegisterTOcubMeter}
\title{Convert ton, register to cubic meter }
\usage{NISTtonRegisterTOcubMeter(tonRegister)}
\description{\code{NISTtonRegisterTOcubMeter} converts from ton, register to cubic meter (m3) }
\arguments{
  \item{tonRegister}{ton, register }
}
\value{cubic meter (m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTtonRegisterTOcubMeter(10)
}
\keyword{programming}