\name{NISTtonShortPerCubYardTOkgPerCubMeter}
\alias{NISTtonShortPerCubYardTOkgPerCubMeter}
\title{Convert ton, short, per cubic yard to kilogram per cubic meter }
\usage{NISTtonShortPerCubYardTOkgPerCubMeter(tonShortPerCubYard)}
\description{\code{NISTtonShortPerCubYardTOkgPerCubMeter} converts from ton, short, per cubic yard to kilogram per cubic meter (kg/m3) }
\arguments{
  \item{tonShortPerCubYard}{ton, short, per cubic yard }
}
\value{kilogram per cubic meter (kg/m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTtonShortPerCubYardTOkgPerCubMeter(10)
}
\keyword{programming}