\name{NISTtonShortPerHourTOkgPerSec}
\alias{NISTtonShortPerHourTOkgPerSec}
\title{Convert ton, short, per hour to kilogram per second }
\usage{NISTtonShortPerHourTOkgPerSec(tonShortPerHour)}
\description{\code{NISTtonShortPerHourTOkgPerSec} converts from ton, short, per hour to kilogram per second (kg/s) }
\arguments{
  \item{tonShortPerHour}{ton, short, per hour }
}
\value{kilogram per second (kg/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTtonShortPerHourTOkgPerSec(10)
}
\keyword{programming}