\name{NISTtonShortTOkg}
\alias{NISTtonShortTOkg}
\title{Convert ton, short  to kilogram }
\usage{NISTtonShortTOkg(tonShort)}
\description{\code{NISTtonShortTOkg} converts from ton, short (2000 lb) to kilogram (kg) }
\arguments{
  \item{tonShort}{ton, short (2000 lb) }
}
\value{kilogram (kg) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTtonShortTOkg(10)
}
\keyword{programming}