\name{NISTtonSItOkg}
\alias{NISTtonSItOkg}
\title{Convert ton, metric  to kilogram }
\usage{NISTtonSItOkg(tonSI)}
\description{\code{NISTtonSItOkg} converts from ton, metric (t) to kilogram (kg) }
\arguments{
  \item{tonSI}{ton, metric (t) }
}
\value{kilogram (kg) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTtonSItOkg(10)
}
\keyword{programming}