\name{NISTtorrTOpascal}
\alias{NISTtorrTOpascal}
\title{Convert torr  to pascal }
\usage{NISTtorrTOpascal(torr)}
\description{\code{NISTtorrTOpascal} converts from torr (Torr) to pascal (Pa) }
\arguments{
  \item{torr}{torr (Torr) }
}
\value{pascal (Pa) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTtorrTOpascal(10)
}
\keyword{programming}