\name{NISTukThUn39FtOjoule}
\alias{NISTukThUn39FtOjoule}
\title{Convert British thermal unit  to joule }
\usage{NISTukThUn39FtOjoule(ukThUn39F)}
\description{\code{NISTukThUn39FtOjoule} converts from British thermal unit (39 F) (Btu) to joule (J) }
\arguments{
  \item{ukThUn39F}{British thermal unit (39 F) (Btu) }
}
\value{joule (J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUn39FtOjoule(10)
}
\keyword{programming}