\name{NISTukThUn59FtOjoule}
\alias{NISTukThUn59FtOjoule}
\title{Convert British thermal unit  to joule }
\usage{NISTukThUn59FtOjoule(ukThUn59F)}
\description{\code{NISTukThUn59FtOjoule} converts from British thermal unit (59 F) (Btu) to joule (J) }
\arguments{
  \item{ukThUn59F}{British thermal unit (59 F) (Btu) }
}
\value{joule (J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUn59FtOjoule(10)
}
\keyword{programming}