\name{NISTukThUn60FtOjoule}
\alias{NISTukThUn60FtOjoule}
\title{Convert British thermal unit  to joule }
\usage{NISTukThUn60FtOjoule(ukThUn60F)}
\description{\code{NISTukThUn60FtOjoule} converts from British thermal unit (60 F) (Btu) to joule (J) }
\arguments{
  \item{ukThUn60F}{British thermal unit (60 F) (Btu) }
}
\value{joule (J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUn60FtOjoule(10)
}
\keyword{programming}