\name{NISTukThUnITFtPerHourSqrFtDegFtOwattPerMeterK}
\alias{NISTukThUnITFtPerHourSqrFtDegFtOwattPerMeterK}
\title{Convert British thermal unitIT foot per hour square foot degree Fahrenheit  to watt per meter kelvin }
\usage{NISTukThUnITFtPerHourSqrFtDegFtOwattPerMeterK(ukThUnITFtPerHourSqrFtDegF)}
\description{\code{NISTukThUnITFtPerHourSqrFtDegFtOwattPerMeterK} converts from British thermal unitIT foot per hour square foot degree Fahrenheit [BtuIT * ft/(h * ft2 * F)] to watt per meter kelvin [W/(m * K)] }
\arguments{
  \item{ukThUnITFtPerHourSqrFtDegF}{British thermal unitIT foot per hour square foot degree Fahrenheit [BtuIT * ft/(h * ft2 * F)] }
}
\value{watt per meter kelvin [W/(m * K)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnITFtPerHourSqrFtDegFtOwattPerMeterK(10)
}
\keyword{programming}