\name{NISTukThUnITInchPerSecSqrFtDegFtOwattPerMeterK}
\alias{NISTukThUnITInchPerSecSqrFtDegFtOwattPerMeterK}
\title{Convert British thermal unitIT inch per second square foot degree Fahrenheit  to watt per meter kelvin }
\usage{NISTukThUnITInchPerSecSqrFtDegFtOwattPerMeterK(ukThUnITInchPerSecSqrFtDegF)}
\description{\code{NISTukThUnITInchPerSecSqrFtDegFtOwattPerMeterK} converts from British thermal unitIT inch per second square foot degree Fahrenheit [BtuIT * in /(s * ft2 * F)] to watt per meter kelvin [W/(m * K)] }
\arguments{
  \item{ukThUnITInchPerSecSqrFtDegF}{British thermal unitIT inch per second square foot degree Fahrenheit [BtuIT * in /(s * ft2 * F)] }
}
\value{watt per meter kelvin [W/(m * K)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnITInchPerSecSqrFtDegFtOwattPerMeterK(10)
}
\keyword{programming}