\name{NISTukThUnITPerDegFtOjoulePerK}
\alias{NISTukThUnITPerDegFtOjoulePerK}
\title{Convert British thermal unitIT per degree Fahrenheit  to joule per kelvin }
\usage{NISTukThUnITPerDegFtOjoulePerK(ukThUnITPerDegF)}
\description{\code{NISTukThUnITPerDegFtOjoulePerK} converts from British thermal unitIT per degree Fahrenheit (BtuIT/F) to joule per kelvin (J/K) }
\arguments{
  \item{ukThUnITPerDegF}{British thermal unitIT per degree Fahrenheit (BtuIT/F) }
}
\value{joule per kelvin (J/K) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnITPerDegFtOjoulePerK(10)
}
\keyword{programming}