\name{NISTukThUnITPerDegRankineTOjoulePerK}
\alias{NISTukThUnITPerDegRankineTOjoulePerK}
\title{Convert British thermal unitIT per degree Rankine  to joule per kelvin }
\usage{NISTukThUnITPerDegRankineTOjoulePerK(ukThUnITPerDegRankine)}
\description{\code{NISTukThUnITPerDegRankineTOjoulePerK} converts from British thermal unitIT per degree Rankine (BtuIT/R) to joule per kelvin (J/K) }
\arguments{
  \item{ukThUnITPerDegRankine}{British thermal unitIT per degree Rankine (BtuIT/R) }
}
\value{joule per kelvin (J/K) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnITPerDegRankineTOjoulePerK(10)
}
\keyword{programming}