\name{NISTukThUnITPerHourSqrFtDegFtOwattPerSqrMeterK}
\alias{NISTukThUnITPerHourSqrFtDegFtOwattPerSqrMeterK}
\title{Convert British thermal unitIT per hour square foot degree Fahrenheit  to watt per square meter kelvin }
\usage{NISTukThUnITPerHourSqrFtDegFtOwattPerSqrMeterK(ukThUnITPerHourSqrFtDegF)}
\description{\code{NISTukThUnITPerHourSqrFtDegFtOwattPerSqrMeterK} converts from British thermal unitIT per hour square foot degree Fahrenheit [BtuIT/(h * ft2 * F)] to watt per square meter kelvin [W/(m2 * K)] }
\arguments{
  \item{ukThUnITPerHourSqrFtDegF}{British thermal unitIT per hour square foot degree Fahrenheit [BtuIT/(h * ft2 * F)] }
}
\value{watt per square meter kelvin [W/(m2 * K)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnITPerHourSqrFtDegFtOwattPerSqrMeterK(10)
}
\keyword{programming}