\name{NISTukThUnITPerHourTOwatt}
\alias{NISTukThUnITPerHourTOwatt}
\title{Convert British thermal unitIT per hour  to watt }
\usage{NISTukThUnITPerHourTOwatt(ukThUnITPerHour)}
\description{\code{NISTukThUnITPerHourTOwatt} converts from British thermal unitIT per hour (BtuIT/h) to watt (W) }
\arguments{
  \item{ukThUnITPerHour}{British thermal unitIT per hour (BtuIT/h) }
}
\value{watt (W) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnITPerHourTOwatt(10)
}
\keyword{programming}