\name{NISTukThUnITPerPoundDegFtOjoulePerKgK}
\alias{NISTukThUnITPerPoundDegFtOjoulePerKgK}
\title{Convert British thermal unitIT per pound degree Fahrenheit  to joule per kilogram kelvin }
\usage{NISTukThUnITPerPoundDegFtOjoulePerKgK(ukThUnITPerPoundDegF)}
\description{\code{NISTukThUnITPerPoundDegFtOjoulePerKgK} converts from British thermal unitIT per pound degree Fahrenheit [BtuIT/(lb * F)] to joule per kilogram kelvin [J/(kg * K)] }
\arguments{
  \item{ukThUnITPerPoundDegF}{British thermal unitIT per pound degree Fahrenheit [BtuIT/(lb * F)] }
}
\value{joule per kilogram kelvin [J/(kg * K)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnITPerPoundDegFtOjoulePerKgK(10)
}
\keyword{programming}