\name{NISTukThUnITPerPoundDegRankineTOjoulePerKgK}
\alias{NISTukThUnITPerPoundDegRankineTOjoulePerKgK}
\title{Convert British thermal unitIT per pound degree Rankine  to joule per kilogram kelvin }
\usage{NISTukThUnITPerPoundDegRankineTOjoulePerKgK(ukThUnITPerPoundDegRankine)}
\description{\code{NISTukThUnITPerPoundDegRankineTOjoulePerKgK} converts from British thermal unitIT per pound degree Rankine [BtuIT/(lb * R)] to joule per kilogram kelvin [J/(kg * K)] }
\arguments{
  \item{ukThUnITPerPoundDegRankine}{British thermal unitIT per pound degree Rankine [BtuIT/(lb * R)] }
}
\value{joule per kilogram kelvin [J/(kg * K)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnITPerPoundDegRankineTOjoulePerKgK(10)
}
\keyword{programming}