\name{NISTukThUnITPerPoundTOjoulePerKg}
\alias{NISTukThUnITPerPoundTOjoulePerKg}
\title{Convert British thermal unitIT per pound  to joule per kilogram }
\usage{NISTukThUnITPerPoundTOjoulePerKg(ukThUnITPerPound)}
\description{\code{NISTukThUnITPerPoundTOjoulePerKg} converts from British thermal unitIT per pound (BtuIT/lb) to joule per kilogram (J/kg) }
\arguments{
  \item{ukThUnITPerPound}{British thermal unitIT per pound (BtuIT/lb) }
}
\value{joule per kilogram (J/kg) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnITPerPoundTOjoulePerKg(10)
}
\keyword{programming}