\name{NISTukThUnITPerSecTOwatt}
\alias{NISTukThUnITPerSecTOwatt}
\title{Convert British thermal unitIT per second  to watt }
\usage{NISTukThUnITPerSecTOwatt(ukThUnITPerSec)}
\description{\code{NISTukThUnITPerSecTOwatt} converts from British thermal unitIT per second (BtuIT/s) to watt (W) }
\arguments{
  \item{ukThUnITPerSec}{British thermal unitIT per second (BtuIT/s) }
}
\value{watt (W) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnITPerSecTOwatt(10)
}
\keyword{programming}