\name{NISTukThUnITPerSqrFtHourTOwattPerSqrMeter}
\alias{NISTukThUnITPerSqrFtHourTOwattPerSqrMeter}
\title{Convert British thermal unitIT per square foot hour  to watt per square meter }
\usage{NISTukThUnITPerSqrFtHourTOwattPerSqrMeter(ukThUnITPerSqrFtHour)}
\description{\code{NISTukThUnITPerSqrFtHourTOwattPerSqrMeter} converts from British thermal unitIT per square foot hour [BtuIT/(ft2 * h)] to watt per square meter (W/m2) }
\arguments{
  \item{ukThUnITPerSqrFtHour}{British thermal unitIT per square foot hour [BtuIT/(ft2 * h)] }
}
\value{watt per square meter (W/m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnITPerSqrFtHourTOwattPerSqrMeter(10)
}
\keyword{programming}