\name{NISTukThUnITPerSqrFtSecTOwattPerSqrMeter}
\alias{NISTukThUnITPerSqrFtSecTOwattPerSqrMeter}
\title{Convert British thermal unitIT per square foot second  to watt per square meter }
\usage{NISTukThUnITPerSqrFtSecTOwattPerSqrMeter(ukThUnITPerSqrFtSec)}
\description{\code{NISTukThUnITPerSqrFtSecTOwattPerSqrMeter} converts from British thermal unitIT per square foot second [BtuIT/(ft2 * s)] to watt per square meter (W/m2) }
\arguments{
  \item{ukThUnITPerSqrFtSec}{British thermal unitIT per square foot second [BtuIT/(ft2 * s)] }
}
\value{watt per square meter (W/m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnITPerSqrFtSecTOwattPerSqrMeter(10)
}
\keyword{programming}