\name{NISTukThUnITPerSqrFtTOjoulePerSqrMeter}
\alias{NISTukThUnITPerSqrFtTOjoulePerSqrMeter}
\title{Convert British thermal unitIT per square foot  to joule per square meter }
\usage{NISTukThUnITPerSqrFtTOjoulePerSqrMeter(ukThUnITPerSqrFt)}
\description{\code{NISTukThUnITPerSqrFtTOjoulePerSqrMeter} converts from British thermal unitIT per square foot (BtuIT/ft2) to joule per square meter (J/m2) }
\arguments{
  \item{ukThUnITPerSqrFt}{British thermal unitIT per square foot (BtuIT/ft2) }
}
\value{joule per square meter (J/m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnITPerSqrFtTOjoulePerSqrMeter(10)
}
\keyword{programming}