\name{NISTukThUnITtOjoule}
\alias{NISTukThUnITtOjoule}
\title{Convert British thermal unitIT 9 to joule }
\usage{NISTukThUnITtOjoule(ukThUnIT)}
\description{\code{NISTukThUnITtOjoule} converts from British thermal unitIT (BtuIT) 9 to joule (J) }
\arguments{
  \item{ukThUnIT}{British thermal unitIT (BtuIT) 9 }
}
\value{joule (J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnITtOjoule(10)
}
\keyword{programming}