\name{NISTukThUnthInchPerHourSqrFtDegFtOwattPerMeterK}
\alias{NISTukThUnthInchPerHourSqrFtDegFtOwattPerMeterK}
\title{Convert British thermal unitth inch per hour square foot degree Fahrenheit  to watt per meter kelvin }
\usage{NISTukThUnthInchPerHourSqrFtDegFtOwattPerMeterK(ukThUnthInchPerHourSqrFtDegF)}
\description{\code{NISTukThUnthInchPerHourSqrFtDegFtOwattPerMeterK} converts from British thermal unitth inch per hour square foot degree Fahrenheit [Btuth * in/(h * ft2 * F)] to watt per meter kelvin [W / (m * K)] }
\arguments{
  \item{ukThUnthInchPerHourSqrFtDegF}{British thermal unitth inch per hour square foot degree Fahrenheit [Btuth * in/(h * ft2 * F)] }
}
\value{watt per meter kelvin [W / (m * K)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnthInchPerHourSqrFtDegFtOwattPerMeterK(10)
}
\keyword{programming}