\name{NISTukThUnthPerCubFtTOjoulePerCubMeter}
\alias{NISTukThUnthPerCubFtTOjoulePerCubMeter}
\title{Convert British thermal unitth per cubic foot  to joule per cubic meter }
\usage{NISTukThUnthPerCubFtTOjoulePerCubMeter(ukThUnthPerCubFt)}
\description{\code{NISTukThUnthPerCubFtTOjoulePerCubMeter} converts from British thermal unitth per cubic foot (Btuth/ft3) to joule per cubic meter (J/m3) }
\arguments{
  \item{ukThUnthPerCubFt}{British thermal unitth per cubic foot (Btuth/ft3) }
}
\value{joule per cubic meter (J/m3) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnthPerCubFtTOjoulePerCubMeter(10)
}
\keyword{programming}