\name{NISTukThUnthPerDegFtOjoulePerK}
\alias{NISTukThUnthPerDegFtOjoulePerK}
\title{Convert British thermal unitth per degree Fahrenheit  to joule per kelvin }
\usage{NISTukThUnthPerDegFtOjoulePerK(ukThUnthPerDegF)}
\description{\code{NISTukThUnthPerDegFtOjoulePerK} converts from British thermal unitth per degree Fahrenheit (Btuth/F) to joule per kelvin (J/K) }
\arguments{
  \item{ukThUnthPerDegF}{British thermal unitth per degree Fahrenheit (Btuth/F) }
}
\value{joule per kelvin (J/K) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnthPerDegFtOjoulePerK(10)
}
\keyword{programming}