\name{NISTukThUnthPerDegRankineTOjoulePerK}
\alias{NISTukThUnthPerDegRankineTOjoulePerK}
\title{Convert British thermal unitth per degree Rankine  to joule per kelvin }
\usage{NISTukThUnthPerDegRankineTOjoulePerK(ukThUnthPerDegRankine)}
\description{\code{NISTukThUnthPerDegRankineTOjoulePerK} converts from British thermal unitth per degree Rankine (Btuth/R) to joule per kelvin (J/K) }
\arguments{
  \item{ukThUnthPerDegRankine}{British thermal unitth per degree Rankine (Btuth/R) }
}
\value{joule per kelvin (J/K) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnthPerDegRankineTOjoulePerK(10)
}
\keyword{programming}