\name{NISTukThUnthPerHourSqrFtDegFtOwattPerSqrMeterK}
\alias{NISTukThUnthPerHourSqrFtDegFtOwattPerSqrMeterK}
\title{Convert British thermal unitth per hour square foot degree Fahrenheit  to watt per square meter kelvin }
\usage{NISTukThUnthPerHourSqrFtDegFtOwattPerSqrMeterK(ukThUnthPerHourSqrFtDegF)}
\description{\code{NISTukThUnthPerHourSqrFtDegFtOwattPerSqrMeterK} converts from British thermal unitth per hour square foot degree Fahrenheit [Btuth/(h * ft2 * F)] to watt per square meter kelvin [W/(m2 * K)] }
\arguments{
  \item{ukThUnthPerHourSqrFtDegF}{British thermal unitth per hour square foot degree Fahrenheit [Btuth/(h * ft2 * F)] }
}
\value{watt per square meter kelvin [W/(m2 * K)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnthPerHourSqrFtDegFtOwattPerSqrMeterK(10)
}
\keyword{programming}