\name{NISTukThUnthPerHourTOwatt}
\alias{NISTukThUnthPerHourTOwatt}
\title{Convert British thermal unitth per hour  to watt }
\usage{NISTukThUnthPerHourTOwatt(ukThUnthPerHour)}
\description{\code{NISTukThUnthPerHourTOwatt} converts from British thermal unitth per hour (Btuth/h) to watt (W) }
\arguments{
  \item{ukThUnthPerHour}{British thermal unitth per hour (Btuth/h) }
}
\value{watt (W) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnthPerHourTOwatt(10)
}
\keyword{programming}