\name{NISTukThUnthPerMinTOwatt}
\alias{NISTukThUnthPerMinTOwatt}
\title{Convert British thermal unitth per minute  to watt }
\usage{NISTukThUnthPerMinTOwatt(ukThUnthPerMin)}
\description{\code{NISTukThUnthPerMinTOwatt} converts from British thermal unitth per minute (Btuth/min) to watt (W) }
\arguments{
  \item{ukThUnthPerMin}{British thermal unitth per minute (Btuth/min) }
}
\value{watt (W) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnthPerMinTOwatt(10)
}
\keyword{programming}