\name{NISTukThUnthPerPoundDegFtOjoulePerKgK}
\alias{NISTukThUnthPerPoundDegFtOjoulePerKgK}
\title{Convert British thermal unitth per pound degree Fahrenheit  to joule per kilogram kelvin }
\usage{NISTukThUnthPerPoundDegFtOjoulePerKgK(ukThUnthPerPoundDegF)}
\description{\code{NISTukThUnthPerPoundDegFtOjoulePerKgK} converts from British thermal unitth per pound degree Fahrenheit [Btuth/(lb * F)] to joule per kilogram kelvin [J/(kg * K)] }
\arguments{
  \item{ukThUnthPerPoundDegF}{British thermal unitth per pound degree Fahrenheit [Btuth/(lb * F)] }
}
\value{joule per kilogram kelvin [J/(kg * K)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnthPerPoundDegFtOjoulePerKgK(10)
}
\keyword{programming}