\name{NISTukThUnthPerPoundDegRankineTOjoulePerKgK}
\alias{NISTukThUnthPerPoundDegRankineTOjoulePerKgK}
\title{Convert British thermal unitth per pound degree Rankine  to joule per kilogram kelvin }
\usage{NISTukThUnthPerPoundDegRankineTOjoulePerKgK(ukThUnthPerPoundDegRankine)}
\description{\code{NISTukThUnthPerPoundDegRankineTOjoulePerKgK} converts from British thermal unitth per pound degree Rankine [Btuth/(lb * R)] to joule per kilogram kelvin [J/(kg * K)] }
\arguments{
  \item{ukThUnthPerPoundDegRankine}{British thermal unitth per pound degree Rankine [Btuth/(lb * R)] }
}
\value{joule per kilogram kelvin [J/(kg * K)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnthPerPoundDegRankineTOjoulePerKgK(10)
}
\keyword{programming}