\name{NISTukThUnthPerPoundTOjoulePerKg}
\alias{NISTukThUnthPerPoundTOjoulePerKg}
\title{Convert British thermal unitth per pound  to joule per kilogram }
\usage{NISTukThUnthPerPoundTOjoulePerKg(ukThUnthPerPound)}
\description{\code{NISTukThUnthPerPoundTOjoulePerKg} converts from British thermal unitth per pound (Btuth/lb) to joule per kilogram (J/kg) }
\arguments{
  \item{ukThUnthPerPound}{British thermal unitth per pound (Btuth/lb) }
}
\value{joule per kilogram (J/kg) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnthPerPoundTOjoulePerKg(10)
}
\keyword{programming}