\name{NISTukThUnthPerSecSqrFtDegFtOwattPerSqrMeterK}
\alias{NISTukThUnthPerSecSqrFtDegFtOwattPerSqrMeterK}
\title{Convert British thermal unitth per second square foot degree Fahrenheit  to watt per square meter kelvin }
\usage{NISTukThUnthPerSecSqrFtDegFtOwattPerSqrMeterK(ukThUnthPerSecSqrFtDegF)}
\description{\code{NISTukThUnthPerSecSqrFtDegFtOwattPerSqrMeterK} converts from British thermal unitth per second square foot degree Fahrenheit [Btuth/(s * ft2 * F)] to watt per square meter kelvin [W/(m2 * K)] }
\arguments{
  \item{ukThUnthPerSecSqrFtDegF}{British thermal unitth per second square foot degree Fahrenheit [Btuth/(s * ft2 * F)] }
}
\value{watt per square meter kelvin [W/(m2 * K)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnthPerSecSqrFtDegFtOwattPerSqrMeterK(10)
}
\keyword{programming}