\name{NISTukThUnthPerSecTOwatt}
\alias{NISTukThUnthPerSecTOwatt}
\title{Convert British thermal unitth per second  to watt }
\usage{NISTukThUnthPerSecTOwatt(ukThUnthPerSec)}
\description{\code{NISTukThUnthPerSecTOwatt} converts from British thermal unitth per second (Btuth/s) to watt (W) }
\arguments{
  \item{ukThUnthPerSec}{British thermal unitth per second (Btuth/s) }
}
\value{watt (W) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnthPerSecTOwatt(10)
}
\keyword{programming}