\name{NISTukThUnthPerSqrFtHourTOwattPerSqrMeter}
\alias{NISTukThUnthPerSqrFtHourTOwattPerSqrMeter}
\title{Convert British thermal unitth per square foot hour  to watt per square meter }
\usage{NISTukThUnthPerSqrFtHourTOwattPerSqrMeter(ukThUnthPerSqrFtHour)}
\description{\code{NISTukThUnthPerSqrFtHourTOwattPerSqrMeter} converts from British thermal unitth per square foot hour [Btuth/(ft2 * h)] to watt per square meter (W/m2) }
\arguments{
  \item{ukThUnthPerSqrFtHour}{British thermal unitth per square foot hour [Btuth/(ft2 * h)] }
}
\value{watt per square meter (W/m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnthPerSqrFtHourTOwattPerSqrMeter(10)
}
\keyword{programming}