\name{NISTukThUnthPerSqrFtMinTOwattPerSqrMeter}
\alias{NISTukThUnthPerSqrFtMinTOwattPerSqrMeter}
\title{Convert British thermal unitth per square foot minute  to watt per square meter }
\usage{NISTukThUnthPerSqrFtMinTOwattPerSqrMeter(ukThUnthPerSqrFtMin)}
\description{\code{NISTukThUnthPerSqrFtMinTOwattPerSqrMeter} converts from British thermal unitth per square foot minute [Btuth/(ft2 * min)] to watt per square meter (W/m2) }
\arguments{
  \item{ukThUnthPerSqrFtMin}{British thermal unitth per square foot minute [Btuth/(ft2 * min)] }
}
\value{watt per square meter (W/m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnthPerSqrFtMinTOwattPerSqrMeter(10)
}
\keyword{programming}