\name{NISTukThUnthPerSqrFtSecTOwattPerSqrMeter}
\alias{NISTukThUnthPerSqrFtSecTOwattPerSqrMeter}
\title{Convert British thermal unitth per square foot second  to watt per square meter }
\usage{NISTukThUnthPerSqrFtSecTOwattPerSqrMeter(ukThUnthPerSqrFtSec)}
\description{\code{NISTukThUnthPerSqrFtSecTOwattPerSqrMeter} converts from British thermal unitth per square foot second [Btuth/(ft2 * s)] to watt per square meter (W/m2) }
\arguments{
  \item{ukThUnthPerSqrFtSec}{British thermal unitth per square foot second [Btuth/(ft2 * s)] }
}
\value{watt per square meter (W/m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnthPerSqrFtSecTOwattPerSqrMeter(10)
}
\keyword{programming}