\name{NISTukThUnthPerSqrFtTOjoulePerSqrMeter}
\alias{NISTukThUnthPerSqrFtTOjoulePerSqrMeter}
\title{Convert British thermal unitth per square foot  to joule per square meter }
\usage{NISTukThUnthPerSqrFtTOjoulePerSqrMeter(ukThUnthPerSqrFt)}
\description{\code{NISTukThUnthPerSqrFtTOjoulePerSqrMeter} converts from British thermal unitth per square foot (Btuth/ft2) to joule per square meter (J/m2) }
\arguments{
  \item{ukThUnthPerSqrFt}{British thermal unitth per square foot (Btuth/ft2) }
}
\value{joule per square meter (J/m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnthPerSqrFtTOjoulePerSqrMeter(10)
}
\keyword{programming}