\name{NISTukThUnthPerSqrInchSecTOwattPerSqrMeter}
\alias{NISTukThUnthPerSqrInchSecTOwattPerSqrMeter}
\title{Convert British thermal unitth per square inch second  to watt per square meter }
\usage{NISTukThUnthPerSqrInchSecTOwattPerSqrMeter(ukThUnthPerSqrInchSec)}
\description{\code{NISTukThUnthPerSqrInchSecTOwattPerSqrMeter} converts from British thermal unitth per square inch second [Btuth/(in2 * s)] to watt per square meter (W/m2) }
\arguments{
  \item{ukThUnthPerSqrInchSec}{British thermal unitth per square inch second [Btuth/(in2 * s)] }
}
\value{watt per square meter (W/m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnthPerSqrInchSecTOwattPerSqrMeter(10)
}
\keyword{programming}