\name{NISTukThUnthTOjoule}
\alias{NISTukThUnthTOjoule}
\title{Convert British thermal unitth 9 to joule }
\usage{NISTukThUnthTOjoule(ukThUnth)}
\description{\code{NISTukThUnthTOjoule} converts from British thermal unitth (Btuth) 9 to joule (J) }
\arguments{
  \item{ukThUnth}{British thermal unitth (Btuth) 9 }
}
\value{joule (J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnthTOjoule(10)
}
\keyword{programming}