\name{NISTukThUnTOjoule}
\alias{NISTukThUnTOjoule}
\title{Convert British thermal unit  to joule }
\usage{NISTukThUnTOjoule(ukThUn)}
\description{\code{NISTukThUnTOjoule} converts from British thermal unit (mean) (Btu) to joule (J) }
\arguments{
  \item{ukThUn}{British thermal unit (mean) (Btu) }
}
\value{joule (J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTukThUnTOjoule(10)
}
\keyword{programming}