\name{NISTunPoleTOweber}
\alias{NISTunPoleTOweber}
\title{Convert unit pole to weber }
\usage{NISTunPoleTOweber(unPole)}
\description{\code{NISTunPoleTOweber} converts from unit pole to weber (Wb) }
\arguments{
  \item{unPole}{unit pole }
}
\value{weber (Wb) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTunPoleTOweber(10)
}
\keyword{programming}