\name{NISTvoltTOabvolt}
\alias{NISTvoltTOabvolt}
\title{Convert volt  to abvolt}
\usage{NISTvoltTOabvolt(volt)}
\description{\code{NISTvoltTOabvolt} converts from volt (V) to abvolt }
\arguments{
  \item{volt}{volt (V) }
}
\value{abvolt }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTvoltTOabvolt(10)
}
\keyword{programming}