\name{NISTvoltToEMUOfElectricPotential}
\alias{NISTvoltToEMUOfElectricPotential}
\title{Convert volt  to EMU of electric potential }
\usage{NISTvoltToEMUOfElectricPotential(volt)}
\description{\code{NISTvoltToEMUOfElectricPotential} converts from volt (V) to EMU of electric potential (abvolt) }
\arguments{
  \item{volt}{volt (V) }
}
\value{EMU of electric potential (abvolt) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTvoltToEMUOfElectricPotential(10)
}
\keyword{programming}