\name{NISTvoltToESUOfElectricPotential}
\alias{NISTvoltToESUOfElectricPotential}
\title{Convert volt  to ESU of electric potential }
\usage{NISTvoltToESUOfElectricPotential(volt)}
\description{\code{NISTvoltToESUOfElectricPotential} converts from volt (V) to ESU of electric potential (statvolt) }
\arguments{
  \item{volt}{volt (V) }
}
\value{ESU of electric potential (statvolt) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTvoltToESUOfElectricPotential(10)
}
\keyword{programming}