\name{NISTvoltTOstatvolt}
\alias{NISTvoltTOstatvolt}
\title{Convert volt  to statvolt}
\usage{NISTvoltTOstatvolt(volt)}
\description{\code{NISTvoltTOstatvolt} converts from volt (V) to statvolt }
\arguments{
  \item{volt}{volt (V) }
}
\value{statvolt }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTvoltTOstatvolt(10)
}
\keyword{programming}