\name{NISTwattHourTOjoule}
\alias{NISTwattHourTOjoule}
\title{Convert watt hour  to joule }
\usage{NISTwattHourTOjoule(wattHour)}
\description{\code{NISTwattHourTOjoule} converts from watt hour (W * h) to joule (J) }
\arguments{
  \item{wattHour}{watt hour (W * h) }
}
\value{joule (J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattHourTOjoule(10)
}
\keyword{programming}