\name{NISTwattPerMeterKtOcalthPerCmSecDegC}
\alias{NISTwattPerMeterKtOcalthPerCmSecDegC}
\title{Convert watt per meter kelvin  to calorieth per centimeter second degree Celsius }
\usage{NISTwattPerMeterKtOcalthPerCmSecDegC(wattPerMeterK)}
\description{\code{NISTwattPerMeterKtOcalthPerCmSecDegC} converts from watt per meter kelvin [W/(m * K)] to calorieth per centimeter second degree Celsius [calth/(cm * s * C)] }
\arguments{
  \item{wattPerMeterK}{watt per meter kelvin [W/(m * K)] }
}
\value{calorieth per centimeter second degree Celsius [calth/(cm * s * C)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattPerMeterKtOcalthPerCmSecDegC(10)
}
\keyword{programming}