\name{NISTwattPerMeterKtOukThUnITFtPerHourSqrFtDegF}
\alias{NISTwattPerMeterKtOukThUnITFtPerHourSqrFtDegF}
\title{Convert watt per meter kelvin  to British thermal unitIT foot per hour square foot degree Fahrenheit }
\usage{NISTwattPerMeterKtOukThUnITFtPerHourSqrFtDegF(wattPerMeterK)}
\description{\code{NISTwattPerMeterKtOukThUnITFtPerHourSqrFtDegF} converts from watt per meter kelvin [W/(m * K)] to British thermal unitIT foot per hour square foot degree Fahrenheit [BtuIT * ft/(h * ft2 * F)] }
\arguments{
  \item{wattPerMeterK}{watt per meter kelvin [W/(m * K)] }
}
\value{British thermal unitIT foot per hour square foot degree Fahrenheit [BtuIT * ft/(h * ft2 * F)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattPerMeterKtOukThUnITFtPerHourSqrFtDegF(10)
}
\keyword{programming}