\name{NISTwattPerMeterKtOukThUnthFtPerHourSqrFtDegF}
\alias{NISTwattPerMeterKtOukThUnthFtPerHourSqrFtDegF}
\title{Convert watt per meter kelvin  to British thermal unitth foot per hour square foot degree Fahrenheit }
\usage{NISTwattPerMeterKtOukThUnthFtPerHourSqrFtDegF(wattPerMeterK)}
\description{\code{NISTwattPerMeterKtOukThUnthFtPerHourSqrFtDegF} converts from watt per meter kelvin [W/(m * K)] to British thermal unitth foot per hour square foot degree Fahrenheit [Btuth * ft/(h * ft2 * F)] }
\arguments{
  \item{wattPerMeterK}{watt per meter kelvin [W/(m * K)] }
}
\value{British thermal unitth foot per hour square foot degree Fahrenheit [Btuth * ft/(h * ft2 * F)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattPerMeterKtOukThUnthFtPerHourSqrFtDegF(10)
}
\keyword{programming}