\name{NISTwattPerMeterKtOukThUnthInchPerSecSqrFtDegF}
\alias{NISTwattPerMeterKtOukThUnthInchPerSecSqrFtDegF}
\title{Convert watt per meter kelvin  to British thermal unitth inch per second square foot degree Fahrenheit }
\usage{NISTwattPerMeterKtOukThUnthInchPerSecSqrFtDegF(wattPerMeterK)}
\description{\code{NISTwattPerMeterKtOukThUnthInchPerSecSqrFtDegF} converts from watt per meter kelvin [W/(m * K)] to British thermal unitth inch per second square foot degree Fahrenheit [Btuth * in /(s * ft2 * F)] }
\arguments{
  \item{wattPerMeterK}{watt per meter kelvin [W/(m * K)] }
}
\value{British thermal unitth inch per second square foot degree Fahrenheit [Btuth * in /(s * ft2 * F)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattPerMeterKtOukThUnthInchPerSecSqrFtDegF(10)
}
\keyword{programming}