\name{NISTwattPerSqrCmTOwattPerSqrMeter}
\alias{NISTwattPerSqrCmTOwattPerSqrMeter}
\title{Convert watt per square centimeter  to watt per square meter }
\usage{NISTwattPerSqrCmTOwattPerSqrMeter(wattPerSqrCm)}
\description{\code{NISTwattPerSqrCmTOwattPerSqrMeter} converts from watt per square centimeter (W/cm2) to watt per square meter (W/m2) }
\arguments{
  \item{wattPerSqrCm}{watt per square centimeter (W/cm2) }
}
\value{watt per square meter (W/m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattPerSqrCmTOwattPerSqrMeter(10)
}
\keyword{programming}