\name{NISTwattPerSqrInchTOwattPerSqrMeter}
\alias{NISTwattPerSqrInchTOwattPerSqrMeter}
\title{Convert watt per square inch  to watt per square meter }
\usage{NISTwattPerSqrInchTOwattPerSqrMeter(wattPerSqrInch)}
\description{\code{NISTwattPerSqrInchTOwattPerSqrMeter} converts from watt per square inch (W/in2) to watt per square meter (W/m2) }
\arguments{
  \item{wattPerSqrInch}{watt per square inch (W/in2) }
}
\value{watt per square meter (W/m2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattPerSqrInchTOwattPerSqrMeter(10)
}
\keyword{programming}