\name{NISTwattPerSqrMeterKtOukThUnITPerSecSqrFtDegF}
\alias{NISTwattPerSqrMeterKtOukThUnITPerSecSqrFtDegF}
\title{Convert watt per square meter kelvin  to British thermal unitIT per second square foot degree Fahrenheit }
\usage{NISTwattPerSqrMeterKtOukThUnITPerSecSqrFtDegF(wattPerSqrMeterK)}
\description{\code{NISTwattPerSqrMeterKtOukThUnITPerSecSqrFtDegF} converts from watt per square meter kelvin [W/(m2 * K)] to British thermal unitIT per second square foot degree Fahrenheit [BtuIT/(s * ft2 * F)] }
\arguments{
  \item{wattPerSqrMeterK}{watt per square meter kelvin [W/(m2 * K)] }
}
\value{British thermal unitIT per second square foot degree Fahrenheit [BtuIT/(s * ft2 * F)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattPerSqrMeterKtOukThUnITPerSecSqrFtDegF(10)
}
\keyword{programming}