\name{NISTwattPerSqrMeterKtOukThUnthPerHourSqrFtDegF}
\alias{NISTwattPerSqrMeterKtOukThUnthPerHourSqrFtDegF}
\title{Convert watt per square meter kelvin  to British thermal unitth per hour square foot degree Fahrenheit }
\usage{NISTwattPerSqrMeterKtOukThUnthPerHourSqrFtDegF(wattPerSqrMeterK)}
\description{\code{NISTwattPerSqrMeterKtOukThUnthPerHourSqrFtDegF} converts from watt per square meter kelvin [W/(m2 * K)] to British thermal unitth per hour square foot degree Fahrenheit [Btuth/(h * ft2 * F)] }
\arguments{
  \item{wattPerSqrMeterK}{watt per square meter kelvin [W/(m2 * K)] }
}
\value{British thermal unitth per hour square foot degree Fahrenheit [Btuth/(h * ft2 * F)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattPerSqrMeterKtOukThUnthPerHourSqrFtDegF(10)
}
\keyword{programming}