\name{NISTwattPerSqrMeterTOcalthPerSqrCmSec}
\alias{NISTwattPerSqrMeterTOcalthPerSqrCmSec}
\title{Convert watt per square meter  to calorieth per square centimeter second }
\usage{NISTwattPerSqrMeterTOcalthPerSqrCmSec(wattPerSqrMeter)}
\description{\code{NISTwattPerSqrMeterTOcalthPerSqrCmSec} converts from watt per square meter (W/m2) to calorieth per square centimeter second [calth/(cm2 * s)] }
\arguments{
  \item{wattPerSqrMeter}{watt per square meter (W/m2) }
}
\value{calorieth per square centimeter second [calth/(cm2 * s)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattPerSqrMeterTOcalthPerSqrCmSec(10)
}
\keyword{programming}