\name{NISTwattPerSqrMeterTOergPerSqrCmSec}
\alias{NISTwattPerSqrMeterTOergPerSqrCmSec}
\title{Convert watt per square meter  to erg per square centimeter second }
\usage{NISTwattPerSqrMeterTOergPerSqrCmSec(wattPerSqrMeter)}
\description{\code{NISTwattPerSqrMeterTOergPerSqrCmSec} converts from watt per square meter (W/m2) to erg per square centimeter second [erg/(cm2 * s)] }
\arguments{
  \item{wattPerSqrMeter}{watt per square meter (W/m2) }
}
\value{erg per square centimeter second [erg/(cm2 * s)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattPerSqrMeterTOergPerSqrCmSec(10)
}
\keyword{programming}