\name{NISTwattPerSqrMeterTOukThUnITPerSqrFtHour}
\alias{NISTwattPerSqrMeterTOukThUnITPerSqrFtHour}
\title{Convert watt per square meter  to British thermal unitIT per square foot hour }
\usage{NISTwattPerSqrMeterTOukThUnITPerSqrFtHour(wattPerSqrMeter)}
\description{\code{NISTwattPerSqrMeterTOukThUnITPerSqrFtHour} converts from watt per square meter (W/m2) to British thermal unitIT per square foot hour [BtuIT/(ft2 * h)] }
\arguments{
  \item{wattPerSqrMeter}{watt per square meter (W/m2) }
}
\value{British thermal unitIT per square foot hour [BtuIT/(ft2 * h)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattPerSqrMeterTOukThUnITPerSqrFtHour(10)
}
\keyword{programming}