\name{NISTwattPerSqrMeterTOukThUnthPerSqrFtMin}
\alias{NISTwattPerSqrMeterTOukThUnthPerSqrFtMin}
\title{Convert watt per square meter  to British thermal unitth per square foot minute }
\usage{NISTwattPerSqrMeterTOukThUnthPerSqrFtMin(wattPerSqrMeter)}
\description{\code{NISTwattPerSqrMeterTOukThUnthPerSqrFtMin} converts from watt per square meter (W/m2) to British thermal unitth per square foot minute [Btuth/(ft2 * min)] }
\arguments{
  \item{wattPerSqrMeter}{watt per square meter (W/m2) }
}
\value{British thermal unitth per square foot minute [Btuth/(ft2 * min)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattPerSqrMeterTOukThUnthPerSqrFtMin(10)
}
\keyword{programming}