\name{NISTwattPerSqrMeterTOukThUnthPerSqrInchSec}
\alias{NISTwattPerSqrMeterTOukThUnthPerSqrInchSec}
\title{Convert watt per square meter  to British thermal unitth per square inch second }
\usage{NISTwattPerSqrMeterTOukThUnthPerSqrInchSec(wattPerSqrMeter)}
\description{\code{NISTwattPerSqrMeterTOukThUnthPerSqrInchSec} converts from watt per square meter (W/m2) to British thermal unitth per square inch second [Btuth/(in2 * s)] }
\arguments{
  \item{wattPerSqrMeter}{watt per square meter (W/m2) }
}
\value{British thermal unitth per square inch second [Btuth/(in2 * s)] }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattPerSqrMeterTOukThUnthPerSqrInchSec(10)
}
\keyword{programming}