\name{NISTwattPerSqrMeterTOwattPerSqrInch}
\alias{NISTwattPerSqrMeterTOwattPerSqrInch}
\title{Convert watt per square meter  to watt per square inch }
\usage{NISTwattPerSqrMeterTOwattPerSqrInch(wattPerSqrMeter)}
\description{\code{NISTwattPerSqrMeterTOwattPerSqrInch} converts from watt per square meter (W/m2) to watt per square inch (W/in2) }
\arguments{
  \item{wattPerSqrMeter}{watt per square meter (W/m2) }
}
\value{watt per square inch (W/in2) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattPerSqrMeterTOwattPerSqrInch(10)
}
\keyword{programming}