\name{NISTwattSecTOjoule}
\alias{NISTwattSecTOjoule}
\title{Convert watt second  to joule }
\usage{NISTwattSecTOjoule(wattSec)}
\description{\code{NISTwattSecTOjoule} converts from watt second (W * s) to joule (J) }
\arguments{
  \item{wattSec}{watt second (W * s) }
}
\value{joule (J) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattSecTOjoule(10)
}
\keyword{programming}