\name{NISTwattTOcalthPerMin}
\alias{NISTwattTOcalthPerMin}
\title{Convert watt  to calorieth per minute }
\usage{NISTwattTOcalthPerMin(watt)}
\description{\code{NISTwattTOcalthPerMin} converts from watt (W) to calorieth per minute (calth/min) }
\arguments{
  \item{watt}{watt (W) }
}
\value{calorieth per minute (calth/min) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattTOcalthPerMin(10)
}
\keyword{programming}