\name{NISTwattTOcalthPerSec}
\alias{NISTwattTOcalthPerSec}
\title{Convert watt  to calorieth per second }
\usage{NISTwattTOcalthPerSec(watt)}
\description{\code{NISTwattTOcalthPerSec} converts from watt (W) to calorieth per second (calth/s) }
\arguments{
  \item{watt}{watt (W) }
}
\value{calorieth per second (calth/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattTOcalthPerSec(10)
}
\keyword{programming}