\name{NISTwattTOergPerSec}
\alias{NISTwattTOergPerSec}
\title{Convert watt  to erg per second }
\usage{NISTwattTOergPerSec(watt)}
\description{\code{NISTwattTOergPerSec} converts from watt (W) to erg per second (erg/s) }
\arguments{
  \item{watt}{watt (W) }
}
\value{erg per second (erg/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattTOergPerSec(10)
}
\keyword{programming}