\name{NISTwattTOftPoundForcePerHour}
\alias{NISTwattTOftPoundForcePerHour}
\title{Convert watt  to foot pound-force per hour }
\usage{NISTwattTOftPoundForcePerHour(watt)}
\description{\code{NISTwattTOftPoundForcePerHour} converts from watt (W) to foot pound-force per hour (ft * lbf/h) }
\arguments{
  \item{watt}{watt (W) }
}
\value{foot pound-force per hour (ft * lbf/h) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattTOftPoundForcePerHour(10)
}
\keyword{programming}