\name{NISTwattTOftPoundForcePerMin}
\alias{NISTwattTOftPoundForcePerMin}
\title{Convert watt  to foot pound-force per minute }
\usage{NISTwattTOftPoundForcePerMin(watt)}
\description{\code{NISTwattTOftPoundForcePerMin} converts from watt (W) to foot pound-force per minute (ft * lbf/min) }
\arguments{
  \item{watt}{watt (W) }
}
\value{foot pound-force per minute (ft * lbf/min) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattTOftPoundForcePerMin(10)
}
\keyword{programming}