\name{NISTwattTOftPoundForcePerSec}
\alias{NISTwattTOftPoundForcePerSec}
\title{Convert watt  to foot pound-force per second }
\usage{NISTwattTOftPoundForcePerSec(watt)}
\description{\code{NISTwattTOftPoundForcePerSec} converts from watt (W) to foot pound-force per second (ft * lbf/s) }
\arguments{
  \item{watt}{watt (W) }
}
\value{foot pound-force per second (ft * lbf/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattTOftPoundForcePerSec(10)
}
\keyword{programming}