\name{NISTwattTOhorsepower}
\alias{NISTwattTOhorsepower}
\title{Convert watt  to horsepower }
\usage{NISTwattTOhorsepower(watt)}
\description{\code{NISTwattTOhorsepower} converts from watt (W) to horsepower (550 ft * lbf/s) (hp) }
\arguments{
  \item{watt}{watt (W) }
}
\value{horsepower (550 ft * lbf/s) (hp) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattTOhorsepower(10)
}
\keyword{programming}