\name{NISTwattTOhorsepowerElectric}
\alias{NISTwattTOhorsepowerElectric}
\title{Convert watt  to horsepower }
\usage{NISTwattTOhorsepowerElectric(watt)}
\description{\code{NISTwattTOhorsepowerElectric} converts from watt (W) to horsepower (electric) }
\arguments{
  \item{watt}{watt (W) }
}
\value{horsepower (electric) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattTOhorsepowerElectric(10)
}
\keyword{programming}