\name{NISTwattTOhorsepowerImperial}
\alias{NISTwattTOhorsepowerImperial}
\title{Convert watt  to horsepower }
\usage{NISTwattTOhorsepowerImperial(watt)}
\description{\code{NISTwattTOhorsepowerImperial} converts from watt (W) to horsepower (U.K.) }
\arguments{
  \item{watt}{watt (W) }
}
\value{horsepower (U.K.) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattTOhorsepowerImperial(10)
}
\keyword{programming}