\name{NISTwattTOhorsepowerSI}
\alias{NISTwattTOhorsepowerSI}
\title{Convert watt  to horsepower }
\usage{NISTwattTOhorsepowerSI(watt)}
\description{\code{NISTwattTOhorsepowerSI} converts from watt (W) to horsepower (metric) }
\arguments{
  \item{watt}{watt (W) }
}
\value{horsepower (metric) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattTOhorsepowerSI(10)
}
\keyword{programming}