\name{NISTwattTOhorsepowerWater}
\alias{NISTwattTOhorsepowerWater}
\title{Convert watt  to horsepower }
\usage{NISTwattTOhorsepowerWater(watt)}
\description{\code{NISTwattTOhorsepowerWater} converts from watt (W) to horsepower (water) }
\arguments{
  \item{watt}{watt (W) }
}
\value{horsepower (water) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattTOhorsepowerWater(10)
}
\keyword{programming}