\name{NISTwattTOkilocalthPerMin}
\alias{NISTwattTOkilocalthPerMin}
\title{Convert watt  to kilocalorieth per minute }
\usage{NISTwattTOkilocalthPerMin(watt)}
\description{\code{NISTwattTOkilocalthPerMin} converts from watt (W) to kilocalorieth per minute (kcalth/min) }
\arguments{
  \item{watt}{watt (W) }
}
\value{kilocalorieth per minute (kcalth/min) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattTOkilocalthPerMin(10)
}
\keyword{programming}