\name{NISTwattTOkilocalthPerSec}
\alias{NISTwattTOkilocalthPerSec}
\title{Convert watt  to kilocalorieth per second }
\usage{NISTwattTOkilocalthPerSec(watt)}
\description{\code{NISTwattTOkilocalthPerSec} converts from watt (W) to kilocalorieth per second (kcalth/s) }
\arguments{
  \item{watt}{watt (W) }
}
\value{kilocalorieth per second (kcalth/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattTOkilocalthPerSec(10)
}
\keyword{programming}