\name{NISTwattTOtonOfRefrigeration}
\alias{NISTwattTOtonOfRefrigeration}
\title{Convert watt  to ton of refrigeration }
\usage{NISTwattTOtonOfRefrigeration(watt)}
\description{\code{NISTwattTOtonOfRefrigeration} converts from watt (W) to ton of refrigeration (12 000 BtuIT/h) }
\arguments{
  \item{watt}{watt (W) }
}
\value{ton of refrigeration (12 000 BtuIT/h) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattTOtonOfRefrigeration(10)
}
\keyword{programming}