\name{NISTwattTOukThUnITPerHour}
\alias{NISTwattTOukThUnITPerHour}
\title{Convert watt  to British thermal unitIT per hour }
\usage{NISTwattTOukThUnITPerHour(watt)}
\description{\code{NISTwattTOukThUnITPerHour} converts from watt (W) to British thermal unitIT per hour (BtuIT/h) }
\arguments{
  \item{watt}{watt (W) }
}
\value{British thermal unitIT per hour (BtuIT/h) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattTOukThUnITPerHour(10)
}
\keyword{programming}