\name{NISTwattTOukThUnITPerSec}
\alias{NISTwattTOukThUnITPerSec}
\title{Convert watt  to British thermal unitIT per second }
\usage{NISTwattTOukThUnITPerSec(watt)}
\description{\code{NISTwattTOukThUnITPerSec} converts from watt (W) to British thermal unitIT per second (BtuIT/s) }
\arguments{
  \item{watt}{watt (W) }
}
\value{British thermal unitIT per second (BtuIT/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattTOukThUnITPerSec(10)
}
\keyword{programming}