\name{NISTwattTOukThUnthPerMin}
\alias{NISTwattTOukThUnthPerMin}
\title{Convert watt  to British thermal unitth per minute }
\usage{NISTwattTOukThUnthPerMin(watt)}
\description{\code{NISTwattTOukThUnthPerMin} converts from watt (W) to British thermal unitth per minute (Btuth/min) }
\arguments{
  \item{watt}{watt (W) }
}
\value{British thermal unitth per minute (Btuth/min) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattTOukThUnthPerMin(10)
}
\keyword{programming}