\name{NISTwattTOukThUnthPerSec}
\alias{NISTwattTOukThUnthPerSec}
\title{Convert watt  to British thermal unitth per second }
\usage{NISTwattTOukThUnthPerSec(watt)}
\description{\code{NISTwattTOukThUnthPerSec} converts from watt (W) to British thermal unitth per second (Btuth/s) }
\arguments{
  \item{watt}{watt (W) }
}
\value{British thermal unitth per second (Btuth/s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTwattTOukThUnthPerSec(10)
}
\keyword{programming}