\name{NISTweberTOmaxwell}
\alias{NISTweberTOmaxwell}
\title{Convert weber  to maxwell }
\usage{NISTweberTOmaxwell(weber)}
\description{\code{NISTweberTOmaxwell} converts from weber (Wb) to maxwell (Mx) }
\arguments{
  \item{weber}{weber (Wb) }
}
\value{maxwell (Mx) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTweberTOmaxwell(10)
}
\keyword{programming}