\name{NISTweberTOunPole}
\alias{NISTweberTOunPole}
\title{Convert weber  to unit pole}
\usage{NISTweberTOunPole(weber)}
\description{\code{NISTweberTOunPole} converts from weber (Wb) to unit pole }
\arguments{
  \item{weber}{weber (Wb) }
}
\value{unit pole }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTweberTOunPole(10)
}
\keyword{programming}