\name{NISTyardTOmeter}
\alias{NISTyardTOmeter}
\title{Convert yard  to meter }
\usage{NISTyardTOmeter(yard)}
\description{\code{NISTyardTOmeter} converts from yard (yd) to meter (m) }
\arguments{
  \item{yard}{yard (yd) }
}
\value{meter (m) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTyardTOmeter(10)
}
\keyword{programming}