\name{NISTyearSiderealTOsec}
\alias{NISTyearSiderealTOsec}
\title{Convert year  to second }
\usage{NISTyearSiderealTOsec(yearSidereal)}
\description{\code{NISTyearSiderealTOsec} converts from year (sidereal) to second (s) }
\arguments{
  \item{yearSidereal}{year (sidereal) }
}
\value{second (s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTyearSiderealTOsec(10)
}
\keyword{programming}