\name{NISTyearTOsec}
\alias{NISTyearTOsec}
\title{Convert year  to second }
\usage{NISTyearTOsec(year)}
\description{\code{NISTyearTOsec} converts from year (365 days) to second (s) }
\arguments{
  \item{year}{year (365 days) }
}
\value{second (s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTyearTOsec(10)
}
\keyword{programming}