\name{NISTyearTropicalTOsec}
\alias{NISTyearTropicalTOsec}
\title{Convert year  to second }
\usage{NISTyearTropicalTOsec(yearTropical)}
\description{\code{NISTyearTropicalTOsec} converts from year (tropical) to second (s) }
\arguments{
  \item{yearTropical}{year (tropical) }
}
\value{second (s) }
\source{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\references{
National Institute of Standards and Technology (NIST), 2014
NIST Guide to SI Units
B.8 Factors for Units Listed Alphabetically
\url{http://physics.nist.gov/Pubs/SP811/appenB8.html}
}
\author{Jose Gama}
\examples{
NISTyearTropicalTOsec(10)
}
\keyword{programming}