\name{ShortenTitles}
\alias{ShortenTitles}
\title{Shorten titles for factor help files}
\usage{ShortenTitles(s)}
\description{\code{ShortenTitles} shortens the titles for factor help files by removing info on parenthesis
}
\arguments{
  \item{s}{ string with the factor name}
}
\value{short string}
\author{Jose Gama}
\examples{
ShortenTitles('degree Fahrenheit (something long here)')
}
\keyword{programming}
